Multi-trait selection index

Multitrait index based on factor analysis and ideotype-design proposed by Rocha et al. (2018).

fai_blup(.data, DI, UI, SI = NULL, mineval = 1, verbose = TRUE)

Arguments

.data
An object of class waasb

DI, UI
A vector of the same length of .data to construct the desirable (DI) and undesirable (UI) ideotypes. For each element of the vector, allowed values are "max", "min", "mean", or a numeric value.

SI
An integer [0-100]. The selection intensity in percentage of the total number of genotypes.

mineval
The minimum value so that an eigenvector is retained in the factor analysis.

verbose
Logical value. If TRUE some results are shown in console.

References

Rocha, J.R.A.S.C.R, J.C. Machado, and P.C.S. Carneiro. 2018. Multitrait index based on factor analysis and ideotype-design: proposal and application on elephant grass breeding for bioenergy. GCB Bioenergy 10:52-60. doi: doi:10.1111/gcbb.12443

Examples

library(metan)
library(dplyr)

mod = waasb(data_ge,
            env = ENV,
            gen = GEN,
            rep = REP,
            resp = c(GY, HM),)

#> Evaluating variable GY 50 % 
#> Evaluating variable HM 100 % 
#> Done!

FAI = fai_blup(mod,
               SI = 15,
               DI = c("max", "max"),
               UI = c("min", "min"))

#> 
#> -----------------------------------------------------------------------------------
#> Principal Component Analysis
#> -----------------------------------------------------------------------------------
#>     eigen.values cumulative.var
#> PC1    1.1046076       55.23038
#> PC2    0.8953924      100.00000
#> 
#> -----------------------------------------------------------------------------------
#> Factor Analysis
#> -----------------------------------------------------------------------------------
#>           FA1 comunalits
#> GY -0.7431714  0.5523038
#> HM  0.7431714  0.5523038
#> 
#> -----------------------------------------------------------------------------------
#> Comunalit Mean: 0.5523038 
#> 
#> -----------------------------------------------------------------------------------
#> Multitrait stability index
#> -----------------------------------------------------------------------------------
#>     G4     G9     G6    G10     G5     G3     G2     G7     G1     G8 
#> 0.6084 0.5552 0.5315 0.5304 0.5237 0.4613 0.4608 0.3661 0.3170 0.0784 
#> 
#> -----------------------------------------------------------------------------------
#> Selection differential
#> -----------------------------------------------------------------------------------
#>    Factor Selection differential (%)
#> GY      1                  -2.993100
#> HM      1                  -0.172027
#> 
#> -----------------------------------------------------------------------------------
#> Selected genotypes
#> G4 G9
#> -----------------------------------------------------------------------------------

# Or using the pipe operator %>%

FAI = data_ge2 %>%
      waasb(ENV, GEN, REP, c(KW, NKE, PH, EH)) %>%
      fai_blup(DI = c("max", "max", "max", "min"),
               UI = c("min", "min", "min", "max"),
               SI = 15)

#> Warning: Model failed to converge with max|grad| = 0.00577455 (tol = 0.002, component 1)

#> Evaluating variable KW 25 % 

#> boundary (singular) fit: see ?isSingular

#> Evaluating variable NKE 50 % 
#> Evaluating variable PH 75 % 
#> Evaluating variable EH 100 % 
#> Done!
#> 
#> -----------------------------------------------------------------------------------
#> Principal Component Analysis
#> -----------------------------------------------------------------------------------
#>     eigen.values cumulative.var
#> PC1   3.14414174       78.60354
#> PC2   0.66041176       95.11384
#> PC3   0.13857767       98.57828
#> PC4   0.05686883      100.00000
#> 
#> -----------------------------------------------------------------------------------
#> Factor Analysis
#> -----------------------------------------------------------------------------------
#>            FA1 comunalits
#> KW  -0.9474560  0.8976729
#> NKE -0.7429934  0.5520392
#> PH  -0.9429971  0.8892436
#> EH  -0.8973216  0.8051861
#> 
#> -----------------------------------------------------------------------------------
#> Comunalit Mean: 0.7860354 
#> 
#> -----------------------------------------------------------------------------------
#> Multitrait stability index
#> -----------------------------------------------------------------------------------
#>    H13     H6     H2     H5     H1     H4     H3     H8     H9    H10     H7 
#> 0.9807 0.8843 0.8413 0.8409 0.8023 0.7972 0.6222 0.2623 0.2354 0.1810 0.1163 
#>    H12    H11 
#> 0.1014 0.0780 
#> 
#> -----------------------------------------------------------------------------------
#> Selection differential
#> -----------------------------------------------------------------------------------
#>     Factor Selection differential (%)
#> KW       1               1.297090e+00
#> NKE      1               5.510697e-08
#> PH       1               9.006835e-02
#> EH       1               2.407567e-08
#> 
#> -----------------------------------------------------------------------------------
#> Selected genotypes
#> H13 H6
#> -----------------------------------------------------------------------------------