Fox's stability function

Performs a stability analysis based on the criteria of Fox et al. (1990), using the statistical "TOP third" only. A stratified ranking of the genotypes at each environment is done. The proportion of locations at which the genotype occurred in the top third are expressed in the output.

Fox(.data, env, gen, resp, verbose = TRUE)

Arguments

.data
The dataset containing the columns related to Environments, Genotypes, replication/block and response variable(s).

env
The name of the column that contains the levels of the environments.

gen
The name of the column that contains the levels of the genotypes.

resp
The response variable(s). To analyze multiple variables in a single procedure use, for example, resp = c(var1, var2, var3).

verbose
Logical argument. If verbose = FALSE the code will run silently.

Value

An object of class Fox, which is a list containing the results for each variable used in the argument resp. For each variable, a tibble with the following columns is returned.

GEN the genotype's code.

mean the mean for the response variable.

TOP The proportion of locations at which the

References

Fox, P.N., B. Skovmand, B.K. Thompson, H.J. Braun, and R. Cormier. 1990. Yield and adaptation of hexaploid spring triticale. Euphytica 47:57-64. doi:10.1007/BF00040364.

Examples

# \donttest{
library(metan)
out <- Fox(data_ge2, ENV, GEN, PH)
print(out)

#> Variable PH 
#> ---------------------------------------------------------------------------
#> Fox TOP third criteria
#> ---------------------------------------------------------------------------
#> # A tibble: 13 x 3
#>    GEN       Y   TOP
#>    <fct> <dbl> <int>
#>  1 H1     2.62     2
#>  2 H10    2.31     0
#>  3 H11    2.39     0
#>  4 H12    2.44     1
#>  5 H13    2.54     3
#>  6 H2     2.60     1
#>  7 H3     2.59     2
#>  8 H4     2.58     1
#>  9 H5     2.57     1
#> 10 H6     2.56     0
#> 11 H7     2.40     0
#> 12 H8     2.33     0
#> 13 H9     2.36     1
#> 
#> 
#> 

# }