Huehn's stability statistics

Performs a stability analysis based on Huehn (1979) statistics. The four nonparametric measures of phenotypic stability are: S1 (mean of the absolute rank differences of a genotype over the n environments), S2 (variance among the ranks over the k environments), S3 (sum of the absolute deviations), and S6 (relative sum of squares of rank for each genotype).

Usage

Huehn(.data, env, gen, resp, verbose = TRUE)

Arguments

.data

The dataset containing the columns related to Environments, Genotypes, replication/block and response variable(s).

env

The name of the column that contains the levels of the environments.

gen

The name of the column that contains the levels of the genotypes.

resp

The response variable(s). To analyze multiple variables in a single procedure use, for example, resp = c(var1, var2, var3).

verbose

Logical argument. If verbose = FALSE the code will run silently.

Value

An object of class Huehn, which is a list containing the results for each variable used in the argument resp. For each variable, a tibble with the following columns is returned.

GEN The genotype's code.

Y The mean for the response variable.

S1 Mean of the absolute rank differences of a genotype over the n environments.

S2 variance among the ranks over the k environments.

S3 Sum of the absolute deviations.

S6 Relative sum of squares of rank for each genotype.

References

Huehn, V.M. 1979. Beitrage zur erfassung der phanotypischen stabilitat. EDV Med. Biol. 10:112.

Author

Tiago Olivoto tiagoolivoto@gmail.com

Examples

# \donttest{
library(metan)
out <- Huehn(data_ge2, ENV, GEN, PH)
#> Evaluating trait PH |============================================| 100% 00:00:00 
print(out)
#> Variable PH 
#> ---------------------------------------------------------------------------
#> Huehn's stability indexes
#> ---------------------------------------------------------------------------
#> # A tibble: 13 × 11
#>    GEN       Y   Y_R    S1  S1_R    S2  S2_R    S3  S3_R    S6  S6_R
#>    <chr> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1 H1     2.62     1 0.167   3.5 16.9      8  5.70     8 1.35    8  
#>  2 H10    2.31    13 0       1.5  8.25     3  7.71     9 2.57   11  
#>  3 H11    2.39    10 0.667  11    6.25     2  4.26     7 1.33    7  
#>  4 H12    2.44     8 0.333   6   20.9     10 10.3     11 2.33    9  
#>  5 H13    2.54     7 1      13   20.2      9  3.33     3 1.11    4.5
#>  6 H2     2.60     2 0.333   6   22.7     11  3.56     5 0.889   1  
#>  7 H3     2.59     3 0       1.5 34.2     13 11.7     12 2.45   10  
#>  8 H4     2.58     4 0.5     9    8.92     4  3.41     4 1.06    3  
#>  9 H5     2.57     5 0.833  12    5.58     1  3.78     6 1.11    4.5
#> 10 H6     2.56     6 0.5     9   11        5  2.62     1 1.23    6  
#> 11 H7     2.40     9 0.167   3.5 11.3      6  2.78     2 1.04    2  
#> 12 H8     2.33    12 0.333   6   28.2     12  9.5     10 3      12  
#> 13 H9     2.36    11 0.5     9   14.2      7 16.4     13 3.2    13  
#> ---------------------------------------------------------------------------
#> 
#> 
# }