Multi-trait stability index

Computes the multitrait stability index proposed by Olivoto et al. (2019)

mtsi(.data, index = "waasb", SI = 15, mineval = 1, verbose = TRUE)

Arguments

.data
An object of class waasb or waas.

index
If index = 'waasb' (Default) the multitrait index will be computed considering the stability of genotypes only. If index = 'waasby' both stability and mean performance are considered. More details can be seen in waasb and waas functions.

SI
An integer [0-100]. The selection intensity in percentage of the total number of genotypes.

mineval
The minimum value so that an eigenvector is retained in the factor analysis.

verbose
If verbose = TRUE (Default) then some results are shown in the console.

References

Olivoto, T., A.D.C. L\'ucio, J.A.G. da silva, B.G. Sari, and M.I. Diel. 2019. Mean performance and stability in multi-environment trials II: Selection based on multiple traits. Agron. J.doi:10.2134/agronj2019.03.0221.

Examples

library(metan)

# Based on stability only, for both GY and HM, higher is better
mtsi_model = waasb(data_ge,
                   env = ENV,
                   gen = GEN,
                   rep = REP,
                   resp = c(GY, HM))

#> Evaluating variable GY 50 % 
#> Evaluating variable HM 100 % 
#> Done!

mtsi_index = mtsi(mtsi_model)

#> 
#> -------------------------------------------------------------------------------
#> Principal Component Analysis
#> -------------------------------------------------------------------------------
#>     Eigenvalues Variance (%) Cum. variance (%)
#> PC1   1.6566277     82.83139          82.83139
#> PC2   0.3433723     17.16861         100.00000
#> -------------------------------------------------------------------------------
#> Factor Analysis - factorial loadings after rotation-
#> -------------------------------------------------------------------------------
#>          FA1 Communality Uniquenesses
#> GY 0.9101175   0.8283139    0.1716861
#> HM 0.9101175   0.8283139    0.1716861
#> -------------------------------------------------------------------------------
#> Comunalit Mean: 0.8283139 
#> -------------------------------------------------------------------------------
#> Multitrait stability index
#> -------------------------------------------------------------------------------
#>     G3     G1     G6     G8     G4     G2     G9     G5     G7    G10 
#> 0.0000 0.2303 0.4846 1.0066 1.2954 1.6022 1.8318 1.8374 1.8450 3.4178 
#> -------------------------------------------------------------------------------
#> Selection differential for the  waasb index
#> -------------------------------------------------------------------------------
#>    Factor        Xo        Xs         SD    SDperc
#> GY    FA1 0.2504736 0.1165028 -0.1339708 -53.48700
#> HM    FA1 0.6141648 0.3732404 -0.2409244 -39.22798
#> ------------------------------------------------------------------------------
#> Mean of selection differential
#> -------------------------------------------------------------------------------
#>          Xo          Xs          SD      SDperc 
#>   0.4323192   0.2448716  -0.1874476 -46.3574863 
#> -------------------------------------------------------------------------------
#> Selection differential for the mean of the variables
#> -------------------------------------------------------------------------------
#>    Factor        Xo       Xs         SD    SDperc
#> GY    FA1  2.674242  2.77951  0.1052684  3.936384
#> HM    FA1 48.088286 47.34179 -0.7465000 -1.552353
#> ------------------------------------------------------------------------------
#> Selected genotypes
#> -------------------------------------------------------------------------------
#> G3 G1
#> -------------------------------------------------------------------------------

# Based on mean performance and stability (using pipe operator %>%)
# GY: higher is better
# HM: lower is better
mtsi_index2 = data_ge %>%
              waasb(ENV, GEN, REP, c(GY, HM), mresp = c(100, 0)) %>%
              mtsi(index = 'waasby')

#> Evaluating variable GY 50 % 
#> Evaluating variable HM 100 % 
#> Done!
#> 
#> -------------------------------------------------------------------------------
#> Principal Component Analysis
#> -------------------------------------------------------------------------------
#>     Eigenvalues Variance (%) Cum. variance (%)
#> PC1   1.5844965     79.22483          79.22483
#> PC2   0.4155035     20.77517         100.00000
#> -------------------------------------------------------------------------------
#> Factor Analysis - factorial loadings after rotation-
#> -------------------------------------------------------------------------------
#>          FA1 Communality Uniquenesses
#> GY 0.8900833   0.7922483    0.2077517
#> HM 0.8900833   0.7922483    0.2077517
#> -------------------------------------------------------------------------------
#> Comunalit Mean: 0.7922483 
#> -------------------------------------------------------------------------------
#> Multitrait stability index
#> -------------------------------------------------------------------------------
#>     G3     G1     G2     G8     G6     G4     G7     G9     G5    G10 
#> 0.4901 1.0853 1.5365 1.6202 2.0663 2.0779 2.1950 2.5718 3.2045 3.9391 
#> -------------------------------------------------------------------------------
#> Selection differential for the  waasby index
#> -------------------------------------------------------------------------------
#>    Factor       Xo       Xs       SD   SDperc
#> GY    FA1 48.34113 76.54450 28.20337 58.34239
#> HM    FA1 54.11473 86.20436 32.08962 59.29924
#> ------------------------------------------------------------------------------
#> Mean of selection differential
#> -------------------------------------------------------------------------------
#>       Xo       Xs       SD   SDperc 
#> 51.22793 81.37443 30.14650 58.82081 
#> -------------------------------------------------------------------------------
#> Selection differential for the mean of the variables
#> -------------------------------------------------------------------------------
#>    Factor        Xo       Xs         SD    SDperc
#> GY    FA1  2.674242  2.77951  0.1052684  3.936384
#> HM    FA1 48.088286 47.34179 -0.7465000 -1.552353
#> ------------------------------------------------------------------------------
#> Selected genotypes
#> -------------------------------------------------------------------------------
#> G3 G1
#> -------------------------------------------------------------------------------