Shukla's stability variance parameter

The function computes the Shukla's stability variance parameter (1972) and uses the Kang's nonparametric stability (rank sum) to imcorporate the mean performance and stability into a single selection criteria.

Shukla(.data, env, gen, rep, resp, verbose = TRUE)

Arguments

.data
The dataset containing the columns related to Environments, Genotypes, replication/block and response variable(s).

env
The name of the column that contains the levels of the environments.

gen
The name of the column that contains the levels of the genotypes.

rep
The name of the column that contains the levels of the replications/blocks.

resp
The response variable(s). To analyze multiple variables in a single procedure use, for example, resp = c(var1, var2, var3).

verbose
Logical argument. If verbose = FALSE the code will run silently.

Value

An object of class Shukla, which is a list containing the results for each variable used in the argument resp. For each variable, a tibble with the following columns is returned.

GEN the genotype's code.

Y the mean for the response variable.

ShuklaVar The Shukla's stability variance parameter.

rMean The rank for Y (decreasing).

rShukaVar The rank for ShukaVar.

ssiShukaVar The simultaneous selection index (\(ssiShukaVar = rMean + rShukaVar\)).

References

Shukla, G.K. 1972. Some statistical aspects of partitioning genotype-environmental components of variability. Heredity. 29:238-245. doi:10.1038/hdy.1972.87.

Kang, M.S., and H.N. Pham. 1991. Simultaneous Selection for High Yielding and Stable Crop Genotypes. Agron. J. 83:161. doi:10.2134/agronj1991.00021962008300010037x.

Examples

# \donttest{
library(metan)
out <- Shukla(data_ge2,
             env = ENV,
             gen = GEN,
             rep = REP,
             resp = PH)
# }