Thennarasu's stability statistics

Performs a stability analysis based on Thennarasu (1995) statistics.

Thennarasu(.data, env, gen, resp, verbose = TRUE)

Arguments

.data
The dataset containing the columns related to Environments, Genotypes, replication/block and response variable(s).

env
The name of the column that contains the levels of the environments.

gen
The name of the column that contains the levels of the genotypes.

resp
The response variable(s). To analyze multiple variables in a single procedure use, for example, resp = c(var1, var2, var3).

verbose
Logical argument. If verbose = FALSE the code will run silently.

Value

An object of class Thennarasu, which is a list containing the results for each variable used in the argument resp. For each variable, a tibble with the columns GEN, N1, N2, N3 and N4 is returned.

References

Thennarasu, K. 1995. On certain nonparametric procedures for studying genotype x environment interactions and yield stability. Ph.D. thesis. P.J. School, IARI, New Delhi, India.

Author

Tiago Olivoto tiagoolivoto@gmail.com

Examples

# \donttest{
library(metan)
out <- Thennarasu(data_ge, ENV, GEN, GY)
print(out)

#> Variable GY 
#> ---------------------------------------------------------------------------
#> Thennarasu's stability indexes
#> ---------------------------------------------------------------------------
#> # A tibble: 10 x 11
#>    GEN       Y   Y_R    N1  N1_R    N2  N2_R    N3  N3_R      N4  N4_R
#>    <chr> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>   <dbl> <dbl>
#>  1 G1     2.60     6  2.57   8   0.367     5 0.429     4 0         1.5
#>  2 G10    2.47    10  3.71  10   0.531     7 0.577     7 0.0032    4.5
#>  3 G2     2.74     3  2.43   5.5 0.540     8 0.634     8 0.014     8  
#>  4 G3     2.96     2  2      2.5 0.667     9 0.862     9 0.0085    6  
#>  5 G4     2.64     5  1.93   1   0.321     1 0.402     2 0.0147    9  
#>  6 G5     2.54     7  2.71   9   0.339     2 0.384     1 0         1.5
#>  7 G6     2.53     8  2.43   5.5 0.347     3 0.411     3 0.0032    4.5
#>  8 G7     2.74     4  2.29   4   0.508     6 0.564     6 0.0022    3  
#>  9 G8     3.00     1  2      2.5 1        10 1.12     10 0.0149   10  
#> 10 G9     2.51     9  2.5    7   0.357     4 0.436     5 0.00960   7  
#> ---------------------------------------------------------------------------
#> 
#> 

# }