Weighted Average of Absolute Scores

Weighted Average of Absolute Scores

Predict the means of a WAAS.AMMI object considering a specific number of axis.

Summary the WAAS.AMMI object in two ways. By default, the results are shown in the R console. The results can also be exported to the directory.

WAAS.AMMI(.data, env, gen, rep, resp, mresp = NULL, wresp = NULL,
  prob = 0.05, naxis = NULL, verbose = TRUE)

WAAS.AMMI(.data, env, gen, rep, resp, mresp = NULL, wresp = NULL,
  prob = 0.05, naxis = NULL, verbose = TRUE)

WAAS.AMMI(.data, env, gen, rep, resp, mresp = NULL, wresp = NULL,
  prob = 0.05, naxis = NULL, verbose = TRUE)

Arguments

.data
The dataset containing the columns related to Environments, Genotypes, replication/block and response variable(s).

env
The name of the column that contains the levels of the environments.

gen
The name of the column that contains the levels of the genotypes.

rep
The name of the column that contains the levels of the replications/blocks.

resp
The response variable(s). To analyze multiple variables in a single procedure a vector of variables may be used. For example resp = c(var1, var2, var3).

mresp
A numeric vector of the same length of resp. The mresp will be the new maximum value after reescaling. By default, all variables in resp are rescaled so that de maximum value is 100 and the minimum value is 0.

wresp
The weight for the response variable(s) for computing the WAASBY index. Must be a numeric vector of the same length of resp. Defatul is 50, i.e., equal weights for stability and mean performance.

prob
The p-value for considering a IPCA significant.

naxis
The number of IPCAs to be used for computing the WAAS index. Default is NULL (Significant IPCAs are used). If values are informed, the number of IPCAS will be used independently on its significance. Note that if two or more variables are included in resp, then naxis must be a vector.

verbose
Logical argument. If verbose = FALSE the code is run silently.

object
An object of class WAAS.AMMI

naxis
The the number of axis to be use in the prediction. If object has more than one variable, then naxis must be a vector.

...
Additional parameter for the function

object
The WAAS.AMMI object

export
A logical argument. If TRUE|T, a *.txt file is exported to the working directory

file.name
The name of the file if export = TRUE

digits
The significant digits to be shown.

...
Other arguments of the function

Details

This function is used to predict the response variable of a two-way table (for examples the yielding of the i-th genotype in the j-th environment) based on AMMI model. This prediction is based on the number of multiplicative terms used. If naxis = 0, only the main effects (AMMI0) are used. In this case, the predicted mean will be the predicted value from OLS estimation. If naxis = 1 the AMMI1 (with one multiplicative term) is used for predicting the response variable. If naxis = min(gen-1;env-1), the AMMIF is fitted and the predicted value will be the cell mean, i.e. the mean of R-replicates of the i-th genotype in the j-th environment. The number of axis to be used must be carrefully chosen. Precures based on Postdictive sucess (such as Gollobs's d.f.) or Predictive sucess (such as cross-validation) should be used to do this. This package provide both. WAAS.AMMI function compute traditional AMMI analysis showing the number of significant axis. On the other hand, cv_ammif function provide a cross-validation, estimating the RMSPD of all AMMI-family models, based on resampling procedures.

See also