Different scenarios of stability and mean performance

This function computes the WAASBY index in mixed-effec model analysis in different combinations of weights for stability and productivity.

This function computes the WAASBY index in AMMI analysis in different combinations of weights for stability and productivity.

Plot heat maps with genotype ranking in two ways.

WAASBYratio(.data, env, gen, rep, resp, mresp = 100, increment = 10,
  saveWAASY = 50, progbar = TRUE)

WAASBYratio(.data, env, gen, rep, resp, mresp = 100, increment = 10,
  saveWAASY = 50, progbar = TRUE)

WAASBYratio(.data, env, gen, rep, resp, mresp = 100, increment = 10,
  saveWAASY = 50, progbar = TRUE)

Arguments

.data
The dataset containing the columns related to Environments, Genotypes, replication/block and response variable(s).

env
The name of the column that contains the levels of the environments.

gen
The name of the column that contains the levels of the genotypes.

rep
The name of the column that contains the levels of the replications/blocks.

resp
The response variable.

mresp
A numeric value that will be the new maximum value after reescaling. By default, the variable in resp is rescaled so that the original maximum and minimum values are 100 and 0, respectively.

increment
The range of the increment for WAASB/GY ratio. Default is 5. The function compute the WAASBY values starting with a weight o 100 for stability and 0 for response variable. With the default, the first scenario will be a WAASB/GY ratio = 100/0. In the next scenario, the WAASBY values are computed based on a WAASB/GY ratio = 95/5.

saveWAASY
Automatically save the WAASY values when the wheight for WAAS (stability) in the WAAS/GY ratio is "saveWAASY". Default is 100. The value of "saveWAASY" must be multiple of "Increment". If this assumption is not valid, an error will be occour.

progbar
A logical argument to define if a progress bar is shown. Default is TRUE.

.data
The dataset containing the columns related to Environments, Genotypes, replication/block and response variable(s).

env
The name of the column that contains the levels of the environments.

gen
The name of the column that contains the levels of the genotypes.

rep
The name of the column that contains the levels of the replications/blocks.

resp
The response variable.

mresp
A numeric value that will be the new maximum value after reescaling. By default, the variable in resp is rescaled so that the original maximum and minimum values are 100 and 0, respectively.

p.valuePC
The p-value for considering the PC significant. Default is 0.05. The number of significant Principal Components to be used for calculating the WAASB will be chosen based on this probability.

increment
The range of the increment for WAAS/GY ratio. Default is 5. The function compute the WAASY values starting with a weight o 100 for stability and 0 for response variable. With the default, the first scenario will be a WAAS/GY ratio = 100/0. In the next scenario, the WAASY values are computed based on a WAAS/GY ratio = 95/5.

saveWAASY
Automatically save the WAASY values when the wheight for WAAS (stability) in the WAAS/GY ratio is "saveWAASY". Default is 100. The value of "saveWAASY" must be multiple of "Increment". If this assumption is not valid, an error will be occour.

progbar
A logical argument to define if a progress bar is shown. Default is TRUE.

x
The WAASBYratio object

type
1 = Heat map Ranks: this graphic shows the genotype ranking considering the WAAS estimated with different numbers of Principal Components; 2 = Heat map WAASY-GY ratio: this graphic shows the genotype ranking considering the different combinations in the WAAS/GY ratio.

export
Export (or not) the plot. Default is TRUE.

file.type
If export = TRUE define the type of file to be exported. Default is pdf, Graphic can also be exported in *.tiff format by declaring file.type = "tiff".

file.name
The name of the file for exportation, default is NULL, i.e. the files are automatically named.

width
The width "inch" of the plot. Default is 8.

height
The height "inch" of the plot. Default is 7.

size.lab
The label size of the plot. It is suggested attribute 1

margins
Numeric vector of length 2 containing the margins for column and row names, respectively. Default is c(5, 4).

y.lab
The label of y axis. Default is "Genotypes".

x.lab
The label of x axis. Default is "Number of axes".

key.lab
The label of color key. Default is "Genotype ranking".

resolution
Valid parameter if file.type = "tiff". Define the resolution of the plot. Default is "300".

...
Other arguments of the function

Value

anova

Joint analysis of variance for the main effects and Principal Component analysis of the interaction effect.

PC

Principal Component Analysis.

MeansGxE

The means of genotypes in the environments, with observed, predicted and residual values.

WAAS

A data frame with the response variable, the scores of all Principal Components, the estimates of Weighted Average of Absolute Scores, and WAASY (the index that consider the weights for stability and productivity in the genotype ranking.

WAASY

The values of the WAASY estimated when the wheight for the stability in the loop match with argument "saveWAASY".

WAASY.values

All the values of WAASY estimated in the different scenarios of WAAS/GY weighting ratio.

Details

This function considers both stability (weighted average of absolute scores based on SVD of BLUP-interaction effects matrix) and productivitye for genotype ranking. This function provide the option of attributing weights for stability and productive in genotype ranking. This is important depending on the goal of a selection strategy. For example, if a a goal of a breeding program is to select a genotype whith high yielding (independeltly on the stability performance), that genotype with the first rank in an WAASB/GY = 0/100 ratio should be selected. The reciprocal is true. Aiming at selecting a high-stable genotype (independentely on the productivity), that genotype with the first rank in an WAASB/GY = 100/0 ratio should be selected. By defalut, the increment on the WAASB/GY ratio is equal to 5. In other words, twenty one different combinations are computed. Each combination, the genotypes are ranked regarding the WAASY value.

This function is very similar to the WAASBYratio. The main difference is that here, the WAASBY values are computed considering a traditional AMMI model.

The first type of heatmap shows the genotype ranking depending on the number of principal component axis used for estimating the WAASB index. An euclidian distance-based dendrogram is used for grouping the genotype ranking for both genotypes and principal component axis. The second type of heatmap shows the genotype ranking depending on the WAASB/GY ratio. The ranks obtained with a ratio of 100/0 considers exclusively the stability for the genotype ranking. On the other hand, a ratio of 0/100 considers exclusively the productivity for the genotype ranking. Four clusters are estimated (1) unproductive and unstable genotypes; (2) productive, but unstable genotypes; (3) stable, but unproductive genotypes; and (4), productive and stable genotypes.

See also