Canonical correlation analysis

Performs canonical correlation analysis with collinearity diagnostic, estimation of canonical loads, canonical scores, and hipotesis testing for correlation pairs.

can_corr(.data = NULL, FG = NULL, SG = NULL, use = "cor",
  test = "Bartlett", prob = 0.05, center = TRUE, stdscores = FALSE,
  verbose = TRUE, collinearity = TRUE)

Arguments

.data
The data to be analyzed. Must be a dataframe containing the numeric variables that will be used in the estimation of the correlations. The data can also be passed directly by the arguments FG and SG. Alternatively, .data may be passed from the function split_factors. In such case, the canonical correlation will be estimated for each level of the grouping variable in that function.

FG
If a dataframe is informed in .data, then FG is a comma-separated list of unquoted variable names that will compose the first (smallest) group of the correlation analysis. FG can also be a cordinate for the variables; for example, FG = data[,1:3].

SG
Similar than FG but for the second group of variables.

use
The matrix to be used. Must be one of 'cor' for analysis using the correlation matrix (default) or 'cov' for analysis using the covariance matrix.

test
The test of significance of the relationship between the FG and SG. Must be one of the 'Bartlett' (default) or 'Rao'.

prob
The probability of error assumed. Set to 0.05.

center
Should the data be centered to compute the scores?

stdscores
Rescale scores to produce scores of unit variance?

verbose
Logical argument. If TRUE (default) then the results are shown in the console.

collinearity
Logical argument. If TRUE (default) then a collinearity diagnostic is performed for each group of variables according to Olivoto et al.(2017).

Value

Matrix

The correlation (or covariance) matrix of the variables

MFG, MSG

The correlation (or covariance) matrix for the variables of the first group or second group, respectively.

MFG_SG

The correlation (or covariance) matrix for the variables of the first group with the second group.

Coef_FG, Coef_SG

Matrix of the canonical coefficients of the first group or second group, respectively.

Loads_FG, Loads_SG

Matrix of the canonical loadings of the first group or second group, respectively.

Crossload_FG, Crossload_FG

Canonical cross-loadings for FG variables on the SG scores, and cross-loadings for SG variables on the FG scores, respectively.

SigTest

A dataframe with the correlation of the canonical pairs and hypothesis testing results.

collinearity

A list with the collinearity diagnostic for each group of variables.

References

Olivoto, T., V.Q. Souza, M. Nardino, I.R. Carvalho, M. Ferrari, A.J. Pelegrin, V.J. Szareski, and D. Schmidt. 2017. Multicollinearity in path analysis: a simple method to reduce its effects. Agron. J. 109:131-142. doi:10.2134/agronj2016.04.0196. 10.2134/agronj2016.04.0196

Examples

# NOT RUN {
library(metan)
library(dplyr)

cc1 = can_corr(data_ge2,
               FG = c(PH, EH, EP),
               SG = c(EL, ED, CL, CD, CW, KW, NR))

cc2 = can_corr(FG = datage_2[, 4:6],
               SG = datage_2[, 7:13],
               verbose = FALSE,
               collinearity = FALSE)

# Canonical correlations for each environment
cc3 = data_ge2 %>%
      split_factors(ENV, REP) %>%
      can_corr(FG = c(PH, EH, EP),
               SG = c(EL, ED, CL, CD, CW, KW, NR))

# }