Cross-validation procedure

Cross-validation for estimation of AMMI models

cv_ammi(.data, env, gen, rep, resp, block = NULL, naxis = 2,
  nboot = 200, design = "RCBD", verbose = TRUE)

Arguments

.data
The dataset containing the columns related to Environments, Genotypes, replication/block and response variable(s).

env
The name of the column that contains the levels of the environments.

gen
The name of the column that contains the levels of the genotypes.

rep
The name of the column that contains the levels of the replications/blocks.

resp
The response variable.

block
Defaults to NULL. In this case, a randomized complete block design is considered. If block is informed, then a resolvable alpha-lattice design (Patterson and Williams, 1976) is employed. All effects are assumed to be fixed.

naxis
The number of axis to be considered for estimation of GE effects.

nboot
The number of resamples to be used in the cross-validation. Defaults to 200.

design
The experimental design. Defaults to RCBD (Randomized complete Block Design). For Completely Randomized Designs inform design = 'CRD'.

verbose
A logical argument to define if a progress bar is shown. Default is TRUE.

Value

An object of class cv_ammi with the following items: * RMSPD: A vector with nboot-estimates of the Root Mean Squared Prediction Difference between predicted and validating data.

RMSPDmean: The mean of RMSPDmean estimates.

Estimated: A data frame that contain the values (predicted, observed, validation) of the last loop.

Modeling: The dataset used as modeling data in the last loop

Testing: The dataset used as testing data in the last loop.

Details

THe original dataset is split into two datasets: training set and validation set. The 'training' set has all combinations (genotype x environment) with N-1 replications. The 'validation' set has the remaining replication. The splitting of the dataset into modeling and validation sets depends on the design informed. For Completely Randomized Block Design (default), and alpha-lattice design (declaring block arguments), complete replicates are selected within environments. The remained replicate serves as validation data. If design = 'RCD' is informed, completely randomly samples are made for each genotype-by-environment combination (Olivoto et al. 2019). The estimated values considering naxis-Interaction Principal Component Axis are compared with the 'validation' data. The Root Mean Square Prediction Difference (RMSPD) is computed. At the end of boots, a list is returned.

References

Olivoto, T., A.D.C. L'ucio, J.A.G. da silva, V.S. Marchioro, V.Q. de Souza, and E. Jost. 2019. Mean performance and stability in multi-environment trials I: Combining features of AMMI and BLUP techniques. Agron. J. doi:10.2134/agronj2019.03.0220

Patterson, H.D., and E.R. Williams. 1976. A new class of resolvable incomplete block designs. Biometrika 63:83-92.

See also

Examples


# \donttest{
library(metan)
model <- cv_ammi(data_ge,
                env = ENV,
                gen = GEN,
                rep = REP,
                resp = GY,
                nboot = 10,
                naxis = 2)

#> 

#> Validating 3 of 10 sets [=======>--------------------]  30% (00:00:00 - 1s left)

#> 

#> Validating 4 of 10 sets [==========>-----------------]  40% (00:00:00 - 1s left)

#> 

#> Validating 5 of 10 sets [=============>--------------]  50% (00:00:00 - 1s left)

#> 

#> Validating 6 of 10 sets [================>-----------]  60% (00:00:00 - 0s left)

#> 

#> Validating 7 of 10 sets [===================>--------]  70% (00:00:00 - 0s left)

#> 

#> Validating 8 of 10 sets [=====================>------]  80% (00:00:00 - 0s left)

#> 

#> Validating 9 of 10 sets [========================>---]  90% (00:00:00 - 0s left)

#> 

#> Validating 10 of 10 sets [===========================] 100% (00:00:00 - 0s left)

#> 


# Alternatively using the pipe operator %>%
model <- data_ge %>%
         cv_ammi(ENV, GEN, REP, GY)

#> 

#> Validating 5 of 200 sets [>--------------------------]   2% (00:00:00 - 8s left)

#> 

#> Validating 6 of 200 sets [>--------------------------]   3% (00:00:00 - 9s left)

#> 

#> Validating 7 of 200 sets [>--------------------------]   4% (00:00:00 - 9s left)

#> 

#> Validating 8 of 200 sets [>--------------------------]   4% (00:00:00 - 9s left)

#> 

#> Validating 9 of 200 sets [>--------------------------]   4% (00:00:00 - 9s left)

#> 

#> Validating 10 of 200 sets [>-------------------------]   5% (00:00:00 -10s left)

#> 

#> Validating 11 of 200 sets [>-------------------------]   6% (00:00:00 -10s left)

#> 

#> Validating 12 of 200 sets [=>------------------------]   6% (00:00:00 -10s left)

#> 

#> Validating 13 of 200 sets [=>------------------------]   6% (00:00:00 -10s left)

#> 

#> Validating 14 of 200 sets [=>------------------------]   7% (00:00:00 -10s left)

#> 

#> Validating 15 of 200 sets [=>------------------------]   8% (00:00:00 -10s left)

#> 

#> Validating 16 of 200 sets [=>------------------------]   8% (00:00:00 -10s left)

#> 

#> Validating 17 of 200 sets [=>------------------------]   8% (00:00:00 -10s left)

#> 

#> Validating 18 of 200 sets [=>------------------------]   9% (00:00:00 -10s left)

#> 

#> Validating 19 of 200 sets [=>------------------------]  10% (00:00:01 -10s left)

#> 

#> Validating 20 of 200 sets [==>-----------------------]  10% (00:00:01 -10s left)

#> 

#> Validating 21 of 200 sets [==>-----------------------]  10% (00:00:01 -10s left)

#> 

#> Validating 22 of 200 sets [==>-----------------------]  11% (00:00:01 -10s left)

#> 

#> Validating 23 of 200 sets [==>-----------------------]  12% (00:00:01 -10s left)

#> 

#> Validating 24 of 200 sets [==>-----------------------]  12% (00:00:01 -10s left)

#> 

#> Validating 25 of 200 sets [==>-----------------------]  12% (00:00:01 -10s left)

#> 

#> Validating 26 of 200 sets [==>-----------------------]  13% (00:00:01 -10s left)

#> 

#> Validating 27 of 200 sets [===>----------------------]  14% (00:00:01 -10s left)

#> 

#> Validating 28 of 200 sets [===>----------------------]  14% (00:00:01 -10s left)

#> 

#> Validating 29 of 200 sets [===>----------------------]  14% (00:00:01 -10s left)

#> 

#> Validating 30 of 200 sets [===>----------------------]  15% (00:00:01 -10s left)

#> 

#> Validating 31 of 200 sets [===>----------------------]  16% (00:00:01 -10s left)

#> 

#> Validating 32 of 200 sets [===>----------------------]  16% (00:00:01 -10s left)

#> 

#> Validating 33 of 200 sets [===>----------------------]  16% (00:00:01 -10s left)

#> 

#> Validating 34 of 200 sets [===>----------------------]  17% (00:00:01 -10s left)

#> 

#> Validating 35 of 200 sets [====>---------------------]  18% (00:00:02 - 9s left)

#> 

#> Validating 36 of 200 sets [====>---------------------]  18% (00:00:02 - 9s left)

#> 

#> Validating 37 of 200 sets [====>---------------------]  18% (00:00:02 - 9s left)

#> 

#> Validating 38 of 200 sets [====>---------------------]  19% (00:00:02 - 9s left)

#> 

#> Validating 39 of 200 sets [====>---------------------]  20% (00:00:02 - 9s left)

#> 

#> Validating 40 of 200 sets [====>---------------------]  20% (00:00:02 -10s left)

#> 

#> Validating 41 of 200 sets [====>---------------------]  20% (00:00:02 - 9s left)

#> 

#> Validating 42 of 200 sets [====>---------------------]  21% (00:00:02 -10s left)

#> 

#> Validating 43 of 200 sets [=====>--------------------]  22% (00:00:02 - 9s left)

#> 

#> Validating 44 of 200 sets [=====>--------------------]  22% (00:00:02 - 9s left)

#> 

#> Validating 45 of 200 sets [=====>--------------------]  22% (00:00:02 - 9s left)

#> 

#> Validating 46 of 200 sets [=====>--------------------]  23% (00:00:02 - 9s left)

#> 

#> Validating 47 of 200 sets [=====>--------------------]  24% (00:00:02 - 9s left)

#> 

#> Validating 48 of 200 sets [=====>--------------------]  24% (00:00:02 - 9s left)

#> 

#> Validating 49 of 200 sets [=====>--------------------]  24% (00:00:02 - 9s left)

#> 

#> Validating 50 of 200 sets [=====>--------------------]  25% (00:00:02 - 9s left)

#> 

#> Validating 51 of 200 sets [======>-------------------]  26% (00:00:03 - 9s left)

#> 

#> Validating 52 of 200 sets [======>-------------------]  26% (00:00:03 - 9s left)

#> 

#> Validating 53 of 200 sets [======>-------------------]  26% (00:00:03 - 9s left)

#> 

#> Validating 54 of 200 sets [======>-------------------]  27% (00:00:03 - 9s left)

#> 

#> Validating 55 of 200 sets [======>-------------------]  28% (00:00:03 - 9s left)

#> 

#> Validating 56 of 200 sets [======>-------------------]  28% (00:00:03 - 9s left)

#> 

#> Validating 57 of 200 sets [======>-------------------]  28% (00:00:03 - 9s left)

#> 

#> Validating 58 of 200 sets [=======>------------------]  29% (00:00:03 - 9s left)

#> 

#> Validating 59 of 200 sets [=======>------------------]  30% (00:00:03 - 9s left)

#> 

#> Validating 60 of 200 sets [=======>------------------]  30% (00:00:03 - 9s left)

#> 

#> Validating 61 of 200 sets [=======>------------------]  30% (00:00:03 - 9s left)

#> 

#> Validating 62 of 200 sets [=======>------------------]  31% (00:00:03 - 9s left)

#> 

#> Validating 63 of 200 sets [=======>------------------]  32% (00:00:03 - 8s left)

#> 

#> Validating 64 of 200 sets [=======>------------------]  32% (00:00:03 - 8s left)

#> 

#> Validating 65 of 200 sets [=======>------------------]  32% (00:00:03 - 8s left)

#> 

#> Validating 66 of 200 sets [========>-----------------]  33% (00:00:04 - 8s left)

#> 

#> Validating 67 of 200 sets [========>-----------------]  34% (00:00:04 - 8s left)

#> 

#> Validating 68 of 200 sets [========>-----------------]  34% (00:00:04 - 8s left)

#> 

#> Validating 69 of 200 sets [========>-----------------]  34% (00:00:04 - 8s left)

#> 

#> Validating 70 of 200 sets [========>-----------------]  35% (00:00:04 - 8s left)

#> 

#> Validating 71 of 200 sets [========>-----------------]  36% (00:00:04 - 8s left)

#> 

#> Validating 72 of 200 sets [========>-----------------]  36% (00:00:04 - 8s left)

#> 

#> Validating 73 of 200 sets [========>-----------------]  36% (00:00:04 - 8s left)

#> 

#> Validating 74 of 200 sets [=========>----------------]  37% (00:00:04 - 8s left)

#> 

#> Validating 75 of 200 sets [=========>----------------]  38% (00:00:04 - 8s left)

#> 

#> Validating 76 of 200 sets [=========>----------------]  38% (00:00:04 - 8s left)

#> 

#> Validating 77 of 200 sets [=========>----------------]  38% (00:00:04 - 8s left)

#> 

#> Validating 78 of 200 sets [=========>----------------]  39% (00:00:04 - 7s left)

#> 

#> Validating 79 of 200 sets [=========>----------------]  40% (00:00:04 - 7s left)

#> 

#> Validating 80 of 200 sets [=========>----------------]  40% (00:00:04 - 7s left)

#> 

#> Validating 81 of 200 sets [==========>---------------]  40% (00:00:05 - 7s left)

#> 

#> Validating 82 of 200 sets [==========>---------------]  41% (00:00:05 - 7s left)

#> 

#> Validating 83 of 200 sets [==========>---------------]  42% (00:00:05 - 7s left)

#> 

#> Validating 84 of 200 sets [==========>---------------]  42% (00:00:05 - 7s left)

#> 

#> Validating 85 of 200 sets [==========>---------------]  42% (00:00:05 - 7s left)

#> 

#> Validating 86 of 200 sets [==========>---------------]  43% (00:00:05 - 7s left)

#> 

#> Validating 87 of 200 sets [==========>---------------]  44% (00:00:05 - 7s left)

#> 

#> Validating 88 of 200 sets [==========>---------------]  44% (00:00:05 - 7s left)

#> 

#> Validating 89 of 200 sets [===========>--------------]  44% (00:00:05 - 7s left)

#> 

#> Validating 90 of 200 sets [===========>--------------]  45% (00:00:05 - 7s left)

#> 

#> Validating 91 of 200 sets [===========>--------------]  46% (00:00:05 - 7s left)

#> 

#> Validating 92 of 200 sets [===========>--------------]  46% (00:00:05 - 7s left)

#> 

#> Validating 93 of 200 sets [===========>--------------]  46% (00:00:05 - 7s left)

#> 

#> Validating 94 of 200 sets [===========>--------------]  47% (00:00:05 - 7s left)

#> 

#> Validating 95 of 200 sets [===========>--------------]  48% (00:00:05 - 7s left)

#> 

#> Validating 96 of 200 sets [===========>--------------]  48% (00:00:05 - 6s left)

#> 

#> Validating 97 of 200 sets [============>-------------]  48% (00:00:06 - 6s left)

#> 

#> Validating 98 of 200 sets [============>-------------]  49% (00:00:06 - 6s left)

#> 

#> Validating 99 of 200 sets [============>-------------]  50% (00:00:06 - 6s left)

#> 

#> Validating 100 of 200 sets [===========>-------------]  50% (00:00:06 - 6s left)

#> 

#> Validating 101 of 200 sets [============>------------]  50% (00:00:06 - 6s left)

#> 

#> Validating 102 of 200 sets [============>------------]  51% (00:00:06 - 6s left)

#> 

#> Validating 103 of 200 sets [============>------------]  52% (00:00:06 - 6s left)

#> 

#> Validating 104 of 200 sets [============>------------]  52% (00:00:06 - 6s left)

#> 

#> Validating 105 of 200 sets [============>------------]  52% (00:00:06 - 6s left)

#> 

#> Validating 106 of 200 sets [============>------------]  53% (00:00:06 - 6s left)

#> 

#> Validating 107 of 200 sets [============>------------]  54% (00:00:06 - 6s left)

#> 

#> Validating 108 of 200 sets [=============>-----------]  54% (00:00:06 - 6s left)

#> 

#> Validating 109 of 200 sets [=============>-----------]  55% (00:00:06 - 6s left)

#> 

#> Validating 110 of 200 sets [=============>-----------]  55% (00:00:06 - 6s left)

#> 

#> Validating 111 of 200 sets [=============>-----------]  56% (00:00:06 - 6s left)

#> 

#> Validating 112 of 200 sets [=============>-----------]  56% (00:00:06 - 5s left)

#> 

#> Validating 113 of 200 sets [=============>-----------]  56% (00:00:06 - 5s left)

#> 

#> Validating 114 of 200 sets [=============>-----------]  57% (00:00:07 - 5s left)

#> 

#> Validating 115 of 200 sets [=============>-----------]  57% (00:00:07 - 5s left)

#> 

#> Validating 116 of 200 sets [=============>-----------]  58% (00:00:07 - 5s left)

#> 

#> Validating 117 of 200 sets [==============>----------]  58% (00:00:07 - 5s left)

#> 

#> Validating 118 of 200 sets [==============>----------]  59% (00:00:07 - 5s left)

#> 

#> Validating 119 of 200 sets [==============>----------]  60% (00:00:07 - 5s left)

#> 

#> Validating 120 of 200 sets [==============>----------]  60% (00:00:07 - 5s left)

#> 

#> Validating 121 of 200 sets [==============>----------]  60% (00:00:07 - 5s left)

#> 

#> Validating 122 of 200 sets [==============>----------]  61% (00:00:07 - 5s left)

#> 

#> Validating 123 of 200 sets [==============>----------]  62% (00:00:07 - 5s left)

#> 

#> Validating 124 of 200 sets [===============>---------]  62% (00:00:07 - 5s left)

#> 

#> Validating 125 of 200 sets [===============>---------]  62% (00:00:07 - 5s left)

#> 

#> Validating 126 of 200 sets [===============>---------]  63% (00:00:07 - 5s left)

#> 

#> Validating 127 of 200 sets [===============>---------]  64% (00:00:07 - 4s left)

#> 

#> Validating 128 of 200 sets [===============>---------]  64% (00:00:07 - 4s left)

#> 

#> Validating 129 of 200 sets [===============>---------]  64% (00:00:07 - 4s left)

#> 

#> Validating 130 of 200 sets [===============>---------]  65% (00:00:08 - 4s left)

#> 

#> Validating 131 of 200 sets [===============>---------]  66% (00:00:08 - 4s left)

#> 

#> Validating 132 of 200 sets [===============>---------]  66% (00:00:08 - 4s left)

#> 

#> Validating 133 of 200 sets [================>--------]  66% (00:00:08 - 4s left)

#> 

#> Validating 134 of 200 sets [================>--------]  67% (00:00:08 - 4s left)

#> 

#> Validating 135 of 200 sets [================>--------]  68% (00:00:08 - 4s left)

#> 

#> Validating 136 of 200 sets [================>--------]  68% (00:00:08 - 4s left)

#> 

#> Validating 137 of 200 sets [================>--------]  68% (00:00:08 - 4s left)

#> 

#> Validating 138 of 200 sets [================>--------]  69% (00:00:08 - 4s left)

#> 

#> Validating 139 of 200 sets [================>--------]  70% (00:00:08 - 4s left)

#> 

#> Validating 140 of 200 sets [=================>-------]  70% (00:00:08 - 4s left)

#> 

#> Validating 141 of 200 sets [=================>-------]  70% (00:00:08 - 4s left)

#> 

#> Validating 142 of 200 sets [=================>-------]  71% (00:00:08 - 4s left)

#> 

#> Validating 143 of 200 sets [=================>-------]  72% (00:00:08 - 4s left)

#> 

#> Validating 144 of 200 sets [=================>-------]  72% (00:00:08 - 3s left)

#> 

#> Validating 145 of 200 sets [=================>-------]  72% (00:00:08 - 3s left)

#> 

#> Validating 146 of 200 sets [=================>-------]  73% (00:00:08 - 3s left)

#> 

#> Validating 147 of 200 sets [=================>-------]  74% (00:00:09 - 3s left)

#> 

#> Validating 148 of 200 sets [=================>-------]  74% (00:00:09 - 3s left)

#> 

#> Validating 149 of 200 sets [==================>------]  74% (00:00:09 - 3s left)

#> 

#> Validating 150 of 200 sets [==================>------]  75% (00:00:09 - 3s left)

#> 

#> Validating 151 of 200 sets [==================>------]  76% (00:00:09 - 3s left)

#> 

#> Validating 152 of 200 sets [==================>------]  76% (00:00:09 - 3s left)

#> 

#> Validating 153 of 200 sets [==================>------]  76% (00:00:09 - 3s left)

#> 

#> Validating 154 of 200 sets [==================>------]  77% (00:00:09 - 3s left)

#> 

#> Validating 155 of 200 sets [==================>------]  78% (00:00:09 - 3s left)

#> 

#> Validating 156 of 200 sets [===================>-----]  78% (00:00:09 - 3s left)

#> 

#> Validating 157 of 200 sets [===================>-----]  78% (00:00:09 - 3s left)

#> 

#> Validating 158 of 200 sets [===================>-----]  79% (00:00:09 - 3s left)

#> 

#> Validating 159 of 200 sets [===================>-----]  80% (00:00:09 - 3s left)

#> 

#> Validating 160 of 200 sets [===================>-----]  80% (00:00:09 - 2s left)

#> 

#> Validating 161 of 200 sets [===================>-----]  80% (00:00:09 - 2s left)

#> 

#> Validating 162 of 200 sets [===================>-----]  81% (00:00:09 - 2s left)

#> 

#> Validating 163 of 200 sets [===================>-----]  82% (00:00:10 - 2s left)

#> 

#> Validating 164 of 200 sets [===================>-----]  82% (00:00:10 - 2s left)

#> 

#> Validating 165 of 200 sets [====================>----]  82% (00:00:10 - 2s left)

#> 

#> Validating 166 of 200 sets [====================>----]  83% (00:00:10 - 2s left)

#> 

#> Validating 167 of 200 sets [====================>----]  84% (00:00:10 - 2s left)

#> 

#> Validating 168 of 200 sets [====================>----]  84% (00:00:10 - 2s left)

#> 

#> Validating 169 of 200 sets [====================>----]  84% (00:00:10 - 2s left)

#> 

#> Validating 170 of 200 sets [====================>----]  85% (00:00:10 - 2s left)

#> 

#> Validating 171 of 200 sets [====================>----]  86% (00:00:10 - 2s left)

#> 

#> Validating 172 of 200 sets [=====================>---]  86% (00:00:10 - 2s left)

#> 

#> Validating 173 of 200 sets [=====================>---]  86% (00:00:10 - 2s left)

#> 

#> Validating 174 of 200 sets [=====================>---]  87% (00:00:10 - 2s left)

#> 

#> Validating 175 of 200 sets [=====================>---]  88% (00:00:10 - 2s left)

#> 

#> Validating 176 of 200 sets [=====================>---]  88% (00:00:10 - 1s left)

#> 

#> Validating 177 of 200 sets [=====================>---]  88% (00:00:10 - 1s left)

#> 

#> Validating 178 of 200 sets [=====================>---]  89% (00:00:10 - 1s left)

#> 

#> Validating 179 of 200 sets [=====================>---]  90% (00:00:10 - 1s left)

#> 

#> Validating 180 of 200 sets [=====================>---]  90% (00:00:11 - 1s left)

#> 

#> Validating 181 of 200 sets [======================>--]  90% (00:00:11 - 1s left)

#> 

#> Validating 182 of 200 sets [======================>--]  91% (00:00:11 - 1s left)

#> 

#> Validating 183 of 200 sets [======================>--]  92% (00:00:11 - 1s left)

#> 

#> Validating 184 of 200 sets [======================>--]  92% (00:00:11 - 1s left)

#> 

#> Validating 185 of 200 sets [======================>--]  92% (00:00:11 - 1s left)

#> 

#> Validating 186 of 200 sets [======================>--]  93% (00:00:11 - 1s left)

#> 

#> Validating 187 of 200 sets [======================>--]  94% (00:00:11 - 1s left)

#> 

#> Validating 188 of 200 sets [=======================>-]  94% (00:00:11 - 1s left)

#> 

#> Validating 189 of 200 sets [=======================>-]  94% (00:00:11 - 1s left)

#> 

#> Validating 190 of 200 sets [=======================>-]  95% (00:00:11 - 1s left)

#> 

#> Validating 191 of 200 sets [=======================>-]  96% (00:00:11 - 1s left)

#> 

#> Validating 192 of 200 sets [=======================>-]  96% (00:00:11 - 0s left)

#> 

#> Validating 193 of 200 sets [=======================>-]  96% (00:00:11 - 0s left)

#> 

#> Validating 194 of 200 sets [=======================>-]  97% (00:00:11 - 0s left)

#> 

#> Validating 195 of 200 sets [=======================>-]  98% (00:00:11 - 0s left)

#> 

#> Validating 196 of 200 sets [=======================>-]  98% (00:00:11 - 0s left)

#> 

#> Validating 197 of 200 sets [========================>]  98% (00:00:12 - 0s left)

#> 

#> Validating 198 of 200 sets [========================>]  99% (00:00:12 - 0s left)

#> 

#> Validating 199 of 200 sets [========================>] 100% (00:00:12 - 0s left)

#> 

#> Validating 200 of 200 sets [=========================] 100% (00:00:12 - 0s left)

#> 

# }