Descriptive statistics

Compute the most used measures of central tendency, position, and dispersion.

desc_stat(
  .data = NULL,
  ...,
  by = NULL,
  values = NULL,
  stats = "main",
  hist = FALSE,
  level = 0.95,
  digits = 4,
  na.rm = FALSE,
  verbose = TRUE,
  plot_theme = theme_metan()
)

Arguments

.data
The data to be analyzed. Must be a dataframe or an object of class split_factors.

...
A single variable name or a comma-separated list of unquoted variables names. If no variable is informed, all the numeric from .data variables will be used.

by
One variable (factor) to split the data into subsets. The function is then applied to each subset and returns a list where each element contains the results for one level of the variable in by. To split the data by more than one factor variable, use the function split_factors to pass subsetted data to .data.

values
An alternative way to pass the data to the function. It must be a numeric vector.

stats
The descriptive statistics to show. Defaults to "main" (main statistics). Set to "all" to compute all the statistics bellow or chose one (or more) of the following: 'AV.dev' (average deviation), 'CI.mean' (confidence interval for the mean), 'CV' (coefficient of variation), 'IQR' (interquartile range), 'gm.mean' (geometric mean), 'hm.mean' (harmonic mean), 'Kurt' (kurtosis), 'mad' (median absolute deviation), 'max' (maximum value), 'mean' (arithmetic mean), 'median' (median), 'min' (minimum value), 'n' (the length of the data), 'Q2.5' (the percentile 2.5%), 'Q25' (the first quartile, Q1), 'Q75' (the third quartile, Q3), 'Q97.5' (the percentile 97.5%), range (The range of data), 'SD.amo' (the sample standard deviation), 'SD.pop' (the population standard deviation), 'SE.mean' (the standard error of the mean), 'skew' (the skewness), sum (the sum of the values), sum.dev (the sum of the absolute deviations), sum.sq.dev (the sum of the squared deviations), valid.n (The size of sample with valid number (not NA), 'var.amo' (the sample variance), 'var.pop' (the population variance). Use a comma-separated vector of names to select the statistics. For example, stats = c("median, mean, CV, n")

hist
Logical argument defaults to FALSE. If hist = TRUE then a histogram is created for each selected variable.

level
The confidence level to compute the confidence interval of mean. Defaults to 0.95.

digits
The number of significant digits.

na.rm
Logical. Should missing values be removed?

verbose
Logical argument. If verbose = FALSE the code is run silently.

plot_theme
The graphical theme of the plot. Default is plot_theme = theme_metan(). For more details, see theme.

Value

A tibble with the statistics in the lines and variables in columns. If .data is an object of class split_factors, then the statistics will be shown for each level of the grouping variable in the function split_factors to pass subsetted data.to pass subsetted data to code.data.to pass subsetted data to code.data.

Details

In cases when the statistics are computed for more than two variables with data coming from the function split_factors to pass subsetted data.to pass subsetted data to code.data.to pass subsetted data to code.data.the results are returned in a long format. Thus, use the function desc_wider to convert it into a wide format (levels of the factors in the rows and statistics in the columns).

Examples

# \donttest{
library(metan)

desc_stat(data_ge2, TKW)

#> # A tibble: 7 x 2
#>   Statistic     TKW
#>   <chr>       <dbl>
#> 1 mean       339.  
#> 2 min        218.  
#> 3 median     342.  
#> 4 max        452.  
#> 5 var.amo   2215.  
#> 6 se.mean      3.77
#> 7 cv          13.9 

# Compute the main statistics
# Use a numeric vector as input data
vect <- data_ge2$TKW
desc_stat(values = vect)

#> # A tibble: 7 x 2
#>   Statistic   value
#>   <chr>       <dbl>
#> 1 mean       339.  
#> 2 min        218.  
#> 3 median     342.  
#> 4 max        452.  
#> 5 var.amo   2215.  
#> 6 se.mean      3.77
#> 7 cv          13.9 

# Select specific statistics
desc_stat(values = c(12, 13, 19, 21, 8, NA, 23, NA),
          na.rm = TRUE,
          stats = c('mean, se.mean, cv, n, valid.n'))

#> # A tibble: 5 x 2
#>   Statistic value
#>   <chr>     <dbl>
#> 1 n          8   
#> 2 valid.n    6   
#> 3 mean      16   
#> 4 se.mean    2.39
#> 5 cv        36.7 

# Compute the main statistics for each level of "ENV"
stats <-
  desc_stat(data_ge2,
            EP, EL, EH, ED, PH, CD,
            by = ENV,
            verbose = FALSE)

# To get a 'wide' format with the statistics of the variable EP above.
desc_wider(stats, PH)

#> # A tibble: 4 x 8
#>   ENV      cv   max  mean median   min se.mean var.amo
#>   <chr> <dbl> <dbl> <dbl>  <dbl> <dbl>   <dbl>   <dbl>
#> 1 A1     4.42  3.04  2.79   2.78  2.55  0.0198  0.0153
#> 2 A2    16.2   3.04  2.46   2.27  1.81  0.0639  0.159 
#> 3 A3    10.3   2.77  2.17   2.10  1.71  0.0358  0.0499
#> 4 A4     6.61  2.81  2.52   2.47  2.11  0.0266  0.0277

# Compute all the statistics for each combination of "ENV" and "GEN"
# All the numeric variables in .data

stats_all <-
  data_ge2 %>%
  split_factors(ENV, GEN) %>%
  desc_stat(stats = "all", verbose = FALSE)

#> Warning: Use 'keep_factors = TRUE' to keep the columns ENV GEN REP in the grouped data. 

desc_wider(stats_all, PH)

#> # A tibble: 52 x 30
#>    ENV   GEN   av.dev ci.mean    cv gm.mean hm.mean    iqr  kurt    mad   max
#>    <chr> <chr>  <dbl>   <dbl> <dbl>   <dbl>   <dbl>  <dbl> <dbl>  <dbl> <dbl>
#>  1 "A1 " " H1" 0.0996  0.333  4.93     2.72    2.72 0.130  -2.33 0.107   2.87
#>  2 "A1 " " H1~ 0.0396  0.136  1.97     2.78    2.78 0.0540 -2.33 0.0563  2.83
#>  3 "A1 " " H1~ 0.0191  0.0675 0.988    2.75    2.75 0.0270 -2.33 0.0326  2.77
#>  4 "A1 " " H1~ 0.0898  0.299  4.47     2.69    2.69 0.116  -2.33 0.0890  2.79
#>  5 "A1 " " H1~ 0.107   0.366  5.31     2.77    2.77 0.145  -2.33 0.148   2.93
#>  6 "A1 " " H2" 0.161   0.521  7.51     2.79    2.78 0.186  -2.33 0.0267  2.92
#>  7 "A1 " " H3" 0.0782  0.281  3.86     2.93    2.93 0.113  -2.33 0.148   3.04
#>  8 "A1 " " H4" 0.102   0.352  4.95     2.87    2.86 0.14   -2.33 0.148   3.02
#>  9 "A1 " " H5" 0.0462  0.154  2.19     2.83    2.83 0.06   -2.33 0.0474  2.90
#> 10 "A1 " " H6" 0.0862  0.291  4.23     2.77    2.77 0.114  -2.33 0.101   2.90
#> # ... with 42 more rows, and 19 more variables: mean <dbl>, median <dbl>,
#> #   min <dbl>, n <dbl>, q2.5 <dbl>, q25 <dbl>, q75 <dbl>, q97.5 <dbl>,
#> #   range <dbl>, sd.amo <dbl>, sd.pop <dbl>, se.mean <dbl>, skew <dbl>,
#> #   sum <dbl>, sum.dev <dbl>, sum.sq.dev <dbl>, valid.n <dbl>, var.amo <dbl>,
#> #   var.pop <dbl>

# }