Multi-trait selection index

Multitrait index based on factor analysis and ideotype-design proposed by Rocha et al. (2018).

Usage

fai_blup(
  .data,
  use_data = "blup",
  DI = NULL,
  UI = NULL,
  SI = 15,
  mineval = 1,
  verbose = TRUE
)

Arguments

.data

An object of class waasb or a two-way table with genotypes in the rows and traits in columns. In the last case the row names must contain the genotypes names.

use_data

Define which data to use If .data is an object of class gamem. Defaults to "blup" (the BLUPs for genotypes). Use "pheno" to use phenotypic means instead BLUPs for computing the index.

DI, UI

A vector of the same length of .data to construct the desirable (DI) and undesirable (UI) ideotypes. For each element of the vector, allowed values are 'max', 'min', 'mean', or a numeric value. Use a comma-separated vector of text. For example, DI = c("max, max, min, min"). By default, DI is set to "max" for all traits and UI is set to "min" for all traits.

SI

An integer (0-100). The selection intensity in percentage of the total number of genotypes. Defaults to 15.

mineval

The minimum value so that an eigenvector is retained in the factor analysis.

verbose

Logical value. If TRUE some results are shown in console.

Value

An object of class fai_blup with the following items:

data The data (BLUPS) used to compute the index.

eigen The eigenvalues and explained variance for each axis.

FA The results of the factor analysis.

canonical_loadings The canonical loadings for each factor retained.

FAI A list with the FAI-BLUP index for each ideotype design.

sel_dif_trait A list with the selection differential for each ideotype design.

sel_gen The selected genotypes.

ideotype_construction A list with the construction of the ideotypes.

total_gain A list with the total gain for variables to be increased or decreased.

References

Rocha, J.R.A.S.C.R, J.C. Machado, and P.C.S. Carneiro. 2018. Multitrait index based on factor analysis and ideotype-design: proposal and application on elephant grass breeding for bioenergy. GCB Bioenergy 10:52-60. doi:10.1111/gcbb.12443

Author

Tiago Olivoto tiagoolivoto@gmail.com

Examples

# \donttest{
library(metan)

mod <- waasb(data_ge,
             env = ENV,
             gen = GEN,
             rep = REP,
             resp = c(GY, HM))
#> Evaluating trait GY |======================                      | 50% 00:00:01 Evaluating trait HM |============================================| 100% 00:00:02 
#> Method: REML/BLUP
#> Random effects: GEN, GEN:ENV
#> Fixed effects: ENV, REP(ENV)
#> Denominador DF: Satterthwaite's method
#> ---------------------------------------------------------------------------
#> P-values for Likelihood Ratio Test of the analyzed traits
#> ---------------------------------------------------------------------------
#>     model       GY       HM
#>  COMPLETE       NA       NA
#>       GEN 1.11e-05 5.07e-03
#>   GEN:ENV 2.15e-11 2.27e-15
#> ---------------------------------------------------------------------------
#> All variables with significant (p < 0.05) genotype-vs-environment interaction

FAI <- fai_blup(mod,
                SI = 15,
                DI = c('max, max'),
                UI = c('min, min'))
#> 
#> -----------------------------------------------------------------------------------
#> Principal Component Analysis
#> -----------------------------------------------------------------------------------
#>     eigen.values cumulative.var
#> PC1          1.1          55.23
#> PC2          0.9         100.00
#> 
#> -----------------------------------------------------------------------------------
#> Factor Analysis
#> -----------------------------------------------------------------------------------
#>      FA1 comunalits
#> GY -0.74       0.55
#> HM  0.74       0.55
#> 
#> -----------------------------------------------------------------------------------
#> Comunalit Mean: 0.5523038 
#> Selection differential
#> -----------------------------------------------------------------------------------
#>   VAR Factor        Xo        Xs          SD     SDperc    sense goal
#> 1  GY      1  2.674242  2.594199 -0.08004274 -2.9931005 increase    0
#> 2  HM      1 48.088286 48.005568 -0.08271774 -0.1720122 increase    0
#> 
#> -----------------------------------------------------------------------------------
#> Selected genotypes
#> G4 G9
#> -----------------------------------------------------------------------------------
# }