Geometric adaptability index

Performs a stability analysis based on the geometric mean (GAI), according to the following model: $$GAI = \sqrt[E]{{{{\bar Y}_{1.}} \cdot {{\bar Y}_{2.}} \cdot ... \cdot {{\bar Y}_{i.}}}}$$

gai(.data, env, gen, rep, resp, verbose = TRUE)

Arguments

.data
The dataset containing the columns related to Environments, Genotypes, replication/block and response variable(s).

env
The name of the column that contains the levels of the environments.

gen
The name of the column that contains the levels of the genotypes.

rep
The name of the column that contains the levels of the replications/blocks.

resp
The response variable(s). To analyze multiple variables in a single procedure use, for example, resp = c(var1, var2, var3).

verbose
Logical argument. If verbose = FALSE the code will run silently.

Value

An object of class gai, which is a list containing the results for each variable used in the argument resp. For each variable, a tibble with the following columns is returned.

GEN the genotype's code.

GAI Geometric adaptability index

GAI_R The rank for the GAI value.

References

Shahbazi, E. 2019. Genotype selection and stability analysis for seed yield of Nigella sativa using parametric and non-parametric statistics. Sci. Hortic. (Amsterdam). 253:172-179. doi:10.1016/j.scienta.2019.04.047.

Examples

library(metan)
out <- gai(data_ge2, ENV, GEN, REP, c(EH, PH, EL, CD, ED, NKE))

#> Evaluating variable EH 0 % 
#> Evaluating variable PH 20 % 
#> Evaluating variable EL 40 % 
#> Evaluating variable CD 60 % 
#> Evaluating variable ED 80 % 
#> Evaluating variable NKE 100 %