Genotype analysis by mixed-effect models

Analysis of genotypes in single experiments using mixed-effect models with estimation of genetic parameters.

gamem(.data, gen, rep, resp, block = NULL, prob = 0.05, verbose = TRUE)

Arguments

.data
The dataset containing the columns related to, Genotypes, replication/block and response variable(s).

gen
The name of the column that contains the levels of the genotypes, that will be treated as random effect.

rep
The name of the column that contains the levels of the replications (assumed to be fixed).

resp
The response variable(s). To analyze multiple variables in a single procedure a vector of variables may be used. For example resp = c(var1, var2, var3). Select helpers are also allowed.

block
Defaults to NULL. In this case, a randomized complete block design is considered. If block is informed, then an alpha-lattice design is employed considering block as random to make use of inter-block information, whereas the complete replicate effect is always taken as fixed, as no inter-replicate information was to be recovered (Mohring et al., 2015).

prob
The probability for estimating confidence interval for BLUP's prediction.

verbose
Logical argument. If verbose = FALSE the code are run silently.

Value

An object of class gamem, which is a list with the following items for each element (variable):

fixed: Test for fixed effects.

random: Variance components for random effects.

LRT: The Likelihood Ratio Test for the random effects.

BLUPgen: The estimated BLUPS for genotypes

ranef: The random effects of the model

Details: A tibble with the following data: Ngen, the number of genotypes; OVmean, the grand mean; Min, the minimum observed (returning the genotype and replication/block); Max the maximum observed, MinGEN the winner genotype, MaxGEN, the loser genotype.

ESTIMATES: A tibble with the values for the genotypic variance, block-within-replicate variance (if an alpha-lattice design is used by informing the block in block), the residual variance and their respective contribution to the phenotypic variance; broad-sence heritability, heritability on the entry-mean basis, genotypic coefficient of variation residual coefficient of variation and ratio between genotypic and residual coefficient of variation.

residuals: The residuals of the model.

Details

gamem analyses data from a one-way genotype testing experiment. By default, a randomized complete block design is used according to the following model: $$Y_{ij} = m + g_i + r_j + e_{ij}$$ where \(Y_{ij}\) is the response variable of the ith genotype in the jth block; m is the grand mean (fixed); \(g_i\) is the effect of the ith genotype (assumed to be random); \(r_j\) is the effect of the jth replicate (assumed to be fixed); and \(e_{ij}\) is the random error.

When block is informed, then a resolvable alpha design is implemented, according to the following model:

$$Y_{ijk} = m + g_i + r_j + b_{jk} + e_{ijk}$$ where where \(y_{ijk}\) is the response variable of the ith genotype in the kth block of the jth replicate; m is the intercept, \(t_i\) is the effect for the ith genotype \(r_j\) is the effect of the jth replicate, \(b_{jk}\) is the effect of the kth incomplete block of the jth replicate, and \(e_{ijk}\) is the plot error effect corresponding to \(y_{ijk}\).

References

Mohring, J., E. Williams, and H.-P. Piepho. 2015. Inter-block information: to recover or not to recover it? TAG. Theor. Appl. Genet. 128:1541-54. doi:10.1007/s00122-015-2530-0

See also

Examples

# \donttest{
library(metan)

# fitting the model considering an RCBD
# Genotype as random effects

rcbd <- gamem(data_g,
             gen = GEN,
             rep = REP,
             resp = c(PH, ED, EL, CL, CW, KW, NR, TKW, NKE))

#> Model:  Y ~ REP + (1 | GEN) 
#> ---------------------------------------------------------------------------
#> P-values for Likelihood Ratio Test of the analyzed traits
#> ---------------------------------------------------------------------------
#>     model    PH       ED    EL       CL       CW     KW     NR     TKW     NKE
#>  Complete    NA       NA    NA       NA       NA     NA     NA      NA      NA
#>  Genotype 0.051 2.73e-05 0.786 2.25e-06 1.24e-05 0.0253 0.0056 0.00955 0.00952
#> ---------------------------------------------------------------------------
#> Variables with nonsignificant Genotype effect
#> PH EL 
#> ---------------------------------------------------------------------------

# Likelihood ratio test for random effects
## Statistic
get_model_data(rcbd, "lrt")

#> Class of the model: gamem

#> Variable extracted: lrt

#> # A tibble: 2 x 10
#>   model       PH    ED      EL    CL    CW    KW    NR   TKW   NKE
#>   <chr>    <dbl> <dbl>   <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1 Complete NA     NA   NA       NA    NA   NA    NA    NA    NA   
#> 2 Genotype  3.81  17.6  0.0735  22.4  19.1  5.00  7.67  6.72  6.72

## P-value
get_model_data(rcbd, "pval_lrt")

#> Class of the model: gamem

#> Variable extracted: pval_lrt

#> # A tibble: 2 x 10
#>   model      PH       ED     EL       CL       CW      KW       NR      TKW
#>   <chr>   <dbl>    <dbl>  <dbl>    <dbl>    <dbl>   <dbl>    <dbl>    <dbl>
#> 1 Comp~ NA      NA       NA     NA       NA       NA      NA       NA      
#> 2 Geno~  0.0510  2.73e-5  0.786  2.25e-6  1.24e-5  0.0253  0.00560  0.00955
#> # ... with 1 more variable: NKE <dbl>

# Variance components
get_model_data(rcbd, "vcomp")

#> Class of the model: gamem

#> Variable extracted: vcomp

#> # A tibble: 2 x 10
#>   Group        PH    ED     EL    CL    CW    KW    NR   TKW   NKE
#>   <chr>     <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1 GEN      0.0171  5.37 0.0472  4.27 18.5   181.  1.18  841. 1982.
#> 2 Residual 0.0328  2.43 0.984   1.41  7.54  280.  1.27 1018. 2399.

# Genetic parameters
get_model_data(rcbd, "genpar")

#> Class of the model: gamem

#> Variable extracted: genpar

#> # A tibble: 11 x 10
#>    Parameters      PH     ED      EL     CL     CW      KW     NR     TKW
#>    <chr>        <dbl>  <dbl>   <dbl>  <dbl>  <dbl>   <dbl>  <dbl>   <dbl>
#>  1 Gen_var     0.0171  5.37   0.0472  4.27  18.5   181.     1.18  8.41e+2
#>  2 Gen (%)    34.3    68.8    4.58   75.1   71.0    39.2   48.2   4.52e+1
#>  3 Res_var     0.0328  2.43   0.984   1.41   7.54  280.     1.27  1.02e+3
#>  4 Res (%)    65.7    31.2   95.4    24.9   29.0    60.8   51.8   5.48e+1
#>  5 Phen_var    0.0498  7.80   1.03    5.68  26.0   461.     2.45  1.86e+3
#>  6 H2          0.343   0.688  0.0458  0.751  0.710   0.392  0.482 4.52e-1
#>  7 H2mg        0.610   0.869  0.126   0.901  0.880   0.659  0.736 7.12e-1
#>  8 Accuracy    0.781   0.932  0.355   0.949  0.938   0.812  0.858 8.44e-1
#>  9 CVg         6.03    4.84   1.48    7.26  20.7     9.16   6.88  9.13e+0
#> 10 CVr         8.35    3.26   6.76    4.18  13.2    11.4    7.14  1.00e+1
#> 11 CV ratio    0.722   1.49   0.219   1.74   1.56    0.803  0.964 9.09e-1
#> # ... with 1 more variable: NKE <dbl>

# random effects
get_model_data(rcbd, "ranef")

#> Class of the model: gamem

#> Variable extracted: ranef

#> $GEN
#> # A tibble: 13 x 10
#>    GEN         PH     ED       EL     CL     CW     KW      NR    TKW    NKE
#>    <fct>    <dbl>  <dbl>    <dbl>  <dbl>  <dbl>  <dbl>   <dbl>  <dbl>  <dbl>
#>  1 H1     0.0188   2.36   0.0208   2.21   5.33    6.60  0.0604  36.4  -30.5 
#>  2 H10   -0.0784  -3.48  -0.0858  -3.31  -7.48  -17.3  -0.332  -50.2   14.9 
#>  3 H11   -0.0398  -0.717 -0.0530  -1.89  -4.20   -4.02  0.355  -20.7   14.1 
#>  4 H12    0.161   -0.115 -0.0895  -2.36  -2.63    1.02  0.355  -24.3   34.9 
#>  5 H13    0.264    2.44   0.0905  -1.05   0.673  22.9   2.02     1.36  79.1 
#>  6 H2    -0.00767  2.40   0.0922   1.53   1.12    8.90 -0.0377  20.1   -1.11
#>  7 H3    -0.0752  -0.696  0.00822  0.109 -2.08   -5.16 -0.725   13.1  -33.4 
#>  8 H4    -0.0715  -1.78   0.109   -0.693 -1.16   -2.21 -1.41    -7.83   4.49
#>  9 H5    -0.0439   1.96   0.00319  1.65   5.23    9.43  1.24    -8.71  48.9 
#> 10 H6    -0.00807  1.85  -0.143    3.11   1.99   -6.43  0.453    7.35 -34.5 
#> 11 H7     0.00657  0.809 -0.0161   1.60   5.74    5.83 -0.136   28.7  -19.4 
#> 12 H8    -0.0406  -1.53   0.00487  0.527  1.11   -3.55 -0.921   21.4  -43.2 
#> 13 H9    -0.0845  -3.49   0.0586  -1.42  -3.64  -16.1  -0.921  -16.6  -34.3 
#> 

# Predicted values
predict(rcbd)

#> # A tibble: 39 x 11
#>    GEN   REP      PH    ED    EL    CL    CW    KW    NR   TKW   NKE
#>    <fct> <fct> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1 H1    1      2.12  50.5  14.9  31.5  26.9  156.  15.8  360.  436.
#>  2 H1    2      2.20  49.5  14.5  29.9  24.4  146.  16.1  343.  428.
#>  3 H1    3      2.24  50.7  14.6  30.6  27.1  159.  15.7  359.  449.
#>  4 H10   1      2.02  44.6  14.8  26.0  14.0  132.  15.4  274.  481.
#>  5 H10   2      2.10  43.7  14.4  24.4  11.6  122.  15.7  257.  473.
#>  6 H10   3      2.14  44.9  14.5  25.1  14.2  135.  15.3  272.  494.
#>  7 H11   1      2.06  47.4  14.9  27.4  17.3  145.  16.1  303.  481.
#>  8 H11   2      2.14  46.4  14.5  25.8  14.9  135.  16.4  286.  472.
#>  9 H11   3      2.18  47.6  14.5  26.5  17.5  148.  16.0  302.  493.
#> 10 H12   1      2.26  48.0  14.8  26.9  18.9  150.  16.1  300.  501.
#> # ... with 29 more rows

# fitting the model considering an alpha-lattice design
# Genotype and block-within-replicate as random effects
# Note that block effect was now informed.

alpha <- gamem(data_alpha,
               gen = GEN,
               rep = REP,
               block = BLOCK,
               resp = YIELD)

#> Model:  Y ~ (1 | GEN) + REP + (1 | REP:BLOCK) 
#> ---------------------------------------------------------------------------
#> P-values for Likelihood Ratio Test of the analyzed traits
#> ---------------------------------------------------------------------------
#>      model    YIELD
#>   Complete       NA
#>   Genotype 1.18e-06
#>  rep:block 3.35e-03
#> ---------------------------------------------------------------------------
#> All variables with significant (p < 0.05) genotype effect

# Genetic parameters
get_model_data(alpha, "genpar")

#> Class of the model: gamem

#> Variable extracted: genpar

#> # A tibble: 13 x 2
#>    Parameters      YIELD
#>    <chr>           <dbl>
#>  1 Gen_var        0.143 
#>  2 Gen (%)       48.5   
#>  3 rep:block_var  0.0702
#>  4 rep:block (%) 23.8   
#>  5 Res_var        0.0816
#>  6 Res (%)       27.7   
#>  7 Phen_var       0.295 
#>  8 H2             0.485 
#>  9 H2mg           0.798 
#> 10 Accuracy       0.893 
#> 11 CVg            8.44  
#> 12 CVr            6.38  
#> 13 CV ratio       1.32  

# Random effects
get_model_data(alpha, "ranef")

#> Class of the model: gamem

#> Variable extracted: ranef

#> $GEN
#> # A tibble: 24 x 2
#>    GEN      YIELD
#>    <fct>    <dbl>
#>  1 G11   -0.196  
#>  2 G04    0.00613
#>  3 G05    0.475  
#>  4 G22    0.0241 
#>  5 G21    0.256  
#>  6 G10   -0.0894 
#>  7 G20   -0.339  
#>  8 G02    0.00496
#>  9 G23   -0.177  
#> 10 G14    0.243  
#> # ... with 14 more rows
#> 
#> $REP_BLOCK
#> # A tibble: 18 x 3
#>    REP   BLOCK    YIELD
#>    <fct> <fct>    <dbl>
#>  1 R1    B1     0.123  
#>  2 R1    B2    -0.141  
#>  3 R1    B3    -0.150  
#>  4 R1    B4    -0.107  
#>  5 R1    B5     0.0737 
#>  6 R1    B6     0.202  
#>  7 R2    B1    -0.533  
#>  8 R2    B2    -0.301  
#>  9 R2    B3     0.243  
#> 10 R2    B4     0.135  
#> 11 R2    B5     0.275  
#> 12 R2    B6     0.180  
#> 13 R3    B1     0.0506 
#> 14 R3    B2    -0.0478 
#> 15 R3    B3     0.151  
#> 16 R3    B4     0.0538 
#> 17 R3    B5    -0.00805
#> 18 R3    B6    -0.200  
#> 
#> $GEN_REP_BLOCK
#> # A tibble: 72 x 4
#>    GEN   REP   BLOCK   YIELD
#>    <fct> <fct> <fct>   <dbl>
#>  1 G11   R1    B1    -0.0733
#>  2 G04   R1    B1     0.129 
#>  3 G05   R1    B1     0.598 
#>  4 G22   R1    B1     0.147 
#>  5 G21   R1    B2     0.115 
#>  6 G10   R1    B2    -0.231 
#>  7 G20   R1    B2    -0.480 
#>  8 G02   R1    B2    -0.136 
#>  9 G23   R1    B3    -0.327 
#> 10 G14   R1    B3     0.0930
#> # ... with 62 more rows
#> 

# }