Adjusted Coefficient of Variation as yield stability index

Performs a stability analysis based on the scale-adjusted coefficient of variation (Doring and Reckling, 2018). For more details see acv()

ge_acv(.data, env, gen, resp, verbose = TRUE)

Arguments

.data
The dataset containing the columns related to Environments, Genotypes and response variable(s).

env
The name of the column that contains the levels of the environments.

gen
The name of the column that contains the levels of the genotypes.

resp
The response variable(s). To analyze multiple variables in a single procedure use, for example, resp = c(var1, var2, var3).

verbose
Logical argument. If verbose = FALSE the code will run silently.

Value

An object of class ge_acv, which is a list containing the results for each variable used in the argument resp. For each variable, a tibble with the following columns is returned.

GEN the genotype's code.

ACV The adjusted coefficient of variation

ACV_R The rank for the ACV value.

References

Doring, T.F., and M. Reckling. 2018. Detecting global trends of cereal yield stability by adjusting the coefficient of variation. Eur. J. Agron. 99: 30-36. doi: 10.1016/j.eja.2018.06.007

Author

Tiago Olivoto tiagoolivoto@gmail.com

Examples

# \donttest{
library(metan)
out <- ge_acv(data_ge2, ENV, GEN, c(EH, PH, EL, CD, ED, NKE))

#> Evaluating trait EH |=======                                     | 17% 00:00:00 Evaluating trait PH |===============                             | 33% 00:00:00 Evaluating trait EL |======================                      | 50% 00:00:00 Evaluating trait CD |=============================               | 67% 00:00:00 Evaluating trait ED |=====================================       | 83% 00:00:00 Evaluating trait NKE |===========================================| 100% 00:00:00 

gmd(out)

#> Class of the model: ge_acv

#> Variable extracted: ACV

#> # A tibble: 13 x 7
#>    GEN      EH    PH    EL    CD    ED   NKE
#>    <chr> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1 H1     22.9 12.5   1.71 0.953  1.65 13.2 
#>  2 H10    23.4 14.1   5.98 4.89   6.05  2.22
#>  3 H11    19.1 12.7   6.68 5.00   3.61  4.19
#>  4 H12    20.8 10.2   5.22 5.05   2.53 10.2 
#>  5 H13    14.7  9.19  4.25 4.63   6.11 17.2 
#>  6 H2     21.3 14.1   3.14 4.37   6.58 14.7 
#>  7 H3     25.7 17.4   8.59 6.74   4.07 14.3 
#>  8 H4     24.9 15.4   4.51 3.99   4.50 12.4 
#>  9 H5     21.1 13.2   4.92 2.19   3.04  2.99
#> 10 H6     14.5 12.4  10.8  8.11   6.33 19.1 
#> 11 H7     17.3 12.2   7.33 6.61   3.43  8.11
#> 12 H8     21.9 14.1   7.69 7.48   4.80 12.3 
#> 13 H9     23.4 15.7   7.02 7.00   4.64 13.2 

# }