Details for genotype-environment trials

Provide details for genotype-environment trials

Usage

ge_details(.data, env, gen, resp)

Arguments

.data

The dataset containing the columns related to Environments, Genotypes, replication/block and response variable(s).

env

The name of the column that contains the levels of the environments.

gen

The name of the column that contains the levels of the genotypes.

resp

The response variable(s). To analyze multiple variables in a single procedure a vector of variables may be used. For example resp = c(var1, var2, var3). Select helpers are also allowed.

Value

A tibble with the following results for each variable:

Mean: The grand mean.

SE: The standard error of the mean.

SD: The standard deviation.

CV: The coefficient of variation.

Min,Max: The minimum and maximum value, indicating the genotype and environment of occurence.

MinENV, MinGEN: The environment and genotype with the lower mean.

MaxENV, MaxGEN: The environment and genotype with the higher mean.

Author

Tiago Olivoto tiagoolivoto@gmail.com

Examples

# \donttest{
library(metan)
details <- ge_details(data_ge2, ENV, GEN, everything())
print(details)
#> # A tibble: 10 × 16
#>    Parameters PH     EH    EP    EL    ED    CL    CD    CW    KW    NR    NKR  
#>    <chr>      <chr>  <chr> <chr> <chr> <chr> <chr> <chr> <chr> <chr> <chr> <chr>
#>  1 Mean       "2.48" "1.3… "0.5… "15.… "49.… "29.… "15.… "24.… "172… "16.… "32.…
#>  2 SE         "0.03" "0.0… "0"   "0.1" "0.2… "0.1… "0.0… "0.5" "2.6… "0.1… "0.2…
#>  3 SD         "0.33" "0.2… "0.0… "1.2… "2.7… "2.3" "1.1… "6.2… "32.… "1.6… "3.4…
#>  4 CV         "13.4… "21.… "10.… "8.2… "5.5… "7.9… "7.3… "25.… "18.… "10.… "10.…
#>  5 Min        "1.71… "0.7… "0.3… "11.… "43.… "23.… "12.… "11.… "105… "12.… "23.…
#>  6 Max        "3.04… "1.8… "0.6… "17.… "54.… "34.… "18.… "38.… "250… "21.… "42 …
#>  7 MinENV     "A3 (… "A3 … "A3 … "A3 … "A3 … "A3 … "A3 … "A3 … "A3 … "A3 … "A3 …
#>  8 MaxENV     "A1 (… "A1 … "A1 … "A1 … "A1 … "A1 … "A1 … "A1 … "A1 … "A1 … "A1 …
#>  9 MinGEN     "H10 … "H8 … "H8 … "H12… "H9 … "H12… "H12… "H11… "H9 … "H9 … "H12…
#> 10 MaxGEN     "H1 (… "H1 … "H1 … "H6 … "H6 … "H6 … "H5 … "H5 … "H6 … "H13… "H4 …
#> # … with 4 more variables: CDED <chr>, PERK <chr>, TKW <chr>, NKE <chr>
# }