Stability analysis and environment estratification

This function computes the stability analysis and environmental stratitication using factor analysis as proposed by Murakami and Cruz (2004).

ge_factanal(.data, env, gen, rep, resp, mineval = 1, verbose = TRUE)

Arguments

.data
The dataset containing the columns related to Environments, Genotypes, replication/block and response variable(s)

env
The name of the column that contains the levels of the environments.

gen
The name of the column that contains the levels of the genotypes.

rep
The name of the column that contains the levels of the replications/blocks

resp
The response variable(s). To analyze multiple variables in a single procedure use, for example, resp = c(var1, var2, var3).

mineval
The minimum value so that an eigenvector is retained in the factor analysis.

verbose
Logical argument. If verbose = FALSE the code will run silently.

References

Murakami, D.M.D., and C.D.C. Cruz. 2004. Proposal of methodologies for environment stratification and analysis of genotype adaptability. Crop Breed. Appl. Biotechnol. 4:7-11.

See also

Examples

library(metan)
model = ge_factanal(data_ge2,
                    env = ENV,
                    gen = GEN,
                    rep = REP,
                    resp = PH)