Power Law Residuals as yield stability index

Performs a stability analysis based on the Power Law Residuals (POLAR) statistics (Doring et al., 2015). POLAR is the residuals from the linear regression of \(log(\sigma^2\)) against \(log(\mu\)) and can be used as a measure of crop stability with lower stability (relative to all samples with that mean yield) indicated by more positive POLAR values, and higher stability (relative to all samples with that mean yield) indicated by more negative POLAR values.

Usage

ge_polar(.data, env, gen, resp, base = 10, verbose = TRUE)

Arguments

.data

The dataset containing the columns related to Environments, Genotypes and response variable(s).

env

The name of the column that contains the levels of the environments.

gen

The name of the column that contains the levels of the genotypes.

resp

The response variable(s). To analyze multiple variables in a single procedure use, for example, resp = c(var1, var2, var3).

base

The base with respect to which logarithms are computed. Defaults to 10.

verbose

Logical argument. If verbose = FALSE the code will run silently.

Value

An object of class ge_acv, which is a list containing the results for each variable used in the argument resp. For each variable, a tibble with the following columns is returned.

GEN the genotype's code.

POLAR The Power Law Residuals

POLAR_R The rank for the ACV value.

References

Doring, T.F., S. Knapp, and J.E. Cohen. 2015. Taylor's power law and the stability of crop yields. F. Crop. Res. 183: 294-302. doi: 10.1016/j.fcr.2015.08.005

Author

Tiago Olivoto tiagoolivoto@gmail.com

Examples

# \donttest{
library(metan)
out <- ge_polar(data_ge2, ENV, GEN, c(EH, PH, EL, CD, ED, NKE))
#> Evaluating trait EH |=======                                     | 17% 00:00:00 Evaluating trait PH |===============                             | 33% 00:00:00 Evaluating trait EL |======================                      | 50% 00:00:00 Evaluating trait CD |=============================               | 67% 00:00:00 Evaluating trait ED |=====================================       | 83% 00:00:00 Evaluating trait NKE |===========================================| 100% 00:00:00 
gmd(out)
#> Class of the model: ge_polar
#> Variable extracted: POLAR
#> # A tibble: 13 x 7
#>    GEN         EH       PH      EL       CD      ED     NKE
#>    <chr>    <dbl>    <dbl>   <dbl>    <dbl>   <dbl>   <dbl>
#>  1 H1     0.0950  -0.0439  -1.01   -1.36    -0.796   0.299 
#>  2 H10    0.115    0.0603   0.0782  0.0570   0.334  -1.25  
#>  3 H11   -0.0634  -0.0287   0.174   0.0756  -0.115  -0.699 
#>  4 H12    0.00927 -0.220   -0.0406  0.0842  -0.424   0.0711
#>  5 H13   -0.292   -0.311   -0.218   0.00976  0.342   0.530 
#>  6 H2     0.0316   0.0582  -0.480  -0.0412   0.407   0.394 
#>  7 H3     0.194    0.246    0.393   0.336   -0.0112  0.367 
#>  8 H4     0.167    0.139   -0.166  -0.120    0.0759  0.248 
#>  9 H5     0.0240   0.00211 -0.0905 -0.643   -0.264  -0.991 
#> 10 H6    -0.303   -0.0475   0.590   0.496    0.373   0.618 
#> 11 H7    -0.147   -0.0672   0.255   0.318   -0.158  -0.124 
#> 12 H8     0.0553   0.0605   0.297   0.425    0.133   0.239 
#> 13 H9     0.114    0.153    0.218   0.367    0.103   0.298 
# }