Eberhart and Russell's regression model

Regression-based stability analysis using the Eberhart and Russell (1966) model.

ge_reg(.data, env, gen, rep, resp, verbose = TRUE)

Arguments

.data
The dataset containing the columns related to Environments, Genotypes, replication/block and response variable(s)

env
The name of the column that contains the levels of the environments.

gen
The name of the column that contains the levels of the genotypes.

rep
The name of the column that contains the levels of the replications/blocks

resp
The response variable(s). To analyze multiple variables in a single procedure use, for example, resp = c(var1, var2, var3).

verbose
Logical argument. If verbose = FALSE the code will run silently.

Value

An object of class ge_reg with the folloing items for each variable:

data

The data with means for genotype and environment combinations and the environment index

anova

The analysis of variance for the regression model.

regression

The estimated coefficients of the regression model.

References

Eberhart, S.A., and W.A. Russell. 1966. Stability parameters for comparing Varieties. Crop Sci. 6:36-40. doi:10.2135/cropsci1966.0011183X000600010011x.

See also

Examples

library(metan)
reg = ge_reg(data_ge2,
            env = ENV,
            gen = GEN,
            rep = REP,
            resp = PH)