Get a distance matrix

Get the distance matrices from objects fitted with the function clustering(). This is especially useful to get distance matrices from several objects to be further analyzed using pairs_mantel().

Usage

get_dist(..., digits = 2)

Arguments

...

Object(s) of class clustering.]

digits

The number of significant figures. Defaults to 2.

Value

A list of class clustering.

Author

Tiago Olivoto tiagoolivoto@gmail.com

Examples

# \donttest{
library(metan)
d <- data_ge2 %>%
      mean_by(GEN) %>%
      column_to_rownames("GEN") %>%
      clustering()
get_dist(d)
#> $d
#>        H1   H10   H11   H12   H13    H2    H3    H4    H5    H6    H7    H8
#> H1   0.00 49.24 36.63 55.94 39.76 22.00 28.60 34.48 42.49 10.77 24.96 50.59
#> H10 49.24  0.00 13.73  8.30 42.95 48.85 29.18 46.05 48.06 51.34 26.89  8.77
#> H11 36.63 13.73  0.00 20.00 40.07 40.23 16.18 41.01 45.28 40.25 13.59 14.45
#> H12 55.94  8.30 20.00  0.00 49.06 56.01 33.96 52.93 54.39 58.52 32.68  8.59
#> H13 39.76 42.95 40.07 49.06  0.00 20.76 48.04  9.11  6.78 32.52 40.82 49.88
#> H2  22.00 48.85 40.23 56.01 20.76  0.00 40.93 14.18 22.03 12.71 34.51 53.61
#> H3  28.60 29.18 16.18 33.96 48.04 40.93  0.00 46.77 52.96 36.07  7.93 26.52
#> H4  34.48 46.05 41.01 52.93  9.11 14.18 46.77  0.00  8.67 26.16 39.70 52.51
#> H5  42.49 48.06 45.28 54.39  6.78 22.03 52.96  8.67  0.00 34.14 45.65 55.21
#> H6  10.77 51.34 40.25 58.52 32.52 12.71 36.07 26.16 34.14  0.00 30.85 54.36
#> H7  24.96 26.89 13.59 32.68 40.82 34.51  7.93 39.70 45.65 30.85  0.00 26.25
#> H8  50.59  8.77 14.45  8.59 49.88 53.61 26.52 52.51 55.21 54.36 26.25  0.00
#> H9  63.45 16.57 27.13  9.65 58.17 64.91 39.28 61.89 63.28 66.73 39.22 13.19
#>        H9
#> H1  63.45
#> H10 16.57
#> H11 27.13
#> H12  9.65
#> H13 58.17
#> H2  64.91
#> H3  39.28
#> H4  61.89
#> H5  63.28
#> H6  66.73
#> H7  39.22
#> H8  13.19
#> H9   0.00
#> 
#> attr(,"class")
#> [1] "clustering"
# }