Linear and Partial Correlation Coefficients

Estimates the linear and partial correlation coefficients using as input a data frame or a correlation matrix.

lpcor(.data, ..., n = NULL, method = "pearson", verbose = TRUE)

Arguments

.data
The data to be analyzed. Must be a symmetric correlation matrix or, a dataframe containing the predictor variables, or an object of class split_factors.

...
Variables to use in the correlation. If ... is null (Default) then all the numeric variables from .data are used. It must be a single variable name or a comma-separated list of unquoted variables names.

n
If a correlation matrix is provided, then n is the number of objects used to compute the correlation coefficients.

method
a character string indicating which correlation coefficient is to be computed. One of 'pearson' (default), 'kendall', or 'spearman'.

verbose
If verbose = TRUE then some results are shown in the console.

Value

If a grouping factor is used then a list is returned with the following values.

linear.mat The matrix of linear correlation.

partial.mat The matrix of partial correlations.

results Hypothesis testing for each pairwise comparison.

Examples

library(metan)
partial1 = lpcor(iris)

#> The factors  where excluded to perform the analysis. If you want to perform an analysis for each level of a factor, use the function 'split_factors() before.' 

#>                         Pairs     linear    partial         t         prob
#> 1  Sepal.Length x Sepal.Width -0.1175698  0.6285707  9.765380 0.000000e+00
#> 2 Sepal.Length x Petal.Length  0.8717538  0.7190656 12.502483 0.000000e+00
#> 3  Sepal.Length x Petal.Width  0.8179411 -0.3396174 -4.362929 2.412876e-05
#> 4  Sepal.Width x Petal.Length -0.4284401 -0.6152919 -9.431189 0.000000e+00
#> 5   Sepal.Width x Petal.Width -0.3661259  0.3526260  4.553279 1.104958e-05
#> 6  Petal.Length x Petal.Width  0.9628654  0.8707698 21.398708 0.000000e+00

# Alternatively using the pipe operator %>%
partial2 = iris %>% lpcor()

#> The factors  where excluded to perform the analysis. If you want to perform an analysis for each level of a factor, use the function 'split_factors() before.' 

#>                         Pairs     linear    partial         t         prob
#> 1  Sepal.Length x Sepal.Width -0.1175698  0.6285707  9.765380 0.000000e+00
#> 2 Sepal.Length x Petal.Length  0.8717538  0.7190656 12.502483 0.000000e+00
#> 3  Sepal.Length x Petal.Width  0.8179411 -0.3396174 -4.362929 2.412876e-05
#> 4  Sepal.Width x Petal.Length -0.4284401 -0.6152919 -9.431189 0.000000e+00
#> 5   Sepal.Width x Petal.Width -0.3661259  0.3526260  4.553279 1.104958e-05
#> 6  Petal.Length x Petal.Width  0.9628654  0.8707698 21.398708 0.000000e+00

# Using a correlation matrix
partial3 = cor(iris[1:4]) %>%
           lpcor(n = nrow(iris),
                 verbose = FALSE)

# Select variables and compute the partial correlation
# For each level of \code{Species}

partial4 = iris %>%
           split_factors(Species) %>%
           lpcor(Sepal.Length, Sepal.Width, Petal.Width)

#> Warning: Use 'keep_factors = TRUE' to keep the columns Species in the grouped data. 

#> 
#> ----------------------------------------------------------------------------
#> Level: setosa 
#> ----------------------------------------------------------------------------
#>                        Pairs    linear    partial         t         prob
#> 1 Sepal.Length x Sepal.Width 0.7425467 0.72558229 7.2287250 3.680601e-09
#> 2 Sepal.Length x Petal.Width 0.2780984 0.16160463 1.1226622 2.672834e-01
#> 3  Sepal.Width x Petal.Width 0.2327520 0.04080217 0.2799587 7.807377e-01
#> 
#> ----------------------------------------------------------------------------
#> Level: versicolor 
#> ----------------------------------------------------------------------------
#>                        Pairs    linear   partial        t         prob
#> 1 Sepal.Length x Sepal.Width 0.5259107 0.2603917 1.848938 7.076217e-02
#> 2 Sepal.Length x Petal.Width 0.5464611 0.3101636 2.236681 3.009043e-02
#> 3  Sepal.Width x Petal.Width 0.6639987 0.5287121 4.270338 9.413348e-05
#> 
#> ----------------------------------------------------------------------------
#> Level: virginica 
#> ----------------------------------------------------------------------------
#>                        Pairs    linear    partial         t         prob
#> 1 Sepal.Length x Sepal.Width 0.4572278 0.37827279 2.8014734 0.0073635897
#> 2 Sepal.Length x Petal.Width 0.2811077 0.04700217 0.3225872 0.7484385758
#> 3  Sepal.Width x Petal.Width 0.5377280 0.47944191 3.7454273 0.0004911884