Mahalanobis distance from designed experiments

Compute the Mahalanobis distance using data from an experiment conducted in a randomized complete block design or completely randomized design.

mahala_design(.data, gen, rep, resp, design = "RCBD",
  return = "distance")

Arguments

.data
The dataset containing the columns related to Genotypes, replication/block and response variables. Alternatively, it is possible to use an object of class 'split_factors' to compute the distance for each level of the grouping factor. See ?split_factors.

gen
The name of the column that contains the levels of the genotypes.

rep
The name of the column that contains the levels of the replications/blocks.

resp
The response variables. For example resp = c(var1, var2, var3).

design
The experimental design. Must be RCBD or CRD.

return
What the function return? Default is 'distance', i.e., the Mahalanobis distance. Alternatively, it is possible to return the matrix of means return = 'means', or the variance-covariance matrix of residuals return = 'covmat'.

Value

A symmetric matrix with the Mahalanobis' distance. If the .data is an object of class split_factors then a list of distances will be returned.

Examples

# \donttest{
library(metan)
maha_group = mahala_design(data_ge,
                           gen = GEN,
                           rep = REP,
                           resp = c(GY, HM))

# Compute one distance for each environment
maha_group = data_ge %>%
            split_factors(ENV, keep_factors = TRUE) %>%
            mahala_design(GEN, REP, everything())

#> [1] "E1"
#> [1] "E10"
#> [1] "E11"
#> [1] "E12"
#> [1] "E13"
#> [1] "E14"
#> [1] "E2"
#> [1] "E3"
#> [1] "E4"
#> [1] "E5"
#> [1] "E6"
#> [1] "E7"
#> [1] "E8"
#> [1] "E9"

# Return the variance-covariance matrix of residuals
cov_mat = mahala_design(data_ge,
                       gen = GEN,
                       rep = REP,
                       resp = c(GY, HM),
                       return = 'covmat')
# }