Make a symetric matrix based on an upper or lower triangular matrix

This function help users to easily make a symmetric matrix using a lower or an upper triangular matrix.

make_sym(.matrix, make = "upper")

Arguments

.matrix
The upper or lower triangular matrix.

make
The triangular to built. Default is "upper". In this case, a symetric matrix will be built based on the values of a lower triangular values.

Examples

m <- matrix(NA,4,4)
m[lower.tri(m)] <- 1:16

#> Warning: número de itens para para substituir não é um múltiplo do comprimento do substituto

m

#>      [,1] [,2] [,3] [,4]
#> [1,]   NA   NA   NA   NA
#> [2,]    1   NA   NA   NA
#> [3,]    2    4   NA   NA
#> [4,]    3    5    6   NA

make_sym(m)

#>      [,1] [,2] [,3] [,4]
#> [1,]   NA    1    2    3
#> [2,]    1   NA    4    5
#> [3,]    2    4   NA    6
#> [4,]    3    5    6   NA