Means by one or more factors

Computes the mean for all numeric variables of a data frame, grouping by one or more factors.

means_by(.data, ...)

Arguments

.data
A data frame

...
One or more categorical variables for grouping the data.

Value

An object of class tbl_df with the computed means by each level of the factor(s) declared in ....

Examples

# \donttest{
library(metan)
means_by(data_ge2, ENV)

#> # A tibble: 4 x 16
#>   ENV      PH    EH    EP    EL    ED    CL    CD    CW    KW    NR   NKR  CDED
#>   <fct> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1 A1     2.79  1.58 0.567  15.6  51.6  29.7  16.4  28.3  199.  16.9  33.9 0.576
#> 2 A2     2.46  1.31 0.527  15.2  48.7  28.5  15.9  23.8  168.  15.8  32.3 0.584
#> 3 A3     2.17  1.08 0.496  14.7  47.9  28.4  15.8  20.8  147.  15.8  30.4 0.595
#> 4 A4     2.52  1.41 0.558  15.1  49.9  29.4  15.8  26.4  177.  16.0  32.5 0.589
#> # … with 3 more variables: PERK <dbl>, TKW <dbl>, NKE <dbl>

means_by(data_ge2, GEN, ENV)

#> # A tibble: 52 x 17
#>    GEN   ENV      PH    EH    EP    EL    ED    CL    CD    CW    KW    NR   NKR
#>    <fct> <fct> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1 H1    A1     2.72 1.68  0.626  15.4  51.1  28.1  15.7  23.5  203.  16.3  33.3
#>  2 H1    A2     2.93 1.80  0.612  15.0  51.9  31.5  15.5  30.2  188.  17.1  31.4
#>  3 H1    A3     2.20 1.10  0.497  14.8  50.6  30.9  15.8  26.8  157.  15.9  28.4
#>  4 H1    A4     2.64 1.44  0.547  15.2  51.2  29.8  15.8  26.4  187.  17.2  35.7
#>  5 H10   A1     2.78 1.62  0.584  16.1  53.2  31.4  16.8  24.6  192.  16.7  31.2
#>  6 H10   A2     2.05 0.987 0.494  15.5  46.7  26.8  16.3  26.3  160.  14    33.5
#>  7 H10   A3     2.04 1.01  0.503  14.0  43.9  24.8  15.2  12.3  121.  15.3  33.3
#>  8 H10   A4     2.39 1.43  0.600  14.9  50.0  30.7  15.3  28.0  183.  16.4  31.6
#>  9 H11   A1     2.75 1.58  0.574  16.6  48.9  29.0  17.2  23.6  188.  15.2  34.6
#> 10 H11   A2     2.15 1.02  0.475  15.1  47.3  27.2  15.7  24.3  164.  13.7  35  
#> # … with 42 more rows, and 4 more variables: CDED <dbl>, PERK <dbl>, TKW <dbl>,
#> #   NKE <dbl>

# }