Deprecated functions in package metan.

The functions listed below are deprecated and will be defunct in the near future. When possible, alternative functions with similar functionality are also mentioned. Help pages for deprecated functions are available at help("<function>-deprecated").

WAAS.AMMI(.data, env, gen, rep, resp, mresp = NULL, wresp = NULL,
  prob = 0.05, naxis = NULL, verbose = TRUE)

WAASBYratio(.data, env, gen, rep, resp, mresp = 100, increment = 10,
  saveWAASY = 50, progbar = TRUE)

WAASratio.AMMI(.data, env, gen, rep, resp, mresp = 100,
  p.valuePC = 0.05, increment = 5, saveWAASY = 50, progbar = TRUE)

# S3 method for WAAS.AMMI
autoplot(x, conf = 0.95, labels = FALSE,
  theme = theme_waasb(), alpha = 0.2, fill.hist = "gray",
  col.hist = "black", col.point = "black", col.line = "red",
  col.lab.out = "red", size.lab.out = 2.5, size.tex.lab = 10,
  size.shape = 1.5, bins = 30, which = c(1:4), mfrow = c(2, 2),
  ...)

# S3 method for WAASBYratio
plot(x, type = 2, export = FALSE,
  file.type = "pdf", file.name = NULL, width = 6, height = 5,
  size.lab = 1, margins = c(5, 4), y.lab = NULL, x.lab = NULL,
  key.lab = "Genotype ranking", resolution = 300, ...)

# S3 method for WAASratio.AMMI
plot(x, type = 2, export = FALSE,
  file.type = "pdf", file.name = NULL, width = 6, height = 5,
  size.lab = 1, margins = c(5, 4), y.lab = NULL, x.lab = NULL,
  key.lab = "Genotype ranking", resolution = 300, ...)

# S3 method for WAAS.AMMI
predict(object, naxis, ...)

# S3 method for WAAS.AMMI
summary(object, export = FALSE, file.name = NULL,
  digits = 4, ...)

WAAS.AMMI

For WAAS.AMMI, use waas.

WAASBYratio

For WAASBYratio, use wsmp.

WAASratio.AMMI

For WAASratio.AMMI, use wsmp.

autoplot.WAAS.AMMI

For autoplot.WAAS.AMMI, use autoplot.waas.

plot.WAASBYratio

For plot.WAASBYratio, use plot.wsmp.

plot.WAASratio.AMMI

For plot.WAASratio.AMMI, use plot.wsmp.

predict.WAAS.AMMI

For predict.WAAS.AMMI, use predict.waas.

summary.WAAS.AMMI

For summary.WAAS.AMMI, use summary.waas.