Path coefficients with minimal multicollinearity

Computes direct and indirect effects in path analysis. An algorithm to select a set of predictors with minimal multicollinearity and high explanatory power is implemented.

path_coeff(.data, resp, pred = NULL, exclude = FALSE,
  correction = NULL, knumber = 50, brutstep = FALSE, maxvif = 10,
  missingval = "pairwise.complete.obs", verbose = TRUE)

Arguments

.data
The data. Must be a dataframe or an object of class split_factors.

resp
The dependent variable.

pred
The predictor variables, set to NULL, i.e., the predictor variables are all the numeric variables in the data except that in resp.

exclude
Logical argument, set to false. If exclude = TRUE, then the variables in pred are deleted from the data, and the analysis will use as predictor those that remained, except that in resp.

correction
Set to NULL. A correction value (k) that will be added into the diagonal elements of the X'X matrix aiming at reducing the harmful problems of the multicollinearity in path analysis (Olivoto et al., 2017)

knumber
When correction = NULL, a plot showing the values of direct effects in a set of different k values (0-1) is produced. knumber is the number of k values used in the range of 0 to 1.

brutstep
Logical argument, set to FALSE. If true, then an algorithm will select a subset of variables with minimal multicollinearity and fit a set of possible models. See the Details section for more information.

maxvif
The maximum value for the Variance Inflation Factor (cut point) that will be accepted. See the Details section for more information.

missingval
How to deal with missing values. For more information, please see ?cor.

verbose
If verbose = TRUE then some results are shown in the console.

Value

An object of class path_coeff, group_path, or brute_path with the following items:

Corr.x A correlation matrix between the predictor variables.

Corr.y A vector of correlations between each predictor variable with the dependent variable.

Coefficients The path coefficients. Direct effects are the diagonal elements, and the indirect effects those in the off-diagonal elements (column)

Eigen Eigenvectors and eigenvalues of the Corr.x.

VIF The Variance Inflation Factors.

plot A ggplot2-based graphic showing the direct effects in 21 different k values.

Predictors The predictor variables used in the model.

CN The Condition Number, i.e., the ratio between the highest and lowest eigenvalue.

Det The matrix determinant of the Corr.x..

R2 The coefficient of determination of the model.

Residual The residual effect of the model.

Response The response variable.

weightvar The order of the predictor variables with the highest weight (highest eigenvector) in the lowest eigenvalue.

If .data is an object of class split_factors then a list is returned with the above items for each level of the grouping variable in the function split_factors.

Details

When brutstep = TRUE, first, the algorithm will select a set of predictors with minimal multicollinearity. The selection is based on the variance inflation factor (VIF). An iterative process is performed until the maximum VIF observed is less than maxvif. The variables selected in this iterative process are then used in a series of stepwise-based regressions. The first model is fitted and p-1 predictor variables are retained (p is the number of variables selected in the iterative process. The second model adjusts a regression considering p-2 selected variables, and so on until the last model, which considers only two variables. Three objects are created. Summary, with the process summary, Models, containing the aforementioned values for all the adjusted models; and Selectedpred, a vector with the name of the selected variables in the iterative process.

References

Olivoto, T., V.Q. Souza, M. Nardino, I.R. Carvalho, M. Ferrari, A.J. Pelegrin, V.J. Szareski, and D. Schmidt. 2017. Multicollinearity in path analysis: a simple method to reduce its effects. Agron. J. 109:131-142. doi:10.2134/agronj2016.04.0196. 10.2134/agronj2016.04.0196.

Examples

library(metan)

# Using KW as the response variable and all other ones as predictors
pcoeff <- path_coeff(data_ge2, resp = KW)

#> The factors ENV GEN REP where excluded to perform the analysis. If you want to perform an analysis for each level of a factor, use the function 'split_factors() before.' 

#> Severe multicollinearity. 
#> Condition Number = 7865.84
#> Please, consider using a correction factor, or use 'brutstep = TRUE'. 

# Declaring the predictors
pcoeff2 <- path_coeff(data_ge2,
                     resp = KW,
                     pred = c(PH, EH, NKE, TKW))

#> The factors ENV GEN REP where excluded to perform the analysis. If you want to perform an analysis for each level of a factor, use the function 'split_factors() before.' 

#> Weak multicollinearity. 
#> Condition Number = 40.232
#> You will probably have path coefficients close to being unbiased. 

# Selecting variables to be excluded from the analysis
pcoeff3 <- path_coeff(data_ge2,
                     resp = KW,
                     pred = c(PH, EH, NKE, TKW),
                     exclude = TRUE)

#> The factors ENV GEN REP where excluded to perform the analysis. If you want to perform an analysis for each level of a factor, use the function 'split_factors() before.' 

#> Severe multicollinearity. 
#> Condition Number = 3880.917
#> Please, consider using a correction factor, or use 'brutstep = TRUE'. 

# Selecting a set of predictors with minimal multicollinearity
# Maximum variance Inflation factor of 5
pcoeff4 <- path_coeff(data_ge2,
                     resp = KW,
                     brutstep = TRUE,
                     maxvif = 5)

#> The factors ENV GEN REP where excluded to perform the analysis. If you want to perform an analysis for each level of a factor, use the function 'split_factors() before.' 

#> --------------------------------------------------------------------------
#> The algorithm has selected a set of 8 predictors with largest VIF = 3.346. 
#> Selected predictors: NR PERK EP CDED EL NKR TKW PH 
#> A forward stepwise-based selection procedure will fit 6 models.
#> --------------------------------------------------------------------------
#> Adjusting the model 1 with 7 predictors (16.67% concluded)
#> Adjusting the model 2 with 6 predictors (33.33% concluded)
#> Adjusting the model 3 with 5 predictors (50% concluded)
#> Adjusting the model 4 with 4 predictors (66.67% concluded)
#> Adjusting the model 5 with 3 predictors (83.33% concluded)
#> Adjusting the model 6 with 2 predictors (100% concluded)
#> Done!
#> --------------------------------------------------------------------------
#> Summary of the adjusted models 
#> --------------------------------------------------------------------------
#>   Model  AIC Numpred    CN Determinant    R2 Residual maxVIF
#>  Model6 1329       2  2.23      0.8555 0.738   0.2616   1.17
#>  Model5 1148       3  1.34      0.9787 0.919   0.0808   1.02
#>  Model4 1138       4  6.40      0.4595 0.925   0.0747   2.13
#>  Model3 1126       5 12.05      0.1989 0.932   0.0683   2.31
#>  Model2 1125       6 12.26      0.1383 0.933   0.0670   2.46
#>  Model1 1127       7 13.67      0.0841 0.933   0.0669   2.59
#> --------------------------------------------------------------------------

# When one analysis should be carried out for each environment
# Using the forward-pipe operator %>%
pcoeff5 <- data_ge2 %>%
           split_factors(ENV) %>%
           path_coeff(resp = KW,
                      pred = c(PH, EH, NKE, TKW))

#> Warning: Use 'keep_factors = TRUE' to keep the columns ENV GEN REP in the grouped data. 

#> 
#> ----------------------------------------------------------------------------
#> Level dfs 
#> ----------------------------------------------------------------------------
#> Weak multicollinearity. 
#> Condition Number = 5.183
#> You will probably have path coefficients close to being unbiased. 
#> 
#> ----------------------------------------------------------------------------
#> Level names 
#> ----------------------------------------------------------------------------
#> Weak multicollinearity. 
#> Condition Number = 81.987
#> You will probably have path coefficients close to being unbiased.