Stability analysis based on Plaisted and Peterson (1959)

The function computes the stability as the arithmetic mean of the variance component of the genotype-environment interaction between environment pairs that includes a given genotype

Usage

plaisted_peterson(.data, env, gen, rep, resp, verbose = TRUE)

Arguments

.data

The dataset containing the columns related to Environments, Genotypes, replication/block and response variable(s).

env

The name of the column that contains the levels of the environments.

gen

The name of the column that contains the levels of the genotypes.

rep

The name of the column that contains the levels of the replications/blocks.

resp

The response variable(s). To analyze multiple variables in a single procedure use, for example, resp = c(var1, var2, var3).

verbose

Logical argument. If verbose = FALSE the code will run silently.

Value

An object of class plaisted_peterson containing the results for each variable used in the argument resp.

References

Plaisted, R.L., and L.C. Peterson. 1959. A technique for evaluating the ability of selections to yield consistently in different locations or seasons. American Potato Journal 36(11): 381–385. doi:10.1007/BF02852735

Author

Tiago Olivoto tiagoolivoto@gmail.com

Examples

# \donttest{
library(metan)
 plaisted_peterson(data_ge, ENV, GEN, REP, GY)
#> Evaluating trait GY |============================================| 100% 00:00:00 
#> Variable GY 
#>               G1        G10         G2            G3           G4           G5
#> G1            NA 0.12503363 0.02061323 -0.0055346368  0.024997141  0.031362699
#> G10  0.125033634         NA 0.14430703  0.0843416461  0.093958529  0.076995313
#> G2   0.020613234 0.14430703         NA  0.0257285799  0.023524884  0.040610553
#> G3  -0.005534637 0.08434165 0.02572858            NA  0.015273161  0.005673726
#> G4   0.024997141 0.09395853 0.02352488  0.0152731612           NA -0.005702818
#> G5   0.031362699 0.07699531 0.04061055  0.0056737264 -0.005702818           NA
#> G6   0.016916927 0.11171495 0.05124422 -0.0099631033  0.005872310 -0.012946149
#> G7   0.028707482 0.19150849 0.05029837  0.0502960198  0.035132557  0.048298953
#> G8  -0.007565760 0.15713144 0.05730481 -0.0002494876  0.071612521  0.055308115
#> G9   0.031977269 0.14509552 0.08525933  0.0375471911  0.145944773  0.106988291
#>               G6         G7            G8         G9      theta theta_prop
#> G1   0.016916927 0.02870748 -0.0075657602 0.03197727 0.02961200   5.221324
#> G10  0.111714955 0.19150849  0.1571314411 0.14509552 0.12556517  22.140231
#> G2   0.051244217 0.05029837  0.0573048092 0.08525933 0.05543233   9.774085
#> G3  -0.009963103 0.05029602 -0.0002494876 0.03754719 0.02256812   3.979315
#> G4   0.005872310 0.03513256  0.0716125211 0.14594477 0.04562367   8.044577
#> G5  -0.012946149 0.04829895  0.0553081149 0.10698829 0.03850985   6.790235
#> G6            NA 0.05019598  0.0266023305 0.10212977 0.03797414   6.695775
#> G7   0.050195980         NA  0.0501701107 0.13919471 0.07153363  12.613140
#> G8   0.026602330 0.05017011            NA 0.02920043 0.04883494   8.610803
#> G9   0.102129773 0.13919471  0.0292004251         NA 0.09148192  16.130515
#> 
#> 
#> 
# }