Print an object of class AMMI_indexes

Print the AMMI_indexes object in two ways. By default, the results are shown in the R console. The results can also be exported to the directory into a *.txt file.

# S3 method for AMMI_indexes
print(x, export = FALSE, file.name = NULL,
  digits = 3, ...)

Arguments

x
An object of class AMMI_indexes.

export
A logical argument. If TRUE, a *.txt file is exported to the working directory.

file.name
The name of the file if export = TRUE

digits
The significant digits to be shown.

...
Options used by the tibble package to format the output. See trunc_mat for more details.

Examples

library(metan)
model <- performs_ammi(data_ge, ENV, GEN, REP, GY) %>%
         AMMI_indexes()

#> All variables with significant (p < 0.05) genotype-vs-environment interaction
#> Done!

print(model)

#> Variable GY 
#> ---------------------------------------------------------------------------
#> AMMI-based stability indexes
#> ---------------------------------------------------------------------------
#> # A tibble: 10 x 8
#>    Code      Y    rY   ASV  SIPC     EV     ZA  WAAS
#>    <fct> <dbl> <dbl> <dbl> <dbl>  <dbl>  <dbl> <dbl>
#>  1 G1     2.60     6 0.346 0.463 0.0149 0.100  0.151
#>  2 G10    2.47    10 1.23  2.07  0.210  0.437  0.652
#>  3 G2     2.74     3 0.249 1.54  0.179  0.216  0.283
#>  4 G3     2.96     2 0.113 0.552 0.0207 0.0797 0.106
#>  5 G4     2.64     5 0.594 1.04  0.0521 0.219  0.326
#>  6 G5     2.54     7 0.430 0.997 0.0433 0.186  0.270
#>  7 G6     2.53     8 0.265 1.14  0.0911 0.172  0.233
#>  8 G7     2.74     4 0.663 1.79  0.191  0.303  0.428
#>  9 G8     3.00     1 0.574 1.18  0.0669 0.225  0.327
#> 10 G9     2.51     9 0.983 1.50  0.131  0.336  0.507
#> ---------------------------------------------------------------------------
#> Ranks for stability indexes
#> ---------------------------------------------------------------------------
#> # A tibble: 10 x 8
#>    Code      Y    rY  rASV rSIPC   rEV   rZA rWAAS
#>    <fct> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1 G1     2.60     6     4     1     1     2     2
#>  2 G10    2.47    10    10    10    10    10    10
#>  3 G2     2.74     3     2     8     8     5     5
#>  4 G3     2.96     2     1     2     2     1     1
#>  5 G4     2.64     5     7     4     4     6     6
#>  6 G5     2.54     7     5     3     3     4     4
#>  7 G6     2.53     8     3     5     6     3     3
#>  8 G7     2.74     4     8     9     9     8     8
#>  9 G8     3.00     1     6     6     5     7     7
#> 10 G9     2.51     9     9     7     7     9     9
#> ---------------------------------------------------------------------------
#> Simultaneous selection indexes
#> ---------------------------------------------------------------------------
#> # A tibble: 10 x 8
#>    Code      Y    rY ssiASV ssiSIPC ssiEV ssiZA ssiWAAS
#>    <fct> <dbl> <dbl>  <dbl>   <dbl> <dbl> <dbl>   <dbl>
#>  1 G1     2.60     6     10       7     7     8       8
#>  2 G10    2.47    10     20      20    20    20      20
#>  3 G2     2.74     3      5      11    11     8       8
#>  4 G3     2.96     2      3       4     4     3       3
#>  5 G4     2.64     5     12       9     9    11      11
#>  6 G5     2.54     7     12      10    10    11      11
#>  7 G6     2.53     8     11      13    14    11      11
#>  8 G7     2.74     4     12      13    13    12      12
#>  9 G8     3.00     1      7       7     6     8       8
#> 10 G9     2.51     9     18      16    16    18      18
#> 
#> 
#>