Print an object of class Annicchiarico

Print the Annicchiarico object in two ways. By default, the results are shown in the R console. The results can also be exported to the directory into a *.txt file.

# S3 method for Annicchiarico
print(x, export = FALSE, file.name = NULL, digits = 3, ...)

Arguments

x
The Annicchiarico x

export
A logical argument. If TRUE, a *.txt file is exported to the working directory.

file.name
The name of the file if export = TRUE

digits
The significant digits to be shown.

...
Options used by the tibble package to format the output. See tibble::print() for more details.

Examples

# \donttest{
library(metan)
Ann <- Annicchiarico(data_ge2,
  env = ENV,
  gen = GEN,
  rep = REP,
  resp = PH
)
print(Ann)

#> Variable PH 
#> ---------------------------------------------------------------------------
#> Environmental index
#> ---------------------------------------------------------------------------
#> # A tibble: 4 x 4
#>   ENV    Mean   index class      
#>   <fct> <dbl>   <dbl> <chr>      
#> 1 A1     2.79  0.308  favorable  
#> 2 A2     2.46 -0.0229 unfavorable
#> 3 A3     2.17 -0.318  unfavorable
#> 4 A4     2.52  0.0329 favorable  
#> ---------------------------------------------------------------------------
#> Analysis for all environments
#> ---------------------------------------------------------------------------
#> # A tibble: 13 x 6
#>    GEN       Y Mean_rp Sd_rp    Wi  rank
#>    <chr> <dbl>   <dbl> <dbl> <dbl> <dbl>
#>  1 H1     2.62   106.   9.38  99.3     2
#>  2 H10    2.31    93.0  6.93  88.3    12
#>  3 H11    2.39    96.1  6.27  91.9     9
#>  4 H12    2.44    98.6 11.2   91.0    10
#>  5 H13    2.54   103.  12.4   94.4     7
#>  6 H2     2.60   105.   9.91  98.1     3
#>  7 H3     2.59   104.  11.2   96.6     6
#>  8 H4     2.58   104.   8.76  97.7     5
#>  9 H5     2.57   103.   5.55  99.5     1
#> 10 H6     2.56   103.   7.83  97.7     4
#> 11 H7     2.40    96.8  6.67  92.3     8
#> 12 H8     2.33    93.6  9.61  87.1    13
#> 13 H9     2.36    94.8  6.85  90.2    11
#> ---------------------------------------------------------------------------
#> Analysis for favorable environments
#> ---------------------------------------------------------------------------
#> # A tibble: 13 x 6
#>    GEN       Y Mean_rp Sd_rp    Wi  rank
#>    <chr> <dbl>   <dbl> <dbl> <dbl> <dbl>
#>  1 H1     2.68   101.  5.09   97.6     9
#>  2 H10    2.59    97.3 3.38   95.0    12
#>  3 H11    2.66   100.  2.41   98.5     7
#>  4 H12    2.61    98.5 3.00   96.5    10
#>  5 H13    2.66   100.  1.63   99.3     5
#>  6 H2     2.66   100.  0.203 100.      3
#>  7 H3     2.70   101.  5.22   97.9     8
#>  8 H4     2.71   102.  0.839 102.      2
#>  9 H5     2.74   103.  2.32  102.      1
#> 10 H6     2.63    98.9 0.292  98.7     6
#> 11 H7     2.65    99.9 0.812  99.3     4
#> 12 H8     2.62    98.9 3.60   96.5    11
#> 13 H9     2.62    98.1 9.33   91.8    13
#> ---------------------------------------------------------------------------
#> Analysis for unfavorable environments
#> ---------------------------------------------------------------------------
#> # A tibble: 13 x 6
#>    GEN       Y Mean_rp Sd_rp    Wi  rank
#>    <chr> <dbl>   <dbl> <dbl> <dbl> <dbl>
#>  1 H1     2.56   110.  12.4  102.      1
#>  2 H10    2.04    88.7  7.65  83.5    12
#>  3 H11    2.12    92.1  6.93  87.4     9
#>  4 H12    2.26    98.6 19.2   85.7    11
#>  5 H13    2.41   105.  20.8   91.2     7
#>  6 H2     2.55   110.  14.3   99.9     2
#>  7 H3     2.49   107.  17.8   94.9     6
#>  8 H4     2.45   105.  14.9   95.1     5
#>  9 H5     2.40   103.   9.32  97.0     4
#> 10 H6     2.49   107.  10.9   99.7     3
#> 11 H7     2.16    93.6  9.69  87.1    10
#> 12 H8     2.03    88.3 12.3   80.0    13
#> 13 H9     2.11    91.5  3.02  89.4     8
#> 
#> 
#> 

# }