Print an object ofclass Huehn

Print the Huehn object in two ways. By default, the results are shown in the R console. The results can also be exported to the directory into a *.txt file.

# S3 method for Huehn
print(x, export = FALSE, file.name = NULL, digits = 3, ...)

Arguments

x
An object of class Huehn.

export
A logical argument. If TRUE, a *.txt file is exported to the working directory.

file.name
The name of the file if export = TRUE

digits
The significant digits to be shown.

...
Options used by the tibble package to format the output. See tibble::print() for more details.

Examples

# \donttest{
library(metan)
model <- Huehn(data_ge2, ENV, GEN, PH)
print(model)

#> Variable PH 
#> ---------------------------------------------------------------------------
#> Huehn's stability indexes
#> ---------------------------------------------------------------------------
#> # A tibble: 13 x 11
#>    GEN       Y   Y_R    S1  S1_R    S2  S2_R    S3  S3_R    S6  S6_R
#>    <chr> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1 H1     2.62     1 0.167   3.5 16.9      8  5.70     8 1.35    8  
#>  2 H10    2.31    13 0       1.5  8.25     3  7.71     9 2.57   11  
#>  3 H11    2.39    10 0.667  11    6.25     2  4.26     7 1.33    7  
#>  4 H12    2.44     8 0.333   6   20.9     10 10.3     11 2.33    9  
#>  5 H13    2.54     7 1      13   20.2      9  3.33     3 1.11    4.5
#>  6 H2     2.60     2 0.333   6   22.7     11  3.56     5 0.889   1  
#>  7 H3     2.59     3 0       1.5 34.2     13 11.7     12 2.45   10  
#>  8 H4     2.58     4 0.5     9    8.92     4  3.41     4 1.06    3  
#>  9 H5     2.57     5 0.833  12    5.58     1  3.78     6 1.11    4.5
#> 10 H6     2.56     6 0.5     9   11        5  2.62     1 1.23    6  
#> 11 H7     2.40     9 0.167   3.5 11.3      6  2.78     2 1.04    2  
#> 12 H8     2.33    12 0.333   6   28.2     12  9.5     10 3      12  
#> 13 H9     2.36    11 0.5     9   14.2      7 16.4     13 3.2    13  
#> ---------------------------------------------------------------------------
#> 
#> 

# }