Print an object of class Shukla

Print the Shukla object in two ways. By default, the results are shown in the R console. The results can also be exported to the directory into a *.txt file.

# S3 method for Shukla
print(x, export = FALSE, file.name = NULL, digits = 3, ...)

Arguments

x
The Shukla x

export
A logical argument. If TRUE, a *.txt file is exported to the working directory.

file.name
The name of the file if export = TRUE

digits
The significant digits to be shown.

...
Options used by the tibble package to format the output. See tibble::print() for more details.

Author

Tiago Olivoto tiagoolivoto@gmail.com

Examples

# \donttest{
library(metan)
eco <- Shukla(data_ge2,
  env = ENV,
  gen = GEN,
  rep = REP,
  resp = PH
)

#> Evaluating trait PH |============================================| 100% 00:00:00 

print(eco)

#> Variable PH 
#> ---------------------------------------------------------------------------
#> Shukla stability variance
#> ---------------------------------------------------------------------------
#> # A tibble: 13 x 6
#>    GEN       Y ShuklaVar rMean rShukaVar ssiShukaVar
#>    <fct> <dbl>     <dbl> <dbl>     <dbl>       <dbl>
#>  1 H1     2.62    0.0602     1         8           9
#>  2 H10    2.31    0.0301    13         4          17
#>  3 H11    2.39    0.0237    10         2          12
#>  4 H12    2.44    0.0769     8        11          19
#>  5 H13    2.54    0.0869     7        13          20
#>  6 H2     2.60    0.0657     2        10          12
#>  7 H3     2.59    0.0819     3        12          15
#>  8 H4     2.58    0.0475     4         7          11
#>  9 H5     2.57    0.0162     5         1           6
#> 10 H6     2.56    0.0396     6         6          12
#> 11 H7     2.40    0.0272     9         3          12
#> 12 H8     2.33    0.0618    12         9          21
#> 13 H9     2.36    0.0321    11         5          16
#> 
#> 
#> 

# }