Print an object of class ecovalence

Print the ecovalence object in two ways. By default, the results are shown in the R console. The results can also be exported to the directory into a *.txt file.

# S3 method for ecovalence
print(x, export = FALSE, file.name = NULL, digits = 3, ...)

Arguments

x
The ecovalence x

export
A logical argument. If TRUE, a *.txt file is exported to the working directory.

file.name
The name of the file if export = TRUE

digits
The significant digits to be shown.

...
Options used by the tibble package to format the output. See tibble::print() for more details.

Author

Tiago Olivoto tiagoolivoto@gmail.com

Examples

# \donttest{
library(metan)
eco <- ecovalence(data_ge2,
                  env = ENV,
                  gen = GEN,
                  rep = REP,
                  resp = PH)
#> Evaluating trait PH |============================================| 100% 00:00:00 
print(eco)
#> Variable PH 
#> ---------------------------------------------------------------------------
#> Genotypic confidence index
#> ---------------------------------------------------------------------------
#> # A tibble: 13 x 8
#>    GEN        A1     A2      A3      A4 Ecoval Ecov_perc  rank
#>    <chr>   <dbl>  <dbl>   <dbl>   <dbl>  <dbl>     <dbl> <dbl>
#>  1 H1    -0.207   0.332 -0.106  -0.0189  0.493      9.13     8
#>  2 H10    0.160  -0.243  0.0414  0.0409  0.264      4.89     4
#>  3 H11    0.0504 -0.220  0.0296  0.140   0.215      3.99     2
#>  4 H12   -0.0531 -0.321  0.311   0.0636  0.620     11.5     11
#>  5 H13   -0.0749 -0.288  0.378  -0.0149  0.696     12.9     13
#>  6 H2    -0.120   0.365 -0.132  -0.113   0.535      9.91    10
#>  7 H3     0.0322  0.369 -0.233  -0.168   0.659     12.2     12
#>  8 H4    -0.0196  0.289 -0.212  -0.0569  0.397      7.35     7
#>  9 H5    -0.0416  0.160 -0.155   0.0371  0.158      2.92     1
#> 10 H6    -0.0976  0.290 -0.0864 -0.106   0.336      6.23     6
#> 11 H7     0.0613 -0.245  0.0913  0.0921  0.241      4.47     3
#> 12 H8     0.0572 -0.344  0.0917  0.195   0.505      9.36     9
#> 13 H9     0.253  -0.142 -0.0181 -0.0926  0.279      5.17     5
#> 
#> 
#> 
# }