Print an object of class env_dissimilarity

Print the env_dissimilarity object in two ways. By default, the results are shown in the R console. The results can also be exported to the directory into a *.txt file.

Usage

# S3 method for env_dissimilarity
print(x, export = FALSE, file.name = NULL, digits = 3, ...)

Arguments

x

An object of class env_dissimilarity.

export

A logical argument. If TRUE, a *.txt file is exported to the working directory.

file.name

The name of the file if export = TRUE

digits

The significant digits to be shown.

...

Currently not used.

Author

Tiago Olivoto tiagoolivoto@gmail.com

Examples

# \donttest{
library(metan)
mod <- env_dissimilarity(data_ge, ENV, GEN, REP, GY)
#> Evaluating trait Y |=============================================| 100% 00:00:00 
print(mod)
#> Variable GY 
#> ----------------------------------------------------------------------
#> Pearson's correlation coefficient
#> ----------------------------------------------------------------------
#>         E1    E10    E11    E12    E13    E14     E2     E3     E4     E5
#> E1   1.000  0.783  0.782  0.869  0.825  0.197  0.227  0.357 -0.011  0.704
#> E10  0.783  1.000  0.917  0.797  0.812  0.232  0.101 -0.136 -0.342  0.676
#> E11  0.782  0.917  1.000  0.749  0.840  0.234  0.438 -0.165 -0.130  0.635
#> E12  0.869  0.797  0.749  1.000  0.740  0.155  0.053  0.249 -0.190  0.567
#> E13  0.825  0.812  0.840  0.740  1.000  0.161  0.246  0.183 -0.045  0.814
#> E14  0.197  0.232  0.234  0.155  0.161  1.000  0.247 -0.427  0.149  0.218
#> E2   0.227  0.101  0.438  0.053  0.246  0.247  1.000 -0.069  0.661  0.210
#> E3   0.357 -0.136 -0.165  0.249  0.183 -0.427 -0.069  1.000  0.409  0.310
#> E4  -0.011 -0.342 -0.130 -0.190 -0.045  0.149  0.661  0.409  1.000  0.206
#> E5   0.704  0.676  0.635  0.567  0.814  0.218  0.210  0.310  0.206  1.000
#> E6   0.580  0.368  0.349  0.315  0.637  0.274  0.289  0.446  0.358  0.810
#> E7   0.454  0.394  0.188  0.350  0.152  0.566 -0.166  0.046  0.028  0.325
#> E8  -0.053  0.025  0.077  0.034  0.122  0.821  0.317 -0.293  0.417  0.216
#> E9   0.219  0.281  0.241  0.216 -0.184  0.323  0.119 -0.375 -0.276 -0.281
#>         E6     E7     E8     E9
#> E1   0.580  0.454 -0.053  0.219
#> E10  0.368  0.394  0.025  0.281
#> E11  0.349  0.188  0.077  0.241
#> E12  0.315  0.350  0.034  0.216
#> E13  0.637  0.152  0.122 -0.184
#> E14  0.274  0.566  0.821  0.323
#> E2   0.289 -0.166  0.317  0.119
#> E3   0.446  0.046 -0.293 -0.375
#> E4   0.358  0.028  0.417 -0.276
#> E5   0.810  0.325  0.216 -0.281
#> E6   1.000  0.241  0.240 -0.237
#> E7   0.241  1.000  0.265  0.456
#> E8   0.240  0.265  1.000 -0.106
#> E9  -0.237  0.456 -0.106  1.000
#> ----------------------------------------------------------------------
#> Minimum correlation =  -0.427 between environments 8 and 6 
#> Maximum correlation =  0.917 between environments 3 and 2 
#> ----------------------------------------------------------------------
#> Mean square GxEjj'
#> ----------------------------------------------------------------------
#>        E1   E10   E11   E12   E13   E14    E2    E3    E4    E5    E6    E7
#> E1  0.000 0.023 0.023 0.014 0.040 0.064 0.071 0.072 0.146 0.045 0.037 0.061
#> E10 0.023 0.000 0.010 0.021 0.044 0.056 0.075 0.119 0.183 0.048 0.047 0.064
#> E11 0.023 0.010 0.000 0.024 0.052 0.036 0.034 0.093 0.126 0.053 0.031 0.067
#> E12 0.014 0.021 0.024 0.000 0.054 0.065 0.083 0.082 0.168 0.063 0.054 0.071
#> E13 0.040 0.044 0.052 0.054 0.000 0.124 0.122 0.145 0.217 0.040 0.075 0.150
#> E14 0.064 0.056 0.036 0.065 0.124 0.000 0.043 0.107 0.097 0.090 0.032 0.038
#> E2  0.071 0.075 0.034 0.083 0.122 0.043 0.000 0.096 0.050 0.099 0.041 0.105
#> E3  0.072 0.119 0.093 0.082 0.145 0.107 0.096 0.000 0.087 0.100 0.047 0.107
#> E4  0.146 0.183 0.126 0.168 0.217 0.097 0.050 0.087 0.000 0.140 0.079 0.141
#> E5  0.045 0.048 0.053 0.063 0.040 0.090 0.099 0.100 0.140 0.000 0.038 0.098
#> E6  0.037 0.047 0.031 0.054 0.075 0.032 0.041 0.047 0.079 0.038 0.000 0.062
#> E7  0.061 0.064 0.067 0.071 0.150 0.038 0.105 0.107 0.141 0.098 0.062 0.000
#> E8  0.097 0.083 0.056 0.086 0.139 0.012 0.048 0.118 0.077 0.099 0.045 0.068
#> E9  0.134 0.120 0.113 0.133 0.275 0.103 0.133 0.229 0.257 0.257 0.160 0.097
#>        E8    E9
#> E1  0.097 0.134
#> E10 0.083 0.120
#> E11 0.056 0.113
#> E12 0.086 0.133
#> E13 0.139 0.275
#> E14 0.012 0.103
#> E2  0.048 0.133
#> E3  0.118 0.229
#> E4  0.077 0.257
#> E5  0.099 0.257
#> E6  0.045 0.160
#> E7  0.068 0.097
#> E8  0.000 0.163
#> E9  0.163 0.000
#> ----------------------------------------------------------------------
#> Total mean square =  8.091 
#> Minimum =  0.01 between environments 3 and 2 
#> Maximum =  0.275 between environments 14 and 5 
#> ----------------------------------------------------------------------
#> % Of the single part of MS GxEjj' (Robertson, 1959)
#> ----------------------------------------------------------------------
#>         E1    E10    E11    E12    E13    E14     E2     E3     E4     E5
#> E1   0.000  0.986 27.318  0.261 28.828 12.522  3.762  0.001  2.496  7.731
#> E10  0.986  0.000 40.737  0.385 34.423  9.953  1.763  0.171  3.118 11.398
#> E11 27.318 40.737  0.000 22.188 66.463  0.344  2.154  6.468 15.337 35.530
#> E12  0.261  0.385 22.188  0.000 23.835 10.808  2.477  0.034  2.638  6.684
#> E13 28.828 34.423 66.463 23.835  0.000 31.325 20.821  8.123  1.030  5.877
#> E14 12.522  9.953  0.344 10.808 31.325  0.000  3.400  7.351 23.171 24.437
#> E2   3.762  1.763  2.154  2.477 20.821  3.400  0.000  2.671 25.107 12.340
#> E3   0.001  0.171  6.468  0.034  8.123  7.351  2.671  0.000  4.323  3.574
#> E4   2.496  3.118 15.337  2.638  1.030 23.171 25.107  4.323  0.000  0.002
#> E5   7.731 11.398 35.530  6.684  5.877 24.437 12.340  3.574  0.002  0.000
#> E6  20.351 11.047  0.250 12.211 50.386  0.017  3.137 16.015 27.630 55.978
#> E7   0.000  0.349  9.145  0.050  7.742 20.888  2.510  0.000  2.611  3.560
#> E8   2.382  1.304  1.656  2.035 17.463 14.768  0.026  1.890 15.194 11.532
#> E9   7.339 10.869 28.075  8.369  0.025 34.705 17.152  4.388  0.588  0.634
#>         E6     E7     E8     E9
#> E1  20.351  0.000  2.382  7.339
#> E10 11.047  0.349  1.304 10.869
#> E11  0.250  9.145  1.656 28.075
#> E12 12.211  0.050  2.035  8.369
#> E13 50.386  7.742 17.463  0.025
#> E14  0.017 20.888 14.768 34.705
#> E2   3.137  2.510  0.026 17.152
#> E3  16.015  0.000  1.890  4.388
#> E4  27.630  2.611 15.194  0.588
#> E5  55.978  3.560 11.532  0.634
#> E6   0.000 12.350  3.465 21.839
#> E7  12.350  0.000  3.358 10.237
#> E8   3.465  3.358  0.000 13.356
#> E9  21.839 10.237 13.356  0.000
#> ----------------------------------------------------------------------
#> Average =  11.464 
#> Minimum =  0 between environments 12 and 12 
#> Maximum =  66.463 between environments 5 and 3 
#> ----------------------------------------------------------------------
#> % Of the complex part of MS GxEjj' (Robertson, 1959)
#> ----------------------------------------------------------------------
#>          E1    E10    E11    E12    E13    E14     E2      E3     E4     E5
#> E1    0.000 99.014 72.682 99.739 71.172 87.478 96.238  99.999 97.504 92.269
#> E10  99.014  0.000 59.263 99.615 65.577 90.047 98.237  99.829 96.882 88.602
#> E11  72.682 59.263  0.000 77.812 33.537 99.656 97.846  93.532 84.663 64.470
#> E12  99.739 99.615 77.812  0.000 76.165 89.192 97.523  99.966 97.362 93.316
#> E13  71.172 65.577 33.537 76.165  0.000 68.675 79.179  91.877 98.970 94.123
#> E14  87.478 90.047 99.656 89.192 68.675  0.000 96.600  92.649 76.829 75.563
#> E2   96.238 98.237 97.846 97.523 79.179 96.600  0.000  97.329 74.893 87.660
#> E3   99.999 99.829 93.532 99.966 91.877 92.649 97.329   0.000 95.677 96.426
#> E4   97.504 96.882 84.663 97.362 98.970 76.829 74.893  95.677  0.000 99.998
#> E5   92.269 88.602 64.470 93.316 94.123 75.563 87.660  96.426 99.998  0.000
#> E6   79.649 88.953 99.750 87.789 49.614 99.983 96.863  83.985 72.370 44.022
#> E7  100.000 99.651 90.855 99.950 92.258 79.112 97.490 100.000 97.389 96.440
#> E8   97.618 98.696 98.344 97.965 82.537 85.232 99.974  98.110 84.806 88.468
#> E9   92.661 89.131 71.925 91.631 99.975 65.295 82.848  95.612 99.412 99.366
#>         E6      E7     E8     E9
#> E1  79.649 100.000 97.618 92.661
#> E10 88.953  99.651 98.696 89.131
#> E11 99.750  90.855 98.344 71.925
#> E12 87.789  99.950 97.965 91.631
#> E13 49.614  92.258 82.537 99.975
#> E14 99.983  79.112 85.232 65.295
#> E2  96.863  97.490 99.974 82.848
#> E3  83.985 100.000 98.110 95.612
#> E4  72.370  97.389 84.806 99.412
#> E5  44.022  96.440 88.468 99.366
#> E6   0.000  87.650 96.535 78.161
#> E7  87.650   0.000 96.642 89.763
#> E8  96.535  96.642  0.000 86.644
#> E9  78.161  89.763 86.644  0.000
#> ----------------------------------------------------------------------
#> Average =  88.536 
#> Minimum =  33.537 between environments 5 and 3 
#> Maximum =  100 between environments 12 and 12 
#> ----------------------------------------------------------------------
#> % Of the single part of MS GxEjj' (Cruz and Castoldi, 1991)
#> ----------------------------------------------------------------------
#>         E1     E10    E11    E12    E13     E14     E2      E3      E4      E5
#> E1   0.000  53.838 66.053 63.833 70.220  21.623 15.404  19.783   1.938  49.824
#> E10 53.838   0.000 82.915 55.158 71.598  21.096  6.873  -6.386 -12.243  49.530
#> E11 66.053  82.915  0.000 61.028 86.585  12.789 26.675  -0.957  10.015  61.032
#> E12 63.833  55.158 61.028  0.000 61.171  18.020  5.082  13.394  -6.217  38.609
#> E13 70.220  71.598 86.585 61.171  0.000  37.079 31.266  16.941  -1.174  59.404
#> E14 21.623  21.096 12.789 18.020 37.079   0.000 16.166 -10.671  29.120  33.173
#> E2  15.404   6.873 26.675  5.082 31.266  16.166  0.000  -0.651  56.404  22.091
#> E3  19.783  -6.386 -0.957 13.394 16.941 -10.671 -0.651   0.000  26.463  19.926
#> E4   1.938 -12.243 10.015 -6.217 -1.174  29.120 56.404  26.463   0.000  10.922
#> E5  49.824  49.530 61.032 38.609 59.404  33.173 22.091  19.926  10.922   0.000
#> E6  48.374  29.311 19.513 27.350 70.128  14.796 18.342  37.471  42.032  80.822
#> E7  26.095  22.442 18.142 19.396 15.031  47.887 -5.293   2.351   3.998  20.747
#> E8  -0.157   2.521  5.536  3.726 22.646  63.892 17.366 -11.548  35.236  21.672
#> E9  18.128  24.437 37.322 18.890 -8.784  46.272 22.252 -12.098 -12.287 -12.467
#>         E6     E7      E8      E9
#> E1  48.374 26.095  -0.157  18.128
#> E10 29.311 22.442   2.521  24.437
#> E11 19.513 18.142   5.536  37.322
#> E12 27.350 19.396   3.726  18.890
#> E13 70.128 15.031  22.646  -8.784
#> E14 14.796 47.887  63.892  46.272
#> E2  18.342 -5.293  17.366  22.252
#> E3  37.471  2.351 -11.548 -12.098
#> E4  42.032  3.998  35.236 -12.287
#> E5  80.822 20.747  21.672 -12.467
#> E6   0.000 23.614  15.820  13.068
#> E7  23.614  0.000  17.172  33.771
#> E8  15.820 17.172   0.000   8.861
#> E9  13.068 33.771   8.861   0.000
#> ----------------------------------------------------------------------
#> Average =  25.478 
#> Minimum =  -12.467 between environments 14 and 10 
#> Maximum =  86.585 between environments 5 and 3 
#> ----------------------------------------------------------------------
#> % Of the complex part of MS GxEjj' (Cruz and Castoldi, 1991)
#> ----------------------------------------------------------------------
#>          E1     E10     E11     E12     E13     E14      E2      E3      E4
#> E1    0.000  46.162  33.947  36.167  29.780  78.377  84.596  80.217  98.062
#> E10  46.162   0.000  17.085  44.842  28.402  78.904  93.127 106.386 112.243
#> E11  33.947  17.085   0.000  38.972  13.415  87.211  73.325 100.957  89.985
#> E12  36.167  44.842  38.972   0.000  38.829  81.980  94.918  86.606 106.217
#> E13  29.780  28.402  13.415  38.829   0.000  62.921  68.734  83.059 101.174
#> E14  78.377  78.904  87.211  81.980  62.921   0.000  83.834 110.671  70.880
#> E2   84.596  93.127  73.325  94.918  68.734  83.834   0.000 100.651  43.596
#> E3   80.217 106.386 100.957  86.606  83.059 110.671 100.651   0.000  73.537
#> E4   98.062 112.243  89.985 106.217 101.174  70.880  43.596  73.537   0.000
#> E5   50.176  50.470  38.968  61.391  40.596  66.827  77.909  80.074  89.078
#> E6   51.626  70.689  80.487  72.650  29.872  85.204  81.658  62.529  57.968
#> E7   73.905  77.558  81.858  80.604  84.969  52.113 105.293  97.649  96.002
#> E8  100.157  97.479  94.464  96.274  77.354  36.108  82.634 111.548  64.764
#> E9   81.872  75.563  62.678  81.110 108.784  53.728  77.748 112.098 112.287
#>          E5     E6      E7      E8      E9
#> E1   50.176 51.626  73.905 100.157  81.872
#> E10  50.470 70.689  77.558  97.479  75.563
#> E11  38.968 80.487  81.858  94.464  62.678
#> E12  61.391 72.650  80.604  96.274  81.110
#> E13  40.596 29.872  84.969  77.354 108.784
#> E14  66.827 85.204  52.113  36.108  53.728
#> E2   77.909 81.658 105.293  82.634  77.748
#> E3   80.074 62.529  97.649 111.548 112.098
#> E4   89.078 57.968  96.002  64.764 112.287
#> E5    0.000 19.178  79.253  78.328 112.467
#> E6   19.178  0.000  76.386  84.180  86.932
#> E7   79.253 76.386   0.000  82.828  66.229
#> E8   78.328 84.180  82.828   0.000  91.139
#> E9  112.467 86.932  66.229  91.139   0.000
#> ----------------------------------------------------------------------
#> Average =  74.522 
#> Minimum =  13.415 between environments 5 and 3 
#> Maximum =  112.467 between environments 14 and 10 
#> ----------------------------------------------------------------------
#> 
#> 
#> 
# }