Print the env_stratification model

Print an object of class ge_factanal in two ways. By default, the results are shown in the R console. The results can also be exported to the directory.

# S3 method for env_stratification
print(x, export = FALSE, file.name = NULL, digits = 3, ...)

Arguments

x
An object of class env_stratification.

export
A logical argument. If TRUE, a *.txt file is exported to the working directory

file.name
The name of the file if export = TRUE

digits
The significant digits to be shown.

...
Currently not used.

Author

Tiago Olivoto tiagoolivoto@gmail.com

Examples

# \donttest{
library(metan)
model <-
env_stratification(data_ge,
                   env = ENV,
                   gen = GEN,
                   resp = GY)
print(model)

#> Variable GY 
#> ------------------------------------------------------------------------------------
#> Environment stratification based on factor analysis
#> ------------------------------------------------------------------------------------
#>  MEGA_ENV                          ENV
#>       ME1 E1 , E10, E11, E12, E13, E5 
#>       ME2                       E3, E6
#>       ME3                     E14, E8 
#>       ME4                       E2, E4
#>       ME5                       E7, E9
#> ------------------------------------------------------------------------------------
#> Statistic by environment 
#> ------------------------------------------------------------------------------------
#>  ENV MEGA_ENV  MEAN   MIN   MAX     CV
#>   E1      ME1 2.521 1.974 2.902 13.304
#>  E10      ME1 2.175 1.536 2.575 14.435
#>  E11      ME1 1.368 0.899 1.683 16.392
#>  E12      ME1 1.609 1.022 2.000 20.309
#>  E13      ME1 2.910 1.831 3.520 16.760
#>   E5      ME1 3.910 3.366 4.812 10.709
#>   E3      ME2 4.064 3.435 4.569  8.223
#>   E6      ME2 2.663 2.336 2.985  7.955
#>  E14      ME3 1.782 1.427 2.060 11.705
#>   E8      ME3 2.536 2.045 2.879 10.545
#>   E2      ME4 3.180 2.613 3.608  8.253
#>   E4      ME4 3.675 3.019 4.271 11.451
#>   E7      ME5 1.989 1.393 2.551 16.850
#>   E9      ME5 3.057 1.943 3.725 15.567
#> ------------------------------------------------------------------------------------
#> Statistic by mega-environment (mean values) 
#> ------------------------------------------------------------------------------------
#>  MEGA_ENV  MEAN   MIN   MAX     CV
#>       ME1 2.415 1.771 2.915 15.318
#>       ME2 3.364 2.885 3.777  8.089
#>       ME3 2.159 1.736 2.469 11.125
#>       ME4 3.428 2.816 3.940  9.852
#>       ME5 2.523 1.668 3.138 16.209
#> ------------------------------------------------------------------------------------
#> Mean = mean; Min = minimum; Max = maximum; CV = coefficient of variation (%)
#> ------------------------------------------------------------------------------------
#> 
#> 

# }