Print an object of class ge_factanal

Print the ge_factanal object in two ways. By default, the results are shown in the R console. The results can also be exported to the directory.

# S3 method for ge_factanal
print(x, export = FALSE, file.name = NULL, digits = 4, ...)

Arguments

x
An object of class ge_factanal.

export
A logical argument. If TRUE, a *.txt file is exported to the working directory

file.name
The name of the file if export = TRUE

digits
The significant digits to be shown.

...
Options used by the tibble package to format the output. See tibble::print() for more details.

Author

Tiago Olivoto tiagoolivoto@gmail.com

Examples

# \donttest{
model <- ge_factanal(data_ge2,
  env = ENV,
  gen = GEN,
  rep = REP,
  resp = PH
)
print(model)

#> Variable PH 
#> ------------------------------------------------------------------------------------
#> Correlation matrix among environments
#> ------------------------------------------------------------------------------------
#> # A tibble: 4 x 5
#>   ENV        A1      A2      A3      A4
#>   <chr>   <dbl>   <dbl>   <dbl>   <dbl>
#> 1 A1     1       0.3919 -0.4723 -0.4983
#> 2 A2     0.3919  1      -0.2203  0.2722
#> 3 A3    -0.4723 -0.2203  1       0.3238
#> 4 A4    -0.4983  0.2722  0.3238  1     
#> ------------------------------------------------------------------------------------
#> Eigenvalues and explained variance
#> ------------------------------------------------------------------------------------
#> # A tibble: 4 x 4
#>   PCA   Eigenvalues Variance Cumul_var
#>   <chr>       <dbl>    <dbl>     <dbl>
#> 1 PC1        1.926    48.15      48.15
#> 2 PC2        1.263    31.58      79.74
#> 3 PC3        0.5928   14.82      94.56
#> 4 PC4        0.2178    5.444    100   
#> ------------------------------------------------------------------------------------
#> Initial loadings
#> ------------------------------------------------------------------------------------
#> # A tibble: 4 x 3
#>   Env       FA1      FA2
#>   <chr>   <dbl>    <dbl>
#> 1 A1    -0.8912  0.08731
#> 2 A2    -0.3718  0.8816 
#> 3 A3     0.7660 -0.04547
#> 4 A4     0.6380  0.6902 
#> ------------------------------------------------------------------------------------
#> Loadings after varimax rotation and commonalities
#> ------------------------------------------------------------------------------------
#> # A tibble: 4 x 5
#>   Env       FA1     FA2 Communality Uniquenesses
#>   <chr>   <dbl>   <dbl>       <dbl>        <dbl>
#> 1 A1    -0.8763  0.1839      0.8018      0.1982 
#> 2 A2    -0.2735  0.9169      0.9154      0.08461
#> 3 A3     0.7564 -0.1287      0.5888      0.4112 
#> 4 A4     0.7095  0.6166      0.8835      0.1165 
#> ------------------------------------------------------------------------------------
#> Environmental stratification based on factor analysis
#> ------------------------------------------------------------------------------------
#> # A tibble: 4 x 6
#>   Env   Factor  Mean   Min   Max     CV
#>   <chr> <chr>  <dbl> <dbl> <dbl>  <dbl>
#> 1 A1    FA1    2.793 2.692 2.935  2.810
#> 2 A3    FA1    2.167 2.028 2.599  7.724
#> 3 A4    FA1    2.518 2.305 2.638  3.655
#> 4 A2    FA2    2.462 1.959 2.947 16.20 
#> ------------------------------------------------------------------------------------
#> Mean = mean; Min = minimum; Max = maximum; CV = coefficient of variation (%)
#> ------------------------------------------------------------------------------------
#> 
#> 
#> 

# }