Print the partial correlation coefficients

Print an object of class lpcor or or lpcor_group in two ways. By default, the results are shown in the R console. The results can also be exported to the directory.

# S3 method for lpcor
print(x, export = FALSE, file.name = NULL, digits = 3, ...)

Arguments

x
An object of class lpcor or lpcor_group.

export
A logical argument. If TRUE, a *.txt file is exported to the working directory

file.name
The name of the file if export = TRUE

digits
The significant digits to be shown.

...
Options used by the tibble package to format the output. See tibble::print() for more details.

Author

Tiago Olivoto tiagoolivoto@gmail.com

Examples

# \donttest{
library(metan)
pcor <- lpcor(data_ge2, NR, NKR, NKE)
print(pcor)

#> # A tibble: 3 x 5
#>   Pairs     linear partial     t  prob
#>   <chr>      <dbl>   <dbl> <dbl> <dbl>
#> 1 NR x NKR  0.0206  -0.767 -14.8     0
#> 2 NR x NKE  0.626    0.866  21.4     0
#> 3 NKR x NKE 0.708    0.891  24.3     0

# Compute the correlations for each level of the factor ENV
lpc2 <- lpcor(data_ge2,
              NR, NKR, NKE,
              by = ENV)
print(lpc2)

#> # A tibble: 420 x 6
#>    ENV   Pairs      linear partial       t   prob
#>    <fct> <chr>       <dbl>   <dbl>   <dbl>  <dbl>
#>  1 A1    PH x EH   0.425     0.981  25.1   0     
#>  2 A1    PH x EP  -0.124    -0.971 -20.1   0     
#>  3 A1    PH x EL  -0.00287  -0.169  -0.841 0.408 
#>  4 A1    PH x ED   0.0729    0.301   1.55  0.135 
#>  5 A1    PH x CL  -0.158    -0.311  -1.60  0.122 
#>  6 A1    PH x CD   0.140     0.183   0.911 0.372 
#>  7 A1    PH x CW  -0.0111    0.155   0.771 0.448 
#>  8 A1    PH x KW   0.0825    0.155   0.770 0.449 
#>  9 A1    PH x NR  -0.0142    0.389   2.07  0.0496
#> 10 A1    PH x NKR  0.140     0.271   1.38  0.181 
#> # ... with 410 more rows

# }