Print an object of class waasb

Print a waasb object in two ways. By default, the results are shown in the R console. The results can also be exported to the directory.

# S3 method for waasb
print(x, export = FALSE, blup = FALSE,
  file.name = NULL, digits = 4, ...)

Arguments

x
An object of class waasb.

export
A logical argument. If TRUE|T, a *.txt file is exported to the working directory

blup
A logical argument. If TRUE|T, the blups are shown.

file.name
The name of the file if export = TRUE

digits
The significant digits to be shown.

...
Options used by the tibble package to format the output. See trunc_mat for more details.

Examples

library(metan)
model <- waasb(data_ge,
  resp = c(GY, HM),
  gen = GEN,
  env = ENV,
  rep = REP
)

#> Error in fortify.merMod(Complete): could not find function "fortify.merMod"

print(model)

#> Variable PH 
#> ---------------------------------------------------------------------------
#> Environmental index
#> ---------------------------------------------------------------------------
#> # A tibble: 4 x 4
#>   ENV    Mean   index class      
#>   <fct> <dbl>   <dbl> <chr>      
#> 1 A1     2.79  0.308  favorable  
#> 2 A2     2.46 -0.0229 unfavorable
#> 3 A3     2.17 -0.318  unfavorable
#> 4 A4     2.52  0.0329 favorable  
#> ---------------------------------------------------------------------------
#> Analysis for all environments
#> ---------------------------------------------------------------------------
#> # A tibble: 13 x 6
#>    Genotype  Mean Mean_rp Sd_rp    Wi  rank
#>    <chr>    <dbl>   <dbl> <dbl> <dbl> <dbl>
#>  1 H1        2.62   106.   9.38  99.3     2
#>  2 H10       2.31    93.0  6.93  88.3    12
#>  3 H11       2.39    96.1  6.27  91.9     9
#>  4 H12       2.44    98.6 11.2   91.0    10
#>  5 H13       2.54   103.  12.4   94.4     7
#>  6 H2        2.60   105.   9.91  98.1     3
#>  7 H3        2.59   104.  11.2   96.6     6
#>  8 H4        2.58   104.   8.76  97.7     5
#>  9 H5        2.57   103.   5.55  99.5     1
#> 10 H6        2.56   103.   7.83  97.7     4
#> 11 H7        2.40    96.8  6.67  92.3     8
#> 12 H8        2.33    93.6  9.61  87.1    13
#> 13 H9        2.36    94.8  6.85  90.2    11
#> ---------------------------------------------------------------------------
#> Analysis for favorable environments
#> ---------------------------------------------------------------------------
#> # A tibble: 13 x 6
#>    Genotype  Mean Mean_rp Sd_rp    Wi  rank
#>    <chr>    <dbl>   <dbl> <dbl> <dbl> <dbl>
#>  1 H1        2.68   101.  5.09   97.6     9
#>  2 H10       2.59    97.3 3.38   95.0    12
#>  3 H11       2.66   100.  2.41   98.5     7
#>  4 H12       2.61    98.5 3.00   96.5    10
#>  5 H13       2.66   100.  1.63   99.3     5
#>  6 H2        2.66   100.  0.203 100.0     3
#>  7 H3        2.70   101.  5.22   97.9     8
#>  8 H4        2.71   102.  0.839 102.      2
#>  9 H5        2.74   103.  2.32  102.      1
#> 10 H6        2.63    98.9 0.292  98.7     6
#> 11 H7        2.65    99.9 0.812  99.3     4
#> 12 H8        2.62    98.9 3.60   96.5    11
#> 13 H9        2.62    98.1 9.33   91.8    13
#> ---------------------------------------------------------------------------
#> Analysis for unfavorable environments
#> ---------------------------------------------------------------------------
#> # A tibble: 13 x 6
#>    Genotype  Mean Mean_rp Sd_rp    Wi  rank
#>    <chr>    <dbl>   <dbl> <dbl> <dbl> <dbl>
#>  1 H1        2.56   110.  12.4  102.      1
#>  2 H10       2.04    88.7  7.65  83.5    12
#>  3 H11       2.12    92.1  6.93  87.4     9
#>  4 H12       2.26    98.6 19.2   85.7    11
#>  5 H13       2.41   105.  20.8   91.2     7
#>  6 H2        2.55   110.  14.3   99.9     2
#>  7 H3        2.49   107.  17.8   94.9     6
#>  8 H4        2.45   105.  14.9   95.1     5
#>  9 H5        2.40   103.   9.32  97.0     4
#> 10 H6        2.49   107.  10.9   99.7     3
#> 11 H7        2.16    93.6  9.69  87.1    10
#> 12 H8        2.03    88.3 12.3   80.0    13
#> 13 H9        2.11    91.5  3.02  89.4     8
#> 
#> 
#> 
#> Variable EH 
#> ---------------------------------------------------------------------------
#> Environmental index
#> ---------------------------------------------------------------------------
#> # A tibble: 4 x 4
#>   ENV    Mean   index class      
#>   <fct> <dbl>   <dbl> <chr>      
#> 1 A1     1.58  0.234  favorable  
#> 2 A2     1.31 -0.0302 unfavorable
#> 3 A3     1.08 -0.265  unfavorable
#> 4 A4     1.41  0.0620 favorable  
#> ---------------------------------------------------------------------------
#> Analysis for all environments
#> ---------------------------------------------------------------------------
#> # A tibble: 13 x 6
#>    Genotype  Mean Mean_rp Sd_rp    Wi  rank
#>    <chr>    <dbl>   <dbl> <dbl> <dbl> <dbl>
#>  1 H1        1.50   112.  16.7  101.      1
#>  2 H10       1.26    93.3 12.7   84.7    11
#>  3 H11       1.27    94.3 11.3   86.6    10
#>  4 H12       1.28    96.4 21.3   82.0    12
#>  5 H13       1.35   102.  15.3   91.4     6
#>  6 H2        1.38   103.  20.6   89.4     8
#>  7 H3        1.41   105.  19.7   91.3     7
#>  8 H4        1.43   105.  14.9   95.2     4
#>  9 H5        1.37   101.   8.93  95.4     3
#> 10 H6        1.41   106.   8.93  99.7     2
#> 11 H7        1.32    98.4  9.61  91.9     5
#> 12 H8        1.21    89.9 17.4   78.1    13
#> 13 H9        1.27    93.6  9.03  87.6     9
#> ---------------------------------------------------------------------------
#> Analysis for favorable environments
#> ---------------------------------------------------------------------------
#> # A tibble: 13 x 6
#>    Genotype  Mean Mean_rp  Sd_rp    Wi  rank
#>    <chr>    <dbl>   <dbl>  <dbl> <dbl> <dbl>
#>  1 H1        1.56   105.   2.82  103.      2
#>  2 H10       1.52   102.   0.882 102.      3
#>  3 H11       1.47    98.6  2.15   97.1     9
#>  4 H12       1.47    98.7  1.40   97.7     8
#>  5 H13       1.50   100.   0.364 100.      6
#>  6 H2        1.36    91.6  9.35   85.3    13
#>  7 H3        1.47    98.5  3.05   96.4    10
#>  8 H4        1.57   105.   1.92  104.      1
#>  9 H5        1.55   104.   4.73  101.      4
#> 10 H6        1.48    99.1  1.00   98.5     7
#> 11 H7        1.51   101.   0.442 101.      5
#> 12 H8        1.44    96.9  5.71   93.1    11
#> 13 H9        1.48    98.7 10.2    91.8    12
#> ---------------------------------------------------------------------------
#> Analysis for unfavorable environments
#> ---------------------------------------------------------------------------
#> # A tibble: 13 x 6
#>    Genotype  Mean Mean_rp Sd_rp    Wi  rank
#>    <chr>    <dbl>   <dbl> <dbl> <dbl> <dbl>
#>  1 H1       1.45    119.  24.8  102.      2
#>  2 H10      0.999    84.5 13.2   75.6    11
#>  3 H11      1.06     89.9 17.5   78.1    10
#>  4 H12      1.09     94.1 36.7   69.3    12
#>  5 H13      1.21    103.  26.4   85.2     7
#>  6 H2       1.40    115.  25.3   97.9     3
#>  7 H3       1.35    111.  31.7   89.4     5
#>  8 H4       1.28    106.  25.8   88.3     6
#>  9 H5       1.19     98.6 13.5   89.4     4
#> 10 H6       1.35    112.   7.91 107.      1
#> 11 H7       1.13     95.4 15.5   84.9     8
#> 12 H8       0.969    82.8 26.0   65.3    13
#> 13 H9       1.06     88.6  6.25  84.4     9
#> 
#> 
#> 
#> Variable EP 
#> ---------------------------------------------------------------------------
#> Environmental index
#> ---------------------------------------------------------------------------
#> # A tibble: 4 x 4
#>   ENV    Mean    index class      
#>   <fct> <dbl>    <dbl> <chr>      
#> 1 A1    0.567  0.0299  favorable  
#> 2 A2    0.527 -0.00996 unfavorable
#> 3 A3    0.496 -0.0411  unfavorable
#> 4 A4    0.558  0.0211  favorable  
#> ---------------------------------------------------------------------------
#> Analysis for all environments
#> ---------------------------------------------------------------------------
#> # A tibble: 13 x 6
#>    Genotype  Mean Mean_rp Sd_rp    Wi  rank
#>    <chr>    <dbl>   <dbl> <dbl> <dbl> <dbl>
#>  1 H1       0.570   106.   8.55 100.      2
#>  2 H10      0.545   101.   5.74  97.5     4
#>  3 H11      0.527    98.2  6.87  93.6    10
#>  4 H12      0.519    96.7 10.4   89.7    12
#>  5 H13      0.532    99.1  3.33  96.8     6
#>  6 H2       0.525    98.1 11.2   90.5    11
#>  7 H3       0.538   100.   8.67  94.4     9
#>  8 H4       0.546   102.   6.60  97.0     5
#>  9 H5       0.530    98.6  3.88  96.0     7
#> 10 H6       0.553   103.   3.64 101.      1
#> 11 H7       0.547   102.   3.20  99.9     3
#> 12 H8       0.514    95.9  9.34  89.6    13
#> 13 H9       0.532    99.0  6.50  94.6     8
#> ---------------------------------------------------------------------------
#> Analysis for favorable environments
#> ---------------------------------------------------------------------------
#> # A tibble: 13 x 6
#>    Genotype  Mean Mean_rp Sd_rp    Wi  rank
#>    <chr>    <dbl>   <dbl> <dbl> <dbl> <dbl>
#>  1 H1       0.586   104.  8.75   98.3     9
#>  2 H10      0.592   105.  3.16  103.      1
#>  3 H11      0.553    98.4 4.19   95.5    12
#>  4 H12      0.564   100.  1.76   99.1     8
#>  5 H13      0.563   100.  0.146 100.      4
#>  6 H2       0.512    91.1 8.95   85.1    13
#>  7 H3       0.546    97.0 1.86   95.8    11
#>  8 H4       0.576   102.  0.719 102.      2
#>  9 H5       0.567   101.  2.18   99.4     7
#> 10 H6       0.564   100.  1.21   99.5     6
#> 11 H7       0.572   102.  1.37  101.      3
#> 12 H8       0.550    97.8 2.81   95.9    10
#> 13 H9       0.565   100.  0.833  99.8     5
#> ---------------------------------------------------------------------------
#> Analysis for unfavorable environments
#> ---------------------------------------------------------------------------
#> # A tibble: 13 x 6
#>    Genotype  Mean Mean_rp Sd_rp    Wi  rank
#>    <chr>    <dbl>   <dbl> <dbl> <dbl> <dbl>
#>  1 H1       0.555   108.  11.3  101.      2
#>  2 H10      0.498    97.6  5.47  93.9     7
#>  3 H11      0.500    98.0 11.1   90.5    10
#>  4 H12      0.474    93.1 16.5   82.0    13
#>  5 H13      0.500    97.9  5.29  94.3     5
#>  6 H2       0.538   105.  10.3   98.1     4
#>  7 H3       0.530   103.  13.5   94.3     6
#>  8 H4       0.515   101.  11.3   93.0     9
#>  9 H5       0.494    96.4  4.60  93.3     8
#> 10 H6       0.541   106.   2.60 104.      1
#> 11 H7       0.522   102.   5.35  98.6     3
#> 12 H8       0.479    93.9 15.4   83.5    12
#> 13 H9       0.500    97.6 10.9   90.3    11
#> 
#> 
#> 
#> Variable EL 
#> ---------------------------------------------------------------------------
#> Environmental index
#> ---------------------------------------------------------------------------
#> # A tibble: 4 x 4
#>   ENV    Mean   index class      
#>   <fct> <dbl>   <dbl> <chr>      
#> 1 A1     15.6  0.458  favorable  
#> 2 A2     15.2  0.0703 favorable  
#> 3 A3     14.7 -0.495  unfavorable
#> 4 A4     15.1 -0.0333 unfavorable
#> ---------------------------------------------------------------------------
#> Analysis for all environments
#> ---------------------------------------------------------------------------
#> # A tibble: 13 x 6
#>    Genotype  Mean Mean_rp Sd_rp    Wi  rank
#>    <chr>    <dbl>   <dbl> <dbl> <dbl> <dbl>
#>  1 H1        15.1    99.6  1.29  98.7     3
#>  2 H10       15.1    99.7  3.51  97.3     7
#>  3 H11       15.2   100.   4.22  97.4     6
#>  4 H12       14.3    94.2  4.40  91.2    12
#>  5 H13       15.0    99.2  5.58  95.5     9
#>  6 H2        15.3   101.   4.50  98.2     4
#>  7 H3        14.5    95.8  8.26  90.3    13
#>  8 H4        15.7   104.   4.33 101.      1
#>  9 H5        15.6   103.   3.69 100.      2
#> 10 H6        15.8   104.   8.75  97.9     5
#> 11 H7        15.4   101.   7.12  96.6     8
#> 12 H8        15.0    99.0  7.98  93.7    11
#> 13 H9        15.0    99.2  7.42  94.2    10
#> ---------------------------------------------------------------------------
#> Analysis for favorable environments
#> ---------------------------------------------------------------------------
#> # A tibble: 13 x 6
#>    Genotype  Mean Mean_rp Sd_rp    Wi  rank
#>    <chr>    <dbl>   <dbl> <dbl> <dbl> <dbl>
#>  1 H1        15.2    98.5 0.357  98.3     7
#>  2 H10       15.8   102.  0.807 102.      3
#>  3 H11       15.9   103.  5.11   99.4     5
#>  4 H12       14.3    92.4 6.70   87.9    13
#>  5 H13       15.2    98.9 5.97   94.9    10
#>  6 H2        15.5   100.  6.34   96.2     8
#>  7 H3        15.4    99.7 1.10   99.0     6
#>  8 H4        16.2   105.  4.35  102.      2
#>  9 H5        16.1   105.  5.16  101.      4
#> 10 H6        17.0   110.  4.58  107.      1
#> 11 H7        15.0    97.4 2.09   96.0     9
#> 12 H8        14.4    93.4 4.97   90.1    11
#> 13 H9        14.5    94.0 7.47   89.0    12
#> ---------------------------------------------------------------------------
#> Analysis for unfavorable environments
#> ---------------------------------------------------------------------------
#> # A tibble: 13 x 6
#>    Genotype  Mean Mean_rp  Sd_rp    Wi  rank
#>    <chr>    <dbl>   <dbl>  <dbl> <dbl> <dbl>
#>  1 H1        15.0   101.   0.687 100.      4
#>  2 H10       14.4    96.9  2.15   95.4     9
#>  3 H11       14.6    97.7  0.833  97.2     8
#>  4 H12       14.3    95.9  1.07   95.2    10
#>  5 H13       14.8    99.6  7.56   94.5    11
#>  6 H2        15.2   102.   4.29   99.1     6
#>  7 H3        13.7    92.0 12.0    83.9    13
#>  8 H4        15.2   102.   5.25   98.7     7
#>  9 H5        15.0   101.   1.05  100.      3
#> 10 H6        14.6    97.7  7.73   92.5    12
#> 11 H7        15.7   106.   9.02   99.4     5
#> 12 H8        15.6   105.   6.25  100.      2
#> 13 H9        15.6   104.   1.67  103.      1
#> 
#> 
#> 
#> Variable ED 
#> ---------------------------------------------------------------------------
#> Environmental index
#> ---------------------------------------------------------------------------
#> # A tibble: 4 x 4
#>   ENV    Mean  index class      
#>   <fct> <dbl>  <dbl> <chr>      
#> 1 A1     51.6  2.07  favorable  
#> 2 A2     48.7 -0.799 unfavorable
#> 3 A3     47.9 -1.66  unfavorable
#> 4 A4     49.9  0.394 favorable  
#> ---------------------------------------------------------------------------
#> Analysis for all environments
#> ---------------------------------------------------------------------------
#> # A tibble: 13 x 6
#>    Genotype  Mean Mean_rp Sd_rp    Wi  rank
#>    <chr>    <dbl>   <dbl> <dbl> <dbl> <dbl>
#>  1 H1        51.2   103.   3.40 101.      2
#>  2 H10       48.4    97.7  4.97  94.3    12
#>  3 H11       48.8    98.5  3.81  95.9    10
#>  4 H12       48.6    98.1  2.00  96.7     9
#>  5 H13       50.6   102.   4.35  99.2     3
#>  6 H2        50.9   103.   5.79  99.0     4
#>  7 H3        49.4    99.8  2.64  98.0     7
#>  8 H4        49.2    99.4  2.99  97.4     8
#>  9 H5        49.9   101.   3.75  98.3     5
#> 10 H6        51.5   104.   2.17 103.      1
#> 11 H7        49.5    99.9  2.43  98.2     6
#> 12 H8        48.4    97.5  3.74  95.0    11
#> 13 H9        47.6    96.0  4.66  92.8    13
#> ---------------------------------------------------------------------------
#> Analysis for favorable environments
#> ---------------------------------------------------------------------------
#> # A tibble: 13 x 6
#>    Genotype  Mean Mean_rp Sd_rp    Wi  rank
#>    <chr>    <dbl>   <dbl> <dbl> <dbl> <dbl>
#>  1 H1        51.1   101.   2.45  99.1     4
#>  2 H10       51.6   102.   2.11 100.      2
#>  3 H11       50.4    99.3  6.34  95.0    13
#>  4 H12       49.9    98.4  2.09  97.0     9
#>  5 H13       50.7    99.7  5.49  96.0    12
#>  6 H2        49.9    98.3  3.31  96.1    11
#>  7 H3        50.2    98.8  2.15  97.4     7
#>  8 H4        50.9   100.   2.84  98.4     6
#>  9 H5        50.7    99.8  4.58  96.8    10
#> 10 H6        52.2   103.   2.80 101.      1
#> 11 H7        50.9   100.   2.36  98.6     5
#> 12 H8        51.0   100.   1.10  99.7     3
#> 13 H9        50.5    99.5  3.69  97.0     8
#> ---------------------------------------------------------------------------
#> Analysis for unfavorable environments
#> ---------------------------------------------------------------------------
#> # A tibble: 13 x 6
#>    Genotype  Mean Mean_rp Sd_rp    Wi  rank
#>    <chr>    <dbl>   <dbl> <dbl> <dbl> <dbl>
#>  1 H1        51.2   106.  0.535 106.      2
#>  2 H10       45.3    93.8 3.00   91.7    12
#>  3 H11       47.2    97.7 0.846  97.1     8
#>  4 H12       47.2    97.8 2.71   96.0     9
#>  5 H13       50.5   104.  1.92  103.      4
#>  6 H2        51.9   107.  2.38  106.      1
#>  7 H3        48.7   101.  3.51   98.4     6
#>  8 H4        47.6    98.5 3.94   95.8    10
#>  9 H5        49.1   102.  4.18   98.9     5
#> 10 H6        50.8   105.  1.04  104.      3
#> 11 H7        48.1    99.5 3.42   97.2     7
#> 12 H8        45.7    94.6 2.45   92.9    11
#> 13 H9        44.6    92.4 1.13   91.6    13
#> 
#> 
#> 
#> Variable CL 
#> ---------------------------------------------------------------------------
#> Environmental index
#> ---------------------------------------------------------------------------
#> # A tibble: 4 x 4
#>   ENV    Mean  index class      
#>   <fct> <dbl>  <dbl> <chr>      
#> 1 A1     29.7  0.716 favorable  
#> 2 A2     28.5 -0.548 unfavorable
#> 3 A3     28.4 -0.562 unfavorable
#> 4 A4     29.4  0.394 favorable  
#> ---------------------------------------------------------------------------
#> Analysis for all environments
#> ---------------------------------------------------------------------------
#> # A tibble: 13 x 6
#>    Genotype  Mean Mean_rp Sd_rp    Wi  rank
#>    <chr>    <dbl>   <dbl> <dbl> <dbl> <dbl>
#>  1 H1        30.1   104.   7.44  98.7     3
#>  2 H10       28.4    97.8  8.78  91.9    13
#>  3 H11       28.3    97.4  4.69  94.2     8
#>  4 H12       28.2    97.3  5.15  93.8    10
#>  5 H13       29.4   101.   9.02  95.3     5
#>  6 H2        29.3   101.  11.6   93.2    11
#>  7 H3        28.5    98.3  5.86  94.4     7
#>  8 H4        28.6    98.6  5.06  95.2     6
#>  9 H5        29.4   101.   6.33  97.0     4
#> 10 H6        30.3   104.   7.39  99.3     1
#> 11 H7        29.5   102.   4.51  98.7     2
#> 12 H8        28.7    98.9  9.54  92.4    12
#> 13 H9        28.6    98.4  6.26  94.1     9
#> ---------------------------------------------------------------------------
#> Analysis for favorable environments
#> ---------------------------------------------------------------------------
#> # A tibble: 13 x 6
#>    Genotype  Mean Mean_rp Sd_rp    Wi  rank
#>    <chr>    <dbl>   <dbl> <dbl> <dbl> <dbl>
#>  1 H1        28.9    97.8 4.88   94.5     9
#>  2 H10       31.0   105.  0.791 104.      1
#>  3 H11       29.7   101.  4.41   97.6     7
#>  4 H12       30.0   101.  1.44  100.      4
#>  5 H13       29.2    98.6 8.90   92.6    11
#>  6 H2        27.1    91.6 3.04   89.6    13
#>  7 H3        27.7    93.6 0.786  93.1    10
#>  8 H4        28.5    96.4 5.93   92.4    12
#>  9 H5        29.9   101.  8.32   95.5     8
#> 10 H6        31.1   105.  2.02  104.      2
#> 11 H7        30.4   103.  0.587 102.      3
#> 12 H8        30.6   104.  7.84   98.3     5
#> 13 H9        30.3   102.  7.10   97.7     6
#> ---------------------------------------------------------------------------
#> Analysis for unfavorable environments
#> ---------------------------------------------------------------------------
#> # A tibble: 13 x 6
#>    Genotype  Mean Mean_rp  Sd_rp    Wi  rank
#>    <chr>    <dbl>   <dbl>  <dbl> <dbl> <dbl>
#>  1 H1        31.2   110.   1.49  109.      1
#>  2 H10       25.8    90.6  4.98   87.3    12
#>  3 H11       26.8    94.1  2.17   92.7    10
#>  4 H12       26.5    93.2  3.45   90.9    11
#>  5 H13       29.6   104.  11.6    96.3     6
#>  6 H2        31.4   110.   6.33  106.      2
#>  7 H3        29.3   103.   3.72  101.      3
#>  8 H4        28.7   101.   4.84   97.5     4
#>  9 H5        28.9   101.   7.15   96.6     5
#> 10 H6        29.4   103.  12.5    94.9     8
#> 11 H7        28.7   101.   7.58   95.8     7
#> 12 H8        26.8    94.2 11.1    86.7    13
#> 13 H9        26.8    94.3  0.232  94.1     9
#> 
#> 
#> 
#> Variable CD 
#> ---------------------------------------------------------------------------
#> Environmental index
#> ---------------------------------------------------------------------------
#> # A tibble: 4 x 4
#>   ENV    Mean   index class      
#>   <fct> <dbl>   <dbl> <chr>      
#> 1 A1     16.4  0.426  favorable  
#> 2 A2     15.9 -0.0828 unfavorable
#> 3 A3     15.8 -0.2    unfavorable
#> 4 A4     15.8 -0.144  unfavorable
#> ---------------------------------------------------------------------------
#> Analysis for all environments
#> ---------------------------------------------------------------------------
#> # A tibble: 13 x 6
#>    Genotype  Mean Mean_rp Sd_rp    Wi  rank
#>    <chr>    <dbl>   <dbl> <dbl> <dbl> <dbl>
#>  1 H1        15.7    98.4  2.08  97.0     6
#>  2 H10       15.9    99.5  3.46  97.2     5
#>  3 H11       16.0   100.   3.19  98.0     4
#>  4 H12       14.8    92.7  3.73  90.2    13
#>  5 H13       15.8    99.1  5.33  95.5     9
#>  6 H2        16.0   100.   5.52  96.7     7
#>  7 H3        15.7    98.3  6.22  94.1    12
#>  8 H4        16.5   103.   4.13 101.      2
#>  9 H5        16.6   104.   2.51 102.      1
#> 10 H6        16.6   104.   7.45  98.7     3
#> 11 H7        16.2   101.   7.01  96.5     8
#> 12 H8        15.9    99.6  7.31  94.7    11
#> 13 H9        15.9    99.5  6.21  95.3    10
#> ---------------------------------------------------------------------------
#> Analysis for favorable environments
#> ---------------------------------------------------------------------------
#> # A tibble: 13 x 6
#>    Genotype  Mean Mean_rp Sd_rp    Wi  rank
#>    <chr>    <dbl>   <dbl> <dbl> <dbl> <dbl>
#>  1 H1        15.7    95.8    NA    NA     1
#>  2 H10       16.8   102.     NA    NA     2
#>  3 H11       17.2   105.     NA    NA     3
#>  4 H12       15.6    95.4    NA    NA     4
#>  5 H13       15.5    94.2    NA    NA     5
#>  6 H2        15.5    94.4    NA    NA     6
#>  7 H3        16.4    99.8    NA    NA     7
#>  8 H4        16.7   102.     NA    NA     8
#>  9 H5        16.7   102.     NA    NA     9
#> 10 H6        17.6   107.     NA    NA    10
#> 11 H7        16.1    98.4    NA    NA    11
#> 12 H8        16.5   100.     NA    NA    12
#> 13 H9        16.9   103.     NA    NA    13
#> ---------------------------------------------------------------------------
#> Analysis for unfavorable environments
#> ---------------------------------------------------------------------------
#> # A tibble: 13 x 6
#>    Genotype  Mean Mean_rp Sd_rp    Wi  rank
#>    <chr>    <dbl>   <dbl> <dbl> <dbl> <dbl>
#>  1 H1        15.7    99.2  1.48  98.2     4
#>  2 H10       15.6    98.6  3.52  96.2     9
#>  3 H11       15.6    98.6  1.09  97.9     5
#>  4 H12       14.5    91.8  3.99  89.1    13
#>  5 H13       16.0   101.   5.16  97.3     6
#>  6 H2        16.2   102.   4.66  99.3     3
#>  7 H3        15.5    97.8  7.51  92.7    12
#>  8 H4        16.5   104.   4.88 101.      2
#>  9 H5        16.6   105.   2.55 103.      1
#> 10 H6        16.2   102.   8.64  96.7     7
#> 11 H7        16.2   102.   8.27  96.6     8
#> 12 H8        15.7    99.3  8.93  93.3    11
#> 13 H9        15.5    98.2  6.92  93.5    10
#> 
#> 
#> 
#> Variable CW 
#> ---------------------------------------------------------------------------
#> Environmental index
#> ---------------------------------------------------------------------------
#> # A tibble: 4 x 4
#>   ENV    Mean index class      
#>   <fct> <dbl> <dbl> <chr>      
#> 1 A1     28.3  3.52 favorable  
#> 2 A2     23.8 -1.06 unfavorable
#> 3 A3     20.8 -4.03 unfavorable
#> 4 A4     26.4  1.58 favorable  
#> ---------------------------------------------------------------------------
#> Analysis for all environments
#> ---------------------------------------------------------------------------
#> # A tibble: 13 x 6
#>    Genotype  Mean Mean_rp Sd_rp    Wi  rank
#>    <chr>    <dbl>   <dbl> <dbl> <dbl> <dbl>
#>  1 H1        26.7   110.  22.3   94.7     4
#>  2 H10       22.8    90.7 23.5   74.9    11
#>  3 H11       22.6    90.8 12.5   82.3    10
#>  4 H12       22.6    90.7  6.71  86.2     8
#>  5 H13       26.0   106.  26.5   87.7     5
#>  6 H2        25.7   105.  32.8   83.2     9
#>  7 H3        22.9    92.5  7.84  87.3     6
#>  8 H4        25.7   104.   9.57  97.1     3
#>  9 H5        27.7   112.  13.9  103.      1
#> 10 H6        27.3   109.  13.9   99.8     2
#> 11 H7        25.9   105.  26.4   87.0     7
#> 12 H8        23.4    93.2 29.5   73.4    13
#> 13 H9        23.2    91.7 25.1   74.8    12
#> ---------------------------------------------------------------------------
#> Analysis for favorable environments
#> ---------------------------------------------------------------------------
#> # A tibble: 13 x 6
#>    Genotype  Mean Mean_rp Sd_rp    Wi  rank
#>    <chr>    <dbl>   <dbl> <dbl> <dbl> <dbl>
#>  1 H1        24.9    91.4 12.0   83.3    11
#>  2 H10       26.3    96.5 13.7   87.2     8
#>  3 H11       25.1    91.9 12.0   83.7    10
#>  4 H12       26.1    95.4  6.91  90.7     7
#>  5 H13       25.1    91.1 27.3   72.7    12
#>  6 H2        22.6    82.8 16.3   71.9    13
#>  7 H3        24.3    88.9  4.15  86.1     9
#>  8 H4        28.2   103.   7.81  97.9     6
#>  9 H5        30.8   113.   1.25 112.      1
#> 10 H6        32.0   117.  13.5  108.      2
#> 11 H7        30.0   110.   7.36 105.      4
#> 12 H8        29.7   108.  14.8   98.0     5
#> 13 H9        30.7   112.   8.70 106.      3
#> ---------------------------------------------------------------------------
#> Analysis for unfavorable environments
#> ---------------------------------------------------------------------------
#> # A tibble: 13 x 6
#>    Genotype  Mean Mean_rp  Sd_rp    Wi  rank
#>    <chr>    <dbl>   <dbl>  <dbl> <dbl> <dbl>
#>  1 H1        28.5   128.   1.44  127.      1
#>  2 H10       19.3    85.0 36.6    60.3    12
#>  3 H11       20.2    89.7 17.8    77.6     9
#>  4 H12       19.2    86.0  0.590  85.6     8
#>  5 H13       27.0   120.  23.0   104.      3
#>  6 H2        28.8   128.  30.6   107.      2
#>  7 H3        21.5    96.2 10.7    89.0     7
#>  8 H4        23.3   104.  14.6    94.2     5
#>  9 H5        24.6   112.  24.0    95.5     4
#> 10 H6        22.5   102.  13.0    92.9     6
#> 11 H7        21.8   100.  44.2    70.3    10
#> 12 H8        17.1    78.5 39.0    52.2    13
#> 13 H9        15.7    71.3 12.4    62.9    11
#> 
#> 
#> 
#> Variable KW 
#> ---------------------------------------------------------------------------
#> Environmental index
#> ---------------------------------------------------------------------------
#> # A tibble: 4 x 4
#>   ENV    Mean  index class      
#>   <fct> <dbl>  <dbl> <chr>      
#> 1 A1     199.  26.5  favorable  
#> 2 A2     168.  -4.50 unfavorable
#> 3 A3     147. -26.1  unfavorable
#> 4 A4     177.   4.13 favorable  
#> ---------------------------------------------------------------------------
#> Analysis for all environments
#> ---------------------------------------------------------------------------
#> # A tibble: 13 x 6
#>    Genotype  Mean Mean_rp Sd_rp    Wi  rank
#>    <chr>    <dbl>   <dbl> <dbl> <dbl> <dbl>
#>  1 H1        184.   106.   4.15 104.      1
#>  2 H10       164.    94.2  8.80  88.2    10
#>  3 H11       167.    96.7  2.21  95.3     6
#>  4 H12       157.    91.3  9.92  84.6    11
#>  5 H13       180.   105.  16.3   93.6     7
#>  6 H2        187.   109.  15.6   98.1     4
#>  7 H3        169.    98.0 12.0   89.9     9
#>  8 H4        184.   106.   8.77 101.      3
#>  9 H5        184.   107.   6.72 102.      2
#> 10 H6        188.   108.  16.7   96.8     5
#> 11 H7        171.    99.3 13.5   90.2     8
#> 12 H8        160.    92.1 17.3   80.4    12
#> 13 H9        153.    87.4 16.0   76.6    13
#> ---------------------------------------------------------------------------
#> Analysis for favorable environments
#> ---------------------------------------------------------------------------
#> # A tibble: 13 x 6
#>    Genotype  Mean Mean_rp Sd_rp    Wi  rank
#>    <chr>    <dbl>   <dbl> <dbl> <dbl> <dbl>
#>  1 H1        195.   104.   2.93 102.      1
#>  2 H10       188.    99.8  4.74  96.6     6
#>  3 H11       182.    97.0  3.69  94.5     8
#>  4 H12       175.    93.2  3.81  90.6    11
#>  5 H13       184.    97.2 17.6   85.3    12
#>  6 H2        184.    97.7  6.38  93.4     9
#>  7 H3        174.    92.0 10.5   84.9    13
#>  8 H4        198.   105.   6.03 101.      2
#>  9 H5        194.   103.   9.34  96.9     5
#> 10 H6        200.   106.  15.1   95.4     7
#> 11 H7        192.   103.  16.8   91.6    10
#> 12 H8        193.   103.   5.92  98.5     3
#> 13 H9        188.    99.8  3.69  97.3     4
#> ---------------------------------------------------------------------------
#> Analysis for unfavorable environments
#> ---------------------------------------------------------------------------
#> # A tibble: 13 x 6
#>    Genotype  Mean Mean_rp  Sd_rp    Wi  rank
#>    <chr>    <dbl>   <dbl>  <dbl> <dbl> <dbl>
#>  1 H1        173.   109.   3.48  107.      3
#>  2 H10       140.    88.5  9.06   82.4    10
#>  3 H11       152.    96.5  0.909  95.9     6
#>  4 H12       140.    89.5 16.3    78.5    11
#>  5 H13       175.   112.  16.5   101.      4
#>  6 H2        190.   120.  14.7   110.      2
#>  7 H3        165.   104.  13.2    95.1     7
#>  8 H4        170.   107.  13.8    98.2     5
#>  9 H5        174.   110.   0.584 110.      1
#> 10 H6        176.   111.  24.3    94.2     8
#> 11 H7        150.    95.7 14.6    85.9     9
#> 12 H8        127.    81.7 20.7    67.7    12
#> 13 H9        117.    75.1 11.8    67.1    13
#> 
#> 
#> 
#> Variable NR 
#> ---------------------------------------------------------------------------
#> Environmental index
#> ---------------------------------------------------------------------------
#> # A tibble: 4 x 4
#>   ENV    Mean  index class      
#>   <fct> <dbl>  <dbl> <chr>      
#> 1 A1     16.9  0.769 favorable  
#> 2 A2     15.8 -0.328 unfavorable
#> 3 A3     15.8 -0.338 unfavorable
#> 4 A4     16.0 -0.103 unfavorable
#> ---------------------------------------------------------------------------
#> Analysis for all environments
#> ---------------------------------------------------------------------------
#> # A tibble: 13 x 6
#>    Genotype  Mean Mean_rp Sd_rp    Wi  rank
#>    <chr>    <dbl>   <dbl> <dbl> <dbl> <dbl>
#>  1 H1        16.6   103.   5.65  99.2     2
#>  2 H10       15.6    96.7  5.81  92.8     9
#>  3 H11       15.6    97.0 10.2   90.2    13
#>  4 H12       16.3   101.   4.59  98.2     3
#>  5 H13       17.4   108.   8.03 103.      1
#>  6 H2        16.7   103.  11.2   95.7     6
#>  7 H3        15.8    97.8  7.96  92.4    10
#>  8 H4        15.5    96.1  7.03  91.3    11
#>  9 H5        16.1   100.  10.3   93.3     8
#> 10 H6        16.3   101.   5.70  97.3     5
#> 11 H7        16.2   100.   3.59  98.0     4
#> 12 H8        15.9    98.8  5.22  95.3     7
#> 13 H9        15.5    96.0  7.25  91.2    12
#> ---------------------------------------------------------------------------
#> Analysis for favorable environments
#> ---------------------------------------------------------------------------
#> # A tibble: 13 x 6
#>    Genotype  Mean Mean_rp Sd_rp    Wi  rank
#>    <chr>    <dbl>   <dbl> <dbl> <dbl> <dbl>
#>  1 H1        16.3    96.3    NA    NA     1
#>  2 H10       16.7    98.7    NA    NA     2
#>  3 H11       15.2    90.0    NA    NA     3
#>  4 H12       17.3   103.     NA    NA     4
#>  5 H13       18.7   111.     NA    NA     5
#>  6 H2        19.2   114.     NA    NA     6
#>  7 H3        18.5   110.     NA    NA     7
#>  8 H4        16.4    97.1    NA    NA     8
#>  9 H5        14.5    86.0    NA    NA     9
#> 10 H6        16.8    99.5    NA    NA    10
#> 11 H7        17.9   106.     NA    NA    11
#> 12 H8        17.2   102.     NA    NA    12
#> 13 H9        14.9    88.4    NA    NA    13
#> ---------------------------------------------------------------------------
#> Analysis for unfavorable environments
#> ---------------------------------------------------------------------------
#> # A tibble: 13 x 6
#>    Genotype  Mean Mean_rp  Sd_rp    Wi  rank
#>    <chr>    <dbl>   <dbl>  <dbl> <dbl> <dbl>
#>  1 H1        16.7   105.   4.16  103.      1
#>  2 H10       15.2    96.0  6.93   91.4    12
#>  3 H11       15.8    99.4 11.1    91.9    11
#>  4 H12       16     101.   5.51   97.1     5
#>  5 H13       17.0   107.   9.64  101.      3
#>  6 H2        15.8    99.8 10.8    92.5    10
#>  7 H3        14.9    93.8  0.669  93.4     9
#>  8 H4        15.2    95.8  8.57   90.0    13
#>  9 H5        16.7   105.   4.95  102.      2
#> 10 H6        16.1   102.   6.85   97.1     6
#> 11 H7        15.6    98.6  0.343  98.4     4
#> 12 H8        15.5    97.8  5.89   93.8     8
#> 13 H9        15.6    98.6  6.31   94.3     7
#> 
#> 
#> 
#> Variable NKR 
#> ---------------------------------------------------------------------------
#> Environmental index
#> ---------------------------------------------------------------------------
#> # A tibble: 4 x 4
#>   ENV    Mean   index class      
#>   <fct> <dbl>   <dbl> <chr>      
#> 1 A1     33.9  1.62   favorable  
#> 2 A2     32.3  0.0321 favorable  
#> 3 A3     30.4 -1.85   unfavorable
#> 4 A4     32.5  0.201  favorable  
#> ---------------------------------------------------------------------------
#> Analysis for all environments
#> ---------------------------------------------------------------------------
#> # A tibble: 13 x 6
#>    Genotype  Mean Mean_rp Sd_rp    Wi  rank
#>    <chr>    <dbl>   <dbl> <dbl> <dbl> <dbl>
#>  1 H1        32.2    99.8  7.21  94.9     8
#>  2 H10       32.4   101.   7.60  95.6     5
#>  3 H11       33     102.   7.14  97.6     3
#>  4 H12       30.4    94.2  3.33  92.0    11
#>  5 H13       31.0    96.2  7.63  91.1    13
#>  6 H2        31.9    98.8  3.35  96.5     4
#>  7 H3        31.4    97.3  2.91  95.3     7
#>  8 H4        35     109.   3.81 106.      1
#>  9 H5        33.9   105.   5.76 101.      2
#> 10 H6        32.8   102.  13.6   92.4    10
#> 11 H7        31.4    97.6  4.98  94.2     9
#> 12 H8        31.3    96.9  8.60  91.1    12
#> 13 H9        32.5   101.   8.20  95.4     6
#> ---------------------------------------------------------------------------
#> Analysis for favorable environments
#> ---------------------------------------------------------------------------
#> # A tibble: 13 x 6
#>    Genotype  Mean Mean_rp Sd_rp    Wi  rank
#>    <chr>    <dbl>   <dbl> <dbl> <dbl> <dbl>
#>  1 H1        33.5   102.   7.16  97.0     5
#>  2 H10       32.1    97.7  5.78  93.8     9
#>  3 H11       33.2   101.   8.05  95.6     6
#>  4 H12       30.6    93.2  3.17  91.0    12
#>  5 H13       30.6    93.1  5.48  89.4    13
#>  6 H2        32.9   100.   2.17  98.8     3
#>  7 H3        31.9    96.9  3.43  94.6     7
#>  8 H4        35.5   108.   4.42 105.      1
#>  9 H5        34.5   105.   6.99 100.      2
#> 10 H6        35.0   106.  11.6   98.7     4
#> 11 H7        32.1    97.6  6.10  93.5    10
#> 12 H8        32.2    98.0 10.2   91.1    11
#> 13 H9        33.2   101.  10.0   94.5     8
#> ---------------------------------------------------------------------------
#> Analysis for unfavorable environments
#> ---------------------------------------------------------------------------
#> # A tibble: 13 x 6
#>    Genotype  Mean Mean_rp Sd_rp    Wi  rank
#>    <chr>    <dbl>   <dbl> <dbl> <dbl> <dbl>
#>  1 H1        28.4    93.4    NA    NA     1
#>  2 H10       33.3   110.     NA    NA     2
#>  3 H11       32.4   107.     NA    NA     3
#>  4 H12       29.6    97.4    NA    NA     4
#>  5 H13       32.1   105.     NA    NA     5
#>  6 H2        28.7    94.5    NA    NA     6
#>  7 H3        29.9    98.5    NA    NA     7
#>  8 H4        33.5   110.     NA    NA     8
#>  9 H5        32.3   106.     NA    NA     9
#> 10 H6        26.4    86.8    NA    NA    10
#> 11 H7        29.6    97.4    NA    NA    11
#> 12 H8        28.5    93.6    NA    NA    12
#> 13 H9        30.4   100      NA    NA    13
#> 
#> 
#> 
#> Variable CDED 
#> ---------------------------------------------------------------------------
#> Environmental index
#> ---------------------------------------------------------------------------
#> # A tibble: 4 x 4
#>   ENV    Mean    index class      
#>   <fct> <dbl>    <dbl> <chr>      
#> 1 A1    0.576 -0.00961 unfavorable
#> 2 A2    0.584 -0.00202 unfavorable
#> 3 A3    0.595  0.00855 favorable  
#> 4 A4    0.589  0.00308 favorable  
#> ---------------------------------------------------------------------------
#> Analysis for all environments
#> ---------------------------------------------------------------------------
#> # A tibble: 13 x 6
#>    Genotype  Mean Mean_rp Sd_rp    Wi  rank
#>    <chr>    <dbl>   <dbl> <dbl> <dbl> <dbl>
#>  1 H1       0.588   100.   3.91  97.6     4
#>  2 H10      0.586   100.0  4.15  97.2     5
#>  3 H11      0.580    98.9  3.61  96.5     7
#>  4 H12      0.582    99.3  5.86  95.3    10
#>  5 H13      0.582    99.4  8.04  94.0    12
#>  6 H2       0.574    98.0  6.86  93.4    13
#>  7 H3       0.578    98.6  4.32  95.7     9
#>  8 H4       0.581    99.1  4.06  96.4     8
#>  9 H5       0.588   100.   2.85  98.5     3
#> 10 H6       0.588   100.   7.70  95.1    11
#> 11 H7       0.597   102.   3.05  99.9     2
#> 12 H8       0.594   101.   6.75  96.7     6
#> 13 H9       0.601   103.   2.20 101.      1
#> ---------------------------------------------------------------------------
#> Analysis for favorable environments
#> ---------------------------------------------------------------------------
#> # A tibble: 13 x 6
#>    Genotype  Mean Mean_rp Sd_rp    Wi  rank
#>    <chr>    <dbl>   <dbl> <dbl> <dbl> <dbl>
#>  1 H1       0.597   101.   2.83  98.9     5
#>  2 H10      0.590    99.7  6.53  95.3    10
#>  3 H11      0.575    97.1  4.11  94.3    11
#>  4 H12      0.573    96.8  8.41  91.1    12
#>  5 H13      0.553    93.5  4.22  90.7    13
#>  6 H2       0.585    98.9  1.74  97.7     8
#>  7 H3       0.585    98.9  4.69  95.7     9
#>  8 H4       0.592   100.   2.48  98.4     7
#>  9 H5       0.608   103.   1.45 102.      2
#> 10 H6       0.622   105.   3.37 103.      1
#> 11 H7       0.606   102.   2.73 100.      4
#> 12 H8       0.605   102.   5.10  98.7     6
#> 13 H9       0.603   102.   1.78 101.      3
#> ---------------------------------------------------------------------------
#> Analysis for unfavorable environments
#> ---------------------------------------------------------------------------
#> # A tibble: 13 x 6
#>    Genotype  Mean Mean_rp  Sd_rp    Wi  rank
#>    <chr>    <dbl>   <dbl>  <dbl> <dbl> <dbl>
#>  1 H1       0.579    99.7  6.04   95.6     8
#>  2 H10      0.582   100.   2.93   98.3     6
#>  3 H11      0.584   101.   3.00   98.7     4
#>  4 H12      0.590   102.   2.80   99.9     3
#>  5 H13      0.611   105.   6.10  101.      1
#>  6 H2       0.564    97.1 11.6    89.2    13
#>  7 H3       0.570    98.2  5.81   94.3     9
#>  8 H4       0.570    98.2  6.33   94.0    10
#>  9 H5       0.569    98.1  0.581  97.7     7
#> 10 H6       0.554    95.5  8.59   89.7    12
#> 11 H7       0.589   102.   4.46   98.6     5
#> 12 H8       0.583   100.  10.4    93.4    11
#> 13 H9       0.598   103.   3.14  101.      2
#> 
#> 
#> 
#> Variable PERK 
#> ---------------------------------------------------------------------------
#> Environmental index
#> ---------------------------------------------------------------------------
#> # A tibble: 4 x 4
#>   ENV    Mean   index class      
#>   <fct> <dbl>   <dbl> <chr>      
#> 1 A1     87.4 -0.0169 unfavorable
#> 2 A2     87.6  0.214  favorable  
#> 3 A3     87.6  0.211  favorable  
#> 4 A4     87.0 -0.409  unfavorable
#> ---------------------------------------------------------------------------
#> Analysis for all environments
#> ---------------------------------------------------------------------------
#> # A tibble: 13 x 6
#>    Genotype  Mean Mean_rp Sd_rp    Wi  rank
#>    <chr>    <dbl>   <dbl> <dbl> <dbl> <dbl>
#>  1 H1        87.2    99.8 2.31   98.2     9
#>  2 H10       87.9   100.  2.42   98.8     6
#>  3 H11       88.2   101.  1.39   99.9     2
#>  4 H12       87.4   100.0 1.46   99.0     4
#>  5 H13       87.3    99.8 3.04   97.8    12
#>  6 H2        87.9   101.  2.44   98.9     5
#>  7 H3        87.8   100.  1.10   99.7     3
#>  8 H4        87.8   100.  0.117 100.      1
#>  9 H5        86.8    99.3 2.01   97.9    11
#> 10 H6        87.2    99.7 2.54   98.0    10
#> 11 H7        86.7    99.2 2.87   97.2    13
#> 12 H8        87.6   100.  2.30   98.7     7
#> 13 H9        86.9    99.4 1.31   98.5     8
#> ---------------------------------------------------------------------------
#> Analysis for favorable environments
#> ---------------------------------------------------------------------------
#> # A tibble: 13 x 6
#>    Genotype  Mean Mean_rp Sd_rp    Wi  rank
#>    <chr>    <dbl>   <dbl> <dbl> <dbl> <dbl>
#>  1 H1        85.8    97.9 0.579  97.5    11
#>  2 H10       88.1   101.  3.94   97.9     9
#>  3 H11       88.4   101.  2.09   99.5     4
#>  4 H12       87.8   100.  2.46   98.5     6
#>  5 H13       86.5    98.7 4.60   95.6    13
#>  6 H2        87.0    99.3 1.55   98.2     7
#>  7 H3        88.4   101.  0.291 101.      1
#>  8 H4        88.0   100.  0.107 100.      2
#>  9 H5        87.5    99.8 2.80   98.0     8
#> 10 H6        88.1   101.  4.04   97.8    10
#> 11 H7        87.1    99.4 3.86   96.8    12
#> 12 H8        88.4   101.  2.80   99.0     5
#> 13 H9        88.0   100.  0.365 100.      3
#> ---------------------------------------------------------------------------
#> Analysis for unfavorable environments
#> ---------------------------------------------------------------------------
#> # A tibble: 13 x 6
#>    Genotype  Mean Mean_rp Sd_rp    Wi  rank
#>    <chr>    <dbl>   <dbl> <dbl> <dbl> <dbl>
#>  1 H1        88.7   102.  1.18  101.      1
#>  2 H10       87.6   100.  1.44   99.4     7
#>  3 H11       87.9   101.  1.16  100.0     4
#>  4 H12       87.0    99.7 0.340  99.5     6
#>  5 H13       88.1   101.  1.30  100.      3
#>  6 H2        88.8   102.  3.00   99.8     5
#>  7 H3        87.3   100.  1.70   98.9     8
#>  8 H4        87.5   100.  0.131 100.      2
#>  9 H5        86.1    98.7 1.69   97.5    12
#> 10 H6        86.2    98.9 0.254  98.7     9
#> 11 H7        86.3    98.9 3.10   96.9    13
#> 12 H8        86.8    99.5 2.45   97.9    10
#> 13 H9        85.7    98.3 0.685  97.8    11
#> 
#> 
#> 
#> Variable TKW 
#> ---------------------------------------------------------------------------
#> Environmental index
#> ---------------------------------------------------------------------------
#> # A tibble: 4 x 4
#>   ENV    Mean  index class      
#>   <fct> <dbl>  <dbl> <chr>      
#> 1 A1     360.  21.7  favorable  
#> 2 A2     334.  -4.85 unfavorable
#> 3 A3     318. -20.9  unfavorable
#> 4 A4     343.   4.13 favorable  
#> ---------------------------------------------------------------------------
#> Analysis for all environments
#> ---------------------------------------------------------------------------
#> # A tibble: 13 x 6
#>    Genotype  Mean Mean_rp Sd_rp    Wi  rank
#>    <chr>    <dbl>   <dbl> <dbl> <dbl> <dbl>
#>  1 H1        365.   108.  11.0  100.      2
#>  2 H10       320.    94.1 11.4   86.4    11
#>  3 H11       333.    98.2  5.27  94.6     7
#>  4 H12       316.    93.2  3.33  90.9    10
#>  5 H13       340.   100.   1.80  99.2     3
#>  6 H2        356.   106.  14.6   95.7     5
#>  7 H3        346.   102.  13.4   93.4     8
#>  8 H4        346.   102.  10.1   95.5     6
#>  9 H5        341.   101.   4.91  97.3     4
#> 10 H6        363.   107.   5.15 104.      1
#> 11 H7        345.   102.  13.4   93.0     9
#> 12 H8        322.    95.0 16.8   83.6    12
#> 13 H9        311.    91.3 17.7   79.3    13
#> ---------------------------------------------------------------------------
#> Analysis for favorable environments
#> ---------------------------------------------------------------------------
#> # A tibble: 13 x 6
#>    Genotype  Mean Mean_rp  Sd_rp    Wi  rank
#>    <chr>    <dbl>   <dbl>  <dbl> <dbl> <dbl>
#>  1 H1        350.    99.5  9.13   93.4     9
#>  2 H10       358.   102.   1.76  101.      3
#>  3 H11       351.    99.7  2.51   98.1     6
#>  4 H12       333.    94.8  3.49   92.5    10
#>  5 H13       357.   102.   0.790 101.      2
#>  6 H2        332.    94.4  4.41   91.4    11
#>  7 H3        323.    92.0  2.60   90.3    12
#>  8 H4        345.    98.1  6.25   93.9     8
#>  9 H5        363.   103.   6.13   99.1     4
#> 10 H6        389.   111.   5.31  107.      1
#> 11 H7        365.   104.  14.5    94.3     7
#> 12 H8        351.    99.8  2.38   98.2     5
#> 13 H9        353.   100.  20.5    86.3    13
#> ---------------------------------------------------------------------------
#> Analysis for unfavorable environments
#> ---------------------------------------------------------------------------
#> # A tibble: 13 x 6
#>    Genotype  Mean Mean_rp  Sd_rp    Wi  rank
#>    <chr>    <dbl>   <dbl>  <dbl> <dbl> <dbl>
#>  1 H1        379.   116.   0.314 116.      1
#>  2 H10       282.    86.2 11.9    78.2    11
#>  3 H11       315.    96.6  8.20   91.1     8
#>  4 H12       298.    91.5  3.21   89.4     9
#>  5 H13       323.    99.3  1.90   98.0     5
#>  6 H2        381.   117.  11.1   109.      2
#>  7 H3        368.   113.   9.95  106.      3
#>  8 H4        348.   106.  14.1    97.0     6
#>  9 H5        319.    97.9  2.51   96.2     7
#> 10 H6        337.   103.   0.291 103.      4
#> 11 H7        325.   100.  17.8    88.1    10
#> 12 H8        292.    90.1 27.3    71.7    13
#> 13 H9        268.    82.4 14.4    72.7    12
#> 
#> 
#> 
#> Variable NKE 
#> ---------------------------------------------------------------------------
#> Environmental index
#> ---------------------------------------------------------------------------
#> # A tibble: 4 x 4
#>   ENV    Mean  index class      
#>   <fct> <dbl>  <dbl> <chr>      
#> 1 A1     558.  46.2  favorable  
#> 2 A2     505.  -6.99 unfavorable
#> 3 A3     468. -43.7  unfavorable
#> 4 A4     516.   4.51 favorable  
#> ---------------------------------------------------------------------------
#> Analysis for all environments
#> ---------------------------------------------------------------------------
#> # A tibble: 13 x 6
#>    Genotype  Mean Mean_rp Sd_rp    Wi  rank
#>    <chr>    <dbl>   <dbl> <dbl> <dbl> <dbl>
#>  1 H1        507.    98.9 10.5   91.8     8
#>  2 H10       504.    98.7  5.25  95.2     5
#>  3 H11       501.    98.0  4.44  95.0     6
#>  4 H12       502.    98.3 10.9   90.9     9
#>  5 H13       538.   105.  16.1   94.5     7
#>  6 H2        527.   103.   7.38  97.7     3
#>  7 H3        491.    95.6  9.41  89.2    12
#>  8 H4        535.   104.   3.73 102.      1
#>  9 H5        542.   106.   8.13 101.      2
#> 10 H6        516.   101.  16.2   89.8    11
#> 11 H7        498.    97.3  2.78  95.4     4
#> 12 H8        496.    96.7  8.78  90.8    10
#> 13 H9        496.    96.9 12.8   88.3    13
#> ---------------------------------------------------------------------------
#> Analysis for favorable environments
#> ---------------------------------------------------------------------------
#> # A tibble: 13 x 6
#>    Genotype  Mean Mean_rp  Sd_rp    Wi  rank
#>    <chr>    <dbl>   <dbl>  <dbl> <dbl> <dbl>
#>  1 H1        558.   104.  14.0    94.9     7
#>  2 H10       508.    94.6  3.13   92.5     9
#>  3 H11       518.    96.6  2.38   95.0     6
#>  4 H12       533.    99.3  0.116  99.2     2
#>  5 H13       524.    97.2 17.5    85.4    13
#>  6 H2        554.   103.  12.1    94.7     8
#>  7 H3        535.    99.2 14.3    89.5    10
#>  8 H4        578.   108.   0.528 107.      1
#>  9 H5        535.    99.7  4.47   96.7     5
#> 10 H6        512.    95.0  9.72   88.5    11
#> 11 H7        532.    99.2  2.39   97.6     3
#> 12 H8        550.   103.   7.86   97.4     4
#> 13 H9        543.   102.  19.9    88.2    12
#> ---------------------------------------------------------------------------
#> Analysis for unfavorable environments
#> ---------------------------------------------------------------------------
#> # A tibble: 13 x 6
#>    Genotype  Mean Mean_rp  Sd_rp    Wi  rank
#>    <chr>    <dbl>   <dbl>  <dbl> <dbl> <dbl>
#>  1 H1        455.    93.4  3.63   91.0     8
#>  2 H10       500.   103.   2.33  101.      3
#>  3 H11       483.    99.5  6.73   94.9     6
#>  4 H12       470.    97.2 18.7    84.6    13
#>  5 H13       551.   114.  14.2   104.      2
#>  6 H2        499.   102.   3.98   99.8     5
#>  7 H3        448.    92.0  2.87   90.1    10
#>  8 H4        492.   101.   0.138 101.      4
#>  9 H5        548.   113.   2.55  111.      1
#> 10 H6        521.   106.  23.7    90.4     9
#> 11 H7        464.    95.3  1.65   94.2     7
#> 12 H8        442.    90.8  5.24   87.2    12
#> 13 H9        449     92.2  3.58   89.8    11
#> 
#> 
#>