Remove NA values

These functions allow you to remove rows or columns with NA values quickly.

has_na(): Check for NA values in the data and return a logical value.

remove_rows_na(): Remove rows with NA values.

remove_cols_na(): Remove columns with NA values.

remove_rows_na(.data, verbose = TRUE)

remove_cols_na(.data, verbose = TRUE)

has_na(.data)

Arguments

.data
A data frame or tibble

verbose
Logical argument. If TRUE (default) shows in console the rows or columns deleted.

Value

A data frame with rows or columns with NA values deleted.

Examples

library(metan)
data_with_na <- data_g
data_with_na[c(1, 5, 10), c(3:5, 10:15)] <- NA
data_with_na

#> # A tibble: 39 x 17
#>    GEN   REP       PH     EH      EP    EL    ED    CL    CD    CW    KW    NR
#>    <fct> <fct>  <dbl>  <dbl>   <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1 H1    1     NA     NA     NA      15.74 49.93 30.53 16.6  NA     NA    NA  
#>  2 H1    2      2.198  1.086  0.4921 13.7  49.20 30.46 14.66 22.34 130.2  16.4
#>  3 H1    3      2.288  1.154  0.5022 15.06 52.64 31.7  16.16 29.61 176.0  15.6
#>  4 H10   1      1.788  0.888  0.5135 13.94 44.06 26.23 15.04 12.86 115.9  14.8
#>  5 H10   2     NA     NA     NA      13.56 43.88 23.49 14.44 NA     NA    NA  
#>  6 H10   3      2.272  1.114  0.4911 14.46 43.68 24.61 16.06 12.50 127.7  15.2
#>  7 H11   1      1.71   0.808  0.4885 15.48 45.21 24.99 16.72 15.17 140.2  15.6
#>  8 H11   2      2.094  1.06   0.5087 12.24 46.91 26.53 14.3  13.49 114.1  16.8
#>  9 H11   3      2.5    1.44   0.5773 15.02 49.03 27.52 15.16 19.36 167.8  16.4
#> 10 H12   1     NA     NA     NA      14.36 49.17 28.41 15    NA     NA    NA  
#> # … with 29 more rows, and 5 more variables: NKR <dbl>, CDED <dbl>, PERK <dbl>,
#> #   TKW <dbl>, NKE <dbl>

has_na(data_with_na)

#> [1] TRUE

remove_cols_na(data_with_na)

#> Warning: Column(s) PH, EH, EP, CW, KW, NR, NKR, CDED, PERK with NA values deleted.

#> # A tibble: 39 x 8
#>    GEN   REP      EL    ED    CL    CD   TKW   NKE
#>    <fct> <fct> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1 H1    1     15.74 49.93 30.53 16.6  347.2 458.4
#>  2 H1    2     13.7  49.20 30.46 14.66 336.9 386.4
#>  3 H1    3     15.06 52.64 31.7  16.16 422.2 430.6
#>  4 H10   1     13.94 44.06 26.23 15.04 258.3 445.6
#>  5 H10   2     13.56 43.88 23.49 14.44 232.6 496.4
#>  6 H10   3     14.46 43.68 24.61 16.06 250.9 524.4
#>  7 H11   1     15.48 45.21 24.99 16.72 263.7 534.6
#>  8 H11   2     12.24 46.91 26.53 14.3  287.6 397  
#>  9 H11   3     15.02 49.03 27.52 15.16 314.6 531.6
#> 10 H12   1     14.36 49.17 28.41 15    291.2 524.8
#> # … with 29 more rows

remove_rows_na(data_with_na)

#> Warning: Row(s) 1, 5, 10 with NA values deleted.

#> # A tibble: 36 x 17
#>    GEN   REP      PH    EH     EP    EL    ED    CL    CD    CW    KW    NR
#>    <fct> <fct> <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1 H1    2     2.198 1.086 0.4921 13.7  49.20 30.46 14.66 22.34 130.2  16.4
#>  2 H1    3     2.288 1.154 0.5022 15.06 52.64 31.7  16.16 29.61 176.0  15.6
#>  3 H10   1     1.788 0.888 0.5135 13.94 44.06 26.23 15.04 12.86 115.9  14.8
#>  4 H10   3     2.272 1.114 0.4911 14.46 43.68 24.61 16.06 12.50 127.7  15.2
#>  5 H11   1     1.71  0.808 0.4885 15.48 45.21 24.99 16.72 15.17 140.2  15.6
#>  6 H11   2     2.094 1.06  0.5087 12.24 46.91 26.53 14.3  13.49 114.1  16.8
#>  7 H11   3     2.5   1.44  0.5773 15.02 49.03 27.52 15.16 19.36 167.8  16.4
#>  8 H12   2     2.766 1.582 0.5720 13.78 46.52 23.78 14.62 16.29 153.1  17.6
#>  9 H12   3     2.002 0.782 0.3856 13.73 47.53 25.29 14.34 18.87 138.9  14.8
#> 10 H13   1     2.52  1.094 0.4344 16.1  51.67 28.15 16.64 23.92 199.1  18  
#> # … with 26 more rows, and 5 more variables: NKR <dbl>, CDED <dbl>, PERK <dbl>,
#> #   TKW <dbl>, NKE <dbl>