Summary an Annicchiarico object

Summary the Annicchiarico object in two ways. By default, the results are shown in the R console. The results can also be exported to the directory into a *.txt file.

# S3 method for Annicchiarico
summary(object, export = FALSE,
  file.name = NULL, digits = 3, ...)

Arguments

object
The Annicchiarico object

export
A logical argument. If TRUE, a *.txt file is exported to the working directory.

file.name
The name of the file if export = TRUE

digits
The significant digits to be shown.

...
Other arguments of the function

Examples

library(metan)
Ann = Annicchiarico(data_ge2,
                    env = ENV,
                    gen = GEN,
                    rep = REP,
                    resp = PH)
summary(Ann)

#> Variable PH 
#> ---------------------------------------------------------------------------
#> Environmental index
#> ---------------------------------------------------------------------------
#>  ENV Mean   index       class
#>   A1 2.79  0.3083   favorable
#>   A2 2.46 -0.0229 unfavorable
#>   A3 2.17 -0.3182 unfavorable
#>   A4 2.52  0.0329   favorable
#> ---------------------------------------------------------------------------
#> Analysis for all environments
#> ---------------------------------------------------------------------------
#>  Genotype Mean Mean_rp Sd_rp   Wi order_Wi
#>        H1 2.62   105.6  9.38 90.2        9
#>       H10 2.31    93.0  6.93 81.6        1
#>       H11 2.39    96.1  6.27 85.8       10
#>       H12 2.44    98.6 11.21 80.1        8
#>       H13 2.54   102.8 12.38 82.4        6
#>        H2 2.60   104.8  9.91 88.5        7
#>        H3 2.59   104.1 11.15 85.8       11
#>        H4 2.58   103.6  8.76 89.2        3
#>        H5 2.57   103.2  5.55 94.1       13
#>        H6 2.56   103.0  7.83 90.1        5
#>        H7 2.40    96.8  6.67 85.8        2
#>        H8 2.33    93.6  9.61 77.8        4
#>        H9 2.36    94.8  6.85 83.5       12
#> ---------------------------------------------------------------------------
#> Analysis for favorable environments
#> ---------------------------------------------------------------------------
#>  Genotype Mean Mean_rp Sd_rp    Wi order_Wi
#>        H1 2.68   101.1 5.085  92.7        8
#>       H10 2.59    97.3 3.378  91.7        6
#>       H11 2.66   100.1 2.405  96.2        9
#>       H12 2.61    98.5 2.999  93.6       11
#>       H13 2.66   100.4 1.627  97.7       10
#>        H2 2.66   100.1 0.203  99.8        5
#>        H3 2.70   101.4 5.219  92.8        3
#>        H4 2.71   102.1 0.839 100.7        4
#>        H5 2.74   103.1 2.325  99.3       12
#>        H6 2.63    98.9 0.292  98.4        7
#>        H7 2.65    99.9 0.812  98.6        1
#>        H8 2.62    98.9 3.604  93.0        2
#>        H9 2.61    98.1 9.332  82.8       13
#> ---------------------------------------------------------------------------
#> Analysis for unfavorable environments
#> ---------------------------------------------------------------------------
#>  Genotype Mean Mean_rp Sd_rp   Wi order_Wi
#>        H1 2.56   110.2 12.44 89.8        1
#>       H10 2.04    88.7  7.65 76.1       10
#>       H11 2.12    92.1  6.93 80.7        9
#>       H12 2.26    98.6 19.18 67.0       13
#>       H13 2.41   105.2 20.84 70.9        6
#>        H2 2.55   109.6 14.33 86.0        3
#>        H3 2.49   106.9 17.77 77.6        8
#>        H4 2.45   105.1 14.86 80.7       11
#>        H5 2.40   103.3  9.32 87.9        7
#>        H6 2.49   107.1 10.85 89.2        2
#>        H7 2.16    93.6  9.69 77.7        5
#>        H8 2.03    88.3 12.26 68.1       12
#>        H9 2.11    91.5  3.02 86.5        4
#> 
#> 
#>