Summary a waasb object

Summary the waasb object in two ways. By default, the results are shown in the R console. The results can also be exported to the directory.

# S3 method for waasb
summary(object, export = FALSE, blup = FALSE,
  file.name = NULL, digits = 4, ...)

Arguments

object
The waasb object

export
A logical argument. If TRUE|T, a *.txt file is exported to the working directory

blup
A logical argument. If TRUE|T, the blups are shown.

file.name
The name of the file if export = TRUE

digits
The significant digits to be shown.

...
Other arguments of the function

Examples

library(metan)
model = waasb(data_ge,
              resp = c(GY, HM),
              gen = GEN,
              env = ENV,
              rep = REP)

#> Evaluating variable GY 50 % 
#> Evaluating variable HM 100 % 
#> Done!

summary(model)

#> Variable GY 
#> ---------------------------------------------------------------------------
#> Individual fixed-model analysis of variance
#> ---------------------------------------------------------------------------
#>  ENV  MEAN     MSB    MSG     MSR     FCB      PRFB    FCG      PRFG     CV
#>   E1 2.521 0.06522 0.3374 0.14401  0.4529 6.428e-01  2.343 5.936e-02 15.055
#>  E10 2.175 0.65367 0.2958 0.02667 24.5080 7.275e-06 11.090 1.097e-05  7.508
#>  E11 1.368 0.37731 0.1509 0.10498  3.5943 4.860e-02  1.438 2.443e-01 23.680
#>  E12 1.609 0.09190 0.3202 0.05352  1.7171 2.077e-01  5.982 6.474e-04 14.382
#>  E13 2.910 0.07674 0.7134 0.09937  0.7722 4.767e-01  7.179 2.095e-04 10.834
#>  E14 1.782 0.10352 0.1305 0.07533  1.3742 2.784e-01  1.732 1.535e-01 15.403
#>   E2 3.180 0.69829 0.2066 0.17851  3.9117 3.885e-02  1.157 3.763e-01 13.287
#>   E3 4.064 0.48864 0.3351 0.17894  2.7308 9.209e-02  1.872 1.230e-01 10.408
#>   E4 3.675 0.11646 0.5313 0.13769  0.8458 4.456e-01  3.859 7.123e-03 10.096
#>   E5 3.910 0.21881 0.5261 0.06636  3.2974 6.024e-02  7.929 1.101e-04  6.587
#>   E6 2.663 0.15985 0.1346 0.05857  2.7294 9.219e-02  2.299 6.345e-02  9.087
#>   E7 1.989 0.38099 0.3369 0.09103  4.1852 3.217e-02  3.701 8.727e-03 15.170
#>   E8 2.536 0.81677 0.2146 0.02781 29.3693 2.149e-06  7.717 1.314e-04  6.575
#>   E9 3.057 0.58259 0.6793 0.11091  5.2526 1.596e-02  6.124 5.622e-04 10.896
#>      h2     AS
#>  0.5732 0.7571
#>  0.9098 0.9538
#>  0.3044 0.5518
#>  0.8328 0.9126
#>  0.8607 0.9277
#>  0.4227 0.6502
#>  0.1360 0.3688
#>  0.4659 0.6826
#>  0.7409 0.8607
#>  0.8739 0.9348
#>  0.5650 0.7517
#>  0.7298 0.8543
#>  0.8704 0.9330
#>  0.8367 0.9147
#> ---------------------------------------------------------------------------
#> Fixed effects
#> ---------------------------------------------------------------------------
#>   Sum Sq Mean Sq NumDF DenDF F value    Pr(>F)
#> 1 66.089  5.0837    13 279.8  52.576 2.954e-67
#> 2  9.662  0.3451    28 252.0   3.569 3.593e-08
#> ---------------------------------------------------------------------------
#> Random effects
#> ---------------------------------------------------------------------------
#>     Group Variance
#>       GEN  0.02803
#>   GEN:ENV  0.05671
#>  Residual  0.09669
#> ---------------------------------------------------------------------------
#> Likelihood ratio test
#> ---------------------------------------------------------------------------
#>   npar logLik   AIC   LRT Df Pr(>Chisq)
#> 1   45 -214.7 519.4    NA NA         NA
#> 2   44 -224.4 536.7 19.32  1  1.108e-05
#> 3   44 -237.1 562.3 44.83  1  2.147e-11
#> ---------------------------------------------------------------------------
#> Variance components and genetic parameters
#> ---------------------------------------------------------------------------
#>             Parameters                                    Values
#>           GEI variance 0.056714 (31.26% of phenotypic variance.)
#>     Genotypic variance 0.028025 (15.45% of phenotypic variance.)
#>      Residual variance 0.096693 (53.29% of phenotypic variance.)
#>    Phenotypic variance                         0.181432000765268
#>           Heritability                         0.154466820776706
#>                  GEIr2                          0.31258908175788
#>  Heribatility of means                         0.815198214388929
#>               Accuracy                         0.902883278385932
#>                    rge                         0.369694636992275
#>                    CVg                          6.25999359149609
#>                    CVr                          11.6277882108427
#>               CV ratio                          0.53836494765692
#> ---------------------------------------------------------------------------
#>  Principal component analysis of the G x E interaction matrix
#> ---------------------------------------------------------------------------
#>   PC Eigenvalue Proportion Accumulated
#>  PC1    1.47222    34.3084       34.31
#>  PC2    1.34670    31.3835       65.69
#>  PC3    0.54788    12.7677       78.46
#>  PC4    0.41666     9.7099       88.17
#>  PC5    0.21262     4.9548       93.12
#>  PC6    0.13973     3.2562       96.38
#>  PC7    0.07912     1.8437       98.22
#>  PC8    0.05673     1.3221       99.55
#>  PC9    0.01947     0.4537      100.00
#> ---------------------------------------------------------------------------
#> Some information regarding the analysis
#> ---------------------------------------------------------------------------
#>  Parameters                       Values
#>        Ngen                           10
#>        Nenv                           14
#>      OVmean                       2.6742
#>         Min 0.8991 (Genotype G10 in E11)
#>         Max   4.8121 (Genotype G8 in E5)
#>      MinENV        Environment 2 (2.471)
#>      MaxENV        Environment 9 (3.004)
#>      MinGEN         Genotype 13 (1.368) 
#>      MaxGEN         Genotype 18 (4.064) 
#> ------------------------------ End of data -------------------------------
#> 
#> 
#> 
#> Variable HM 
#> ---------------------------------------------------------------------------
#> Individual fixed-model analysis of variance
#> ---------------------------------------------------------------------------
#>  ENV  MEAN     MSB    MSG   MSR     FCB      PRFB    FCG      PRFG    CV     h2
#>   E1 47.42  3.1063  5.221 1.852  1.6773 2.148e-01  2.819 2.928e-02 2.870 0.6453
#>  E10 44.30  9.1000 28.774 2.174  4.1857 3.216e-02 13.235 3.024e-06 3.328 0.9244
#>  E11 54.20  2.8101  6.057 3.087  0.9103 4.202e-01  1.962 1.069e-01 3.242 0.4903
#>  E12 49.60  2.7661 10.405 5.322  0.5198 6.033e-01  1.955 1.081e-01 4.651 0.4885
#>  E13 46.56 25.4523 13.883 1.342 18.9637 3.705e-05 10.344 1.798e-05 2.488 0.9033
#>  E14 41.03 10.0151  5.278 1.217  8.2304 2.894e-03  4.338 3.935e-03 2.688 0.7695
#>   E2 44.12  6.6986 13.168 3.049  2.1970 1.400e-01  4.319 4.025e-03 3.958 0.7685
#>   E3 52.90  7.3000  8.004 2.781  2.6245 9.996e-02  2.877 2.692e-02 3.153 0.6525
#>   E4 50.00  2.1000  7.926 1.248  1.6825 2.139e-01  6.350 4.510e-04 2.234 0.8425
#>   E5 52.23  6.8602 11.834 6.371  1.0768 3.616e-01  1.858 1.260e-01 4.833 0.4617
#>   E6 45.89 12.8640 15.835 2.122  6.0628 9.706e-03  7.463 1.633e-04 3.174 0.8660
#>   E7 48.49  0.7052  4.997 4.738  0.1488 8.627e-01  1.055 4.382e-01 4.489 0.0518
#>   E8 45.16 11.9541 11.567 2.168  5.5137 1.356e-02  5.335 1.266e-03 3.260 0.8126
#>   E9 51.33  5.7333  9.333 2.289  2.5049 1.097e-01  4.078 5.408e-03 2.947 0.7548
#>      AS
#>  0.8033
#>  0.9615
#>  0.7002
#>  0.6989
#>  0.9504
#>  0.8772
#>  0.8766
#>  0.8078
#>  0.9179
#>  0.6795
#>  0.9306
#>  0.2276
#>  0.9014
#>  0.8688
#> ---------------------------------------------------------------------------
#> Fixed effects
#> ---------------------------------------------------------------------------
#>   Sum Sq Mean Sq NumDF DenDF F value    Pr(>F)
#> 1 1166.0  89.689    13 257.4  31.580 5.684e-46
#> 2  214.9   7.676    28 252.0   2.703 2.197e-05
#> ---------------------------------------------------------------------------
#> Random effects
#> ---------------------------------------------------------------------------
#>     Group Variance
#>       GEN   0.4899
#>   GEN:ENV   2.1892
#>  Residual   2.8401
#> ---------------------------------------------------------------------------
#> Likelihood ratio test
#> ---------------------------------------------------------------------------
#>   npar logLik  AIC    LRT Df Pr(>Chisq)
#> 1   45 -862.7 1815     NA NA         NA
#> 2   44 -866.6 1821  7.855  1  5.068e-03
#> 3   44 -894.1 1876 62.819  1  2.266e-15
#> ---------------------------------------------------------------------------
#> Variance components and genetic parameters
#> ---------------------------------------------------------------------------
#>             Parameters                                    Values
#>           GEI variance 2.189183 (39.67% of phenotypic variance.)
#>     Genotypic variance  0.489906 (8.88% of phenotypic variance.)
#>      Residual variance 2.840058 (51.46% of phenotypic variance.)
#>    Phenotypic variance                          5.51914705328806
#>           Heritability                        0.0887647617601055
#>                  GEIr2                         0.396652462685586
#>  Heribatility of means                         0.686242077696183
#>               Accuracy                         0.828397294597334
#>                    rge                         0.435290961148237
#>                    CVg                          1.45551599851127
#>                    CVr                          3.50448584438064
#>               CV ratio                         0.415329398703424
#> ---------------------------------------------------------------------------
#>  Principal component analysis of the G x E interaction matrix
#> ---------------------------------------------------------------------------
#>   PC Eigenvalue Proportion Accumulated
#>  PC1    61.9620     33.835       33.83
#>  PC2    51.9046     28.343       62.18
#>  PC3    19.8166     10.821       73.00
#>  PC4    13.8954      7.588       80.59
#>  PC5    11.5038      6.282       86.87
#>  PC6    10.3748      5.665       92.53
#>  PC7     8.3910      4.582       97.11
#>  PC8     4.5825      2.502       99.62
#>  PC9     0.7015      0.383      100.00
#> ---------------------------------------------------------------------------
#> Some information regarding the analysis
#> ---------------------------------------------------------------------------
#>  Parameters                       Values
#>        Ngen                           10
#>        Nenv                           14
#>      OVmean                      48.0883
#>         Min   39.35 (Genotype G2 in E14)
#>         Max 56.3333 (Genotype G6 in E11)
#>      MinENV        Environment 3 (46.66)
#>      MaxENV       Environment 6 (49.298)
#>      MinGEN        Genotype 16 (41.032) 
#>      MaxGEN        Genotype 13 (54.198) 
#> ------------------------------ End of data -------------------------------
#> 
#> 
#>