Summary a ge_factanal object

Summary the ge_factanal object in two ways. By default, the results are shown in the R console. The results can also be exported to the directory.

# S3 method for ge_factanal
summary(object, export = FALSE, file.name = NULL,
  digits = 4, ...)

Arguments

object
The ge_factanal object

export
A logical argument. If TRUE, a *.txt file is exported to the working directory

file.name
The name of the file if export = TRUE

digits
The significant digits to be shown.

...
Other arguments of the function

Examples

model = ge_factanal(data_ge2,
                    env = ENV,
                    gen = GEN,
                    rep = REP,
                    resp = PH)
summary(model)

#> Variable PH 
#> ------------------------------------------------------------------------------------
#> Correlation matrix among environments
#> ------------------------------------------------------------------------------------
#>         A1      A2      A3      A4
#> A1  1.0000  0.3919 -0.4723 -0.4983
#> A2  0.3919  1.0000 -0.2203  0.2722
#> A3 -0.4723 -0.2203  1.0000  0.3238
#> A4 -0.4983  0.2722  0.3238  1.0000
#> ------------------------------------------------------------------------------------
#> Eigenvalues and explained variance
#> ------------------------------------------------------------------------------------
#>  PCA Eigenvalues Variance Cumul_var
#>  PC1      1.9262   48.154     48.15
#>  PC2      1.2633   31.582     79.74
#>  PC3      0.5928   14.820     94.56
#>  PC4      0.2178    5.444    100.00
#> ------------------------------------------------------------------------------------
#> Initial loadings
#> ------------------------------------------------------------------------------------
#>  Env     PC1      PC2
#>   A1 -0.8912  0.08731
#>   A2 -0.3718  0.88158
#>   A3  0.7660 -0.04547
#>   A4  0.6380  0.69022
#> ------------------------------------------------------------------------------------
#> Loadings after varimax rotation and commonalities
#> ------------------------------------------------------------------------------------
#>  Env     FA1     FA2 Communality Uniquenesses
#>   A1 -0.8763  0.1839      0.8018      0.19821
#>   A2 -0.2735  0.9169      0.9154      0.08461
#>   A3  0.7564 -0.1287      0.5888      0.41123
#>   A4  0.7095  0.6166      0.8835      0.11651
#> ------------------------------------------------------------------------------------
#> Environmental stratification based on factor analysis
#> ------------------------------------------------------------------------------------
#>  Env Factor  Mean   Min   Max     CV
#>   A1    FA1 2.793 2.692 2.935  2.810
#>   A3    FA1 2.167 2.028 2.599  7.724
#>   A4    FA1 2.518 2.305 2.638  3.655
#>   A2    FA2 2.462 1.959 2.947 16.202
#> ------------------------------------------------------------------------------------
#> Mean = mean; Min = minimum; Max = maximum; CV = coefficient of variation (%)
#> The summary statistics are based on the men values of  3 replicates
#> ------------------------------------------------------------------------------------