Summary a ge_reg object

Summary the ge_reg object in two ways. By default, the results are shown in the R console. The results can also be exported to the directory into a *.txt file.

# S3 method for ge_reg
summary(object, export = FALSE, file.name = NULL,
  digits = 3, ...)

Arguments

object
The ge_reg object

export
A logical argument. If TRUE, a *.txt file is exported to the working directory.

file.name
The name of the file if export = TRUE

digits
The significant digits to be shown.

...
Other arguments of the function

Examples

library(metan)
model = ge_reg(data_ge2, ENV, GEN, REP, PH)
summary(model)

#> Variable PH 
#> ---------------------------------------------------------------------------
#> Joint-regression Analysis of variance
#> ---------------------------------------------------------------------------
#> # A tibble: 20 x 6
#>    SV                       Df `Sum Sq` `Mean Sq` `F value`    `Pr(>F)`
#>    <chr>                 <dbl>    <dbl>     <dbl>     <dbl>       <dbl>
#>  1 Total                    51   15.0      0.294     NA     NA         
#>  2 GEN                      12    1.86     0.155      0.870  0.585     
#>  3 ENV + (GEN x ENV)        39   13.1      0.336     NA     NA         
#>  4 ENV (linear)              1    7.72     7.72      NA     NA         
#>  5 " GEN x ENV (linear)"    12    0.755    0.0629     0.352  0.969     
#>  6 Pooled deviation         26    4.64     0.179     NA     NA         
#>  7 H1                        2    0.471    0.235     11.4    0.0000343 
#>  8 H10                       2    0.236    0.118      5.69   0.00452   
#>  9 H11                       2    0.211    0.106      5.11   0.00767   
#> 10 H12                       2    0.450    0.225     10.9    0.0000515 
#> 11 H13                       2    0.410    0.205      9.91   0.000115  
#> 12 H2                        2    0.534    0.267     12.9    0.00000994
#> 13 H3                        2    0.584    0.292     14.1    0.00000381
#> 14 H4                        2    0.354    0.177      8.55   0.000365  
#> 15 H5                        2    0.140    0.0701     3.39   0.0376    
#> 16 H6                        2    0.334    0.167      8.06   0.000559  
#> 17 H7                        2    0.241    0.121      5.83   0.00398   
#> 18 H8                        2    0.505    0.253     12.2    0.0000174 
#> 19 H9                        2    0.172    0.0861     4.16   0.0183    
#> 20 Pooled error            104    2.15     0.0207    NA     NA         
#> ---------------------------------------------------------------------------
#> Regression parameters
#> ---------------------------------------------------------------------------
#>  GEN Mean   bij   sdij  RMSE    R2
#>   H1 2.62 0.806 0.0710 0.198 0.451
#>  H10 2.31 1.218 0.0318 0.140 0.789
#>  H11 2.39 1.080 0.0278 0.133 0.766
#>  H12 2.44 0.465 0.0676 0.194 0.222
#>  H13 2.54 0.306 0.0609 0.185 0.119
#>   H2 2.60 0.963 0.0815 0.211 0.508
#>   H3 2.59 1.355 0.0898 0.221 0.651
#>   H4 2.58 1.268 0.0515 0.172 0.730
#>   H5 2.57 1.172 0.0159 0.108 0.853
#>   H6 2.56 0.936 0.0481 0.167 0.609
#>   H7 2.40 0.992 0.0328 0.142 0.708
#>   H8 2.33 1.014 0.0767 0.205 0.547
#>   H9 2.36 1.424 0.0212 0.120 0.875
#> ---------------------------------------------------------------------------
#> 
#> 
#>