Summary a ge_stats object

Summary the ge_stats object in two ways. By default, the results are shown in the R console. The results can also be exported to the directory into a *.txt file.

# S3 method for ge_stats
summary(object, export = FALSE, file.name = NULL,
  digits = 3, ...)

Arguments

object
The ge_stats object

export
A logical argument. If TRUE, a *.txt file is exported to the working directory.

file.name
The name of the file if export = TRUE

digits
The significant digits to be shown.

...
Other arguments of the function

Examples

library(metan)
model = ge_stats(.data = data_ge2,
                 env = ENV,
                 gen = GEN,
                 rep = REP,
                 resp = PH)
summary(model)

#> Variable PH 
#> ---------------------------------------------------------------------------
#> Individual analysis of variance
#> ---------------------------------------------------------------------------

#> Warning: Unknown or uninitialised column: 'individual'.

#> data frame with 0 columns and 0 rows
#> ---------------------------------------------------------------------------
#> Genotype-vs-environment mean
#> ---------------------------------------------------------------------------
#>       A1   A2   A3   A4
#> H1  2.72 2.93 2.20 2.64
#> H10 2.78 2.05 2.04 2.39
#> H11 2.75 2.15 2.10 2.56
#> H12 2.69 2.09 2.43 2.53
#> H13 2.77 2.23 2.60 2.56
#> H2  2.79 2.95 2.15 2.52
#> H3  2.94 2.94 2.04 2.46
#> H4  2.87 2.85 2.05 2.56
#> H5  2.83 2.70 2.09 2.64
#> H6  2.77 2.82 2.15 2.49
#> H7  2.77 2.14 2.18 2.53
#> H8  2.69 1.96 2.10 2.55
#> H9  2.93 2.20 2.03 2.30
#> ---------------------------------------------------------------------------
#> Genotype-vs-environment effects
#> ---------------------------------------------------------------------------
#>          A1     A2      A3      A4
#> H1  -0.2069  0.332 -0.1058 -0.0189
#> H10  0.1602 -0.243  0.0414  0.0409
#> H11  0.0504 -0.220  0.0296  0.1404
#> H12 -0.0531 -0.321  0.3107  0.0636
#> H13 -0.0749 -0.288  0.3782 -0.0149
#> H2  -0.1204  0.365 -0.1319 -0.1131
#> H3   0.0322  0.369 -0.2333 -0.1676
#> H4  -0.0196  0.289 -0.2124 -0.0569
#> H5  -0.0416  0.160 -0.1551  0.0371
#> H6  -0.0976  0.290 -0.0864 -0.1056
#> H7   0.0613 -0.245  0.0913  0.0921
#> H8   0.0572 -0.344  0.0917  0.1955
#> H9   0.2527 -0.142 -0.0181 -0.0926
#> ---------------------------------------------------------------------------
#> Genotype + Genotype-vs-environment effects
#> ---------------------------------------------------------------------------
#>           A1     A2      A3       A4
#> H1  -0.07040  0.468  0.0308  0.11761
#> H10 -0.00973 -0.413 -0.1286 -0.12906
#> H11 -0.04440 -0.315 -0.0652  0.04561
#> H12 -0.10040 -0.369  0.2634  0.01628
#> H13 -0.02107 -0.235  0.4321  0.03894
#> H2  -0.00107  0.485 -0.0126  0.00628
#> H3   0.14227  0.479 -0.1232 -0.05759
#> H4   0.07560  0.384 -0.1172  0.03828
#> H5   0.04160  0.243 -0.0719  0.12028
#> H6  -0.02440  0.363 -0.0132 -0.03239
#> H7  -0.01920 -0.325  0.0108  0.01161
#> H8  -0.10107 -0.503 -0.0666  0.03721
#> H9   0.13227 -0.263 -0.1386 -0.21306
#> ---------------------------------------------------------------------------
#> Genotype-vs-environment statistics
#> ---------------------------------------------------------------------------
#>     Mean Variance SQE_Gi
#> H1  2.62    0.286  0.857
#> H10 2.31    0.372  1.116
#> H11 2.39    0.301  0.904
#> H12 2.44    0.193  0.579
#> H13 2.54    0.155  0.466
#> H2  2.60    0.362  1.085
#> H3  2.59    0.558  1.674
#> H4  2.58    0.436  1.309
#> H5  2.57    0.319  0.956
#> H6  2.56    0.285  0.854
#> H7  2.40    0.275  0.826
#> H8  2.33    0.372  1.116
#> H9  2.36    0.459  1.376
#> ---------------------------------------------------------------------------
#> Regression-based stability (anova)
#> ---------------------------------------------------------------------------
#>                     SV  Df Sum Sq Mean Sq F value   Pr(>F)
#> 1                Total  51 14.981  0.2937      NA       NA
#> 2                  GEN  12  1.865  0.1554   0.870 5.85e-01
#> 3    ENV + (GEN x ENV)  39 13.116  0.3363      NA       NA
#> 4         ENV (linear)   1  7.719  7.7185      NA       NA
#> 5   GEN x ENV (linear)  12  0.755  0.0629   0.352 9.69e-01
#> 6     Pooled deviation  26  4.642  0.1785      NA       NA
#> 7                   H1   2  0.471  0.2353  11.367 3.43e-05
#> 8                  H10   2  0.236  0.1178   5.690 4.52e-03
#> 9                  H11   2  0.211  0.1057   5.106 7.67e-03
#> 10                 H12   2  0.450  0.2251  10.874 5.15e-05
#> 11                 H13   2  0.410  0.2052   9.911 1.15e-04
#> 12                  H2   2  0.534  0.2669  12.895 9.94e-06
#> 13                  H3   2  0.584  0.2919  14.102 3.81e-06
#> 14                  H4   2  0.354  0.1770   8.550 3.65e-04
#> 15                  H5   2  0.140  0.0701   3.387 3.76e-02
#> 16                  H6   2  0.334  0.1668   8.057 5.59e-04
#> 17                  H7   2  0.241  0.1207   5.830 3.98e-03
#> 18                  H8   2  0.505  0.2526  12.202 1.74e-05
#> 19                  H9   2  0.172  0.0861   4.161 1.83e-02
#> 20        Pooled error 104  2.153  0.0207      NA       NA
#> ---------------------------------------------------------------------------
#> Regression-based stability (parameters)
#> ---------------------------------------------------------------------------
#>  GEN Mean   bij   sdij  RMSE    R2
#>   H1 2.62 0.806 0.0710 0.198 0.451
#>  H10 2.31 1.218 0.0318 0.140 0.789
#>  H11 2.39 1.080 0.0278 0.133 0.766
#>  H12 2.44 0.465 0.0676 0.194 0.222
#>  H13 2.54 0.306 0.0609 0.185 0.119
#>   H2 2.60 0.963 0.0815 0.211 0.508
#>   H3 2.59 1.355 0.0898 0.221 0.651
#>   H4 2.58 1.268 0.0515 0.172 0.730
#>   H5 2.57 1.172 0.0159 0.108 0.853
#>   H6 2.56 0.936 0.0481 0.167 0.609
#>   H7 2.40 0.992 0.0328 0.142 0.708
#>   H8 2.33 1.014 0.0767 0.205 0.547
#>   H9 2.36 1.424 0.0212 0.120 0.875
#> ---------------------------------------------------------------------------
#> Genotypic confidence index (all environments)
#> ---------------------------------------------------------------------------
#>  Genotype Mean Mean_rp Sd_rp   Wi rank
#>        H1 2.62   105.6  9.38 90.2    2
#>       H10 2.31    93.0  6.93 81.6   11
#>       H11 2.39    96.1  6.27 85.8    8
#>       H12 2.44    98.6 11.21 80.1   12
#>       H13 2.54   102.8 12.38 82.4   10
#>        H2 2.60   104.8  9.91 88.5    5
#>        H3 2.59   104.1 11.15 85.8    6
#>        H4 2.58   103.6  8.76 89.2    4
#>        H5 2.57   103.2  5.55 94.1    1
#>        H6 2.56   103.0  7.83 90.1    3
#>        H7 2.40    96.8  6.67 85.8    7
#>        H8 2.33    93.6  9.61 77.8   13
#>        H9 2.36    94.8  6.85 83.5    9
#> ---------------------------------------------------------------------------
#> Wricke's Ecovalence
#> ---------------------------------------------------------------------------
#>          A1     A2      A3      A4 Ecoval Ecov_perc rank
#> H1  -0.2069  0.332 -0.1058 -0.0189  0.493      9.13    8
#> H10  0.1602 -0.243  0.0414  0.0409  0.264      4.89    4
#> H11  0.0504 -0.220  0.0296  0.1404  0.215      3.99    2
#> H12 -0.0531 -0.321  0.3107  0.0636  0.620     11.49   11
#> H13 -0.0749 -0.288  0.3782 -0.0149  0.696     12.90   13
#> H2  -0.1204  0.365 -0.1319 -0.1131  0.535      9.91   10
#> H3   0.0322  0.369 -0.2333 -0.1676  0.659     12.20   12
#> H4  -0.0196  0.289 -0.2124 -0.0569  0.397      7.35    7
#> H5  -0.0416  0.160 -0.1551  0.0371  0.158      2.92    1
#> H6  -0.0976  0.290 -0.0864 -0.1056  0.336      6.23    6
#> H7   0.0613 -0.245  0.0913  0.0921  0.241      4.47    3
#> H8   0.0572 -0.344  0.0917  0.1955  0.505      9.36    9
#> H9   0.2527 -0.142 -0.0181 -0.0926  0.279      5.17    5
#> ---------------------------------------------------------------------------
#> Nonparametric Superiority index
#> ---------------------------------------------------------------------------
#>  Genotypes Pi_all Or_a Pi_favorable Or_f Pi_unfavorable Or_u
#>         H1 0.0258    1      0.01131    8         0.0403    1
#>        H10 0.1506   12      0.02132   12         0.2799   12
#>        H11 0.1160   10      0.01010    7         0.2218   11
#>        H12 0.1033    8      0.01743   11         0.1892    8
#>        H13 0.0688    7      0.00832    4         0.1294    7
#>         H2 0.0289    2      0.00839    5         0.0494    2
#>         H3 0.0425    6      0.00791    3         0.0771    4
#>         H4 0.0404    5      0.00279    2         0.0780    5
#>         H5 0.0403    4      0.00253    1         0.0781    6
#>         H6 0.0330    3      0.01277    9         0.0533    3
#>         H7 0.1089    9      0.00947    6         0.2084    9
#>         H8 0.1612   13      0.01653   10         0.3059   13
#>         H9 0.1244   11      0.02780   13         0.2210   10
#> ------------------------------- End of data -----------------------------
#> 
#> 
#>