Summary a superiority object

Summary the superiority object in two ways. By default, the results are shown in the R console. The results can also be exported to the directory into a *.txt file.

# S3 method for superiority
summary(object, export = FALSE, file.name = NULL,
  digits = 3, ...)

Arguments

object
The superiority object

export
A logical argument. If TRUE, a *.txt file is exported to the working directory.

file.name
The name of the file if export = TRUE

digits
The significant digits to be shown.

...
Other arguments of the function

Examples

library(metan)
model = superiority(data_ge2, ENV, GEN, REP, PH)
summary(model)

#> Variable PH 
#> ---------------------------------------------------------------------------
#> Superiority index considering all, favorable and unfavorable environments
#> ---------------------------------------------------------------------------
#>  Genotypes Pi_all Or_a Pi_favorable Or_f Pi_unfavorable Or_u
#>         H1 0.0258    1      0.01131    9         0.0403    1
#>        H10 0.1506    6      0.02132    8         0.2799    6
#>        H11 0.1160   10      0.01010    7         0.2218   10
#>        H12 0.1033    9      0.01743    5         0.1892    7
#>        H13 0.0688    8      0.00832    6         0.1294    8
#>         H2 0.0289    7      0.00839   11         0.0494    9
#>         H3 0.0425    5      0.00791    3         0.0771    5
#>         H4 0.0404    4      0.00279    1         0.0780    4
#>         H5 0.0403   11      0.00253   10         0.0781   11
#>         H6 0.0330    3      0.01277   12         0.0533   13
#>         H7 0.1089   13      0.00947    4         0.2084    3
#>         H8 0.1612    2      0.01653    2         0.3059    2
#>         H9 0.1244   12      0.02780   13         0.2210   12
#> ---------------------------------------------------------------------------
#> 
#> 
#>