Summarise a waas object

Summary the waas object in two ways. By default, the results are shown in the R console. The results can also be exported to the directory.

# S3 method for waas
summary(object, export = FALSE, file.name = NULL,
  digits = 4, ...)

Arguments

object
The waas object

export
A logical argument. If TRUE, a *.txt file is exported to the working directory

file.name
The name of the file if export = TRUE

digits
The significant digits to be shown.

...
Other arguments of the function

Examples

library(metan)
model = waas(data_ge,
             resp = c(GY, HM),
             gen = GEN,
             env = ENV,
             rep = REP)

#> Evaluating variable GY 50 % 
#> Evaluating variable HM 100 % 
#> Done!

summary(model)

#> Variable GY 
#> ---------------------------------------------------------------------------
#> Individual analysis of variance
#> ---------------------------------------------------------------------------
#>  ENV  MEAN     MSB    MSG     MSR     FCB      PRFB    FCG      PRFG     CV
#>   E1 2.521 0.06522 0.3374 0.14401  0.4529 6.428e-01  2.343 5.936e-02 15.055
#>  E10 2.175 0.65367 0.2958 0.02667 24.5080 7.275e-06 11.090 1.097e-05  7.508
#>  E11 1.368 0.37731 0.1509 0.10498  3.5943 4.860e-02  1.438 2.443e-01 23.680
#>  E12 1.609 0.09190 0.3202 0.05352  1.7171 2.077e-01  5.982 6.474e-04 14.382
#>  E13 2.910 0.07674 0.7134 0.09937  0.7722 4.767e-01  7.179 2.095e-04 10.834
#>  E14 1.782 0.10352 0.1305 0.07533  1.3742 2.784e-01  1.732 1.535e-01 15.403
#>   E2 3.180 0.69829 0.2066 0.17851  3.9117 3.885e-02  1.157 3.763e-01 13.287
#>   E3 4.064 0.48864 0.3351 0.17894  2.7308 9.209e-02  1.872 1.230e-01 10.408
#>   E4 3.675 0.11646 0.5313 0.13769  0.8458 4.456e-01  3.859 7.123e-03 10.096
#>   E5 3.910 0.21881 0.5261 0.06636  3.2974 6.024e-02  7.929 1.101e-04  6.587
#>   E6 2.663 0.15985 0.1346 0.05857  2.7294 9.219e-02  2.299 6.345e-02  9.087
#>   E7 1.989 0.38099 0.3369 0.09103  4.1852 3.217e-02  3.701 8.727e-03 15.170
#>   E8 2.536 0.81677 0.2146 0.02781 29.3693 2.149e-06  7.717 1.314e-04  6.575
#>   E9 3.057 0.58259 0.6793 0.11091  5.2526 1.596e-02  6.124 5.622e-04 10.896
#>      h2     AS
#>  0.5732 0.7571
#>  0.9098 0.9538
#>  0.3044 0.5518
#>  0.8328 0.9126
#>  0.8607 0.9277
#>  0.4227 0.6502
#>  0.1360 0.3688
#>  0.4659 0.6826
#>  0.7409 0.8607
#>  0.8739 0.9348
#>  0.5650 0.7517
#>  0.7298 0.8543
#>  0.8704 0.9330
#>  0.8367 0.9147
#> ---------------------------------------------------------------------------
#> AMMI analysis table
#> ---------------------------------------------------------------------------
#>            Df   Sum Sq  Mean Sq F value    Pr(>F) Percent Accumul
#> ENV        13 279.5736 21.50566  62.325 0.000e+00       .       .
#> REP(ENV)   28   9.6615  0.34505   3.569 3.593e-08       .       .
#> GEN         9  12.9950  1.44389  14.933 2.190e-19       .       .
#> ENV:GEN   117  31.2196  0.26683   2.760 1.005e-11       .       .
#> PC1        21  10.7491  0.51186   5.290 0.000e+00    34.4    34.4
#> PC2        19   9.9239  0.52231   5.400 0.000e+00    31.8    66.2
#> PC3        17   4.0392  0.23760   2.460 1.400e-03    12.9    79.2
#> PC4        15   3.0738  0.20492   2.120 9.600e-03     9.8      89
#> PC5        13   1.4464  0.11126   1.150 3.176e-01     4.6    93.6
#> PC6        11   0.9322  0.08475   0.880 5.606e-01       3    96.6
#> PC7         9   0.5667  0.06297   0.650 7.535e-01     1.8    98.4
#> PC8         7   0.3623  0.05176   0.540 8.037e-01     1.2    99.6
#> PC9         5   0.1259  0.02517   0.260 9.345e-01     0.4     100
#> Residuals 252  24.3667  0.09669      NA        NA       .       .
#> Total     419 357.8164  0.85398      NA        NA       .       .
#> ---------------------------------------------------------------------------
#> Weighted average of the absolute scores
#> ---------------------------------------------------------------------------
#>  type Code     Y      PC1       PC2      PC3      PC4      PC5      PC6
#>   GEN   G1 2.604  0.31663 -0.044167 -0.03600 -0.06595 -0.31254  0.42717
#>   GEN  G10 2.471 -1.00067 -0.571801 -0.16517 -0.33087 -0.12425 -0.10639
#>   GEN   G2 2.744  0.13897  0.198761 -0.73307  0.47347 -0.04816 -0.28406
#>   GEN   G3 2.955  0.04340 -0.102838  0.22842  0.17689 -0.12702 -0.14002
#>   GEN   G4 2.642 -0.32509  0.478228 -0.09073  0.14171 -0.19242  0.35496
#>   GEN   G5 2.537 -0.32598  0.246084  0.24522  0.17941  0.46625  0.03315
#>   GEN   G6 2.534 -0.09836  0.242895  0.56074  0.23766  0.05094 -0.10107
#>   GEN   G7 2.741  0.28485  0.587113 -0.20680 -0.70852  0.23147 -0.08406
#>   GEN   G8 3.004  0.49948 -0.191580  0.31911 -0.16764 -0.32615 -0.28860
#>   GEN   G9 2.510  0.46676 -0.842694 -0.12170  0.06385  0.38188  0.18893
#>   ENV   E1 2.521  0.24645  0.197617 -0.23207  0.09863 -0.08083 -0.17681
#>   ENV  E10 2.175  0.21715  0.369574  0.11449  0.02614 -0.03517  0.07490
#>   ENV  E11 1.368  0.08229  0.166002  0.15276  0.22582  0.04647  0.05809
#>   ENV  E12 1.609  0.21476  0.271276 -0.18166  0.06975  0.35483 -0.29468
#>   ENV  E13 2.910  0.72492  0.131291  0.20689  0.13186  0.02075 -0.10466
#>   ENV  E14 1.782 -0.29124  0.003197  0.28701 -0.24709  0.19202  0.11024
#>   ENV   E2 3.180 -0.22883 -0.238363  0.24460  0.50779 -0.09064  0.03071
#>   ENV   E3 4.064  0.06575 -0.419723 -0.75335  0.01080  0.26949  0.17138
#>   ENV   E4 3.675 -0.32474 -0.816280 -0.00214  0.16477 -0.23757 -0.29854
#>   ENV   E5 3.910  0.51053 -0.131258  0.09311 -0.23980 -0.39991  0.12709
#>   ENV   E6 2.663  0.05806 -0.191721 -0.03688 -0.07973 -0.02537  0.46646
#>   ENV   E7 1.989 -0.26583  0.188743 -0.18712 -0.61813 -0.23761 -0.21011
#>   ENV   E8 2.536 -0.26420 -0.226459  0.46033 -0.28536  0.37817 -0.04370
#>   ENV   E9 3.057 -0.74508  0.696104 -0.16598  0.23455 -0.15461  0.08961
#>       PC7        PC8       PC9   WAAS PctResp PctWAAS PesRes PesWAAS OrResp
#>  -0.14794  0.2543204  0.166939 0.1508  24.878   91.83     50      50      6
#>  -0.03451  0.0687604 -0.011832 0.6522   0.000    0.00     50      50     10
#>  -0.07719  0.0865631 -0.004828 0.2834  51.262   67.54     50      50      3
#>   0.45681 -0.0965877  0.227167 0.1062  90.929  100.00     50      50      2
#>   0.08926 -0.2919830 -0.174399 0.3256  32.060   59.81     50      50      5
#>  -0.29224 -0.1211984  0.188311 0.2695  12.420   70.09     50      50      7
#>   0.05895  0.3273159 -0.201044 0.2325  11.797   76.87     50      50      8
#>   0.12904  0.0465900 -0.009254 0.4284  50.654   41.00     50      50      4
#>  -0.27388 -0.2222337 -0.055783 0.3266 100.000   59.64     50      50      1
#>   0.09169 -0.0515470 -0.125277 0.5067   7.320   26.64     50      50      9
#>   0.25101 -0.1693333 -0.149897 0.2106  42.744   78.13     50      50      9
#>  -0.25075  0.2373433  0.095571 0.2357  29.930   72.92     50      50     10
#>  -0.14181  0.2321324 -0.183742 0.1383   0.000   93.13     50      50     14
#>  -0.22242 -0.2157789  0.031020 0.2142   8.911   77.39     50      50     13
#>   0.33444  0.1040241  0.154623 0.3720  57.171   44.65     50      50      6
#>   0.15171 -0.0786837 -0.217194 0.1827  15.343   83.92     50      50     12
#>   0.01471  0.0494096 -0.097777 0.2653  67.198   66.79     50      50      4
#>  -0.03065  0.1216531 -0.021908 0.2861 100.000   62.47     50      50      1
#>  -0.03194 -0.0007188  0.095362 0.4361  85.569   31.35     50      50      3
#>  -0.26857 -0.2331498 -0.054109 0.2845  94.294   62.81     50      50      2
#>   0.14340 -0.0933340  0.124391 0.1052  48.031  100.00     50      50      7
#>   0.08166  0.2174456 -0.019025 0.2657  23.020   66.71     50      50     11
#>  -0.04918 -0.0477973  0.115412 0.2815  43.326   63.43     50      50      8
#>   0.01841 -0.1232121  0.127273 0.5873  62.622    0.00     50      50      5
#>  OrWAAS OrPC1 WAASY OrWAASY
#>       2     5 58.36       4
#>      10    10  0.00      10
#>       5     3 59.40       3
#>       1     1 95.46       1
#>       6     6 45.94       5
#>       4     7 41.25       8
#>       3     2 44.33       7
#>       8     4 45.83       6
#>       7     9 79.82       2
#>       9     8 16.98       9
#>       4     7 60.44       5
#>       6     5 51.43       8
#>       2     3 46.57      11
#>       5     4 43.15      13
#>      12    13 50.91       9
#>       3    10 49.63      10
#>       7     6 66.99       4
#>      11     2 81.24       1
#>      13    11 58.46       6
#>      10    12 78.55       2
#>       1     1 74.02       3
#>       8     9 44.86      12
#>       9     8 53.38       7
#>      14    14 31.31      14
#> ---------------------------------------------------------------------------
#> Some information regarding the analysis
#> ---------------------------------------------------------------------------
#>  Parameters                       Values
#>        Ngen                           10
#>        Nenv                           14
#>      OVmean                       2.6742
#>         Min 0.8991 (Genotype G10 in E11)
#>         Max   4.8121 (Genotype G8 in E5)
#>      MinENV        Environment 2 (2.471)
#>      MaxENV        Environment 9 (3.004)
#>      MinGEN         Genotype 13 (1.368) 
#>      MaxGEN         Genotype 18 (4.064) 
#>       SigPC                            4
#> ------------------------------- End of data -------------------------------
#> 
#> 
#> 
#> Variable HM 
#> ---------------------------------------------------------------------------
#> Individual analysis of variance
#> ---------------------------------------------------------------------------
#>  ENV  MEAN     MSB    MSG   MSR     FCB      PRFB    FCG      PRFG    CV     h2
#>   E1 47.42  3.1063  5.221 1.852  1.6773 2.148e-01  2.819 2.928e-02 2.870 0.6453
#>  E10 44.30  9.1000 28.774 2.174  4.1857 3.216e-02 13.235 3.024e-06 3.328 0.9244
#>  E11 54.20  2.8101  6.057 3.087  0.9103 4.202e-01  1.962 1.069e-01 3.242 0.4903
#>  E12 49.60  2.7661 10.405 5.322  0.5198 6.033e-01  1.955 1.081e-01 4.651 0.4885
#>  E13 46.56 25.4523 13.883 1.342 18.9637 3.705e-05 10.344 1.798e-05 2.488 0.9033
#>  E14 41.03 10.0151  5.278 1.217  8.2304 2.894e-03  4.338 3.935e-03 2.688 0.7695
#>   E2 44.12  6.6986 13.168 3.049  2.1970 1.400e-01  4.319 4.025e-03 3.958 0.7685
#>   E3 52.90  7.3000  8.004 2.781  2.6245 9.996e-02  2.877 2.692e-02 3.153 0.6525
#>   E4 50.00  2.1000  7.926 1.248  1.6825 2.139e-01  6.350 4.510e-04 2.234 0.8425
#>   E5 52.23  6.8602 11.834 6.371  1.0768 3.616e-01  1.858 1.260e-01 4.833 0.4617
#>   E6 45.89 12.8640 15.835 2.122  6.0628 9.706e-03  7.463 1.633e-04 3.174 0.8660
#>   E7 48.49  0.7052  4.997 4.738  0.1488 8.627e-01  1.055 4.382e-01 4.489 0.0518
#>   E8 45.16 11.9541 11.567 2.168  5.5137 1.356e-02  5.335 1.266e-03 3.260 0.8126
#>   E9 51.33  5.7333  9.333 2.289  2.5049 1.097e-01  4.078 5.408e-03 2.947 0.7548
#>      AS
#>  0.8033
#>  0.9615
#>  0.7002
#>  0.6989
#>  0.9504
#>  0.8772
#>  0.8766
#>  0.8078
#>  0.9179
#>  0.6795
#>  0.9306
#>  0.2276
#>  0.9014
#>  0.8688
#> ---------------------------------------------------------------------------
#> AMMI analysis table
#> ---------------------------------------------------------------------------
#>            Df   Sum Sq  Mean Sq F value    Pr(>F) Percent Accumul
#> ENV        13 5710.317 439.2551  57.224 1.110e-16       .       .
#> REP(ENV)   28  214.931   7.6761   2.703 2.197e-05       .       .
#> GEN         9  269.811  29.9790  10.556 7.415e-14       .       .
#> ENV:GEN   117 1100.734   9.4080   3.313 1.064e-15       .       .
#> PC1        21  381.128  18.1490   6.390 0.000e+00    34.6    34.6
#> PC2        19  319.433  16.8123   5.920 0.000e+00      29    63.6
#> PC3        17  114.264   6.7214   2.370 2.100e-03    10.4      74
#> PC4        15   81.962   5.4641   1.920 2.180e-02     7.4    81.5
#> PC5        13   68.115   5.2396   1.840 3.770e-02     6.2    87.7
#> PC6        11   59.075   5.3704   1.890 4.100e-02     5.4      93
#> PC7         9   46.694   5.1882   1.830 6.330e-02     4.2    97.3
#> PC8         7   26.654   3.8077   1.340 2.318e-01     2.4    99.7
#> PC9         5    3.409   0.6817   0.240 9.445e-01     0.3     100
#> Residuals 252  715.685   2.8400      NA        NA       .       .
#> Total     419 8011.478  19.1205      NA        NA       .       .
#> ---------------------------------------------------------------------------
#> Weighted average of the absolute scores
#> ---------------------------------------------------------------------------
#>  type Code     Y      PC1      PC2       PC3       PC4      PC5       PC6
#>   GEN   G1 47.08  0.28004  0.46346  0.173960 -1.369245 -1.13517  0.036576
#>   GEN  G10 48.51 -1.77853  1.86581 -0.006219  0.921879  0.10958 -0.009745
#>   GEN   G2 46.66  1.56334  0.55184 -0.935711  0.491256  0.28430  1.183833
#>   GEN   G3 47.60  0.34171 -0.20118 -0.800098  0.375281 -0.49786 -1.294456
#>   GEN   G4 48.03 -0.20204 -1.84111  0.280131  0.005954  0.82010  0.273370
#>   GEN   G5 49.30  1.57950  1.03001  1.077784 -0.278911  1.00505 -0.736800
#>   GEN   G6 48.73  0.54738 -0.24525  0.532363  0.460274 -1.00827  0.586133
#>   GEN   G7 47.97 -1.21787 -0.46798  1.254429 -0.054819 -0.03429  0.336625
#>   GEN   G8 49.10 -0.04176 -1.24123 -0.410457  0.639371 -0.17852 -0.514885
#>   GEN   G9 47.90 -1.07177  0.08563 -1.166181 -1.191040  0.63507  0.139349
#>   ENV   E1 47.42  0.13608 -0.16908  0.495099 -0.541029  0.48252  0.086059
#>   ENV  E10 44.30  2.49354  1.01178  0.356768 -0.494268 -0.22549 -0.217574
#>   ENV  E11 54.20 -0.46534 -0.59877 -0.423673 -0.036626 -1.22740  0.808240
#>   ENV  E12 49.60 -0.74418 -0.33767  0.404642 -0.247003 -1.26082 -0.906877
#>   ENV  E13 46.56 -1.10518 -0.03001  0.235896  0.267836  0.58842 -1.088070
#>   ENV  E14 41.03 -0.73055  0.12973  0.478829 -0.439206  0.09735  0.369507
#>   ENV   E2 44.12  0.49108  0.18732  1.389295  1.201258 -0.26148  0.430604
#>   ENV   E3 52.90 -0.69105  0.52932  0.268717  0.060742  0.76957  0.793984
#>   ENV   E4 50.00 -0.42446  0.21439  0.203973  0.367306  0.28614 -0.418461
#>   ENV   E5 52.23  0.33452  1.26609 -1.439597  1.002510 -0.07259 -0.363093
#>   ENV   E6 45.89  0.91446 -1.79070 -0.347140 -0.170210  0.33968 -0.170344
#>   ENV   E7 48.49 -0.32854 -0.07161 -0.729576  0.298759  0.04263  0.655897
#>   ENV   E8 45.16  0.58302 -1.50123 -0.484574  0.017460  0.40377 -0.037154
#>   ENV   E9 51.33 -0.46339  1.16043 -0.408660 -1.287528  0.03771  0.057282
#>        PC7       PC8       PC9   WAAS PctResp PctWAAS PesRes PesWAAS OrResp
#>  -0.266822  0.337572  0.439998 0.4549   15.92   95.88     50      50      9
#>   0.235071  0.077996  0.327796 1.3254   70.01    0.00     50      50      4
#>  -0.722696  0.033680 -0.007948 0.9851    0.00   37.48     50      50     10
#>  -0.498944 -0.880325 -0.049847 0.4176   35.79  100.00     50      50      8
#>   0.241158 -0.437741  0.582344 0.7516   52.02   63.20     50      50      5
#>   0.275998  0.162378 -0.104567 1.1613  100.00   18.07     50      50      1
#>   1.196184 -0.400868 -0.280420 0.4775   78.54   93.39     50      50      3
#>  -1.089955 -0.107564 -0.414067 0.7655   49.53   61.67     50      50      6
#>   0.096046  1.298947 -0.104294 0.5412   92.60   86.38     50      50      2
#>   0.533959 -0.084076 -0.388996 0.7011   47.15   68.77     50      50      7
#>   0.399378 -0.024035 -0.023974 0.2389   48.51  100.00     50      50      8
#>   0.135002 -0.005999 -0.168531 1.3501   24.82    0.00     50      50     12
#>   0.062126  0.260919 -0.574572 0.5389  100.00   73.01     50      50      1
#>   0.273153 -0.578775  0.352277 0.5838   65.05   68.97     50      50      6
#>  -0.177418  0.774048 -0.234689 0.5706   41.99   70.15     50      50      9
#>  -1.422565 -0.353155  0.015005 0.4287    0.00   82.92     50      50     14
#>  -0.211027  0.517609  0.219801 0.5345   23.46   73.40     50      50     13
#>   0.360456 -0.802945 -0.064551 0.5544   90.14   71.60     50      50      2
#>   0.817326 -0.405620 -0.324392 0.3202   68.11   92.69     50      50      5
#>  -0.528568 -0.337561  0.017366 0.7859   85.06   50.77     50      50      3
#>  -0.260876 -0.269925  0.226574 0.9835   36.90   32.99     50      50     10
#>   0.657687  0.436631  0.516351 0.2908   56.68   95.33     50      50      7
#>  -0.106597  0.185296 -0.137266 0.7697   31.37   52.23     50      50     11
#>   0.001925  0.603513  0.180602 0.6882   78.24   59.56     50      50      4
#>  OrWAAS OrPC1 WAASY OrWAASY
#>       2     3 55.90       7
#>      10    10 35.00       9
#>       8     8 18.74      10
#>       1     4 67.89       3
#>       6     2 57.61       6
#>       9     9 59.04       4
#>       3     5 85.96       2
#>       7     7 55.60       8
#>       4     1 89.49       1
#>       5     6 57.96       5
#>       1     1 74.25       5
#>      14    14 12.41      14
#>       6     6 86.50       1
#>       9    11 67.01       8
#>       8    13 56.07       9
#>       4    10 41.46      12
#>       5     7 48.43      10
#>       7     9 80.87       2
#>       3     4 80.40       3
#>      12     3 67.91       7
#>      13    12 34.95      13
#>       2     2 76.01       4
#>      11     8 41.80      11
#>      10     5 68.90       6
#> ---------------------------------------------------------------------------
#> Some information regarding the analysis
#> ---------------------------------------------------------------------------
#>  Parameters                       Values
#>        Ngen                           10
#>        Nenv                           14
#>      OVmean                      48.0883
#>         Min   39.35 (Genotype G2 in E14)
#>         Max 56.3333 (Genotype G6 in E11)
#>      MinENV        Environment 3 (46.66)
#>      MaxENV       Environment 6 (49.298)
#>      MinGEN        Genotype 16 (41.032) 
#>      MaxGEN        Genotype 13 (54.198) 
#>       SigPC                            6
#> ------------------------------- End of data -------------------------------
#> 
#> 
#>