Utilities for handling with rows and columns

add_cols(): Add one or more columns to an existing data frame. If specified before or after columns does not exist, columns are appended at the end of the data. Return a data frame with all the original columns in .data plus the columns declared in ....

concatenate(): Concatenate either two columns of a data frame or a column and specified values. Return a data frame with all the original columns in .data plus the concatenated variable, after the last column.

column_exists(): Checks if a column exists in a data frame. Return a logical value.

get_levels(): Get the levels of a factor variable.

get_level_size(): Get the size of each level of a factor variable.

get_all_pairs(): Get all the possible pairs between the levels of a factor.

select_numeric_cols(): Select all the numeric columns of a data frame.

select_non_numeric_cols(): Select all the non-numeric columns of a data frame.

add_columns(.data, ..., before = NULL, after = NULL)

add_rows(.data, ..., before = NULL, after = NULL)

concatenate(.data, col_1, col_2, var_name = new_var, sep = "_")

column_exists(.data, cols)

get_levels(.data, group)

get_level_size(.data, group)

all_pairs(.data, levels)

select_numeric_cols(.data)

select_non_numeric_cols(.data)

Arguments

.data
A data frame

...
Name-value pairs. All values must have one element for each row in .data when using add_cols or one element for each column in .data when using add_rows. Values of length 1 will be recycled when using add_cols.

before, after
For add_cols, one-based column index or column name where to add the new columns, default: after last column. For add_rows, one-based row index where to add the new rows, default: after last row.

col_1, col_2
An unquoted variable name for the first and second columns to concatenate, respectively. If both columns are in .data, the values of those columns will be concatenated. Else, col_1 or col_2 must be a numeric or a character vector of length equals to the number of rows in .data. Vectors of length 1 will be recycled.

var_name
The name of the new variable containing the concatenated values from col_1 and col_2. Defaults to new_var.

sep
The separator to be between values of col_1 and col_2. Defaults to "_".

cols
A quoted variable name to check if it exists in .data

group
A factor variable to get the levels.

levels
The levels of a factor or a numeric vector.

Examples

# \donttest{
library(metan)

################# Adding columns #################
# Variables x and y after last column
data_ge %>%
  add_columns(x = 10,
              y = 30)

#> # A tibble: 420 x 7
#>    ENV   GEN   REP      GY    HM     x     y
#>    <fct> <fct> <fct> <dbl> <dbl> <dbl> <dbl>
#>  1 E1    G1    1      2.17  44.9    10    30
#>  2 E1    G1    2      2.50  46.9    10    30
#>  3 E1    G1    3      2.43  47.8    10    30
#>  4 E1    G2    1      3.21  45.2    10    30
#>  5 E1    G2    2      2.93  45.3    10    30
#>  6 E1    G2    3      2.56  45.5    10    30
#>  7 E1    G3    1      2.77  46.7    10    30
#>  8 E1    G3    2      3.62  43.2    10    30
#>  9 E1    G3    3      2.28  47.8    10    30
#> 10 E1    G4    1      2.36  47.9    10    30
#> # … with 410 more rows

# Variables x and y before the variable GEN
data_ge %>%
  add_columns(x = 10,
              y = 30,
              before = "GEN")

#> # A tibble: 420 x 7
#>    ENV       x     y GEN   REP      GY    HM
#>    <fct> <dbl> <dbl> <fct> <fct> <dbl> <dbl>
#>  1 E1       10    30 G1    1      2.17  44.9
#>  2 E1       10    30 G1    2      2.50  46.9
#>  3 E1       10    30 G1    3      2.43  47.8
#>  4 E1       10    30 G2    1      3.21  45.2
#>  5 E1       10    30 G2    2      2.93  45.3
#>  6 E1       10    30 G2    3      2.56  45.5
#>  7 E1       10    30 G3    1      2.77  46.7
#>  8 E1       10    30 G3    2      3.62  43.2
#>  9 E1       10    30 G3    3      2.28  47.8
#> 10 E1       10    30 G4    1      2.36  47.9
#> # … with 410 more rows

########### Concatenating columns ################
# Both variables in data
concatenate(data_ge, ENV, GEN)

#> # A tibble: 420 x 6
#>    ENV   GEN   REP      GY    HM new_var
#>    <fct> <fct> <fct> <dbl> <dbl> <chr>  
#>  1 E1    G1    1      2.17  44.9 E1_G1  
#>  2 E1    G1    2      2.50  46.9 E1_G1  
#>  3 E1    G1    3      2.43  47.8 E1_G1  
#>  4 E1    G2    1      3.21  45.2 E1_G2  
#>  5 E1    G2    2      2.93  45.3 E1_G2  
#>  6 E1    G2    3      2.56  45.5 E1_G2  
#>  7 E1    G3    1      2.77  46.7 E1_G3  
#>  8 E1    G3    2      3.62  43.2 E1_G3  
#>  9 E1    G3    3      2.28  47.8 E1_G3  
#> 10 E1    G4    1      2.36  47.9 E1_G4  
#> # … with 410 more rows

# One variable in data
concatenate(data_ge, ENV, rep(1:14, each=30))

#> # A tibble: 420 x 6
#>    ENV   GEN   REP      GY    HM new_var
#>    <fct> <fct> <fct> <dbl> <dbl> <chr>  
#>  1 E1    G1    1      2.17  44.9 E1_1   
#>  2 E1    G1    2      2.50  46.9 E1_1   
#>  3 E1    G1    3      2.43  47.8 E1_1   
#>  4 E1    G2    1      3.21  45.2 E1_1   
#>  5 E1    G2    2      2.93  45.3 E1_1   
#>  6 E1    G2    3      2.56  45.5 E1_1   
#>  7 E1    G3    1      2.77  46.7 E1_1   
#>  8 E1    G3    2      3.62  43.2 E1_1   
#>  9 E1    G3    3      2.28  47.8 E1_1   
#> 10 E1    G4    1      2.36  47.9 E1_1   
#> # … with 410 more rows

# One variable in data, recycling the second one.
concatenate(data_ge, ENV, 1)

#> # A tibble: 420 x 6
#>    ENV   GEN   REP      GY    HM new_var
#>    <fct> <fct> <fct> <dbl> <dbl> <chr>  
#>  1 E1    G1    1      2.17  44.9 E1_1   
#>  2 E1    G1    2      2.50  46.9 E1_1   
#>  3 E1    G1    3      2.43  47.8 E1_1   
#>  4 E1    G2    1      3.21  45.2 E1_1   
#>  5 E1    G2    2      2.93  45.3 E1_1   
#>  6 E1    G2    3      2.56  45.5 E1_1   
#>  7 E1    G3    1      2.77  46.7 E1_1   
#>  8 E1    G3    2      3.62  43.2 E1_1   
#>  9 E1    G3    3      2.28  47.8 E1_1   
#> 10 E1    G4    1      2.36  47.9 E1_1   
#> # … with 410 more rows

# None variables in data
# Set name of the concatenate variable
# Set the separator symbol
concatenate(data_ge,
            col_1 = rep("ENV", 420),
            col_2 = rep(1:14, each=30),
            var_name = ENVIRONMENT,
            sep = "-")

#> # A tibble: 420 x 6
#>    ENV   GEN   REP      GY    HM ENVIRONMENT
#>    <fct> <fct> <fct> <dbl> <dbl> <chr>      
#>  1 E1    G1    1      2.17  44.9 ENV-1      
#>  2 E1    G1    2      2.50  46.9 ENV-1      
#>  3 E1    G1    3      2.43  47.8 ENV-1      
#>  4 E1    G2    1      3.21  45.2 ENV-1      
#>  5 E1    G2    2      2.93  45.3 ENV-1      
#>  6 E1    G2    3      2.56  45.5 ENV-1      
#>  7 E1    G3    1      2.77  46.7 ENV-1      
#>  8 E1    G3    2      3.62  43.2 ENV-1      
#>  9 E1    G3    3      2.28  47.8 ENV-1      
#> 10 E1    G4    1      2.36  47.9 ENV-1      
#> # … with 410 more rows

################### Adding rows ##################
data_ge %>%
  add_rows(ENV = "E_TEST",
           GEN = "G_TEST",
           REP = 3,
           GY = 10.3,
           HM = 100.11,
           after = 1)

#> # A tibble: 421 x 5
#>    ENV    GEN    REP      GY    HM
#>    <fct>  <fct>  <fct> <dbl> <dbl>
#>  1 E1     G1     1      2.17  44.9
#>  2 E_TEST G_TEST 3     10.3  100. 
#>  3 E1     G1     2      2.50  46.9
#>  4 E1     G1     3      2.43  47.8
#>  5 E1     G2     1      3.21  45.2
#>  6 E1     G2     2      2.93  45.3
#>  7 E1     G2     3      2.56  45.5
#>  8 E1     G3     1      2.77  46.7
#>  9 E1     G3     2      3.62  43.2
#> 10 E1     G3     3      2.28  47.8
#> # … with 411 more rows

########## checking if a column exists ###########
column_exists(data_g, "GEN")

#> [1] TRUE

####### get the levels and size of levels ########
get_levels(data_g, GEN)

#>  [1] "H1"  "H10" "H11" "H12" "H13" "H2"  "H3"  "H4"  "H5"  "H6"  "H7"  "H8" 
#> [13] "H9" 

get_level_size(data_g, GEN)

#>  H1 H10 H11 H12 H13  H2  H3  H4  H5  H6  H7  H8  H9 
#>   3   3   3   3   3   3   3   3   3   3   3   3   3 

############## all possible pairs ################
all_pairs(data_g, GEN)

#>     [,1]  [,2] 
#> V1  "H1"  "H10"
#> V2  "H1"  "H11"
#> V3  "H1"  "H12"
#> V4  "H1"  "H13"
#> V5  "H1"  "H2" 
#> V6  "H1"  "H3" 
#> V7  "H1"  "H4" 
#> V8  "H1"  "H5" 
#> V9  "H1"  "H6" 
#> V10 "H1"  "H7" 
#> V11 "H1"  "H8" 
#> V12 "H1"  "H9" 
#> V13 "H10" "H11"
#> V14 "H10" "H12"
#> V15 "H10" "H13"
#> V16 "H10" "H2" 
#> V17 "H10" "H3" 
#> V18 "H10" "H4" 
#> V19 "H10" "H5" 
#> V20 "H10" "H6" 
#> V21 "H10" "H7" 
#> V22 "H10" "H8" 
#> V23 "H10" "H9" 
#> V24 "H11" "H12"
#> V25 "H11" "H13"
#> V26 "H11" "H2" 
#> V27 "H11" "H3" 
#> V28 "H11" "H4" 
#> V29 "H11" "H5" 
#> V30 "H11" "H6" 
#> V31 "H11" "H7" 
#> V32 "H11" "H8" 
#> V33 "H11" "H9" 
#> V34 "H12" "H13"
#> V35 "H12" "H2" 
#> V36 "H12" "H3" 
#> V37 "H12" "H4" 
#> V38 "H12" "H5" 
#> V39 "H12" "H6" 
#> V40 "H12" "H7" 
#> V41 "H12" "H8" 
#> V42 "H12" "H9" 
#> V43 "H13" "H2" 
#> V44 "H13" "H3" 
#> V45 "H13" "H4" 
#> V46 "H13" "H5" 
#> V47 "H13" "H6" 
#> V48 "H13" "H7" 
#> V49 "H13" "H8" 
#> V50 "H13" "H9" 
#> V51 "H2"  "H3" 
#> V52 "H2"  "H4" 
#> V53 "H2"  "H5" 
#> V54 "H2"  "H6" 
#> V55 "H2"  "H7" 
#> V56 "H2"  "H8" 
#> V57 "H2"  "H9" 
#> V58 "H3"  "H4" 
#> V59 "H3"  "H5" 
#> V60 "H3"  "H6" 
#> V61 "H3"  "H7" 
#> V62 "H3"  "H8" 
#> V63 "H3"  "H9" 
#> V64 "H4"  "H5" 
#> V65 "H4"  "H6" 
#> V66 "H4"  "H7" 
#> V67 "H4"  "H8" 
#> V68 "H4"  "H9" 
#> V69 "H5"  "H6" 
#> V70 "H5"  "H7" 
#> V71 "H5"  "H8" 
#> V72 "H5"  "H9" 
#> V73 "H6"  "H7" 
#> V74 "H6"  "H8" 
#> V75 "H6"  "H9" 
#> V76 "H7"  "H8" 
#> V77 "H7"  "H9" 
#> V78 "H8"  "H9" 

########## select numeric variables only #########
select_numeric_cols(data_g)

#> # A tibble: 39 x 15
#>       PH    EH    EP    EL    ED    CL    CD    CW    KW    NR   NKR  CDED  PERK
#>    <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1  2.11 1.05  0.497  15.7  49.9  30.5  16.6  28.6  164.  15.6  31.2 0.612  85.1
#>  2  2.20 1.09  0.492  13.7  49.2  30.5  14.7  22.3  130.  16.4  24.8 0.619  85.2
#>  3  2.29 1.15  0.502  15.1  52.6  31.7  16.2  29.6  176.  15.6  29.2 0.603  85.9
#>  4  1.79 0.888 0.514  13.9  44.1  26.2  15.0  12.9  116.  14.8  33   0.596  89.8
#>  5  2.05 1.03  0.504  13.6  43.9  23.5  14.4  11.5  118.  16    32.4 0.535  91.1
#>  6  2.27 1.11  0.491  14.5  43.7  24.6  16.1  12.5  128.  15.2  34.6 0.566  90.7
#>  7  1.71 0.808 0.489  15.5  45.2  25.0  16.7  15.2  140.  15.6  36   0.552  90.3
#>  8  2.09 1.06  0.509  12.2  46.9  26.5  14.3  13.5  114.  16.8  26.2 0.566  89.3
#>  9  2.5  1.44  0.577  15.0  49.0  27.5  15.2  19.4  168.  16.4  35   0.562  89.6
#> 10  2.52 1.52  0.601  14.4  49.2  28.4  15    18.2  153.  16.4  32   0.577  89.2
#> # … with 29 more rows, and 2 more variables: TKW <dbl>, NKE <dbl>

select_non_numeric_cols(data_g)

#> # A tibble: 39 x 2
#>    GEN   REP  
#>    <fct> <fct>
#>  1 H1    1    
#>  2 H1    2    
#>  3 H1    3    
#>  4 H10   1    
#>  5 H10   2    
#>  6 H10   3    
#>  7 H11   1    
#>  8 H11   2    
#>  9 H11   3    
#> 10 H12   1    
#> # … with 29 more rows

# }