Random Sampling

sample_random() performs Simple Random Sampling or Stratified Random Sampling

sample_systematic() performs systematic sampling. In this case, a regular interval of size k (k = floor(N/n)) is generated considering the population size (N) and desired sample size (n). Then, the starting member (r) is randomly chosen between 1-k. The second element is r + k, and so on.

Usage

sample_random(data, n, prop, by = NULL, weight = NULL)

sample_systematic(data, n, r = NULL, by = NULL)

Arguments

data

A data frame. If data is a grouped_df, the operation will be performed on each group (stratified).

n, prop

Provide either n, the number of rows, or prop, the proportion of rows to select. If neither are supplied, n = 1 will be used.

by

A categorical variable to compute the sample by. It is a shortcut to dplyr::group_by() that allows to group the data by one categorical variable. If more than one grouping variable needs to be used, use dplyr::group_by() to pass the data grouped.

weight

Sampling weights. This must evaluate to a vector of non-negative numbers the same length as the input. Weights are automatically standardised to sum to 1.

r

The starting element. By default, r is randomly selected between 1:k

Value

An object of the same type as data.

Examples

library(metan)
sample_random(data_ge, n = 5)
#> # A tibble: 5 × 5
#>   ENV   GEN   REP      GY    HM
#>   <fct> <fct> <fct> <dbl> <dbl>
#> 1 E8    G1    2      2.38  43  
#> 2 E14   G2    2      1.64  39.0
#> 3 E10   G10   3      1.17  42  
#> 4 E6    G5    3      2.53  47.6
#> 5 E9    G10   1      2.85  54  
sample_random(data_ge,
              n = 3,
              by = ENV)
#> # A tibble: 42 × 5
#>    ENV   GEN   REP      GY    HM
#>    <fct> <fct> <fct> <dbl> <dbl>
#>  1 E1    G2    1      3.21  45.2
#>  2 E1    G3    2      3.62  43.2
#>  3 E1    G10   3      2.48  48.8
#>  4 E10   G6    3      1.82  47  
#>  5 E10   G10   1      1.70  40  
#>  6 E10   G8    1      2.65  42  
#>  7 E11   G6    1      1.33  56  
#>  8 E11   G1    1      1.02  54  
#>  9 E11   G7    3      1.00  53  
#> 10 E12   G9    2      1.37  48.4
#> # ℹ 32 more rows

sample_systematic(data_g, n = 6)
#> k = 6
#> # A tibble: 6 × 18
#>     .id GEN   REP      PH    EH    EP    EL    ED    CL    CD    CW    KW    NR
#>   <dbl> <fct> <fct> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1     5 H10   2      2.05 1.03  0.504  13.6  43.9  23.5  14.4  11.5  118.  16  
#> 2    11 H12   2      2.77 1.58  0.572  13.8  46.5  23.8  14.6  16.3  153.  17.6
#> 3    17 H2    2      2.10 0.984 0.467  15.9  50.2  30.3  17.0  20.0  166.  16  
#> 4    23 H4    2      2.14 1.03  0.482  16.8  46.2  27.9  17.2  21.9  159.  13.2
#> 5    29 H6    2      2.09 1.06  0.508  13.8  49.4  31.5  15.0  21.1  131.  17.2
#> 6    35 H8    2      2.15 1.11  0.514  14.6  45.1  28.0  15.6  19.1  131.  14.4
#> # ℹ 5 more variables: NKR <dbl>, CDED <dbl>, PERK <dbl>, TKW <dbl>, NKE <dbl>