Weighted Average of Absolute Scores

Compute the Weighted Average of Absolute Scores for AMMI analysis (Olivoto et al., 2019).

waas(.data, env, gen, rep, resp, mresp = NULL, wresp = NULL,
  prob = 0.05, naxis = NULL, ind_anova = TRUE, verbose = TRUE)

Arguments

.data
The dataset containing the columns related to Environments, Genotypes, replication/block and response variable(s).

env
The name of the column that contains the levels of the environments.

gen
The name of the column that contains the levels of the genotypes.

rep
The name of the column that contains the levels of the replications/blocks.

resp
The response variable(s). To analyze multiple variables in a single procedure a vector of variables may be used. For example resp = c(var1, var2, var3).

mresp
A numeric vector of the same length of resp. The mresp will be the new maximum value after rescaling. By default, all variables in resp are rescaled so that de maximum value is 100 and the minimum value is 0.

wresp
The weight for the response variable(s) for computing the WAASBY index. Must be a numeric vector of the same length of resp. Defaults to 50, i.e., equal weights for stability and mean performance.

prob
The p-value for considering an interaction principal component axis significant.

naxis
The number of IPCAs to be used for computing the WAAS index. Default is NULL (Significant IPCAs are used). If values are informed, the number of IPCAS will be used independently on its significance. Note that if two or more variables are included in resp, then naxis must be a vector.

ind_anova
Logical argument set to TRUE. If FALSE the within-environment ANOVA is not performed.

verbose
Logical argument. If verbose = FALSE the code is run silently.

Value

An object of class waas with the following items for each variable:

individual

A within-environments ANOVA considering a fixed-effect model.

model

A data frame with the response variable, the scores of all Principal Components, the estimates of Weighted Average of Absolute Scores, and WAASY (the index that consider the weights for stability and productivity in the genotype ranking.

MeansGxE

The means of genotypes in the environments, with observed, predicted and residual values.

PCA

Principal Component Analysis.

anova

Joint analysis of variance for the main effects and Principal Component analysis of the interaction effect.

Details

A list summarizing the results. The following information are showed. WgtResponse, the weight for the response variable in estimating WAASB, WgtWAAS the weight for stability, Ngen the number of genotypes, Nenv the number of environments, OVmean the overall mean, Min the minimum observed (returning the genotype and environment), Max the maximum observed, Max the maximum observed, MinENV the environment with the lower mean, MaxENV the environment with the larger mean observed, MinGEN the genotype with the lower mean, MaxGEN the genotype with the larger.

residuals

The residuals of the model.

probint

The p-value for the genotype-vs-environment interaction.

Details

This function compute the weighted average of absolute scores, estimated as follows:

$$ WAAS_i = \sum_{k = 1}^{p} |IPCA_{ik} \times EP_k|/ \sum_{k = 1}^{p}EP_k$$

where \(WAAS_i\) is the weighted average of absolute scores of the ith genotype; \(PCA_{ik}\) is the score of the ith genotype in the kth IPCA; and \(EP_k\) is the explained variance of the *k*th IPCA for k = 1,2,..,p, considering p the number of significant PCAs, or a declared number of PCAs. For example if prob = 0.05, all axis that are significant considering this probability level are used. The number of axis can be also informed by declaring naxis = x. This comand ignores the prob argument.

References

Olivoto, T., A.D.C. L\'ucio, J.A.G. da silva, V.S. Marchioro, V.Q. de Souza, and E. Jost. 2019. Mean performance and stability in multi-environment trials I: Combining features of AMMI and BLUP techniques. Agron. J. (in press).

See also

Examples

library(metan)

# Considering p-value <= 0.05 to compute the WAAS

model <- waas(data_ge,
              env = ENV,
              gen = GEN,
              rep = REP,
              resp = GY)

#> Done!

# Declaring the number of axis to be used for computing WAAS
# and assigning a larger weight for the response variable when
# computing the WAASBY index.

model2 <- waas(data_ge,
               env = ENV,
               gen = GEN,
               rep = REP,
               resp = GY,
               naxis = 3,
               wresp = 60)

#> Done!

# Analyzing multiple variables (GY and HM) at the same time
# considering that smaller values of HM are better and higher
# values of GY are better, assigning a larger weight for the GY
# and a smaller weight for HM when computing WAASBY index.

model3 <- waas(data_ge,
               env = ENV,
               gen = GEN,
               rep = REP,
               resp = c(GY, HM),
               mresp = c(100, 0),
               wresp = c(60, 40))

#> Evaluating variable GY 50 % 
#> Evaluating variable HM 100 % 
#> Done!