Weighted Average of Absolute Scores

Compute the Weighted Average of Absolute Scores (Olivoto et al., 2019) based on means for genotype-environment data as follows:

waas_means(
  .data,
  env,
  gen,
  resp,
  mresp = NULL,
  wresp = NULL,
  min_expl_var = 85,
  verbose = TRUE
)

Arguments

.data
The dataset containing the columns related to Environments, Genotypes, replication/block and response variable(s).

env
The name of the column that contains the levels of the environments.

gen
The name of the column that contains the levels of the genotypes.

resp
The response variable(s). To analyze multiple variables in a single procedure a vector of variables may be used. For example resp = c(var1, var2, var3). Select helpers are also allowed.

mresp
A numeric vector of the same length of resp. The mresp will be the new maximum value after rescaling. By default, all variables in resp are rescaled so that de maximum value is 100 and the minimum value is 0.

wresp
The weight for the response variable(s) for computing the WAASBY index. Must be a numeric vector of the same length of resp. Defaults to 50, i.e., equal weights for stability and mean performance.

min_expl_var
The minimum explained variance. Defaults to 85. Interaction Principal Compoment Axis are iteractively retained up to the explained variance (eigenvalues in the singular value decomposition of the matrix with the interaction effects) be greather than or equal to min_expl_var. For example, if the explained variance (in percentage) in seven possible IPCAs are 56, 21, 9, 6, 4, 3, 1, resulting in a cumulative proportion of 56, 77, 86, 92, 96, 99, 100, then p = 3, i.e., three IPCAs will be used to compute the index WAAS.

verbose
Logical argument. If verbose = FALSE the code is run silently.

Value

An object of class waas_means with the following items for each variable:

model A data frame with the response variable, the scores of all Principal Components, the estimates of Weighted Average of Absolute Scores, and WAASY (the index that consider the weights for stability and productivity in the genotype ranking.

ge_eff A gxe matrix containing the genotype-environment effects.

eigenvalues The eigenvalues from the singular value decomposition of the matrix withe the genotype-environment interaction effects.

proportion The proportion of the variance explained by each IPCA.

cum_proportion The cumulative proportion of the variance explained.

Details

$$ WAAS_i = \sum_{k = 1}^{p} |IPCA_{ik} \times EP_k|/ \sum_{k = 1}^{p}EP_k$$

where \(WAAS_i\) is the weighted average of absolute scores of the ith genotype; \(PCA_{ik}\) is the score of the ith genotype in the kth IPCA; and \(EP_k\) is the explained variance of the kth IPCA for k = 1,2,..,p, where p is the number of IPCAs that explain at least an amount of the genotype-interaction variance declared in the argument min_expl_var.

References

Olivoto, T., A.D.C. L\'ucio, J.A.G. da silva, V.S. Marchioro, V.Q. de Souza, and E. Jost. 2019a. Mean performance and stability in multi-environment trials I: Combining features of AMMI and BLUP techniques. Agron. J. 111:2949-2960. doi:10.2134/agronj2019.03.0220

See also

Examples

# \donttest{
library(metan)
# Data with replicates
model <- waas(data_ge,
              env = ENV,
              gen = GEN,
              rep = REP,
              resp = everything())

#> Evaluating variable GY 0 % 
#> Evaluating variable HM 100 % 
#> All variables with significant (p < 0.05) genotype-vs-environment interaction
#> Done!

# Based on means of genotype-environment data
data_means <- means_by(data_ge, ENV, GEN)
model2 <- waas_means(data_ge,
                     env = ENV,
                     gen = GEN,
                     resp = everything())

#> Evaluating variable GY 0 % 
#> Evaluating variable HM 100 % 

# The index WAAS
get_model_data(model, what = "OrWAASB")

#> # A tibble: 10 x 3
#>    gen      GY    HM
#>    <fct> <dbl> <dbl>
#>  1 G1        2     2
#>  2 G10      10    10
#>  3 G2        5     8
#>  4 G3        1     1
#>  5 G4        6     6
#>  6 G5        4     9
#>  7 G6        3     3
#>  8 G7        8     7
#>  9 G8        7     4
#> 10 G9        9     5

get_model_data(model2, what = "OrWAASB")

#> # A tibble: 10 x 3
#>    gen      GY    HM
#>    <fct> <dbl> <dbl>
#>  1 G1        2     3
#>  2 G10      10    10
#>  3 G2        5     8
#>  4 G3        1     1
#>  5 G4        6     6
#>  6 G5        4     9
#>  7 G6        3     2
#>  8 G7        8     7
#>  9 G8        7     4
#> 10 G9        9     5

# }