Weighted Average of Absolute Scores

Compute the Weighted Average of Absolute Scores for quantifying the stability in multi-environment trials using mixed-effect models (Olivoto et al., 2019).

waasb(
  .data,
  env,
  gen,
  rep,
  resp,
  mresp = NULL,
  wresp = NULL,
  random = "gen",
  prob = 0.05,
  ind_anova = TRUE,
  verbose = TRUE
)

Arguments

.data
The dataset containing the columns related to Environments, Genotypes, replication/block and response variable(s).

env
The name of the column that contains the levels of the environments.

gen
The name of the column that contains the levels of the genotypes.

rep
The name of the column that contains the levels of the replications/blocks.

resp
The response variable(s). To analyze multiple variables in a single procedure a vector of variables may be used. For example resp = c(var1, var2, var3).

mresp
A numeric vector of the same length of resp. The mresp will be the new maximum value after rescaling. By default, all variables in resp are rescaled so that de maximum value is 100 and the minimum value is 0.

wresp
The weight for the response variable(s) for computing the WAASBY index. Must be a numeric vector of the same length of resp. Defaults to 50, i.e., equal weights for stability and mean performance.

random
The effects of the model assumed to be random. Default is random = "gen" (genotype and genotype-vs-environment as random effects. Other values allowed are random = "env" (environment, genotype-vs-environment and block-within-environment random effects) or random = "all" all effects except the intercept are assumed to be random effects.

prob
The probability for estimating confidence interval for BLUP's prediction.

ind_anova
Logical argument set to TRUE. If FALSE the within-environment ANOVA is not performed.

verbose
Logical argument. If verbose = FALSE the code are run silently.

Value

An object of class waasb with the following items for each variable: * individual A within-environments ANOVA considering a fixed-effect model.

fixed Test for fixed effects.

random Variance components for random effects.

LRT The Likelihood Ratio Test for the random effects.

model A data frame with the response variable, the scores of all Principal Components, the estimates of Weighted Average of Absolute Scores, and WAASY (the index that consider the weights for stability and productivity in the genotype ranking.

blupGEN The estimated BLUPS for genotypes (If random = "gen" or random = "all")

BLUPenv The estimated BLUPS for environments, (If random = "env" or random = "all").

BLUPge The estimated BLUPS of all genotypes in all environments "BLUPij".

PCA The results of Principal Component Analysis with eigenvalues and explained variance of BLUP-interaction matrix.

MeansGxE The phenotypic means of genotypes in the environments, with observed, predicted and OLS residual prediction.

Details A list summarizing the results. The following information are showed. WgtResponse, the weight for the response variable in estimating WAASB, WgtWAAS the weight for stability, Ngen the number of genotypes, Nenv the number of environments, OVmean the overall mean, Min the minimum observed (returning the genotype and environment), Max the maximum observed, Max the maximum observed, MinENV the environment with the lower mean, MaxENV the environment with the larger mean observed, MinGEN the genotype with the lower mean, MaxGEN the genotype with the larger.

ESTIMATES A list with the following values: GEV the genotype-by-environment variance (and percentage of phenotypic variance); GV the genotypic variance (and percentage of phenotypic variance); EV the environmental variance;RV the residual variance (and percentage of phenotypic variance); FV the phenotypic variance; h2g the heritability of the trait; GEr2 the coefficient of determination of the interaction effects; h2mg the heritability of the mean; AccuGen the selective accuracy; rge the genotype-environment correlation; CVg the genotypic coefficient of variation; CVr the residual coefficient of variation; CVratio the ratio between genotypic and residual coefficient of variation * residuals The residuals of the model.

Details

This function compute the weighted average of absolute scores considering all principal component axis from the Singular Value Decomposition (SVD) of the BLUP'S GxE effects matrix generated by a linear mixed-effect model.

References

Olivoto, T., A.D.C. L\'ucio, J.A.G. da silva, V.S. Marchioro, V.Q. de Souza, and E. Jost. 2019. Mean performance and stability in multi-environment trials I: Combining features of AMMI and BLUP techniques. Agron. J. doi:10.2134/agronj2019.03.0220

See also

Examples

# \donttest{
library(metan)

# Genotypes as random effects
# Equal weights for response variable and stability

model <- waasb(data_ge,
              env = ENV,
              gen = GEN,
              rep = REP,
              resp = GY)

#> All variables with significant (p < 0.05) genotype-vs-environment interaction
#> Done!

# Higher weight for response variable

model2 <- waasb(data_ge,
               env = ENV,
               gen = GEN,
               rep = REP,
               resp = GY,
               wresp = 65)

#> All variables with significant (p < 0.05) genotype-vs-environment interaction
#> Done!

# Environment as random effects analyzing more than one variable
# Smaller values of HM are better and higher
# Higher values of GY are better
# Larger weight for the GY (60%)
# Smaller weight for HM (40%)

model3 <- waasb(data_ge,
                env = ENV,
                gen = GEN,
                rep = REP,
                resp = c(GY, HM),
                random = "env",
                mresp = c(100, 0),
                wresp = c(60, 40))

#> Evaluating variable GY 0 % 
#> Evaluating variable HM 100 % 
#> All variables with significant (p < 0.05) genotype-vs-environment interaction
#> Done!

# }