Weighting between stability and mean performance

This function computes the WAASY or WAASBY indexes (Olivoto et al., 2019) considering different scenarios of weights for stability and mean performance.

After fitting a model with the functions waas() or waasb() it is possible to compute the superiority indexes WAASY or WAASBY in different scenarios of weights for stability and mean performance. The number of scenarios is defined by the arguments increment. By default, twenty-one different scenarios are computed. In this case, the the superiority index is computed considering the following weights: stability (waasb or waas) = 100; mean performance = 0. In other words, only stability is considered for genotype ranking. In the next iteration, the weights becomes 95/5 (since increment = 5). In the third scenario, the weights become 90/10, and so on up to these weights become 0/100. In the last iteration, the genotype ranking for WAASY or WAASBY matches perfectly with the ranks of the response variable.

Usage

wsmp(
  model,
  mresp = 100,
  increment = 5,
  saveWAASY = 50,
  prob = 0.05,
  progbar = TRUE
)

Arguments

model

An object computed with waas(), waasb(), or mps().

mresp

A numeric value that will be the new maximum value after rescaling. By default, the variable in resp is rescaled so that the original maximum and minimum values are 100 and 0, respectively. Let us consider that for a specific trait, say, lodging incidence, lower values are better. In this case, you should use mresp = 0 to rescale the response variable so that the lowest values will become 100 and the highest values 0.

increment

The increment in the weight ratio for stability and mean performance. Se the Details section for more information.

saveWAASY

Automatically save the WAASY values when the weight for stability is saveWAASY.

prob

The p-value for considering an interaction principal component axis significant. must be multiple of increment. If this assumption is not valid, an error will be occur.

progbar

A logical argument to define if a progress bar is shown. Default is TRUE.

Value

An object of class wsmp with the following items for each variable

When computed with waas() or waasb().

scenarios A list with the model for all computed scenarios.

WAASY The values of the WAASY estimated when the weight for the stability in the loop match with argument saveWAASY.

hetdata, hetcomb The data used to produce the heatmaps.

Ranks All the values of WAASY estimated in the different scenarios of WAAS/GY weighting ratio.

When computed with mps()

hetcomb showing the rank for mean performance and stability in the different weights.

References

Olivoto, T., A.D.C. L\'ucio, J.A.G. da silva, V.S. Marchioro, V.Q. de Souza, and E. Jost. 2019. Mean performance and stability in multi-environment trials I: Combining features of AMMI and BLUP techniques. Agron. J. doi:10.2134/agronj2019.03.0220

See also

Author

Tiago Olivoto tiagoolivoto@gmail.com

Examples

# \donttest{
library(metan)
# using the WAASB as statistic and BLUP as mean performance
# the same as using waasb()

model <- mps(data_ge2,
             env = ENV,
             gen = GEN,
             rep = REP,
             resp = PH)
#> Evaluating trait PH |============================================| 100% 00:00:00 
#> Method: REML/BLUP
#> Random effects: GEN, GEN:ENV
#> Fixed effects: ENV, REP(ENV)
#> Denominador DF: Satterthwaite's method
#> ---------------------------------------------------------------------------
#> P-values for Likelihood Ratio Test of the analyzed traits
#> ---------------------------------------------------------------------------
#>     model       PH
#>  COMPLETE       NA
#>       GEN 9.39e-01
#>   GEN:ENV 1.09e-13
#> ---------------------------------------------------------------------------
#> All variables with significant (p < 0.05) genotype-vs-environment interaction
#> Mean performance: blupg
#> Stability: waasb
scenarios <- wsmp(model)
# }