Within-environment analysis of variance

Performs a within-environment analysis of variance in randomized complete block or alpha-lattice designs and returns values such as Mean Squares, p-values, coefficient of variation, heritability, and accuracy of selection.

anova_ind(.data, env, gen, rep, resp, block = NULL, verbose = TRUE)

Arguments

.data
The dataset containing the columns related to Environments, Genotypes, replication/block and response variable(s).

env
The name of the column that contains the levels of the environments. The analysis of variance is computed for each level of this factor.

gen
The name of the column that contains the levels of the genotypes.

rep
The name of the column that contains the levels of the replications/blocks.

resp
The response variable(s). To analyze multiple variables in a single procedure a vector of variables may be used. For example resp = c(var1, var2, var3).

block
Defaults to NULL. In this case, a randomized complete block design is considered. If block is informed, then a resolvable alpha-lattice design (Patterson and Williams, 1976) is employed. All effects, except the error, are assumed to be fixed.

verbose
Logical argument. If verbose = FALSE the code will run silently.

Value

A list where each element is the result for one variable containing:

individual: A tidy tbl_df with the results of the individual analysis of variance with the following column names:

For analysis in alpha-lattice designs: ENV: The environment code; MEAN: The grand mean; MSG, MSCR, MSIB_R: The mean squares for genotype, replicates and incomplete blocks within replicates, respectively. FCG, FCR, FCIB_R: The F-calculated for genotype, replicates and incomplete blocks within replicates, respectively.PFG, PFCR, PFIB_R: The P-values for genotype, replicates and incomplete blocks within replicates, respectively.

For analysis in randomized complete block design: MSG, MSB: The mean squares for genotype and blocks, respectively. FCG, FCB: The F-calculated for genotype and blocks, respectively. PFG, PFB: The P-values for genotype and blocks, respectively.

MSRatio The ratio between the higher and lower residual mean square.

References

Patterson, H.D., and E.R. Williams. 1976. A new class of resolvable incomplete block designs. Biometrika 63:83-92. doi:10.1093/biomet/63.1.83

Examples

# \donttest{
library(metan)
# ANOVA for all variables in data
ind_an <- anova_ind(data_ge,
                    env = ENV,
                    gen = GEN,
                    rep = REP,
                    resp = everything())

#> variable GY 
#> ---------------------------------------------------------------------------
#> Within-environment ANOVA results
#> ---------------------------------------------------------------------------
#>  ENV MEAN   MSG   FCG      PFG    MSB    FCB      PFB    MSE    CV    h2    AS
#>   E1 2.52 0.337  2.34 0.059357 0.0652  0.453 6.43e-01 0.1440 15.05 0.573 0.757
#>  E10 2.18 0.296 11.09 0.000011 0.6537 24.508 7.28e-06 0.0267  7.51 0.910 0.954
#>  E11 1.37 0.151  1.44 0.244312 0.3773  3.594 4.86e-02 0.1050 23.68 0.304 0.552
#>  E12 1.61 0.320  5.98 0.000647 0.0919  1.717 2.08e-01 0.0535 14.38 0.833 0.913
#>  E13 2.91 0.713  7.18 0.000210 0.0767  0.772 4.77e-01 0.0994 10.83 0.861 0.928
#>  E14 1.78 0.131  1.73 0.153470 0.1035  1.374 2.78e-01 0.0753 15.40 0.423 0.650
#>   E2 3.18 0.207  1.16 0.376322 0.6983  3.912 3.88e-02 0.1785 13.29 0.136 0.369
#>   E3 4.06 0.335  1.87 0.123033 0.4886  2.731 9.21e-02 0.1789 10.41 0.466 0.683
#>   E4 3.68 0.531  3.86 0.007123 0.1165  0.846 4.46e-01 0.1377 10.10 0.741 0.861
#>   E5 3.91 0.526  7.93 0.000110 0.2188  3.297 6.02e-02 0.0664  6.59 0.874 0.935
#>   E6 2.66 0.135  2.30 0.063454 0.1599  2.729 9.22e-02 0.0586  9.09 0.565 0.752
#>   E7 1.99 0.337  3.70 0.008727 0.3810  4.185 3.22e-02 0.0910 15.17 0.730 0.854
#>   E8 2.54 0.215  7.72 0.000131 0.8168 29.369 2.15e-06 0.0278  6.57 0.870 0.933
#>   E9 3.06 0.679  6.12 0.000562 0.5826  5.253 1.60e-02 0.1109 10.90 0.837 0.915
#> ---------------------------------------------------------------------------
#> 
#> variable HM 
#> ---------------------------------------------------------------------------
#> Within-environment ANOVA results
#> ---------------------------------------------------------------------------
#>  ENV MEAN   MSG   FCG      PFG    MSB    FCB      PFB  MSE   CV     h2    AS
#>   E1 47.4  5.22  2.82 2.93e-02  3.106  1.677 2.15e-01 1.85 2.87 0.6453 0.803
#>  E10 44.3 28.77 13.24 3.02e-06  9.100  4.186 3.22e-02 2.17 3.33 0.9244 0.961
#>  E11 54.2  6.06  1.96 1.07e-01  2.810  0.910 4.20e-01 3.09 3.24 0.4903 0.700
#>  E12 49.6 10.40  1.96 1.08e-01  2.766  0.520 6.03e-01 5.32 4.65 0.4885 0.699
#>  E13 46.6 13.88 10.34 1.80e-05 25.452 18.964 3.71e-05 1.34 2.49 0.9033 0.950
#>  E14 41.0  5.28  4.34 3.93e-03 10.015  8.230 2.89e-03 1.22 2.69 0.7695 0.877
#>   E2 44.1 13.17  4.32 4.02e-03  6.699  2.197 1.40e-01 3.05 3.96 0.7685 0.877
#>   E3 52.9  8.00  2.88 2.69e-02  7.300  2.625 1.00e-01 2.78 3.15 0.6525 0.808
#>   E4 50.0  7.93  6.35 4.51e-04  2.100  1.682 2.14e-01 1.25 2.23 0.8425 0.918
#>   E5 52.2 11.83  1.86 1.26e-01  6.860  1.077 3.62e-01 6.37 4.83 0.4617 0.679
#>   E6 45.9 15.84  7.46 1.63e-04 12.864  6.063 9.71e-03 2.12 3.17 0.8660 0.931
#>   E7 48.5  5.00  1.05 4.38e-01  0.705  0.149 8.63e-01 4.74 4.49 0.0518 0.228
#>   E8 45.2 11.57  5.34 1.27e-03 11.954  5.514 1.36e-02 2.17 3.26 0.8126 0.901
#>   E9 51.3  9.33  4.08 5.41e-03  5.733  2.505 1.10e-01 2.29 2.95 0.7548 0.869
#> ---------------------------------------------------------------------------
#> 

# mean for each environment
get_model_data(ind_an)

#> Class of the model: anova_ind

#> Variable extracted: MEAN

#> # A tibble: 14 x 3
#>    ENV      GY    HM
#>    <chr> <dbl> <dbl>
#>  1 E1     2.52  47.4
#>  2 E10    2.18  44.3
#>  3 E11    1.37  54.2
#>  4 E12    1.61  49.6
#>  5 E13    2.91  46.6
#>  6 E14    1.78  41.0
#>  7 E2     3.18  44.1
#>  8 E3     4.06  52.9
#>  9 E4     3.68  50  
#> 10 E5     3.91  52.2
#> 11 E6     2.66  45.9
#> 12 E7     1.99  48.5
#> 13 E8     2.54  45.2
#> 14 E9     3.06  51.3

# P-value for genotype effect
get_model_data(ind_an, "PFG")

#> Class of the model: anova_ind

#> Variable extracted: PFG

#> # A tibble: 14 x 3
#>    ENV          GY         HM
#>    <chr>     <dbl>      <dbl>
#>  1 E1    0.0594    0.0293    
#>  2 E10   0.0000110 0.00000302
#>  3 E11   0.244     0.107     
#>  4 E12   0.000647  0.108     
#>  5 E13   0.000210  0.0000180 
#>  6 E14   0.153     0.00393   
#>  7 E2    0.376     0.00402   
#>  8 E3    0.123     0.0269    
#>  9 E4    0.00712   0.000451  
#> 10 E5    0.000110  0.126     
#> 11 E6    0.0635    0.000163  
#> 12 E7    0.00873   0.438     
#> 13 E8    0.000131  0.00127   
#> 14 E9    0.000562  0.00541   

# }