Within-environment analysis of variance

Performs a within-environment analysis of variance in randomized complete block or alpha-lattice designs and returns values such as Mean Squares, p-values, coefficient of variation, heritability, and accuracy of selection.

anova_ind(.data, env, gen, rep, resp, block = NULL)

Arguments

.data
The dataset containing the columns related to Environments, Genotypes, replication/block and response variable(s).

env
The name of the column that contains the levels of the environments. The analysis of variance is computed for each level of this factor.

gen
The name of the column that contains the levels of the genotypes.

rep
The name of the column that contains the levels of the replications/blocks.

resp
The response variable(s). To analyze multiple variables in a single procedure a vector of variables may be used. For example resp = c(var1, var2, var3).

block
Defaults to NULL. In this case, a randomized complete block design is considered. If block is informed, then a resolvable alpha-lattice design (Patterson and Williams, 1976) is employed. All effects, except the error, are assumed to be fixed.

Value

A list where each element is the result for one variable containing:

individual: A tidy tbl_df with the results of the individual analysis of variance with the following column names:

For analysis in alpha-lattice designs: ENV: The environment code; MEAN: The grand mean; MSG, MSCR, MSIB_R: The mean squares for genotype, replicates and incomplete blocks within replicates, respectively. FCG, FCR, FCIB_R: The F-calculated for genotype, replicates and incomplete blocks within replicates, respectively.PFG, PFCR, PFIB_R: The P-values for genotype, replicates and incomplete blocks within replicates, respectively. MSE: The mean square error. CV: coefficient of variation. h2: broad-sense heritability. AS: accuracy of selection (square root of h2)

For analysis in randomized complete block design: MSG, MSB: The mean squares for genotype and blocks, respectively. FCG, FCB: The F-calculated for genotype and blocks, respectively. PFG, PFB: The P-values for genotype and blocks, respectively. MSE: The mean square error. CV: coefficient of variation. h2: broad-sense heritability. AS: accuracy of selection (square root of h2)

MSRatio The ratio between the higher and lower residual mean square.

References

Patterson, H.D., and E.R. Williams. 1976. A new class of resolvable incomplete block designs. Biometrika 63:83-92.

Author

Tiago Olivoto tiagoolivoto@gmail.com

Examples

# \donttest{
library(metan)
# ANOVA for all variables in data
ind_an <- anova_ind(data_ge,
                    env = ENV,
                    gen = GEN,
                    rep = REP,
                    resp = everything())
# mean for each environment
get_model_data(ind_an)

#> Class of the model: anova_ind

#> Variable extracted: MEAN

#> # A tibble: 14 x 3
#>    ENV      GY    HM
#>    <chr> <dbl> <dbl>
#>  1 E1     2.52  47.4
#>  2 E10    2.18  44.3
#>  3 E11    1.37  54.2
#>  4 E12    1.61  49.6
#>  5 E13    2.91  46.6
#>  6 E14    1.78  41.0
#>  7 E2     3.18  44.1
#>  8 E3     4.06  52.9
#>  9 E4     3.68  50  
#> 10 E5     3.91  52.2
#> 11 E6     2.66  45.9
#> 12 E7     1.99  48.5
#> 13 E8     2.54  45.2
#> 14 E9     3.06  51.3

# P-value for genotype effect
get_model_data(ind_an, "PFG")

#> Class of the model: anova_ind

#> Variable extracted: PFG

#> # A tibble: 14 x 3
#>    ENV          GY         HM
#>    <chr>     <dbl>      <dbl>
#>  1 E1    0.0594    0.0293    
#>  2 E10   0.0000110 0.00000302
#>  3 E11   0.244     0.107     
#>  4 E12   0.000647  0.108     
#>  5 E13   0.000210  0.0000180 
#>  6 E14   0.153     0.00393   
#>  7 E2    0.376     0.00402   
#>  8 E3    0.123     0.0269    
#>  9 E4    0.00712   0.000451  
#> 10 E5    0.000110  0.126     
#> 11 E6    0.0635    0.000163  
#> 12 E7    0.00873   0.438     
#> 13 E8    0.000131  0.00127   
#> 14 E9    0.000562  0.00541   

# }