Collinearity Diagnostics

Perform a (multi)collinearity diagnostic of a correlation matrix of predictor variables using several indicators, as shown by Olivoto et al. (2017).

colindiag(.data, ..., by = NULL, n = NULL, verbose = TRUE)

Arguments

.data
The data to be analyzed. It must be a symmetric correlation matrix, or a data frame, possible with grouped data passed from group_by().

...
Variables to use in the correlation. If ... is null then all the numeric variables from .data are used. It must be a single variable name or a comma-separated list of unquoted variables names.

by
One variable (factor) to compute the function by. It is a shortcut to group_by(). To compute the statistics by more than one grouping variable use that function.

n
If a correlation matrix is provided, then n is the number of objects used to compute the correlation coefficients.

verbose
If verbose = TRUE then some results are shown in the console.

Value

If .data is a grouped data passed from group_by() then the results will be returned into a list-column of data frames.

cormat A symmetric Pearson's coefficient correlation matrix between the variables

corlist A hypothesis testing for each of the correlation coefficients

evalevet The eigenvalues with associated eigenvectors of the correlation matrix

VIF The Variance Inflation Factors, being the diagonal elements of the inverse of the correlation matrix.

CN The Condition Number of the correlation matrix, given by the ratio between the largest and smallest eigenvalue.

det The determinant of the correlation matrix.

largest_corr The largest correlation (in absolute value) observed.

smallest_corr The smallest correlation (in absolute value) observed.

weight_var The variables with largest eigenvector (largest weight) in the eigenvalue of smallest value, sorted in decreasing order.

References

Olivoto, T., V.Q. Souza, M. Nardino, I.R. Carvalho, M. Ferrari, A.J. Pelegrin, V.J. Szareski, and D. Schmidt. 2017. Multicollinearity in path analysis: a simple method to reduce its effects. Agron. J. 109:131-142. doi:10.2134/agronj2016.04.0196. doi:10.2134/agronj2016.04.0196

Olivoto, T., M. Nardino, I.I.R. Carvalho, D.N. Follmann, M. Ferrari, A.J. de Pelegrin, V.J. Szareski, A.C. de Oliveira, B.O. Caron, and V.Q. de Souza. 2017. Optimal sample size and data arrangement method in estimating correlation matrices with lesser collinearity: A statistical focus in maize breeding. African J. Agric. Res. 12:93-103. doi:10.5897/AJAR2016.11799.

Examples

# \donttest{
# Using the correlation matrix
library(metan)

cor_iris <- cor(iris[,1:4])
n <- nrow(iris)

col_diag <- colindiag(cor_iris, n = n)

#> The multicollinearity in the matrix should be investigated.
#> NC = 140.889
#> Largest VIF = 31.2614977749215
#> Matrix determinant: 0.0081096 
#> Largest correlation: Petal.Length x Petal.Width = 0.963 
#> Smallest correlation: Sepal.Length x Sepal.Width = -0.118 
#> Number of VIFs > 10: 2 
#> Number of correlations with r >= |0.8|: 3 
#> Variables with largest weight in the last eigenvalues: 
#> Petal.Length > Petal.Width > Sepal.Length > Sepal.Width 
#> 

# Using a data frame
col_diag_gen <- data_ge2 %>%
                group_by(GEN) %>%
                colindiag()

#> Weak multicollinearity in the matrix
#> NC = -19711624781580504
#> Matrix determinant: 0 
#> Largest correlation: CDED x PERK = -0.964 
#> Smallest correlation: PERK x TKW = 0.007 
#> Number of VIFs > 10: 13 
#> Number of correlations with r >= |0.8|: 8 
#> Variables with largest weight in the last eigenvalues: 
#> EH > CDED > CL > PH > NKE > EL > EP > KW > PERK > NKR > CD > CW > NR > ED > TKW 
#> 
#> Weak multicollinearity in the matrix
#> NC = -49819780311919456
#> Matrix determinant: 0 
#> Largest correlation: ED x CL = 0.94 
#> Smallest correlation: EP x PERK = 0.001 
#> Number of VIFs > 10: 7 
#> Number of correlations with r >= |0.8|: 13 
#> Variables with largest weight in the last eigenvalues: 
#> EL > ED > CD > PH > CL > PERK > EP > CW > NKE > TKW > NR > KW > EH > NKR > CDED 
#> 
#> Weak multicollinearity in the matrix
#> NC = -17948962129147206
#> Matrix determinant: 0 
#> Largest correlation: PH x EH = 0.953 
#> Smallest correlation: EP x NKR = 0.023 
#> Number of VIFs > 10: 15 
#> Number of correlations with r >= |0.8|: 5 
#> Variables with largest weight in the last eigenvalues: 
#> NKE > CW > PH > EL > NR > KW > TKW > EP > CL > ED > NKR > EH > CDED > PERK > CD 
#> 
#> Weak multicollinearity in the matrix
#> NC = -3960161670914828
#> Matrix determinant: 0 
#> Largest correlation: EL x CD = 0.974 
#> Smallest correlation: EL x PERK = -0.007 
#> Number of VIFs > 10: 12 
#> Number of correlations with r >= |0.8|: 13 
#> Variables with largest weight in the last eigenvalues: 
#> EH > PH > EP > CW > PERK > NKE > TKW > KW > CD > CDED > ED > EL > NKR > CL > NR 
#> 
#> Weak multicollinearity in the matrix
#> NC = -32183322396391852
#> Matrix determinant: 0 
#> Largest correlation: CL x CW = 0.942 
#> Smallest correlation: EH x NR = -0.013 
#> Number of VIFs > 10: 14 
#> Number of correlations with r >= |0.8|: 7 
#> Variables with largest weight in the last eigenvalues: 
#> PERK > KW > CL > NR > CW > NKR > CD > EP > NKE > ED > EL > EH > CDED > TKW > PH 
#> 
#> Weak multicollinearity in the matrix
#> NC = -97256891576118080
#> Matrix determinant: 0 
#> Largest correlation: PH x EH = 0.897 
#> Smallest correlation: EP x NKR = 0.022 
#> Number of VIFs > 10: 14 
#> Number of correlations with r >= |0.8|: 10 
#> Variables with largest weight in the last eigenvalues: 
#> CW > PERK > CL > CDED > ED > EH > KW > CD > EL > NKE > EP > NR > TKW > NKR > PH 
#> 
#> Weak multicollinearity in the matrix
#> NC = -89458046759948240
#> Matrix determinant: 0 
#> Largest correlation: PH x EH = 0.909 
#> Smallest correlation: CL x CD = 0.002 
#> Number of VIFs > 10: 13 
#> Number of correlations with r >= |0.8|: 7 
#> Variables with largest weight in the last eigenvalues: 
#> NKE > TKW > CW > PERK > CD > EH > CDED > KW > EL > EP > CL > ED > NR > NKR > PH 
#> 
#> Weak multicollinearity in the matrix
#> NC = -6743433537859199
#> Matrix determinant: 0 
#> Largest correlation: PH x EH = 0.979 
#> Smallest correlation: NKR x TKW = -0.011 
#> Number of VIFs > 10: 10 
#> Number of correlations with r >= |0.8|: 8 
#> Variables with largest weight in the last eigenvalues: 
#> EH > PH > EP > CL > CW > NKE > NKR > CDED > ED > NR > PERK > CD > TKW > EL > KW 
#> 
#> Weak multicollinearity in the matrix
#> NC = -8468241968641812
#> Matrix determinant: 0 
#> Largest correlation: PH x EH = 0.95 
#> Smallest correlation: NKR x PERK = 0.013 
#> Number of VIFs > 10: 14 
#> Number of correlations with r >= |0.8|: 4 
#> Variables with largest weight in the last eigenvalues: 
#> PH > EH > KW > EP > PERK > CW > TKW > CL > EL > CDED > ED > CD > NR > NKR > NKE 
#> 
#> Weak multicollinearity in the matrix
#> NC = -9544826917169058
#> Matrix determinant: 0 
#> Largest correlation: EL x CD = 0.961 
#> Smallest correlation: NR x TKW = -0.001 
#> Number of VIFs > 10: 15 
#> Number of correlations with r >= |0.8|: 23 
#> Variables with largest weight in the last eigenvalues: 
#> EH > PH > EP > CD > NKE > KW > ED > PERK > CL > CW > EL > TKW > CDED > NKR > NR 
#> 
#> Weak multicollinearity in the matrix
#> NC = -59612499100712640
#> Matrix determinant: 0 
#> Largest correlation: PH x EH = 0.968 
#> Smallest correlation: EL x NR = 0.009 
#> Number of VIFs > 10: 8 
#> Number of correlations with r >= |0.8|: 12 
#> Variables with largest weight in the last eigenvalues: 
#> CW > CD > ED > EL > PERK > EP > EH > KW > NKR > NKE > TKW > CL > NR > CDED > PH 
#> 
#> Weak multicollinearity in the matrix
#> NC = -10560192828246224
#> Matrix determinant: 0 
#> Largest correlation: PH x EH = 0.959 
#> Smallest correlation: NR x PERK = -0.043 
#> Number of VIFs > 10: 13 
#> Number of correlations with r >= |0.8|: 22 
#> Variables with largest weight in the last eigenvalues: 
#> EH > PH > CW > EP > KW > CL > NR > TKW > EL > ED > CDED > NKE > CD > PERK > NKR 
#> 
#> Weak multicollinearity in the matrix
#> NC = -24787230690432888
#> Matrix determinant: 0 
#> Largest correlation: CW x KW = 0.966 
#> Smallest correlation: KW x CDED = -0.016 
#> Number of VIFs > 10: 12 
#> Number of correlations with r >= |0.8|: 18 
#> Variables with largest weight in the last eigenvalues: 
#> PH > EH > CL > ED > KW > NKE > NKR > EP > CW > CDED > EL > CD > NR > PERK > TKW 
#> 

# Diagnostic by levels of a factor
# For variables with "N" in variable name
col_diag_gen <- data_ge2 %>%
                group_by(GEN) %>%
                colindiag(contains("N"))

#> Weak multicollinearity in the matrix
#> NC = 29.925
#> Matrix determinant: 0.1126898 
#> Largest correlation: NKR x NKE = 0.885 
#> Smallest correlation: NR x NKR = 0.379 
#> Number of VIFs > 10: 0 
#> Number of correlations with r >= |0.8|: 1 
#> Variables with largest weight in the last eigenvalues: 
#> NKE > NKR > NR 
#> 
#> Weak multicollinearity in the matrix
#> NC = 5.825
#> Matrix determinant: 0.4941951 
#> Largest correlation: NR x NKR = -0.59 
#> Smallest correlation: NR x NKE = 0.182 
#> Number of VIFs > 10: 0 
#> Number of correlations with r >= |0.8|: 0 
#> Variables with largest weight in the last eigenvalues: 
#> NKR > NR > NKE 
#> 
#> Weak multicollinearity in the matrix
#> NC = 35.813
#> Matrix determinant: 0.1034821 
#> Largest correlation: NKR x NKE = 0.683 
#> Smallest correlation: NR x NKE = 0.18 
#> Number of VIFs > 10: 0 
#> Number of correlations with r >= |0.8|: 0 
#> Variables with largest weight in the last eigenvalues: 
#> NKR > NKE > NR 
#> 
#> Weak multicollinearity in the matrix
#> NC = 16.453
#> Matrix determinant: 0.2049328 
#> Largest correlation: NR x NKE = 0.819 
#> Smallest correlation: NR x NKR = 0.277 
#> Number of VIFs > 10: 0 
#> Number of correlations with r >= |0.8|: 1 
#> Variables with largest weight in the last eigenvalues: 
#> NKE > NR > NKR 
#> 
#> The multicollinearity in the matrix should be investigated.
#> NC = 483.199
#> Largest VIF = 113.960479386239
#> Matrix determinant: 0.0050956 
#> Largest correlation: NR x NKE = 0.914 
#> Smallest correlation: NR x NKR = 0.648 
#> Number of VIFs > 10: 3 
#> Number of correlations with r >= |0.8|: 2 
#> Variables with largest weight in the last eigenvalues: 
#> NKE > NR > NKR 
#> 
#> Weak multicollinearity in the matrix
#> NC = 91.002
#> Matrix determinant: 0.0407597 
#> Largest correlation: NR x NKE = 0.886 
#> Smallest correlation: NR x NKR = 0.269 
#> Number of VIFs > 10: 2 
#> Number of correlations with r >= |0.8|: 1 
#> Variables with largest weight in the last eigenvalues: 
#> NKE > NR > NKR 
#> 
#> Weak multicollinearity in the matrix
#> NC = 43.111
#> Matrix determinant: 0.0757936 
#> Largest correlation: NKR x NKE = 0.803 
#> Smallest correlation: NR x NKR = 0.412 
#> Number of VIFs > 10: 1 
#> Number of correlations with r >= |0.8|: 1 
#> Variables with largest weight in the last eigenvalues: 
#> NKE > NKR > NR 
#> 
#> Weak multicollinearity in the matrix
#> NC = 60.331
#> Matrix determinant: 0.0624676 
#> Largest correlation: NR x NKE = 0.719 
#> Smallest correlation: NR x NKR = -0.212 
#> Number of VIFs > 10: 2 
#> Number of correlations with r >= |0.8|: 0 
#> Variables with largest weight in the last eigenvalues: 
#> NKE > NR > NKR 
#> 
#> Weak multicollinearity in the matrix
#> NC = 19.403
#> Matrix determinant: 0.1695391 
#> Largest correlation: NKR x NKE = 0.523 
#> Smallest correlation: NR x NKR = -0.375 
#> Number of VIFs > 10: 0 
#> Number of correlations with r >= |0.8|: 0 
#> Variables with largest weight in the last eigenvalues: 
#> NKE > NKR > NR 
#> 
#> Weak multicollinearity in the matrix
#> NC = 97.814
#> Matrix determinant: 0.0370602 
#> Largest correlation: NKR x NKE = 0.929 
#> Smallest correlation: NR x NKR = 0.336 
#> Number of VIFs > 10: 2 
#> Number of correlations with r >= |0.8|: 1 
#> Variables with largest weight in the last eigenvalues: 
#> NKE > NKR > NR 
#> 
#> Weak multicollinearity in the matrix
#> NC = 27.928
#> Matrix determinant: 0.1331732 
#> Largest correlation: NKR x NKE = 0.802 
#> Smallest correlation: NR x NKR = 0.062 
#> Number of VIFs > 10: 0 
#> Number of correlations with r >= |0.8|: 1 
#> Variables with largest weight in the last eigenvalues: 
#> NKE > NKR > NR 
#> 
#> Weak multicollinearity in the matrix
#> NC = 69.823
#> Matrix determinant: 0.0492349 
#> Largest correlation: NKR x NKE = 0.873 
#> Smallest correlation: NR x NKR = 0.379 
#> Number of VIFs > 10: 1 
#> Number of correlations with r >= |0.8|: 1 
#> Variables with largest weight in the last eigenvalues: 
#> NKE > NKR > NR 
#> 
#> Weak multicollinearity in the matrix
#> NC = 26.565
#> Matrix determinant: 0.1385938 
#> Largest correlation: NKR x NKE = 0.717 
#> Smallest correlation: NR x NKR = -0.119 
#> Number of VIFs > 10: 0 
#> Number of correlations with r >= |0.8|: 0 
#> Variables with largest weight in the last eigenvalues: 
#> NKE > NKR > NR 
#> 

# }