Collinearity Diagnostics

Perform a (multi)collinearity diagnostic of a correlation matrix of predictor variables based several indicators, as shown by Olivoto et al. (2017).

colindiag(.data, n = NULL, verbose = TRUE)

Arguments

.data
The data to be analyzed. Must be a symmetric correlation matrix or, a dataframe containing the predictor variables, or an object of class split_factors.

n
If a correlation matrix is provided, then n is the number of objects used to compute the correlation coefficients.

verbose
If verbose = TRUE then some results are shown in the console.

Value

The following values are returned. Please, note that if a grouping variable is used, then the results are returned into a list.

cormat

A symmetric Pearson's coefficient correlation matrix between the variables

corlist

A hypothesis testing for each of the correlation coefficients

evalevet

The eigenvalues with associated eigenvectors of the correlation matrix

VIF

The Variance Inflation Factors, being the diagonal elements of the inverse of the correlation matrix.

CN

The Condition Number of the correlation matrix, given by the ratio between the largest and smallest eigenvalue.

det

The determinant of the correlation matrix.

largest_corr

The largest correlation (in absolute value) observed.

smallest_corr

The smallest correlation (in absolute value) observed.

weight_var

The variables with largest eigenvector (largest weight) in the eigenvalue of smallest value, sorted in decreasing order.

References

Olivoto, T., V.Q. Souza, M. Nardino, I.R. Carvalho, M. Ferrari, A.J. Pelegrin, V.J. Szareski, and D. Schmidt. 2017. Multicollinearity in path analysis: a simple method to reduce its effects. Agron. J. 109:131-142. doi:10.2134/agronj2016.04.0196. doi:10.2134/agronj2016.04.0196

Examples

# Using the correlation matrix
library(metan)
cor_iris = cor(iris[,1:4])
n = nrow(iris)
colindiag(cor_iris, n = n, verbose = FALSE)

# Using the pipe operator %>%
library(dplyr)
cor_iris %>% colindiag(n = n, verbose = FALSE)

# Diagnostic by species, storing into an object
col_diag_spec = iris %>%
                split_factors(Species) %>%
                colindiag()

#> The columns Species where deleted. Use 'keep_factors = TRUE' to keep this columns in the grouped data. 

#> 
#> ----------------------------------------------------------------------------
#> Level: setosa 
#> ----------------------------------------------------------------------------
#> Weak multicollinearity in the matrix
#> NC = 8.186
#> Matrix determinant: 0.3533595 
#> Largest correlation: Sepal.Length x Sepal.Width = 0.743 
#> Smallest correlation: Sepal.Width x Petal.Length = 0.178 
#> Number of correlations with r >= |0.8|: 0 
#> Variables with largest weight in the last eigenvalues: 
#> Sepal.Length > Sepal.Width > Petal.Length > Petal.Width 
#> 
#> ----------------------------------------------------------------------------
#> Level: versicolor 
#> ----------------------------------------------------------------------------
#> Weak multicollinearity in the matrix
#> NC = 22.104
#> Matrix determinant: 0.0835942 
#> Largest correlation: Petal.Length x Petal.Width = 0.787 
#> Smallest correlation: Sepal.Length x Sepal.Width = 0.526 
#> Number of correlations with r >= |0.8|: 0 
#> Variables with largest weight in the last eigenvalues: 
#> Petal.Length > Petal.Width > Sepal.Length > Sepal.Width 
#> 
#> ----------------------------------------------------------------------------
#> Level: virginica 
#> ----------------------------------------------------------------------------
#> Weak multicollinearity in the matrix
#> NC = 19.136
#> Matrix determinant: 0.1373902 
#> Largest correlation: Sepal.Length x Petal.Length = 0.864 
#> Smallest correlation: Sepal.Length x Petal.Width = 0.281 
#> Number of correlations with r >= |0.8|: 1 
#> Variables with largest weight in the last eigenvalues: 
#> Sepal.Length > Petal.Length > Sepal.Width > Petal.Width