Collinearity Diagnostics

Perform a (multi)collinearity diagnostic of a correlation matrix of predictor variables using several indicators, as shown by Olivoto et al. (2017).

colindiag(.data, ..., n = NULL, verbose = TRUE)

Arguments

.data
The data to be analyzed. Must be a symmetric correlation matrix or, a dataframe containing the predictor variables, or an object of class split_factors.

...
Variables to use in the correlation. If ... is null then all the numeric variables from .data are used. It must be a single variable name or a comma-separated list of unquoted variables names.

n
If a correlation matrix is provided, then n is the number of objects used to compute the correlation coefficients.

verbose
If verbose = TRUE then some results are shown in the console.

Value

The following values are returned. Please, note that if a grouping variable is used, then the results are returned into a list.

cormat A symmetric Pearson's coefficient correlation matrix between the variables

corlist A hypothesis testing for each of the correlation coefficients

evalevet The eigenvalues with associated eigenvectors of the correlation matrix

VIF The Variance Inflation Factors, being the diagonal elements of the inverse of the correlation matrix.

CN The Condition Number of the correlation matrix, given by the ratio between the largest and smallest eigenvalue.

det The determinant of the correlation matrix.

largest_corr The largest correlation (in absolute value) observed.

smallest_corr The smallest correlation (in absolute value) observed.

weight_var The variables with largest eigenvector (largest weight) in the eigenvalue of smallest value, sorted in decreasing order.

References

Olivoto, T., V.Q. Souza, M. Nardino, I.R. Carvalho, M. Ferrari, A.J. Pelegrin, V.J. Szareski, and D. Schmidt. 2017. Multicollinearity in path analysis: a simple method to reduce its effects. Agron. J. 109:131-142. doi:10.2134/agronj2016.04.0196. doi:10.2134/agronj2016.04.0196

Olivoto, T., M. Nardino, I.I.R. Carvalho, D.N. Follmann, M. Ferrari, A.J. de Pelegrin, V.J. Szareski, A.C. de Oliveira, B.O. Caron, and V.Q. de Souza. 2017. Optimal sample size and data arrangement method in estimating correlation matrices with lesser collinearity: A statistical focus in maize breeding. African J. Agric. Res. 12:93-103. doi:10.5897/AJAR2016.11799.

Examples

# Using the correlation matrix
library(metan)

cor_iris <- cor(iris[,1:4])
n <- nrow(iris)

col_diag <- colindiag(cor_iris, n = n)

#> The multicollinearity in the matrix should be investigated.
#> NC = 140.889
#> Largest VIF = 31.2614977749215
#> Matrix determinant: 0.0081096 
#> Largest correlation: Petal.Length x Petal.Width = 0.963 
#> Smallest correlation: Sepal.Length x Sepal.Width = -0.118 
#> Number of VIFs > 10: 2 
#> Number of correlations with r >= |0.8|: 3 
#> Variables with largest weight in the last eigenvalues: 
#> Petal.Length > Petal.Width > Sepal.Length > Sepal.Width 

# Using a data frame
col_diag_gen <- data_ge2 %>%
                split_factors(GEN) %>%
                colindiag()

#> Warning: Use 'keep_factors = TRUE' to keep the columns ENV GEN REP in the grouped data. 

#> 
#> ----------------------------------------------------------------------------
#> Level: H1 
#> ----------------------------------------------------------------------------
#> Weak multicollinearity in the matrix
#> NC = -19711624781580504
#> Matrix determinant: 0 
#> Largest correlation: CDED x PERK = -0.964 
#> Smallest correlation: PERK x TKW = 0.007 
#> Number of VIFs > 10: 13 
#> Number of correlations with r >= |0.8|: 8 
#> Variables with largest weight in the last eigenvalues: 
#> EH > CDED > CL > PH > NKE > EL > EP > KW > PERK > NKR > CD > CW > NR > ED > TKW 
#> 
#> ----------------------------------------------------------------------------
#> Level: H10 
#> ----------------------------------------------------------------------------
#> Weak multicollinearity in the matrix
#> NC = -49819780311919456
#> Matrix determinant: 0 
#> Largest correlation: ED x CL = 0.94 
#> Smallest correlation: EP x PERK = 0.001 
#> Number of VIFs > 10: 7 
#> Number of correlations with r >= |0.8|: 13 
#> Variables with largest weight in the last eigenvalues: 
#> EL > ED > CD > PH > CL > PERK > EP > CW > NKE > TKW > NR > KW > EH > NKR > CDED 
#> 
#> ----------------------------------------------------------------------------
#> Level: H11 
#> ----------------------------------------------------------------------------
#> Weak multicollinearity in the matrix
#> NC = -17948962129147206
#> Matrix determinant: 0 
#> Largest correlation: PH x EH = 0.953 
#> Smallest correlation: EP x NKR = 0.023 
#> Number of VIFs > 10: 15 
#> Number of correlations with r >= |0.8|: 5 
#> Variables with largest weight in the last eigenvalues: 
#> NKE > CW > PH > EL > NR > KW > TKW > EP > CL > ED > NKR > EH > CDED > PERK > CD 
#> 
#> ----------------------------------------------------------------------------
#> Level: H12 
#> ----------------------------------------------------------------------------
#> Weak multicollinearity in the matrix
#> NC = -3960161670914828
#> Matrix determinant: 0 
#> Largest correlation: EL x CD = 0.974 
#> Smallest correlation: EL x PERK = -0.007 
#> Number of VIFs > 10: 12 
#> Number of correlations with r >= |0.8|: 13 
#> Variables with largest weight in the last eigenvalues: 
#> EH > PH > EP > CW > PERK > NKE > TKW > KW > CD > CDED > ED > EL > NKR > CL > NR 
#> 
#> ----------------------------------------------------------------------------
#> Level: H13 
#> ----------------------------------------------------------------------------
#> Weak multicollinearity in the matrix
#> NC = -32183322396391852
#> Matrix determinant: 0 
#> Largest correlation: CL x CW = 0.942 
#> Smallest correlation: EH x NR = -0.013 
#> Number of VIFs > 10: 14 
#> Number of correlations with r >= |0.8|: 7 
#> Variables with largest weight in the last eigenvalues: 
#> PERK > KW > CL > NR > CW > NKR > CD > EP > NKE > ED > EL > EH > CDED > TKW > PH 
#> 
#> ----------------------------------------------------------------------------
#> Level: H2 
#> ----------------------------------------------------------------------------
#> Weak multicollinearity in the matrix
#> NC = -97256891576118080
#> Matrix determinant: 0 
#> Largest correlation: PH x EH = 0.897 
#> Smallest correlation: EP x NKR = 0.022 
#> Number of VIFs > 10: 14 
#> Number of correlations with r >= |0.8|: 10 
#> Variables with largest weight in the last eigenvalues: 
#> CW > PERK > CL > CDED > ED > EH > KW > CD > EL > NKE > EP > NR > TKW > NKR > PH 
#> 
#> ----------------------------------------------------------------------------
#> Level: H3 
#> ----------------------------------------------------------------------------
#> Weak multicollinearity in the matrix
#> NC = -89458046759948240
#> Matrix determinant: 0 
#> Largest correlation: PH x EH = 0.909 
#> Smallest correlation: CL x CD = 0.002 
#> Number of VIFs > 10: 13 
#> Number of correlations with r >= |0.8|: 7 
#> Variables with largest weight in the last eigenvalues: 
#> NKE > TKW > CW > PERK > CD > EH > CDED > KW > EL > EP > CL > ED > NR > NKR > PH 
#> 
#> ----------------------------------------------------------------------------
#> Level: H4 
#> ----------------------------------------------------------------------------
#> Weak multicollinearity in the matrix
#> NC = -6743433537859199
#> Matrix determinant: 0 
#> Largest correlation: PH x EH = 0.979 
#> Smallest correlation: NKR x TKW = -0.011 
#> Number of VIFs > 10: 10 
#> Number of correlations with r >= |0.8|: 8 
#> Variables with largest weight in the last eigenvalues: 
#> EH > PH > EP > CL > CW > NKE > NKR > CDED > ED > NR > PERK > CD > TKW > EL > KW 
#> 
#> ----------------------------------------------------------------------------
#> Level: H5 
#> ----------------------------------------------------------------------------
#> Weak multicollinearity in the matrix
#> NC = -8468241968641812
#> Matrix determinant: 0 
#> Largest correlation: PH x EH = 0.95 
#> Smallest correlation: NKR x PERK = 0.013 
#> Number of VIFs > 10: 14 
#> Number of correlations with r >= |0.8|: 4 
#> Variables with largest weight in the last eigenvalues: 
#> PH > EH > KW > EP > PERK > CW > TKW > CL > EL > CDED > ED > CD > NR > NKR > NKE 
#> 
#> ----------------------------------------------------------------------------
#> Level: H6 
#> ----------------------------------------------------------------------------
#> Weak multicollinearity in the matrix
#> NC = -9544826917169058
#> Matrix determinant: 0 
#> Largest correlation: EL x CD = 0.961 
#> Smallest correlation: NR x TKW = -0.001 
#> Number of VIFs > 10: 15 
#> Number of correlations with r >= |0.8|: 23 
#> Variables with largest weight in the last eigenvalues: 
#> EH > PH > EP > CD > NKE > KW > ED > PERK > CL > CW > EL > TKW > CDED > NKR > NR 
#> 
#> ----------------------------------------------------------------------------
#> Level: H7 
#> ----------------------------------------------------------------------------
#> Weak multicollinearity in the matrix
#> NC = -59612499100712640
#> Matrix determinant: 0 
#> Largest correlation: PH x EH = 0.968 
#> Smallest correlation: EL x NR = 0.009 
#> Number of VIFs > 10: 8 
#> Number of correlations with r >= |0.8|: 12 
#> Variables with largest weight in the last eigenvalues: 
#> CW > CD > ED > EL > PERK > EP > EH > KW > NKR > NKE > TKW > CL > NR > CDED > PH 
#> 
#> ----------------------------------------------------------------------------
#> Level: H8 
#> ----------------------------------------------------------------------------
#> Weak multicollinearity in the matrix
#> NC = -10560192828246224
#> Matrix determinant: 0 
#> Largest correlation: PH x EH = 0.959 
#> Smallest correlation: NR x PERK = -0.043 
#> Number of VIFs > 10: 13 
#> Number of correlations with r >= |0.8|: 22 
#> Variables with largest weight in the last eigenvalues: 
#> EH > PH > CW > EP > KW > CL > NR > TKW > EL > ED > CDED > NKE > CD > PERK > NKR 
#> 
#> ----------------------------------------------------------------------------
#> Level: H9 
#> ----------------------------------------------------------------------------
#> Weak multicollinearity in the matrix
#> NC = -24787230690432888
#> Matrix determinant: 0 
#> Largest correlation: CW x KW = 0.966 
#> Smallest correlation: KW x CDED = -0.016 
#> Number of VIFs > 10: 12 
#> Number of correlations with r >= |0.8|: 18 
#> Variables with largest weight in the last eigenvalues: 
#> PH > EH > CL > ED > KW > NKE > NKR > EP > CW > CDED > EL > CD > NR > PERK > TKW 

# Diagnostic by levels of a factor selecting desired variables
col_diag_gen <- data_ge2 %>%
                split_factors(GEN) %>%
                colindiag(EH, PH, CD, CL)

#> Warning: Use 'keep_factors = TRUE' to keep the columns ENV GEN REP in the grouped data. 

#> 
#> ----------------------------------------------------------------------------
#> Level: H1 
#> ----------------------------------------------------------------------------
#> Weak multicollinearity in the matrix
#> NC = 57.499
#> Matrix determinant: 0.0606324 
#> Largest correlation: EH x PH = 0.959 
#> Smallest correlation: EH x CD = -0.034 
#> Number of VIFs > 10: 2 
#> Number of correlations with r >= |0.8|: 1 
#> Variables with largest weight in the last eigenvalues: 
#> EH > PH > CL > CD 
#> 
#> ----------------------------------------------------------------------------
#> Level: H10 
#> ----------------------------------------------------------------------------
#> Weak multicollinearity in the matrix
#> NC = 82.334
#> Matrix determinant: 0.0195163 
#> Largest correlation: EH x PH = 0.924 
#> Smallest correlation: EH x CD = 0.207 
#> Number of VIFs > 10: 2 
#> Number of correlations with r >= |0.8|: 2 
#> Variables with largest weight in the last eigenvalues: 
#> EH > PH > CL > CD 
#> 
#> ----------------------------------------------------------------------------
#> Level: H11 
#> ----------------------------------------------------------------------------
#> Weak multicollinearity in the matrix
#> NC = 59.979
#> Matrix determinant: 0.0450618 
#> Largest correlation: EH x PH = 0.953 
#> Smallest correlation: CD x CL = 0.25 
#> Number of VIFs > 10: 2 
#> Number of correlations with r >= |0.8|: 1 
#> Variables with largest weight in the last eigenvalues: 
#> PH > EH > CL > CD 
#> 
#> ----------------------------------------------------------------------------
#> Level: H12 
#> ----------------------------------------------------------------------------
#> Weak multicollinearity in the matrix
#> NC = 77.032
#> Matrix determinant: 0.0280568 
#> Largest correlation: EH x PH = 0.956 
#> Smallest correlation: PH x CL = 0.345 
#> Number of VIFs > 10: 2 
#> Number of correlations with r >= |0.8|: 1 
#> Variables with largest weight in the last eigenvalues: 
#> EH > PH > CD > CL 
#> 
#> ----------------------------------------------------------------------------
#> Level: H13 
#> ----------------------------------------------------------------------------
#> Weak multicollinearity in the matrix
#> NC = 21.274
#> Matrix determinant: 0.1324678 
#> Largest correlation: EH x PH = 0.883 
#> Smallest correlation: EH x CL = -0.126 
#> Number of VIFs > 10: 0 
#> Number of correlations with r >= |0.8|: 1 
#> Variables with largest weight in the last eigenvalues: 
#> EH > PH > CL > CD 
#> 
#> ----------------------------------------------------------------------------
#> Level: H2 
#> ----------------------------------------------------------------------------
#> Weak multicollinearity in the matrix
#> NC = 66.704
#> Matrix determinant: 0.0288844 
#> Largest correlation: EH x PH = 0.897 
#> Smallest correlation: PH x CD = -0.051 
#> Number of VIFs > 10: 2 
#> Number of correlations with r >= |0.8|: 1 
#> Variables with largest weight in the last eigenvalues: 
#> EH > PH > CL > CD 
#> 
#> ----------------------------------------------------------------------------
#> Level: H3 
#> ----------------------------------------------------------------------------
#> Weak multicollinearity in the matrix
#> NC = 28.096
#> Matrix determinant: 0.1314002 
#> Largest correlation: EH x PH = 0.909 
#> Smallest correlation: CD x CL = 0.002 
#> Number of VIFs > 10: 0 
#> Number of correlations with r >= |0.8|: 1 
#> Variables with largest weight in the last eigenvalues: 
#> EH > PH > CL > CD 
#> 
#> ----------------------------------------------------------------------------
#> Level: H4 
#> ----------------------------------------------------------------------------
#> The multicollinearity in the matrix should be investigated.
#> NC = 103.689
#> Largest VIF = 24.865096078672
#> Matrix determinant: 0.0356218 
#> Largest correlation: EH x PH = 0.979 
#> Smallest correlation: CD x CL = -0.145 
#> Number of VIFs > 10: 2 
#> Number of correlations with r >= |0.8|: 1 
#> Variables with largest weight in the last eigenvalues: 
#> EH > PH > CL > CD 
#> 
#> ----------------------------------------------------------------------------
#> Level: H5 
#> ----------------------------------------------------------------------------
#> The multicollinearity in the matrix should be investigated.
#> NC = 118.615
#> Largest VIF = 27.8406872982704
#> Matrix determinant: 0.022266 
#> Largest correlation: EH x PH = 0.95 
#> Smallest correlation: EH x CD = 0.053 
#> Number of VIFs > 10: 2 
#> Number of correlations with r >= |0.8|: 1 
#> Variables with largest weight in the last eigenvalues: 
#> PH > EH > CD > CL 
#> 
#> ----------------------------------------------------------------------------
#> Level: H6 
#> ----------------------------------------------------------------------------
#> Weak multicollinearity in the matrix
#> NC = 58.953
#> Matrix determinant: 0.0154075 
#> Largest correlation: EH x PH = 0.931 
#> Smallest correlation: EH x CL = -0.277 
#> Number of VIFs > 10: 1 
#> Number of correlations with r >= |0.8|: 3 
#> Variables with largest weight in the last eigenvalues: 
#> PH > EH > CL > CD 
#> 
#> ----------------------------------------------------------------------------
#> Level: H7 
#> ----------------------------------------------------------------------------
#> The multicollinearity in the matrix should be investigated.
#> NC = 102.265
#> Largest VIF = 20.5580961120105
#> Matrix determinant: 0.0310978 
#> Largest correlation: EH x PH = 0.968 
#> Smallest correlation: CD x CL = 0.296 
#> Number of VIFs > 10: 2 
#> Number of correlations with r >= |0.8|: 1 
#> Variables with largest weight in the last eigenvalues: 
#> EH > PH > CL > CD 
#> 
#> ----------------------------------------------------------------------------
#> Level: H8 
#> ----------------------------------------------------------------------------
#> The multicollinearity in the matrix should be investigated.
#> NC = 168.221
#> Largest VIF = 28.9258799972658
#> Matrix determinant: 0.0064985 
#> Largest correlation: EH x PH = 0.959 
#> Smallest correlation: PH x CD = 0.66 
#> Number of VIFs > 10: 2 
#> Number of correlations with r >= |0.8|: 1 
#> Variables with largest weight in the last eigenvalues: 
#> EH > PH > CD > CL 
#> 
#> ----------------------------------------------------------------------------
#> Level: H9 
#> ----------------------------------------------------------------------------
#> Weak multicollinearity in the matrix
#> NC = 75.267
#> Matrix determinant: 0.0106539 
#> Largest correlation: EH x PH = 0.933 
#> Smallest correlation: EH x CD = 0.325 
#> Number of VIFs > 10: 1 
#> Number of correlations with r >= |0.8|: 3 
#> Variables with largest weight in the last eigenvalues: 
#> PH > CL > EH > CD