Confidence interval for correlation coefficient

Computes the half-width confidence interval for correlation coefficient using the nonparametric method proposed by Olivoto et al. (2018).

corr_ci(.data = NA, ..., r = NULL, n = NULL, verbose = TRUE)

Arguments

.data
A dataset containing variables only or a symmetric correlation matrix.

...
Variables to compute the confidence interval. If not informed, all the numeric variables from .data are used.

r
If data is not available, provide the value for correlation coefficient.

n
The sample size if data is a correlation matrix or if r is informed.

verbose
If verbose = TRUE then some results are shown in the console.

Value

A tibble containing the values of the correlation, confidence interval, upper and lower limits for all combination of variables.

Details

The half-width confidence interval is computed according to the following equation: $$CI_w = 0.45304^r \times 2.25152 \times n^{-0.50089}$$

where \(n\) is the sample size and r is the correlation coefficient.

References

Olivoto, T., A.D.C. Lucio, V.Q. Souza, M. Nardino, M.I. Diel, B.G. Sari, D.. K. Krysczun, D. Meira, and C. Meier. 2018. Confidence interval width for Pearson's correlation coefficient: a Gaussian-independent estimator based on sample size and strength of association. Agron. J. 110:1-8. 10.2134/agronj2016.04.0196

Examples

library(metan)

CI1 <- corr_ci(data_ge)

#> The factors ENV GEN REP where excluded to perform the analysis. If you want to perform an analysis for each level of a factor, use the function 'split_factors() before.' 

CI2 <- data_ge2 %>%
split_factors(ENV) %>%
corr_ci(CD, TKW, NKE)

#> Warning: Use 'keep_factors = TRUE' to keep the columns ENV GEN REP in the grouped data.