Cross-validation procedure

Cross-validation for estimation of AMMI models

cv_ammi(.data, env, gen, rep, resp, nboot = 100, nrepval, naxis,
  design = "RCBD", verbose = TRUE)

Arguments

.data
The dataset containing the columns related to Environments, Genotypes, replication/block and response variable(s).

env
The name of the column that contains the levels of the environments.

gen
The name of the column that contains the levels of the genotypes.

rep
The name of the column that contains the levels of the replications/blocks.

resp
The response variable.

nboot
The number of resamples to be used in the cross-validation. Defaults to 100.

nrepval
The number of replicates (r) from total number of replicates (R) to be used in the modeling dataset. Only one replicate is used as validating data each step, so, Nrepval must be equal R-1

naxis
The number of axis to be considered for estimation of GE effects.

design
The experimental desig to be considered. Default is RCBD (Randomized complete Block Design). For Completely Randomized Designs inform design = 'CRD'.

verbose
A logical argument to define if a progress bar is shown. Default is TRUE.

Value

An object of class cv_ammi with the following items:

RMSPD

A vector with nboot-estimates of the root mean squared prediction difference between predicted and validating data.

RMSPDmean

The mean of RMSPDmean estimates.

Estimated

A data frame that contain the values (predicted, observed, validation) of the last loop.

Modeling

The dataset used as modeling data in the last loop.

Testing

The dataset used as testing data in the last loop.

Details

For each iteration, the original dataset is split into two datasets: modeling and validating data. The dataset 'modeling' has all combinations (genotype x environment) with the number of replications informed in nrepval. The dataset 'validating' has one replication. The splitting of the dataset into modeling and validating data depends on the design informed. For Completely Randomized Block Design (default), completely blocks are selected within environments. The remained block serves validation data. If design = 'RCD' is informed, completely randomly samples are made for each genotype-by-environment combination. The estimated values (depending on NAXIS informed) are compared with the 'validating' data. the Root Means Square error is computed. At the end of boots, a list is returned with the following values.

See also

Examples

# \donttest{
library(metan)
model <- cv_ammi(data_ge,
                env = ENV,
                gen = GEN,
                rep = REP,
                resp = GY,
                nboot = 100,
                nrepval = 2,
                naxis = 2)

# Alternatively using the pipe operator %>%
model <- data_ge %>%
         cv_ammi(ENV, GEN, REP, GY, 100, 2, 2)
# }