Descriptive statistics

Compute the most used measures of central tendency, position, and dispersion.

desc_stat(.data = NULL, ..., values = NULL, stats = NULL,
  level = 0.95, digits = 4, verbose = TRUE)

Arguments

.data
The data to be analyzed. Must be a dataframe or an object of class split_factors.

...
A single variable name or a comma-separated list of unquoted variables names.

values
An alternative way to pass the data to the function. It must be a numeric vector.

stats
The descriptive statistics to show. Defaults to NULL (all statistics shown). It must be one of the 'AV.dev' (average deviation), 'CI.mean' (confidence interval for the mean), 'CV' (coefficient of variation), 'IQR' (interquartile range), 'Kurt' (kurtosis), 'max' (maximum value), 'mean' (arithmetic mean), 'median' (median), 'min' (minimum value), 'n' (the length of the data), 'norm.pval' (the p-value for the Shapiro-Wilk test), 'norm.stat' (the statistic for the Shapiro-Wilk test), 'Q2.5' (the percentile 2.5 'Q25' (the first quartile, Q1), 'Q75' (the third quartile, Q3), 'Q97.5' (the percentile 97.5 'SD.amo' (the sample standard deviation), 'SD.pop' (the population standard deviation), 'SE.mean' (the standard error of the mean), 'skew' (the skewness), 'var.amo' (the sample variance), 'var.pop' (the population variance). Use a comma-separated vector of names to select the statistics. For example, stats = c("media, mean, CV, n").

level
The confidence level to compute the confidence interval of mean. Defaults to 0.95.

digits
The number of significant figures.

verbose
Logical argument. If verbose = FALSE the code is run silently.

Value

A tibble with the statistics in the lines and variables in columns. If .data is an object of class split_factors, then the statistics will be shown for each level of the grouping variable in the function split_factors().

Examples

library(metan)
library(dplyr)
desc1 <- data_ge2 %>% desc_stat(TKW)

#> # A tibble: 22 x 2
#>    Statistic   TKW
#>    <chr>     <dbl>
#>  1 n         156  
#>  2 mean      339. 
#>  3 min       218. 
#>  4 Q2.5      233. 
#>  5 Q25       310. 
#>  6 median    342. 
#>  7 Q75       367. 
#>  8 Q97.5     423. 
#>  9 max       452. 
#> 10 IQR        57.8
#> # ... with 12 more rows

vect <- data_ge2 %>%
  select(TKW) %>%
  pull()

desc2 <- desc_stat(values = vect)

#> # A tibble: 22 x 2
#>    Statistic    ``
#>    <chr>     <dbl>
#>  1 n         156  
#>  2 mean      339. 
#>  3 min       218. 
#>  4 Q2.5      233. 
#>  5 Q25       310. 
#>  6 median    342. 
#>  7 Q75       367. 
#>  8 Q97.5     423. 
#>  9 max       452. 
#> 10 IQR        57.8
#> # ... with 12 more rows

desc3 <-
data_ge2 %>%
  split_factors(ENV) %>%
  desc_stat(EP, EL, PH, CL, CW, NR, NKR,
  stats = c('mean, SE.mean, CV'),
  verbose = FALSE)

#> The columns ENV GEN REP where deleted. Use 'keep_factors = TRUE' to keep this columns in the grouped data.