Descriptive statistics

Compute the most used measures of central tendency, position, and dispersion.

desc_stat(.data = NULL, ..., values = NULL, stats = "main",
  hist = FALSE, level = 0.95, digits = 4, na.rm = FALSE,
  verbose = TRUE)

Arguments

.data
The data to be analyzed. Must be a dataframe or an object of class split_factors.

...
A single variable name or a comma-separated list of unquoted variables names.

values
An alternative way to pass the data to the function. It must be a numeric vector.

stats
The descriptive statistics to show. Defaults to "main" (main statistics). Set to "all" to compute all the statistics bellow or chose one (or more) of the following: 'AV.dev' (average deviation), 'CI.mean' (confidence interval for the mean), 'CV' (coefficient of variation), 'IQR' (interquartile range), 'Kurt' (kurtosis), 'mad' (median absolute deviation), 'max' (maximum value), 'mean' (arithmetic mean), 'median' (median), 'min' (minimum value), 'n' (the length of the data), 'norm.pval' (the p-value for the Shapiro-Wilk test), 'norm.stat' (the statistic for the Shapiro-Wilk test), 'Q2.5' (the percentile 2.5 'Q75' (the third quartile, Q3), 'Q97.5' (the percentile 97.5 standard deviation), 'SD.pop' (the population standard deviation), 'SE.mean' (the standard error of the mean), 'skew' (the skewness), sum (the sum of the values), sum.dev (the sum of the absolute deviations), sum.sq.dev (the sum of the squared deviations), valid.n (The size of sample with valid number (not NA), 'var.amo' (the sample variance), 'var.pop' (the population variance). Use a comma-separated vector of names to select the statistics. For example, stats = c("median, mean, CV, n").

hist
Logical argument defaults to FALSE. If hist = TRUE then a histogram is created for each selected variable.

level
The confidence level to compute the confidence interval of mean. Defaults to 0.95.

digits
The number of significant digits.

na.rm
Logical. Should missing values be removed?

verbose
Logical argument. If verbose = FALSE the code is run silently.

Value

A tibble with the statistics in the lines and variables in columns. If .data is an object of class split_factors, then the statistics will be shown for each level of the grouping variable in the function split_factors.

Details

In cases when the statistics are computed for more than two variables with data coming from the function split_factors the results are returned in a long format. Thus, use the function desc_wider to convert it into a wide format (levels of the factors in the rows and statistics in the columns).

Examples

library(metan)

desc_stat(data_ge2, TKW)

#> # A tibble: 7 x 2
#>   Statistic     TKW
#>   <chr>       <dbl>
#> 1 mean       339.  
#> 2 min        218.  
#> 3 median     342.  
#> 4 max        452.  
#> 5 var.amo   2215.  
#> 6 SE.mean      3.77
#> 7 CV          13.9 

# Compute all statistics
# Use a numeric vector as input data
vect <- data_ge2$TKW
desc_stat(values = vect, stats = "all")

#> # A tibble: 28 x 2
#>    Statistic value
#>    <chr>     <dbl>
#>  1 n          156 
#>  2 valid.n    156 
#>  3 mean       339.
#>  4 range      233.
#>  5 min        218.
#>  6 Q2.5       233.
#>  7 Q25        310.
#>  8 median     342.
#>  9 Q75        367.
#> 10 Q97.5      423.
#> # ... with 18 more rows

# Select specific statistics
desc_stat(values = c(12, 13, 19, 21, 8, NA, 23, NA),
          na.rm = TRUE,
          stats = c('mean, SE.mean, CV, n, valid.n'))

#> # A tibble: 5 x 2
#>   Statistic value
#>   <chr>     <dbl>
#> 1 n          8   
#> 2 valid.n    6   
#> 3 mean      16   
#> 4 SE.mean    2.39
#> 5 CV        36.7 

# Compute the statistics for each level of "ENV"
stats <-
data_ge2 %>%
  split_factors(GEN) %>%
  desc_stat(EP, EL, EH, ED, PH, CD,
  stats = c('mean, min, max, median, SE.mean, CI.mean, n, CV'),
  verbose = FALSE)

#> Warning: The columns ENV GEN REP where deleted. Use 'keep_factors = TRUE' to keep this columns in the grouped data. 

# To get a 'wide' format with the statistics of the variable EP above.
desc_wider(stats, PH)

#> # A tibble: 13 x 9
#>    GEN   CI.mean    CV   max  mean median   min     n SE.mean
#>    <chr>   <dbl> <dbl> <dbl> <dbl>  <dbl> <dbl> <dbl>   <dbl>
#>  1 H1      0.195 11.7   3.00  2.62   2.70  2.11    12  0.0885
#>  2 H10     0.217 14.8   2.83  2.31   2.27  1.79    12  0.0987
#>  3 H11     0.218 14.4   2.77  2.39   2.42  1.71    12  0.0991
#>  4 H12     0.191 12.4   2.79  2.44   2.52  1.95    12  0.0870
#>  5 H13     0.151  9.39  2.93  2.54   2.55  2.19    12  0.0688
#>  6 H2      0.217 13.1   3.03  2.60   2.62  2.05    12  0.0988
#>  7 H3      0.254 15.4   3.04  2.59   2.68  1.96    12  0.115 
#>  8 H4      0.226 13.8   3.02  2.58   2.67  1.99    12  0.103 
#>  9 H5      0.200 12.3   2.90  2.57   2.73  2.01    12  0.0910
#> 10 H6      0.190 11.7   2.94  2.56   2.60  2.09    12  0.0863
#> 11 H7      0.191 12.5   2.87  2.40   2.21  2.07    12  0.0868
#> 12 H8      0.210 14.2   2.76  2.33   2.31  1.81    12  0.0953
#> 13 H9      0.233 15.5   3.00  2.36   2.25  1.98    12  0.106