Multi-trait selection index

Multitrait index based on factor analysis and ideotype-design proposed by Rocha et al. (2018).

fai_blup(.data, DI, UI, SI = NULL, mineval = 1, verbose = TRUE)

Arguments

.data
An object of class waasb or a two-way table with genotypes in the rows and traits in columns. In the last case the row names must contain the genotypes names.

DI, UI
A vector of the same length of .data to construct the desirable (DI) and undesirable (UI) ideotypes. For each element of the vector, allowed values are 'max', 'min', 'mean', or a numeric value. Use a comma-separated vector of text. For example, DI = c("max, max, min, min").

SI
An integer (0-100). The selection intensity in percentage of the total number of genotypes.

mineval
The minimum value so that an eigenvector is retained in the factor analysis.

verbose
Logical value. If TRUE some results are shown in console.

Value

An object of class fai_blup with the following items:

data The data (BLUPS) used to compute the index.

FA The results of the factor analysis.

canonical.loadings The canonical loadings for each factor retained.

FAI A list with the FAI-BLUP index for each ideotype design.

selection.diferential A list with the selection differential for each ideotype design.

References

Rocha, J.R.A.S.C.R, J.C. Machado, and P.C.S. Carneiro. 2018. Multitrait index based on factor analysis and ideotype-design: proposal and application on elephant grass breeding for bioenergy. GCB Bioenergy 10:52-60. doi: doi:10.1111/gcbb.12443

Examples

# \donttest{
library(metan)

mod <- waasb(data_ge,
             env = ENV,
             gen = GEN,
             rep = REP,
             resp = c(GY, HM))

#> Model:  Y ~ ENV/REP + (1 | GEN) + (1 | GEN:ENV) 
#> ---------------------------------------------------------------------------
#> P-values for Likelihood Ratio Test of the analyzed traits
#> ---------------------------------------------------------------------------
#>     model       GY       HM
#>  COMPLETE       NA       NA
#>       GEN 1.11e-05 5.07e-03
#>   GEN:ENV 2.15e-11 2.27e-15
#> ---------------------------------------------------------------------------
#> All variables with significant (p < 0.05) genotype-vs-environment interaction

FAI <- fai_blup(mod,
                SI = 15,
                DI = c('max, max'),
                UI = c('min, min'))

#> 
#> -----------------------------------------------------------------------------------
#> Principal Component Analysis
#> -----------------------------------------------------------------------------------
#>     eigen.values cumulative.var
#> PC1    1.1046076       55.23038
#> PC2    0.8953924      100.00000
#> 
#> -----------------------------------------------------------------------------------
#> Factor Analysis
#> -----------------------------------------------------------------------------------
#>           FA1 comunalits
#> GY -0.7431714  0.5523038
#> HM  0.7431714  0.5523038
#> 
#> -----------------------------------------------------------------------------------
#> Comunalit Mean: 0.5523038 
#> 
#> -----------------------------------------------------------------------------------
#> FAI-BLUP index
#> -----------------------------------------------------------------------------------
#>     G4     G9     G6    G10     G5     G3     G2     G7     G1     G8 
#> 0.6084 0.5552 0.5315 0.5304 0.5237 0.4613 0.4608 0.3661 0.3170 0.0784 
#> 
#> -----------------------------------------------------------------------------------
#> Selection differential
#> -----------------------------------------------------------------------------------
#>    Factor        Xo        Xs          SD    SDperc
#> GY      1  2.674242  2.594199 -0.08004274 -2.993100
#> HM      1 48.088286 48.005561 -0.08272485 -0.172027
#> 
#> -----------------------------------------------------------------------------------
#> Selected genotypes
#> G4 G9
#> -----------------------------------------------------------------------------------

# Or using the pipe operator %>%

FAI <- data_ge2 %>%
       waasb(ENV, GEN, REP, c(KW, NKE, PH, EH)) %>%
       fai_blup(DI = c('max, max, max, min'),
                UI = c('min, min, min, max'),
                SI = 15)

#> Model:  Y ~ ENV/REP + (1 | GEN) + (1 | GEN:ENV) 
#> ---------------------------------------------------------------------------
#> P-values for Likelihood Ratio Test of the analyzed traits
#> ---------------------------------------------------------------------------
#>     model       KW     NKE       PH       EH
#>  COMPLETE       NA      NA       NA       NA
#>       GEN 6.21e-01 1.00000 9.39e-01 1.00e+00
#>   GEN:ENV 4.92e-07 0.00101 1.09e-13 8.12e-12
#> ---------------------------------------------------------------------------
#> All variables with significant (p < 0.05) genotype-vs-environment interaction
#> 
#> -----------------------------------------------------------------------------------
#> Principal Component Analysis
#> -----------------------------------------------------------------------------------
#>     eigen.values cumulative.var
#> PC1    3.1441417       78.60354
#> PC2    0.6604119       95.11384
#> PC3    0.1385777       98.57828
#> PC4    0.0568688      100.00000
#> 
#> -----------------------------------------------------------------------------------
#> Factor Analysis
#> -----------------------------------------------------------------------------------
#>           FA1 comunalits
#> KW  0.9474560  0.8976729
#> NKE 0.7429933  0.5520391
#> PH  0.9429971  0.8892436
#> EH  0.8973216  0.8051860
#> 
#> -----------------------------------------------------------------------------------
#> Comunalit Mean: 0.7860354 
#> 
#> -----------------------------------------------------------------------------------
#> FAI-BLUP index
#> -----------------------------------------------------------------------------------
#>    H13     H6     H2     H5     H1     H4     H3     H8     H9    H10     H7 
#> 0.9807 0.8843 0.8413 0.8409 0.8023 0.7972 0.6222 0.2623 0.2354 0.1810 0.1163 
#>    H12    H11 
#> 0.1014 0.0780 
#> 
#> -----------------------------------------------------------------------------------
#> Selection differential
#> -----------------------------------------------------------------------------------
#>     Factor         Xo         Xs           SD       SDperc
#> KW       1 172.938867 175.182040 2.243173e+00 1.297090e+00
#> NKE      1 511.643590 511.643590 2.819513e-07 5.510697e-08
#> PH       1   2.484813   2.487051 2.238030e-03 9.006835e-02
#> EH       1   1.343326   1.343326 3.234133e-10 2.407557e-08
#> 
#> -----------------------------------------------------------------------------------
#> Selected genotypes
#> H13 H6
#> -----------------------------------------------------------------------------------

# }