Genotype analysis by mixed-effect models

Analysis of genotypes in single experiments using mixed-effect models with estimation of genetic parameters.

gamem(.data, gen, rep, resp, block = NULL, prob = 0.05,
  verbose = TRUE)

Arguments

.data
The dataset containing the columns related to, Genotypes, replication/block and response variable(s).

gen
The name of the column that contains the levels of the genotypes, that will be treated as random effect.

rep
The name of the column that contains the levels of the replications (assumed to be fixed).

resp
The response variable(s). To analyze multiple variables in a single procedure a vector of variables may be used. For example resp = c(var1, var2, var3).

block
Defaults to NULL. In this case, a randomized complete block design is considered. If block is informed, then an alpha-lattice design is employed considering block as random to make use of inter-block information, whereas the complete replicate effect is always taken as fixed, as no inter-replicate information was to be recovered (Mohring et al., 2015).

prob
The probability for estimating confidence interval for BLUP's prediction.

verbose
Logical argument. If verbose = FALSE the code are run silently.

Value

An object of class gamem, which is a list with the following items for each element (variable):

fixed: Test for fixed effects.

random: Variance components for random effects.

LRT: The Likelihood Ratio Test for the random effects.

blupGEN: The estimated BLUPS for genotypes

Details: A tibble with the following data: Ngen, the number of genotypes; OVmean, the grand mean; Min, the minimum observed (returning the genotype and replication/block); Max the maximum observed, MinGEN the winner genotype, MaxGEN, the loser genotype.

ESTIMATES: A tibble with the values for the genotypic variance, block-within-replicate variance (if an alpha-lattice design is used by informing the block in block), the residual variance and their respective contribution to the phenotypic variance; broad-sence heritability, heritability on the entry-mean basis, genotypic coefficient of variation residual coefficient of variation and ratio between genotypic and residual coefficient of variation.

residuals: The residuals of the model.

Details

gamem analyses data from a one-way genotype testing experiment. By default, a randomized complete block design is used according to the following model: $$Y_{ij} = m + g_i + r_j + e_{ij}$$ where \(Y_{ij}\) is the response variable of the ith genotype in the jth block; m is the grand mean (fixed); \(g_i\) is the effect of the ith genotype (assumed to be random); \(r_j\) is the effect of the jth replicate (assumed to be fixed); and \(e_{ij}\) is the random error.

When block is informed, then a resolvable alpha design is implemented, according to the following model:

$$Y_{ijk} = m + g_i + r_j + b_{jk} + e_{ijk}$$ where where \(y_{ijk}\) is the response variable of the ith genotype in the kth block of the jth replicate; m is the intercept, \(t_i\) is the effect for the ith genotype \(r_j\) is the effect of the jth replicate, \(b_{jk}\) is the effect of the kth incomplete block of the jth replicate, and \(e_{ijk}\) is the plot error effect corresponding to \(y_{ijk}\).

References

Mohring, J., E. Williams, and H.-P. Piepho. 2015. Inter-block information: to recover or not to recover it? TAG. Theor. Appl. Genet. 128:1541-54. doi:10.1007/s00122-015-2530-0

See also

Examples

# \donttest{
library(metan)

# fitting the model considering an RCBD
# Genotype as random effects

rcbd <- gamem(data_g,
             gen = GEN,
             rep = REP,
             resp = c(PH, ED, EL, CL, CW, KW, NR, TKW, NKE))

#> Evaluating variable PH 11.1 % 
#> Evaluating variable ED 22.2 % 
#> Evaluating variable EL 33.3 % 
#> Evaluating variable CL 44.4 % 
#> Evaluating variable CW 55.6 % 
#> Evaluating variable KW 66.7 % 
#> Evaluating variable NR 77.8 % 
#> Evaluating variable TKW 88.9 % 
#> Evaluating variable NKE 100 % 
#> ------------------------------------------------------------
#> Variables with nonsignificant genotype effect
#> PH EL 
#> ------------------------------------------------------------
#> Done!

# Likelihood ratio test for random effects
## Statistic
get_model_data(rcbd, "lrt")

#> # A tibble: 2 x 10
#>   model       PH    ED      EL    CL    CW    KW    NR   TKW   NKE
#>   <chr>    <dbl> <dbl>   <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1 Complete NA     NA   NA       NA    NA   NA    NA    NA    NA   
#> 2 Genotype  3.81  17.6  0.0735  22.4  19.1  5.00  7.67  6.72  6.72

## P-value
get_model_data(rcbd, "pval_lrt")

#> # A tibble: 2 x 10
#>   model      PH       ED     EL       CL       CW      KW       NR      TKW
#>   <chr>   <dbl>    <dbl>  <dbl>    <dbl>    <dbl>   <dbl>    <dbl>    <dbl>
#> 1 Comp~ NA      NA       NA     NA       NA       NA      NA       NA      
#> 2 Geno~  0.0510  2.73e-5  0.786  2.25e-6  1.24e-5  0.0253  0.00560  0.00955
#> # ... with 1 more variable: NKE <dbl>

# Variance components
get_model_data(rcbd, "vcomp")

#> # A tibble: 2 x 10
#>   Group        PH    ED     EL    CL    CW    KW    NR   TKW   NKE
#>   <chr>     <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1 GEN      0.0171  5.37 0.0472  4.27 18.5   181.  1.18  841. 1982.
#> 2 Residual 0.0328  2.43 0.984   1.41  7.54  280.  1.27 1018. 2399.

# Genetic parameters
get_model_data(rcbd, "genpar")

#> # A tibble: 11 x 10
#>    Parameters      PH     ED      EL     CL     CW      KW     NR     TKW
#>    <chr>        <dbl>  <dbl>   <dbl>  <dbl>  <dbl>   <dbl>  <dbl>   <dbl>
#>  1 Gen_var     0.0171  5.37   0.0472  4.27  18.5   181.     1.18  8.41e+2
#>  2 Gen (%)    34.3    68.8    4.58   75.1   71.0    39.2   48.2   4.52e+1
#>  3 Res_var     0.0328  2.43   0.984   1.41   7.54  280.     1.27  1.02e+3
#>  4 Res (%)    65.7    31.2   95.4    24.9   29.0    60.8   51.8   5.48e+1
#>  5 Phen_var    0.0498  7.80   1.03    5.68  26.0   461.     2.45  1.86e+3
#>  6 H2          0.343   0.688  0.0458  0.751  0.710   0.392  0.482 4.52e-1
#>  7 H2mg        0.610   0.869  0.126   0.901  0.880   0.659  0.736 7.12e-1
#>  8 Accuracy    0.781   0.932  0.355   0.949  0.938   0.812  0.858 8.44e-1
#>  9 CVg         6.03    4.84   1.48    7.26  20.7     9.16   6.88  9.13e+0
#> 10 CVr         8.35    3.26   6.76    4.18  13.2    11.4    7.14  1.00e+1
#> 11 CV ratio    0.722   1.49   0.219   1.74   1.56    0.803  0.964 9.09e-1
#> # ... with 1 more variable: NKE <dbl>

# BLUPs for genotypes
get_model_data(rcbd)

#> setting the argument 'what' to 'blupg'.

#> # A tibble: 13 x 10
#>    gen      PH    ED    EL    CL    CW    KW    NR   TKW   NKE
#>    <chr> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1 H1     2.19  50.2  14.7  30.7  26.1  153.  15.8  354.  437.
#>  2 H10    2.09  44.4  14.6  25.1  13.3  129.  15.5  268.  483.
#>  3 H11    2.13  47.2  14.6  26.6  16.6  143.  16.1  297.  482.
#>  4 H12    2.33  47.8  14.6  26.1  18.2  148.  16.1  293.  503.
#>  5 H13    2.43  50.3  14.8  27.4  21.5  170.  17.8  319.  547.
#>  6 H2     2.16  50.3  14.8  30.0  21.9  156.  15.7  338.  467.
#>  7 H3     2.09  47.2  14.7  28.6  18.7  142.  15.1  331.  434.
#>  8 H4     2.10  46.1  14.8  27.8  19.6  145.  14.4  310.  472.
#>  9 H5     2.12  49.8  14.7  30.1  26.0  156.  17.0  309.  517.
#> 10 H6     2.16  49.7  14.5  31.6  22.8  140.  16.2  325.  433.
#> 11 H7     2.17  48.7  14.7  30.0  26.5  153.  15.6  346.  448.
#> 12 H8     2.13  46.3  14.7  29.0  21.9  143.  14.9  339.  425.
#> 13 H9     2.08  44.4  14.7  27.0  17.1  131.  14.9  301.  434.

# fitting the model considering an alpha-lattice design
# Genotype and block-within-replicate as random effects
# Note that block effect was now informed.

alpha <- gamem(data_alpha,
               gen = GEN,
               rep = REP,
               block = BLOCK,
               resp = YIELD)

#> Done!

# Use the function  get_model_data() to easely extract the model values.
get_model_data(alpha, "genpar")

#> # A tibble: 13 x 2
#>    Parameters      YIELD
#>    <chr>           <dbl>
#>  1 Gen_var        0.143 
#>  2 Gen (%)       48.5   
#>  3 rep:block_var  0.0702
#>  4 rep:block (%) 23.8   
#>  5 Res_var        0.0816
#>  6 Res (%)       27.7   
#>  7 Phen_var       0.295 
#>  8 H2             0.485 
#>  9 H2mg           0.798 
#> 10 Accuracy       0.893 
#> 11 CVg            8.44  
#> 12 CVr            6.38  
#> 13 CV ratio       1.32  

# }