Statistics for genotype-vs-environment interaction

Computes (i) within-environment analysis of variance, GEI effect, GEI means, and genotype plus GEI effects; (ii) parametric statistics including AMMI-based indexes, Annicchiarico's genotypic confidence index (1992), Ecovalence (Wricke, 1965), regression-based stability (Eberhart and Russell., 1966), Shukla's stability variance parameter (1972); and (iii) nonparametric statistics including Fox's stability function (Fox et al. 1990), superiority index (Lin and Binns, 1988), Huehn's stability statistics (Huehn, 1979), and Thennarasu (1995) statistics.

ge_stats(.data, env, gen, rep, resp, verbose = TRUE, prob = 0.05)

Arguments

.data
The dataset containing the columns related to Environments, Genotypes, replication/block and response variable(s).

env
The name of the column that contains the levels of the environments.

gen
The name of the column that contains the levels of the genotypes.

rep
The name of the column that contains the levels of the replications/blocks.

resp
The response variable(s). To analyze multiple variables in a single procedure use, for example, resp = c(var1, var2, var3).

verbose
Logical argument. If verbose = FALSE the code will run silently.

prob
The probability error assumed.

Value

An object of class ge_stats which is a list with one data frame for each variable containing the computed indexes.

Details

The function computes the statistics and ranks for the following stability indexes. "Y" (Response variable), "Var" (Genotype's variance), "Shukla" (Shukla's variance, calling Shukla internally), "Wi_g", "Wi_f", "Wi_u" (Annichiarrico's genotypic confidence index for all, favorable and unfavorable environments, respectively, calling Annicchiarico internally ), "Ecoval" (Wricke's ecovalence, ecovalence internally), "Sij" (Deviations from the joint-regression analysis) and "R2" (R-squared from the joint-regression analysis, calling ge_reg internally), "ASV" (AMMI-stability value), "SIPC" (sum of the absolute values of the IPCA scores), "EV" (Average of the squared eigenvector values), "ZA" (Absolute values of the relative contributions of the IPCAs to the interaction), and "WAAS" (Weighted Average of Absolute Scores), by calling AMMI_indexes internally; "HMGV" (Harmonic mean of the genotypic value), "RPGV" (Relative performance of the genotypic values), "HMRPGV" (Harmonic mean of the relative performance of the genotypic values), by calling Resende_indexes internally; "Pi_a", "Pi_f", "Pi_u" (Superiority indexes for all, favorable and unfavorable environments, respectively, calling superiority internally), "Gai" (Geometric adaptability index, calling gai internally), "S1" (mean of the absolute rank differences of a genotype over the n environments), "S2" (variance among the ranks over the k environments), "S3" (sum of the absolute deviations), "S6" (relative sum of squares of rank for each genotype), by calling Huehn internally; and "N1", "N2", "N3", "N4" (Thennarasu"s statistics, calling Thennarasu internally ).

References

Annicchiarico, P. 1992. Cultivar adaptation and recommendation from alfalfa trials in Northern Italy. Journal of Genetic \& Breeding, 46:269-278

Eberhart, S.A., and W.A. Russell. 1966. Stability parameters for comparing Varieties. Crop Sci. 6:36-40. doi:10.2135/cropsci1966.0011183X000600010011x

Fox, P.N., B. Skovmand, B.K. Thompson, H.J. Braun, and R. Cormier. 1990. Yield and adaptation of hexaploid spring triticale. Euphytica 47:57-64. doi:10.1007/BF00040364.

Huehn, V.M. 1979. Beitrage zur erfassung der phanotypischen stabilitat. EDV Med. Biol. 10:112.

Kang, M.S., and H.N. Pham. 1991. Simultaneous Selection for High Yielding and Stable Crop Genotypes. Agron. J. 83:161. doi:10.2134/agronj1991.00021962008300010037x.

Lin, C.S., and M.R. Binns. 1988. A superiority measure of cultivar performance for cultivar x location data. Can. J. Plant Sci. 68:193-198. doi:10.4141/cjps88-018

Olivoto, T., A.D.C. L\'ucio, J.A.G. da silva, V.S. Marchioro, V.Q. de Souza, and E. Jost. 2019a. Mean performance and stability in multi-environment trials I: Combining features of AMMI and BLUP techniques. Agron. J. doi:10.2134/agronj2019.03.0220

Shahbazi, E. 2019. Genotype selection and stability analysis for seed yield of Nigella sativa using parametric and non-parametric statistics. Sci. Hortic. (Amsterdam). 253:172-179. doi:10.1016/j.scienta.2019.04.047.

Shukla, G.K. 1972. Some statistical aspects of partitioning genotype-environmental components of variability. Heredity. 29:238-245. doi:10.1038/hdy.1972.87.

Thennarasu, K. 1995. On certain nonparametric procedures for studying genotype x environment interactions and yield stability. Ph.D. thesis. P.J. School, IARI, New Delhi, India.

Wricke, G. 1965. Zur berechnung der okovalenz bei sommerweizen und hafer. Z. Pflanzenzuchtg 52:127-138.

Examples

# \donttest{
library(metan)

model <- ge_stats(data_ge, ENV, GEN, REP, GY)
get_model_data(model, "stats")

#> # A tibble: 10 x 32
#>    var   gen       Y   Var Shukla  Wi_g  Wi_f  Wi_u Ecoval   bij      Sij    R2
#>    <chr> <chr> <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl>  <dbl> <dbl>    <dbl> <dbl>
#>  1 GY    G1     2.60 10.9  0.0280  84.4  89.2  81.1  1.22  1.06  -0.00142 0.966
#>  2 GY    G10    2.47 14.2  0.244   59.2  64.6  54.4  7.96  1.12   0.177   0.823
#>  3 GY    G2     2.74 11.3  0.0861  82.8  95.3  75.6  3.03  1.05   0.0497  0.913
#>  4 GY    G3     2.96 10.1  0.0121 104.   99.7 107.   0.725 1.03  -0.0128  0.977
#>  5 GY    G4     2.64  8.93 0.0640  85.9  79.5  91.9  2.34  0.937  0.0298  0.917
#>  6 GY    G5     2.54  7.82 0.0480  82.7  82.2  82.4  1.84  0.887  0.00902 0.937
#>  7 GY    G6     2.53  7.34 0.0468  83.0  83.7  81.8  1.81  0.861  0.00304 0.942
#>  8 GY    G7     2.74  7.33 0.122   83.9  77.6  93.4  4.16  0.819  0.0579  0.852
#>  9 GY    G8     3.00 10.8  0.0712  98.8  90.5 107.   2.57  1.03   0.0382  0.922
#> 10 GY    G9     2.51 14.7  0.167   68.8  68.9  70.3  5.56  1.19   0.0938  0.897
#> # … with 20 more variables: ASV <dbl>, SIPC <dbl>, EV <dbl>, ZA <dbl>,
#> #   WAAS <dbl>, HMGV <dbl>, RPGV <dbl>, HMRPGV <dbl>, Pi_a <dbl>, Pi_f <dbl>,
#> #   Pi_u <dbl>, Gai <dbl>, S1 <dbl>, S2 <dbl>, S3 <dbl>, S6 <dbl>, N1 <dbl>,
#> #   N2 <dbl>, N3 <dbl>, N4 <dbl>

# }