Easily get data from a model

Easily get data from a model of class waas or waasb (Olivoto et al., 2019a, 2019b)

get_model_data(x, what = "Y", type = "GEN")

Arguments

x
An object created with the functions waas or waasb.

what
What should be captured from the model. See more in Details section.

type
Chose if the statistics must be show by genotype (type = "GEN", default) or environment (type = "ENV"), when possible.

Value

A tibble showing the values of the variable chosen in argument what.

Details

The next options are valid in the argument what for both waas and waasb objects.

"Y" for raw means (Default) .

"PCn" for the scores of the nth Interaction Principal Component Axis (IPCA) . where * n is the desired IPCA (e.g. "PC3") .

"WAASB" The weighted average of the absolute scores .

"PctResp" The rescaled values of the response variable .

"PctWAASB" The rescaled values of the WAASB .

"wResp" The weight for the response variable .

"wWAASB" The weight for the stability .

"OrResp" The ranking regarding the response variable .

"OrWAASB" The ranking regarding the WAASB .

"OrPC1" The ranking regarding the first principal component axix .

"WAASBY" The superiority index WAASBY .

"OrWAASBY" The ranking regarding the superiority index .

The next options are only allowed if the object is of class waasb.

"blupg" For genotype's predicted mean.

"blupge" for genotype-vs-environment's predicted mean.

"genpar" Genetic parameters.

"lrt" The statistic for the likelihood-ratio test for random effects.

"pval_lrt" The p-values for the likelihood-ratio test.

"vcomp" The variance components for random effects.

References

Olivoto, T., A.D.C. L'ucio, J.A.G. da silva, V.S. Marchioro, V.Q. de Souza, and E. Jost. 2019a. Mean performance and stability in multi-environment trials I: Combining features of AMMI and BLUP techniques. Agron. J. doi:10.2134/agronj2019.03.0220

Olivoto, T., A.D.C. L'ucio, J.A.G. da silva, B.G. Sari, and M.I. Diel. 2019b. Mean performance and stability in multi-environment trials II: Selection based on multiple traits. Agron. J. (in press)

Examples

library(metan)
# Fit an AMMI model for 5 variables.
AMMI <- waas(data_ge2, ENV, GEN, REP,
             resp = c(PH, ED, TKW, NKR))

#> Evaluating variable PH 25 % 
#> Evaluating variable ED 50 % 
#> Evaluating variable TKW 75 % 
#> Evaluating variable NKR 100 % 
#> Done!

# Getting the weighted average of absolute scores
get_model_data(AMMI, what = "WAASB")

#> # A tibble: 13 x 5
#>    gen      PH    ED   TKW   NKR
#>    <fct> <dbl> <dbl> <dbl> <dbl>
#>  1 H1    0.318 0.667 2.72  0.929
#>  2 H10   0.230 0.973 2.15  0.506
#>  3 H11   0.201 0.649 1.26  0.836
#>  4 H12   0.364 0.315 0.558 0.228
#>  5 H13   0.363 0.838 0.514 0.946
#>  6 H2    0.342 1.08  4.41  0.404
#>  7 H3    0.374 0.486 4.10  0.252
#>  8 H4    0.294 0.378 3.07  0.281
#>  9 H5    0.168 0.567 0.738 0.611
#> 10 H6    0.270 0.409 1.64  1.60 
#> 11 H7    0.228 0.384 3.44  0.518
#> 12 H8    0.315 0.653 4.91  0.941
#> 13 H9    0.146 0.907 5.50  0.888

# And the rank for the WAASB inde.
get_model_data(AMMI, what = "OrWAASB")

#> # A tibble: 13 x 5
#>    gen      PH    ED   TKW   NKR
#>    <fct> <dbl> <dbl> <dbl> <dbl>
#>  1 H1        9     9     7    10
#>  2 H10       5    12     6     5
#>  3 H11       3     7     4     8
#>  4 H12      12     1     2     1
#>  5 H13      11    10     1    12
#>  6 H2       10    13    11     4
#>  7 H3       13     5    10     2
#>  8 H4        7     2     8     3
#>  9 H5        2     6     3     7
#> 10 H6        6     4     5    13
#> 11 H7        4     3     9     6
#> 12 H8        8     8    12    11
#> 13 H9        1    11    13     9

# Fitting a mixed-effect model
blup <- waasb(data_ge2, ENV, GEN, REP,
              resp = c(PH, ED, TKW, NKR))

#> Evaluating variable PH 25 % 
#> Evaluating variable ED 50 % 
#> Evaluating variable TKW 75 % 
#> Evaluating variable NKR 100 % 
#> Done!

# Getting p-values for likelihood-ratio test
get_model_data(blup, what = "pval_lrt")

#> # A tibble: 3 x 5
#>   model           PH            ED       TKW      NKR
#>   <chr>        <dbl>         <dbl>     <dbl>    <dbl>
#> 1 Complete NA        NA            NA        NA      
#> 2 Genotype  9.39e- 1  0.299         1.00e+ 0  0.787  
#> 3 Gen:Env   1.09e-13  0.0000000169  4.21e-10  0.00404

# Getting the variance components
get_model_data(blup, what = "vcomp")

#> # A tibble: 3 x 5
#>   Group          PH    ED     TKW   NKR
#>   <chr>       <dbl> <dbl>   <dbl> <dbl>
#> 1 GEN      0.000456 0.556 8.27e-7 0.187
#> 2 GEN:ENV  0.0425   2.82  1.15e+3 2.96 
#> 3 Residual 0.0224   2.59  9.19e+2 7.85 

# Getting the genetic parameters
get_model_data(blup, what = "genpar")

#> # A tibble: 15 x 5
#>    Parameters                   PH      ED      TKW     NKR
#>    <chr>                     <dbl>   <dbl>    <dbl>   <dbl>
#>  1 GEI variance           0.0425    2.82   1.15e+ 3  2.96  
#>  2 GEI (%)               65.0      47.2    5.55e+ 1 26.9   
#>  3 Genotypic variance     0.000456  0.556  8.27e- 7  0.187 
#>  4 Gen (%)                0.698     9.32   4.00e- 8  1.70  
#>  5 Residual variance      0.0224    2.59   9.19e+ 2  7.85  
#>  6 Res (%)               34.3      43.4    4.45e+ 1 71.4   
#>  7 Phenotypic variance    0.0654    5.97   2.06e+ 3 11.0   
#>  8 Heritability           0.00698   0.0932 4.00e-10  0.0170
#>  9 GEIr2                  0.650     0.472  5.55e- 1  0.269 
#> 10 Heribatility of means  0.0352    0.376  2.28e- 9  0.118 
#> 11 Accuracy               0.188     0.614  4.77e- 5  0.344 
#> 12 rge                    0.655     0.521  5.55e- 1  0.274 
#> 13 CVg                    0.860     1.51   2.68e- 4  1.34  
#> 14 CVr                    6.03      3.25   8.95e+ 0  8.69  
#> 15 CV ratio               0.143     0.463  3.00e- 5  0.154