Get data from a model easily

Easily get data from some objects generated in the metan package such as the WAASB and WAASBY indexes (Olivoto et al., 2019a, 2019b) BLUPs, variance components, details of AMMI models and AMMI-based stability statistics.

get_model_data(x, what = NULL, type = "GEN")

Arguments

x
An object created with the functions AMMI_indexes, ecovalence, Fox, gai, gamem, ge_means, ge_reg, performs_ammi, Resende_indexes, Shukla, superiority, waas or waasb.

what
What should be captured from the model. See more in section Details.

type
Chose if the statistics must be show by genotype (type = "GEN", default) or environment (type = "ENV"), when possible.

Value

A tibble showing the values of the variable chosen in argument what.

Details

Bellow are listed the options allowed in the argument what depending on the class of the object

Objects of class AMMI_indexes:

"ASV" AMMI stability value.

"EV" Averages of the squared eigenvector values.

"SIPC" Sums of the absolute value of the IPCA scores.

"WAAS" Weighted average of absolute scores (default).

"ZA" Absolute value of the relative contribution of IPCAs to the interaction.

Objects of class Annicchiarico and Schmildt:

"Sem_rp" The standard error of the relative mean performance (Schmildt).

"Mean_rp" The relative performance of the mean.

"rank" The rank for genotypic confidence index.

"Wi" The genotypic confidence index.

Objects of class ecovalence:

"Ecoval" Ecovalence value (default).

"Ecov_perc" Ecovalence in percentage value.

"rank" Rank for ecovalence.

Objects of class ge_reg:

"deviations" The deviations from regression.

"RMSE" The Root Mean Square Error.

"R2" The r-square of the regression.

"slope" The sloop of the regression (default).

Objects of class ge_effects:

For objects of class ge_effects no argument what is required.

Objects of class ge_means:

"ge_means" Genotype-environment interaction means (default).

"env_means" Environment means.

"gen_means" Genotype means.

Objects of class Shukla:

"rMean" Rank for the mean.

"ShuklaVar" Shukla's stablity variance (default).

"rShukaVar" Rank for Shukla's stablity variance.

"ssiShukaVar" Simultaneous selection index.

Objects of class Fox:

"TOP" The proportion of locations at which the genotype occurred in the top third (default).

Objects of class gai:

"GAI" The geometric adaptability index (default).

"GAI_R" The rank for the GAI values.

Objects of class superiority:

"Pi_a" The superiority measure for all environments (default).

"R_a" The rank for Pi_a.

"Pi_f" The superiority measure for favorable environments.

"R_f" The rank for Pi_f.

"Pi_u" The superiority measure for unfavorable environments.

"R_u" The rank for Pi_u.

Objects of class Huehn:

"S1" Mean of the absolute rank differences of a genotype over the n environments (default).

"S2" variance among the ranks over the k environments.

"S3" Sum of the absolute deviations.

"S6" Relative sum of squares of rank for each genotype.

"S1_R", "S2_R", "S3_R", and "S6_R", the ranks for S1, S2, S3, and S6, respectively.

Objects of class Thennarasu:

"N1" First statistic (default).

"N2" Second statistic.

"N3" Third statistic.

"N4" Fourth statistic.

"N1_R", "N2_R", "N3_R", and "N4_R", The ranks for the statistics.

Objects of class performs_ammi:

"PC1", "PC2", ..., "PCn" The values for the nth interaction principal component axis.

"ipca_ss" Sum of square for each IPCA.

"ipca_ms" Mean square for each IPCA.

"ipca_fval" F value for each IPCA.

"ipca_pval" P-value for for each IPCA.

"ipca_expl" Explained sum of square for each IPCA (default).

"ipca_accum" Accumulated explained sum of square.

Objects of class waas, waas_means, and waasb:

"PC1", "PC2", ..., "PCn" The values for the nth interaction principal component axis.

"WAASB" The weighted average of the absolute scores (default for objects of class waas).

"PctResp" The rescaled values of the response variable.

"PctWAASB" The rescaled values of the WAASB.

"wResp" The weight for the response variable.

"wWAASB" The weight for the stability.

"OrResp" The ranking regarding the response variable.

"OrWAASB" The ranking regarding the WAASB.

"OrPC1" The ranking regarding the first principal component axix.

"WAASBY" The superiority index WAASBY.

"OrWAASBY" The ranking regarding the superiority index.

Objects of class waasb or gamem:

"blupg" For genotype's predicted mean.

"blupge" for genotype-vs-environment's predicted mean (only for objects of class waasb).

"genpar" Genetic parameters (default).

"lrt" The statistic for the likelihood-ratio test for random effects.

"pval_lrt" The p-values for the likelihood-ratio test.

"vcomp" The variance components for random effects.

Objects of class Res_ind

"HMGV" For harmonic mean of genotypic values.

"RPGV or RPGV_Y" For relative performance of genotypic values

"HMRPGV" For harmonic mean of relative performance of genotypic values

References

Annicchiarico, P. 1992. Cultivar adaptation and recommendation from alfalfa trials in Northern Italy. J. Genet. Breed. 46:269-278.

Dias, P.C., A. Xavier, M.D.V. de Resende, M.H.P. Barbosa, F.A. Biernaski, R.A. Estopa. 2018. Genetic evaluation of Pinus taeda clones from somatic embryogenesis and their genotype x environment interaction. Crop Breed. Appl. Biotechnol. 18:55-64. doi:10.1590/1984-70332018v18n1a8

Azevedo Peixoto, L. de, P.E. Teodoro, L.A. Silva, E.V. Rodrigues, B.G. Laviola, and L.L. Bhering. 2018. Jatropha half-sib family selection with high adaptability and genotypic stability. PLoS One 13:e0199880. doi:10.1371/journal.pone.0199880

Eberhart, S.A., and W.A. Russell. 1966. Stability parameters for comparing Varieties. Crop Sci. 6:36-40. doi:10.2135/cropsci1966.0011183X000600010011x.

Fox, P.N., B. Skovmand, B.K. Thompson, H.J. Braun, and R. Cormier. 1990. Yield and adaptation of hexaploid spring triticale. Euphytica 47:57-64. doi:10.1007/BF00040364.

Huehn, V.M. 1979. Beitrage zur erfassung der phanotypischen stabilitat. EDV Med. Biol. 10:112.

Olivoto, T., A.D.C. L\'ucio, J.A.G. da silva, V.S. Marchioro, V.Q. de Souza, and E. Jost. 2019a. Mean performance and stability in multi-environment trials I: Combining features of AMMI and BLUP techniques. Agron. J. 111:2949-2960. doi:10.2134/agronj2019.03.0220

Olivoto, T., A.D.C. L\'ucio, J.A.G. da silva, B.G. Sari, and M.I. Diel. 2019b. Mean performance and stability in multi-environment trials II: Selection based on multiple traits. Agron. J. 111:2961-2969. doi:10.2134/agronj2019.03.0221

Purchase, J.L., H. Hatting, and C.S. van Deventer. 2000. Genotype vs environment interaction of winter wheat (Triticum aestivum L.) in South Africa: II. Stability analysis of yield performance. South African J. Plant Soil 17:101-107. doi:10.1080/02571862.2000.10634878

Resende MDV (2007) Matematica e estatistica na analise de experimentos e no melhoramento genetico. Embrapa Florestas, Colombo

Sneller, C.H., L. Kilgore-Norquest, and D. Dombek. 1997. Repeatability of Yield Stability Statistics in Soybean. Crop Sci. 37:383-390. doi:10.2135/cropsci1997.0011183X003700020013x

Shahbazi, E. 2019. Genotype selection and stability analysis for seed yield of Nigella sativa using parametric and non-parametric statistics. Sci. Hortic. (Amsterdam). 253:172-179. doi:10.1016/j.scienta.2019.04.047.

Wricke, G. 1965. Zur berechnung der okovalenz bei sommerweizen und hafer. Z. Pflanzenzuchtg 52:127-138.

Zali, H., E. Farshadfar, S.H. Sabaghpour, and R. Karimizadeh. 2012. Evaluation of genotype vs environment interaction in chickpea using measures of stability from AMMI model. Ann. Biol. Res. 3:3126-3136. http://eprints.icrisat.ac.in/id/eprint/7173

Examples

# \donttest{
library(metan)

#################### joint-regression analysis #####################
ge_r <- ge_reg(data_ge2, ENV, GEN, REP,
               resp = c(PH, EH, CD, CL, ED))

#> Evaluating variable PH 0 % 
#> Evaluating variable EH 25 % 
#> Evaluating variable CD 50 % 
#> Evaluating variable CL 75 % 
#> Evaluating variable ED 100 % 

get_model_data(ge_r)

#> # A tibble: 13 x 6
#>    gen      PH    EH      CD     CL      ED
#>    <chr> <dbl> <dbl>   <dbl>  <dbl>   <dbl>
#>  1 H1    0.806 1.06  -0.0594 -2.19   0.0280
#>  2 H10   1.22  1.30   2.31    4.63   2.47  
#>  3 H11   1.08  1.00   2.74    2.37   0.707 
#>  4 H12   0.465 0.590  1.73    2.94   0.808 
#>  5 H13   0.306 0.575 -0.661   0.129  0.636 
#>  6 H2    0.963 0.525 -1.26   -3.86  -0.0534
#>  7 H3    1.35  1.15   1.40   -1.33   0.987 
#>  8 H4    1.27  1.41   0.555  -0.393  1.20  
#>  9 H5    1.17  1.30   0.356   0.486  0.197 
#> 10 H6    0.936 0.780  3.00    1.60   0.940 
#> 11 H7    0.992 0.950 -0.0386  1.48   0.797 
#> 12 H8    1.01  0.886  0.903   3.73   1.89  
#> 13 H9    1.42  1.48   2.04    3.42   2.40  

get_model_data(ge_r, "deviations")

#> # A tibble: 13 x 6
#>    gen       PH       EH      CD     CL     ED
#>    <chr>  <dbl>    <dbl>   <dbl>  <dbl>  <dbl>
#>  1 H1    0.0710  0.0637  -0.268  -0.176 -0.448
#>  2 H10   0.0318  0.0303  -0.0566  0.576 -0.506
#>  3 H11   0.0278  0.0256  -0.275   0.984  4.35 
#>  4 H12   0.0676  0.0861   0.152   0.300  0.475
#>  5 H13   0.0609  0.0298   0.450   9.10   5.48 
#>  6 H2    0.0815  0.0981   0.237   1.18   6.78 
#>  7 H3    0.0898  0.0870   1.13    0.327  1.65 
#>  8 H4    0.0515  0.0335   0.317   1.50   2.16 
#>  9 H5    0.0159  0.00449 -0.117   4.37   1.77 
#> 10 H6    0.0481  0.0106   1.31    5.90   0.844
#> 11 H7    0.0328  0.0155   1.41    1.79   1.12 
#> 12 H8    0.0767  0.0696   1.72    6.05   1.03 
#> 13 H9    0.0212 -0.00212  1.03    0.713 -0.779

#################### AMMI model #####################
# Fit an AMMI model for 7 variables.
AMMI <- data_ge2 %>%
 performs_ammi(ENV, GEN, REP,
               resp = c(PH, ED, TKW, NKR, CD, CL, CW))

#> variable PH 
#> ---------------------------------------------------------------------------
#> AMMI analysis table
#> ---------------------------------------------------------------------------
#>     Source  Df Sum Sq Mean Sq F value   Pr(>F) Percent Accumul
#>        ENV   3  7.719  2.5728 127.913 4.25e-07       .       .
#>   REP(ENV)   8  0.161  0.0201   0.897 5.22e-01       .       .
#>   GEN       12  1.865  0.1554   6.929 6.89e-09       .       .
#>    ENV:GEN  36  5.397  0.1499   6.686 5.01e-14       .       .
#>        PC1  14  4.466  0.3190  14.230 0.00e+00    82.8    82.8
#>        PC2  12  0.653  0.0545   2.430 8.40e-03    12.1    94.9
#>        PC3  10  0.277  0.0277   1.240 2.76e-01     5.1     100
#>  Residuals  96  2.153  0.0224      NA       NA       .       .
#>      Total 155 17.294  0.1116      NA       NA       .       .
#> ---------------------------------------------------------------------------
#> 
#> variable ED 
#> ---------------------------------------------------------------------------
#> AMMI analysis table
#> ---------------------------------------------------------------------------
#>     Source  Df Sum Sq Mean Sq F value   Pr(>F) Percent Accumul
#>        ENV   3  306.0  101.99  43.386 2.70e-05       .       .
#>   REP(ENV)   8   18.8    2.35   0.906 5.15e-01       .       .
#>   GEN       12  212.9   17.74   6.838 8.95e-09       .       .
#>    ENV:GEN  36  398.2   11.06   4.263 7.60e-09       .       .
#>        PC1  14  212.2   15.16   5.840 0.00e+00    53.3    53.3
#>        PC2  12  134.7   11.23   4.330 0.00e+00    33.8    87.1
#>        PC3  10   51.3    5.13   1.980 4.38e-02    12.9     100
#>  Residuals  96  249.1    2.59      NA       NA       .       .
#>      Total 155 1185.0    7.64      NA       NA       .       .
#> ---------------------------------------------------------------------------
#> 
#> variable TKW 
#> ---------------------------------------------------------------------------
#> AMMI analysis table
#> ---------------------------------------------------------------------------
#>     Source  Df Sum Sq Mean Sq F value   Pr(>F) Percent Accumul
#>        ENV   3  37013   12338   11.13 3.16e-03       .       .
#>   REP(ENV)   8   8869    1109    1.21 3.03e-01       .       .
#>   GEN       12  44633    3719    4.05 4.41e-05       .       .
#>    ENV:GEN  36 164572    4571    4.98 1.73e-10       .       .
#>        PC1  14 104276    7448    8.11 0.00e+00    63.4    63.4
#>        PC2  12  33361    2780    3.03 1.20e-03    20.3    83.6
#>        PC3  10  26935    2694    2.93 3.00e-03    16.4     100
#>  Residuals  96  88171     918      NA       NA       .       .
#>      Total 155 343257    2215      NA       NA       .       .
#> ---------------------------------------------------------------------------
#> 
#> variable NKR 
#> ---------------------------------------------------------------------------
#> AMMI analysis table
#> ---------------------------------------------------------------------------
#>     Source  Df Sum Sq Mean Sq F value   Pr(>F) Percent Accumul
#>        ENV   3  237.0   79.01  15.843 0.000997       .       .
#>   REP(ENV)   8   39.9    4.99   0.635 0.746348       .       .
#>   GEN       12  227.8   18.99   2.418 0.008726       .       .
#>    ENV:GEN  36  602.7   16.74   2.132 0.001839       .       .
#>        PC1  14  337.4   24.10   3.070 0.000600      56      56
#>        PC2  12  192.2   16.02   2.040 0.028500    31.9    87.9
#>        PC3  10   73.1    7.31   0.930 0.509500    12.1     100
#>  Residuals  96  753.7    7.85      NA       NA       .       .
#>      Total 155 1861.1   12.01      NA       NA       .       .
#> ---------------------------------------------------------------------------
#> 
#> variable CD 
#> ---------------------------------------------------------------------------
#> AMMI analysis table
#> ---------------------------------------------------------------------------
#>     Source  Df Sum Sq Mean Sq F value   Pr(>F) Percent Accumul
#>        ENV   3   9.72   3.241   6.921 0.012981       .       .
#>   REP(ENV)   8   3.75   0.468   0.519 0.839734       .       .
#>   GEN       12  31.41   2.617   2.899 0.001851       .       .
#>    ENV:GEN  36  81.59   2.266   2.511 0.000194       .       .
#>        PC1  14  48.25   3.446   3.820 0.000000    59.1    59.1
#>        PC2  12  20.72   1.727   1.910 0.042300    25.4    84.5
#>        PC3  10  12.62   1.262   1.400 0.192000    15.5     100
#>  Residuals  96  86.66   0.903      NA       NA       .       .
#>      Total 155 213.12   1.375      NA       NA       .       .
#> ---------------------------------------------------------------------------
#> 
#> variable CL 
#> ---------------------------------------------------------------------------
#> AMMI analysis table
#> ---------------------------------------------------------------------------
#>     Source  Df Sum Sq Mean Sq F value   Pr(>F) Percent Accumul
#>        ENV   3   50.1   16.70    5.73 2.16e-02       .       .
#>   REP(ENV)   8   23.3    2.91    1.76 9.41e-02       .       .
#>   GEN       12   65.7    5.47    3.31 4.81e-04       .       .
#>    ENV:GEN  36  527.3   14.65    8.86 7.16e-18       .       .
#>        PC1  14  320.6   22.90   13.86 0.00e+00    60.8    60.8
#>        PC2  12  129.1   10.76    6.51 0.00e+00    24.5    85.3
#>        PC3  10   77.5    7.75    4.69 0.00e+00    14.7     100
#>  Residuals  96  158.6    1.65      NA       NA       .       .
#>      Total 155  825.0    5.32      NA       NA       .       .
#> ---------------------------------------------------------------------------
#> 
#> variable CW 
#> ---------------------------------------------------------------------------
#> AMMI analysis table
#> ---------------------------------------------------------------------------
#>     Source  Df Sum Sq Mean Sq F value   Pr(>F) Percent Accumul
#>        ENV   3 1258.2  419.41  45.225 2.31e-05       .       .
#>   REP(ENV)   8   74.2    9.27   0.772 6.28e-01       .       .
#>   GEN       12  531.6   44.30   3.686 1.43e-04       .       .
#>    ENV:GEN  36 3056.0   84.89   7.064 9.61e-15       .       .
#>        PC1  14 2223.8  158.84  13.220 0.00e+00    72.8    72.8
#>        PC2  12  502.5   41.88   3.480 3.00e-04    16.4    89.2
#>        PC3  10  329.7   32.97   2.740 5.20e-03    10.8     100
#>  Residuals  96 1153.7   12.02      NA       NA       .       .
#>      Total 155 6073.7   39.19      NA       NA       .       .
#> ---------------------------------------------------------------------------
#> 
#> All variables with significant (p < 0.05) genotype-vs-environment interaction
#> Done!

# Sum of squares
get_model_data(AMMI, "ipca_ss")

#> # A tibble: 3 x 9
#>   PC       DF    PH    ED     TKW   NKR    CD    CL    CW
#>   <chr> <dbl> <dbl> <dbl>   <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1 PC1      14 4.47  212.  104276. 337.   48.2 321.  2224.
#> 2 PC2      12 0.653 135.   33361. 192.   20.7 129.   503.
#> 3 PC3      10 0.277  51.3  26935.  73.1  12.6  77.5  330.

# Mean squares
get_model_data(AMMI, "ipca_ms")

#> # A tibble: 3 x 9
#>   PC       DF     PH    ED   TKW   NKR    CD    CL    CW
#>   <chr> <dbl>  <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1 PC1      14 0.319  15.2  7448. 24.1   3.45 22.9  159. 
#> 2 PC2      12 0.0545 11.2  2780. 16.0   1.73 10.8   41.9
#> 3 PC3      10 0.0277  5.13 2694.  7.31  1.26  7.75  33.0

# Examine the significance (p-value) of the IPCAs
get_model_data(AMMI, "ipca_pval")

#> # A tibble: 3 x 9
#>   PC       DF     PH     ED     TKW      NKR     CD    CL       CW
#>   <chr> <dbl>  <dbl>  <dbl>   <dbl>    <dbl>  <dbl> <dbl>    <dbl>
#> 1 PC1      14 0      0      0       0.000600 0          0 0       
#> 2 PC2      12 0.0084 0      0.00120 0.0285   0.0423     0 0.000300
#> 3 PC3      10 0.276  0.0438 0.003   0.509    0.192      0 0.0052  

# Explained sum of square for each IPCA
get_model_data(AMMI)

#> # A tibble: 3 x 9
#>   PC       DF    PH    ED   TKW   NKR    CD    CL    CW
#>   <chr> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1 PC1      14  82.8  53.3  63.4  56    59.1  60.8  72.8
#> 2 PC2      12  12.1  33.8  20.3  31.9  25.4  24.5  16.4
#> 3 PC3      10   5.1  12.9  16.4  12.1  15.5  14.7  10.8

# Accumulated sum of square
get_model_data(AMMI, "ipca_accum")

#> # A tibble: 3 x 9
#>   PC       DF    PH    ED   TKW   NKR    CD    CL    CW
#>   <chr> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1 PC1      14  82.8  53.3  63.4  56    59.1  60.8  72.8
#> 2 PC2      12  94.9  87.1  83.6  87.9  84.5  85.3  89.2
#> 3 PC3      10 100   100   100   100   100   100   100  

### AMMI-based stability statistics ###
# Get the AMMI stability value
AMMI %>%
AMMI_indexes() %>%
get_model_data("ASV")

#> # A tibble: 13 x 8
#>    gen      PH    ED    TKW   NKR    CD    CL     CW
#>    <fct> <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl>  <dbl>
#>  1 H1    2.27  1.41   7.76  2.01  0.314 2.75   6.24 
#>  2 H10   1.61  2.11   6.58  1.00  0.725 2.67   4.63 
#>  3 H11   1.48  1.31   3.55  2.05  0.505 0.918  3.16 
#>  4 H12   2.62  0.610  0.981 0.547 1.08  1.36   0.881
#>  5 H13   2.53  1.56   2.45  1.86  1.55  2.11   5.63 
#>  6 H2    2.56  2.58  19.0   0.793 1.39  4.47  12.0  
#>  7 H3    2.80  0.863 17.3   0.513 1.45  2.19   2.97 
#>  8 H4    2.22  0.540 13.0   0.565 1.31  1.58   2.59 
#>  9 H5    1.27  1.22   0.984 1.59  0.744 1.04   3.68 
#> 10 H6    2.00  0.848  5.73  3.11  1.39  1.93   5.14 
#> 11 H7    1.72  0.717  9.61  1.41  1.78  1.25   7.93 
#> 12 H8    2.38  1.54  19.8   2.51  2.29  3.25  10.5  
#> 13 H9    0.872 1.95  21.3   2.44  1.81  2.06   8.42 

#################### WAASB model #####################
# Fitting the WAAS index
AMMI <- waas(data_ge2, ENV, GEN, REP,
             resp = c(PH, ED, TKW, NKR))

#> variable PH 
#> ---------------------------------------------------------------------------
#> AMMI analysis table
#> ---------------------------------------------------------------------------
#>     Source  Df Sum Sq Mean Sq F value   Pr(>F) Percent Accumul
#>        ENV   3  7.719  2.5728 127.913 4.25e-07       .       .
#>   REP(ENV)   8  0.161  0.0201   0.897 5.22e-01       .       .
#>   GEN       12  1.865  0.1554   6.929 6.89e-09       .       .
#>    ENV:GEN  36  5.397  0.1499   6.686 5.01e-14       .       .
#>        PC1  14  4.466  0.3190  14.230 0.00e+00    82.8    82.8
#>        PC2  12  0.653  0.0545   2.430 8.40e-03    12.1    94.9
#>        PC3  10  0.277  0.0277   1.240 2.76e-01     5.1     100
#>  Residuals  96  2.153  0.0224      NA       NA       .       .
#>      Total 155 17.294  0.1116      NA       NA       .       .
#> ---------------------------------------------------------------------------
#> 
#> variable ED 
#> ---------------------------------------------------------------------------
#> AMMI analysis table
#> ---------------------------------------------------------------------------
#>     Source  Df Sum Sq Mean Sq F value   Pr(>F) Percent Accumul
#>        ENV   3  306.0  101.99  43.386 2.70e-05       .       .
#>   REP(ENV)   8   18.8    2.35   0.906 5.15e-01       .       .
#>   GEN       12  212.9   17.74   6.838 8.95e-09       .       .
#>    ENV:GEN  36  398.2   11.06   4.263 7.60e-09       .       .
#>        PC1  14  212.2   15.16   5.840 0.00e+00    53.3    53.3
#>        PC2  12  134.7   11.23   4.330 0.00e+00    33.8    87.1
#>        PC3  10   51.3    5.13   1.980 4.38e-02    12.9     100
#>  Residuals  96  249.1    2.59      NA       NA       .       .
#>      Total 155 1185.0    7.64      NA       NA       .       .
#> ---------------------------------------------------------------------------
#> 
#> variable TKW 
#> ---------------------------------------------------------------------------
#> AMMI analysis table
#> ---------------------------------------------------------------------------
#>     Source  Df Sum Sq Mean Sq F value   Pr(>F) Percent Accumul
#>        ENV   3  37013   12338   11.13 3.16e-03       .       .
#>   REP(ENV)   8   8869    1109    1.21 3.03e-01       .       .
#>   GEN       12  44633    3719    4.05 4.41e-05       .       .
#>    ENV:GEN  36 164572    4571    4.98 1.73e-10       .       .
#>        PC1  14 104276    7448    8.11 0.00e+00    63.4    63.4
#>        PC2  12  33361    2780    3.03 1.20e-03    20.3    83.6
#>        PC3  10  26935    2694    2.93 3.00e-03    16.4     100
#>  Residuals  96  88171     918      NA       NA       .       .
#>      Total 155 343257    2215      NA       NA       .       .
#> ---------------------------------------------------------------------------
#> 
#> variable NKR 
#> ---------------------------------------------------------------------------
#> AMMI analysis table
#> ---------------------------------------------------------------------------
#>     Source  Df Sum Sq Mean Sq F value   Pr(>F) Percent Accumul
#>        ENV   3  237.0   79.01  15.843 0.000997       .       .
#>   REP(ENV)   8   39.9    4.99   0.635 0.746348       .       .
#>   GEN       12  227.8   18.99   2.418 0.008726       .       .
#>    ENV:GEN  36  602.7   16.74   2.132 0.001839       .       .
#>        PC1  14  337.4   24.10   3.070 0.000600      56      56
#>        PC2  12  192.2   16.02   2.040 0.028500    31.9    87.9
#>        PC3  10   73.1    7.31   0.930 0.509500    12.1     100
#>  Residuals  96  753.7    7.85      NA       NA       .       .
#>      Total 155 1861.1   12.01      NA       NA       .       .
#> ---------------------------------------------------------------------------
#> 
#> All variables with significant (p < 0.05) genotype-vs-environment interaction
#> Done!

# Getting the weighted average of absolute scores
get_model_data(AMMI, what = "WAASB")

#> # A tibble: 13 x 5
#>    gen      PH    ED   TKW   NKR
#>    <fct> <dbl> <dbl> <dbl> <dbl>
#>  1 H1    0.318 0.667 2.72  0.929
#>  2 H10   0.230 0.973 2.15  0.506
#>  3 H11   0.201 0.649 1.26  0.836
#>  4 H12   0.364 0.315 0.558 0.228
#>  5 H13   0.363 0.838 0.514 0.946
#>  6 H2    0.342 1.08  4.41  0.404
#>  7 H3    0.374 0.486 4.10  0.252
#>  8 H4    0.294 0.378 3.07  0.281
#>  9 H5    0.168 0.567 0.738 0.611
#> 10 H6    0.270 0.409 1.64  1.60 
#> 11 H7    0.228 0.384 3.44  0.518
#> 12 H8    0.315 0.653 4.91  0.941
#> 13 H9    0.146 0.907 5.50  0.888

# And the rank for the WAASB index.
get_model_data(AMMI, what = "OrWAASB")

#> # A tibble: 13 x 5
#>    gen      PH    ED   TKW   NKR
#>    <fct> <dbl> <dbl> <dbl> <dbl>
#>  1 H1        9     9     7    10
#>  2 H10       5    12     6     5
#>  3 H11       3     7     4     8
#>  4 H12      12     1     2     1
#>  5 H13      11    10     1    12
#>  6 H2       10    13    11     4
#>  7 H3       13     5    10     2
#>  8 H4        7     2     8     3
#>  9 H5        2     6     3     7
#> 10 H6        6     4     5    13
#> 11 H7        4     3     9     6
#> 12 H8        8     8    12    11
#> 13 H9        1    11    13     9

#################### BLUP model #####################
# Fitting a mixed-effect model
blup <- waasb(data_ge2, ENV, GEN, REP,
              resp = c(PH, ED, TKW, NKR))

#> Evaluating variable PH 0 % 
#> Evaluating variable ED 33.3 % 
#> Evaluating variable TKW 66.7 % 
#> Evaluating variable NKR 100 % 
#> All variables with significant (p < 0.05) genotype-vs-environment interaction
#> Done!

# Getting p-values for likelihood-ratio test
get_model_data(blup, what = "pval_lrt")

#> # A tibble: 3 x 5
#>   model           PH            ED       TKW      NKR
#>   <chr>        <dbl>         <dbl>     <dbl>    <dbl>
#> 1 Complete NA        NA            NA        NA      
#> 2 Genotype  9.39e- 1  0.299         1.00e+ 0  0.787  
#> 3 Gen:Env   1.09e-13  0.0000000169  4.21e-10  0.00404

# Getting the variance components
get_model_data(blup, what = "vcomp")

#> # A tibble: 3 x 5
#>   Group          PH    ED     TKW   NKR
#>   <chr>       <dbl> <dbl>   <dbl> <dbl>
#> 1 GEN      0.000456 0.556 8.27e-7 0.187
#> 2 GEN:ENV  0.0425   2.82  1.15e+3 2.96 
#> 3 Residual 0.0224   2.59  9.19e+2 7.85 

# Getting the genetic parameters
get_model_data(blup)

#> # A tibble: 15 x 5
#>    Parameters                   PH      ED      TKW     NKR
#>    <chr>                     <dbl>   <dbl>    <dbl>   <dbl>
#>  1 GEI variance           0.0425    2.82   1.15e+ 3  2.96  
#>  2 GEI (%)               65.0      47.2    5.55e+ 1 26.9   
#>  3 Genotypic variance     0.000456  0.556  8.27e- 7  0.187 
#>  4 Gen (%)                0.698     9.32   4.00e- 8  1.70  
#>  5 Residual variance      0.0224    2.59   9.19e+ 2  7.85  
#>  6 Res (%)               34.3      43.4    4.45e+ 1 71.4   
#>  7 Phenotypic variance    0.0654    5.97   2.06e+ 3 11.0   
#>  8 Heritability           0.00698   0.0932 4.00e-10  0.0170
#>  9 GEIr2                  0.650     0.472  5.55e- 1  0.269 
#> 10 Heribatility of means  0.0352    0.376  2.28e- 9  0.118 
#> 11 Accuracy               0.188     0.614  4.77e- 5  0.344 
#> 12 rge                    0.655     0.521  5.55e- 1  0.274 
#> 13 CVg                    0.860     1.51   2.68e- 4  1.34  
#> 14 CVr                    6.03      3.25   8.95e+ 0  8.69  
#> 15 CV ratio               0.143     0.463  3.00e- 5  0.154 

### BLUP-based stability indexes ###
blup %>%
Resende_indexes() %>%
get_model_data()

#> # A tibble: 13 x 5
#>    gen      PH    ED   TKW   NKR
#>    <chr> <dbl> <dbl> <dbl> <dbl>
#>  1 H1    1.04  1.03  1.06  0.998
#>  2 H10   0.937 0.979 0.946 1.00 
#>  3 H11   0.964 0.986 0.984 1.01 
#>  4 H12   0.981 0.984 0.946 0.966
#>  5 H13   1.02  1.02  1.00  0.976
#>  6 H2    1.04  1.02  1.03  0.993
#>  7 H3    1.03  0.998 1.01  0.984
#>  8 H4    1.03  0.994 1.01  1.05 
#>  9 H5    1.03  1.01  1.00  1.03 
#> 10 H6    1.02  1.03  1.05  1.01 
#> 11 H7    0.970 0.999 1.01  0.985
#> 12 H8    0.939 0.978 0.945 0.980
#> 13 H9    0.953 0.965 0.916 1.00 

#################### Stability indexes #####################
stats_ge <- ge_stats(data_ge, ENV, GEN, REP, everything())

#> Evaluating variable GY 0 % 
#> Evaluating variable HM 100 % 

get_model_data(stats_ge)

#> # A tibble: 20 x 33
#>    var   gen       Y    CV    Var Shukla  Wi_g  Wi_f  Wi_u  Ecoval   bij
#>    <chr> <chr> <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl>   <dbl> <dbl>
#>  1 GY    G1     2.60 35.2   10.9  0.0280  84.4  89.2  81.1   1.22  1.06 
#>  2 GY    G10    2.47 42.3   14.2  0.244   59.2  64.6  54.4   7.96  1.12 
#>  3 GY    G2     2.74 34.0   11.3  0.0861  82.8  95.3  75.6   3.03  1.05 
#>  4 GY    G3     2.96 29.9   10.1  0.0121 104.   99.7 107.    0.725 1.03 
#>  5 GY    G4     2.64 31.4    8.93 0.0640  85.9  79.5  91.9   2.34  0.937
#>  6 GY    G5     2.54 30.6    7.82 0.0480  82.7  82.2  82.4   1.84  0.887
#>  7 GY    G6     2.53 29.7    7.34 0.0468  83.0  83.7  81.8   1.81  0.861
#>  8 GY    G7     2.74 27.4    7.33 0.122   83.9  77.6  93.4   4.16  0.819
#>  9 GY    G8     3.00 30.4   10.8  0.0712  98.8  90.5 107.    2.57  1.03 
#> 10 GY    G9     2.51 42.4   14.7  0.167   68.8  68.9  70.3   5.56  1.19 
#> 11 HM    G1    47.1   8.47 207.   1.54    93.7  93.6  93.4  60.2   0.991
#> 12 HM    G10   48.5  11.0  370.   6.96    92.3  99.9  88.2 229.    1.27 
#> 13 HM    G2    46.7   8.70 214.   4.03    90.6  90.1  90.7 138.    0.942
#> 14 HM    G3    47.6   8.45 210.   1.37    95.0  94.4  95.4  55.0   1.00 
#> 15 HM    G4    48.0   7.86 185.   3.51    93.8  94.3  93.9 122.    0.880
#> 16 HM    G5    49.3   7.20 164.   4.82    95.3  96.0  96.2 163.    0.788
#> 17 HM    G6    48.7   8.55 226.   1.48    97.0  98.0  95.9  58.3   1.04 
#> 18 HM    G7    48.0   7.98 191.   2.89    93.9  95.3  92.8 102.    0.911
#> 19 HM    G8    49.1   7.93 197.   2.06    97.2  97.5  97.2  76.5   0.951
#> 20 HM    G9    47.9  10.1  306.   2.70    93.8  98.1  91.4  96.4   1.22 
#> # ... with 22 more variables: Sij <dbl>, R2 <dbl>, ASV <dbl>, SIPC <dbl>,
#> #   EV <dbl>, ZA <dbl>, WAAS <dbl>, HMGV <dbl>, RPGV <dbl>, HMRPGV <dbl>,
#> #   Pi_a <dbl>, Pi_f <dbl>, Pi_u <dbl>, Gai <dbl>, S1 <dbl>, S2 <dbl>,
#> #   S3 <dbl>, S6 <dbl>, N1 <dbl>, N2 <dbl>, N3 <dbl>, N4 <dbl>

# }