Easily get data from a model

Easily get data from a model of class waas or waasb (Olivoto et al., 2019a, 2019b)

get_model_data(x, what = "Y", type = "GEN")

Arguments

x
An object created with the functions waas or waasb.

what
What should be captured from the model. The following options are allowed: "blupg" for genotype's predicted mean and "blupge" for genotype-vs-environment's predicted mean (if the object is of class waasb only); "Y" for raw means (Default); "PCn" for the scores of the nth Interaction Principal Component Axis (IPCA), where n is the desired IPCA (e.g. "PC3"); "WAASB" for the weighted average of the absolute scores; "PctResp" and "PctWAASB" for the rescaled values of the response variable and WAASB, respectively; "wResp" and "wWAASB" for weight of the response variable and stability, respectively; "OrResp" and "OrWAASB" for the ranking regarding the response variable and WAASB, respectively; "OrPC1" for the ranking regarding the first principal component axix; "WAASBY" for the superiority index, "OrWAASBY" for the ranking regarding the superiority index.

type
Chose if the statistics must be show by genotype (type = "GEN", default) or environment (type = "ENV"), when possible.

References

Olivoto, T., A.D.C. L\'ucio, J.A.G. da silva, V.S. Marchioro, V.Q. de Souza, and E. Jost. 2019a. Mean performance and stability in multi-environment trials I: Combining features of AMMI and BLUP techniques. Agron. J. doi:10.2134/agronj2019.03.0220.

Olivoto, T., A.D.C. L\'ucio, J.A.G. da silva, B.G. Sari, and M.I. Diel. 2019b. Mean performance and stability in multi-environment trials II: Selection based on multiple traits. Agron. J.doi:10.2134/agronj2019.03.0221.

Examples

library(metan)

model = waas(data_ge2,
             env = ENV,
             gen = GEN,
             rep = REP,
             resp = c(PH, ED, TKW, NKR))

#> Evaluating variable PH 25 % 
#> Evaluating variable ED 50 % 
#> Evaluating variable TKW 75 % 
#> Evaluating variable NKR 100 % 
#> Done!

get_model_data(model, what = "OrWAASBY")

#> # A tibble: 13 x 5
#>    gen      PH    ED   TKW   NKR
#>    <fct> <dbl> <dbl> <dbl> <dbl>
#>  1 H1        3     2     2    10
#>  2 H10      11    12    11     3
#>  3 H11       7     9     5     6
#>  4 H12      12     6     7     8
#>  5 H13      10     8     3    12
#>  6 H2        6    10     8     5
#>  7 H3        9     7    10     4
#>  8 H4        4     4     6     1
#>  9 H5        1     5     4     2
#> 10 H6        2     1     1    13
#> 11 H7        8     3     9     7
#> 12 H8       13    11    12    11
#> 13 H9        5    13    13     9

waasb(data_ge2, ENV, GEN, REP,
      resp = c(PH, ED, TKW, NKR)) %>%
get_model_data(what = "blupg")

#> Evaluating variable PH 25 % 
#> Evaluating variable ED 50 % 
#> Evaluating variable TKW 75 % 
#> Evaluating variable NKR 100 % 
#> Done!

#> # A tibble: 13 x 5
#>    gen      PH    ED   TKW   NKR
#>    <fct> <dbl> <dbl> <dbl> <dbl>
#>  1 H1     2.49  50.2  339.  32.2
#>  2 H10    2.48  49.1  339.  32.3
#>  3 H11    2.48  49.2  339.  32.3
#>  4 H12    2.48  49.2  339.  32.0
#>  5 H13    2.49  49.9  339.  32.1
#>  6 H2     2.49  50.1  339.  32.2
#>  7 H3     2.49  49.5  339.  32.1
#>  8 H4     2.49  49.4  339.  32.6
#>  9 H5     2.49  49.7  339.  32.4
#> 10 H6     2.49  50.3  339.  32.3
#> 11 H7     2.48  49.5  339.  32.2
#> 12 H8     2.48  49.1  339.  32.1
#> 13 H9     2.48  48.8  339.  32.3