Easily get data from a model

Easily get data from some objects generated in the metan package such as the WAASB and WAASBY indexes (Olivoto et al., 2019a, 2019b) BLUPs, variance components, details of AMMI models and AMMI-based stability statistics

get_model_data(x, what = "Y", type = "GEN")

Arguments

x
An object created with the functions AMMI_indexes, ecovalence, Fox, gai, ge_reg, gamem, performs_ammi, Resende_indexes, Shukla, superiority, waas or waasb.

what
What should be captured from the model. See more in Details section.

type
Chose if the statistics must be show by genotype (type = "GEN", default) or environment (type = "ENV"), when possible.

Value

A tibble showing the values of the variable chosen in argument what.

Details

Bellow are listed the options allowed in the argument what depending on the class of the object

Objects of class AMMI_indexes.

"Y" for raw means (Default).

"ASV" AMMI stability value

"SIPC" Sums of the absolute value of the IPCA scores

"EV" Averages of the squared eigenvector values

"ZA" Absolute value of the relative contribution of IPCAs to the interaction

"WAAS" P-value for for each IPCA.

Objects of class Annicchiarico and Schmildt.

"Mean_rp" The relative performance of the mean.

"Sem_rp" The standard error of the relative mean performance (Schmildt).

"Sd_rp" The standard deviation of the relative performance (Annicchiarico).

"Wi" The genotypic confidence index.

"rank" The rank for genotypic confidence index.

Objects of class ecovalence.

"Ecoval" Ecovalence value.

"Ecov_perc" Ecovalence in percentage value.

"rank" Rank for ecovalence

Objects of class ge_reg.

"Y" for raw means (Default).

"slope" The sloop of the regression

"deviations" The deviations from regression

"RMSE" The Root Mean Square Error

"R2" The r-square of the regression

Objects of class ge_effects.

For objects of class ge_effects no argument what is required

Objects of class Shukla.

"Y" for raw means (Default).

"rMean" Rank for the mean.

"ShuklaVar" Shukla's stablity variance.

"rShukaVar" Rank for Shukla's stablity variance.

"ssiShukaVar" Simultaneous selection index.

Objects of class Fox.

"Y" for raw means (Default).

"TOP" The proportion of locations at which the genotype occurred in the top third

Objects of class gai.

"GAI" The geometric adaptability index.

"GAI_R" The rank for the GAI values.

Objects of class superiority.

"Y" for raw means (Default).

"Pi_a" The superiority measure for all environments.

"R_a" The rank for Pi_a.

"Pi_f" The superiority measure for favorable environments.

"R_f" The rank for Pi_f.

"Pi_u" The superiority measure for unfavorable environments.

"R_u" The rank for Pi_u.

Objects of class Huehn.

"Y" for raw means (Default).

"S1" Mean of the absolute rank differences of a genotype over the n environments.

"S1_R" The rank for S1.

"S2" variance among the ranks over the k environments.

"S2_R" The rank for S2.

"S3" Sum of the absolute deviations.

"S3_R" The rank for S3.

"S6" Relative sum of squares of rank for each genotype.

"S6_R" The rank for S6.

Objects of class Thennarasu.

"Y" for raw means (Default).

"N1" First statistic

"N1_R" The rank for S1.

"N2" Second statistic

"N2_R" The rank for S2.

"N3" Third statistic

"N3_R" The rank for S3.

"N4" Fourth statistic

"N4_R" The rank for S6.

Objects of class performs_ammi.

"Y" for raw means (Default).

"ipca_ss" Sum of square for each IPCA.

"ipca_ms" Mean square for each IPCA.

"ipca_fval" F value for each IPCA.

"ipca_pval" P-value for for each IPCA.

"ipca_expl" Explained sum of square for each IPCA.

"ipca_accum" Accumulated explained sum of square.

Objects of class waas and waasb.

"Y" for raw means (Default).

"PCn" for the scores of the nth Interaction Principal Component Axis (IPCA). where * n is the desired IPCA (e.g. "PC3").

"WAASB" The weighted average of the absolute scores.

"PctResp" The rescaled values of the response variable.

"PctWAASB" The rescaled values of the WAASB.

"wResp" The weight for the response variable.

"wWAASB" The weight for the stability.

"OrResp" The ranking regarding the response variable.

"OrWAASB" The ranking regarding the WAASB.

"OrPC1" The ranking regarding the first principal component axix.

"WAASBY" The superiority index WAASBY.

"OrWAASBY" The ranking regarding the superiority index.

Objects of class waasb or gamem.

"HMGV" For harmonic mean of genotypic values.

"RPGV or RPGV_Y" For relative performance of genotypic values

"HMRPGV" For harmonic mean of relative performance of genotypic values

Objects of class Res_ind

"blupg" For genotype's predicted mean.

"blupge" for genotype-vs-environment's predicted mean (only for objects of class waasb).

"genpar" Genetic parameters.

"lrt" The statistic for the likelihood-ratio test for random effects.

"pval_lrt" The p-values for the likelihood-ratio test.

"vcomp" The variance components for random effects.

References

Olivoto, T., A.D.C. L'ucio, J.A.G. da silva, V.S. Marchioro, V.Q. de Souza, and E. Jost. 2019a. Mean performance and stability in multi-environment trials I: Combining features of AMMI and BLUP techniques. Agron. J. doi:10.2134/agronj2019.03.0220

Olivoto, T., A.D.C. L'ucio, J.A.G. da silva, B.G. Sari, and M.I. Diel. 2019b. Mean performance and stability in multi-environment trials II: Selection based on multiple traits. Agron. J. doi:10.2134/agronj2019.03.0221

Examples

# \donttest{
library(metan)
#################### joint-regression analysis #####################
ge_r <- ge_reg(data_ge2, ENV, GEN, REP,
               resp = c(PH, EH, CD, CL, ED))

#> Evaluating variable PH 0 % 
#> Evaluating variable EH 25 % 
#> Evaluating variable CD 50 % 
#> Evaluating variable CL 75 % 
#> Evaluating variable ED 100 % 

get_model_data(ge_r, "slope")

#> # A tibble: 13 x 6
#>    gen      PH    EH      CD     CL      ED
#>    <chr> <dbl> <dbl>   <dbl>  <dbl>   <dbl>
#>  1 H1    0.806 1.06  -0.0594 -2.19   0.0280
#>  2 H10   1.22  1.30   2.31    4.63   2.47  
#>  3 H11   1.08  1.00   2.74    2.37   0.707 
#>  4 H12   0.465 0.590  1.73    2.94   0.808 
#>  5 H13   0.306 0.575 -0.661   0.129  0.636 
#>  6 H2    0.963 0.525 -1.26   -3.86  -0.0534
#>  7 H3    1.35  1.15   1.40   -1.33   0.987 
#>  8 H4    1.27  1.41   0.555  -0.393  1.20  
#>  9 H5    1.17  1.30   0.356   0.486  0.197 
#> 10 H6    0.936 0.780  3.00    1.60   0.940 
#> 11 H7    0.992 0.950 -0.0386  1.48   0.797 
#> 12 H8    1.01  0.886  0.903   3.73   1.89  
#> 13 H9    1.42  1.48   2.04    3.42   2.40  

get_model_data(ge_r, "deviations")

#> # A tibble: 13 x 6
#>    gen       PH       EH      CD     CL     ED
#>    <chr>  <dbl>    <dbl>   <dbl>  <dbl>  <dbl>
#>  1 H1    0.0710  0.0637  -0.268  -0.176 -0.448
#>  2 H10   0.0318  0.0303  -0.0566  0.576 -0.506
#>  3 H11   0.0278  0.0256  -0.275   0.984  4.35 
#>  4 H12   0.0676  0.0861   0.152   0.300  0.475
#>  5 H13   0.0609  0.0298   0.450   9.10   5.48 
#>  6 H2    0.0815  0.0981   0.237   1.18   6.78 
#>  7 H3    0.0898  0.0870   1.13    0.327  1.65 
#>  8 H4    0.0515  0.0335   0.317   1.50   2.16 
#>  9 H5    0.0159  0.00449 -0.117   4.37   1.77 
#> 10 H6    0.0481  0.0106   1.31    5.90   0.844
#> 11 H7    0.0328  0.0155   1.41    1.79   1.12 
#> 12 H8    0.0767  0.0696   1.72    6.05   1.03 
#> 13 H9    0.0212 -0.00212  1.03    0.713 -0.779

#################### AMMI model #####################
# Fit an AMMI model for 7 variables.
AMMI <- data_ge2 %>%
 performs_ammi(ENV, GEN, REP,
               resp = c(PH, ED, TKW, NKR, CD, CL, CW))

#> Evaluating variable PH 0 % 
#> Evaluating variable ED 16.7 % 
#> Evaluating variable TKW 33.3 % 
#> Evaluating variable NKR 50 % 
#> Evaluating variable CD 66.7 % 
#> Evaluating variable CL 83.3 % 
#> Evaluating variable CW 100 % 
#> All variables with significant (p < 0.05) genotype-vs-environment interaction
#> Done!

# Sum of squares
get_model_data(AMMI, "ipca_ss")

#> # A tibble: 3 x 9
#>   PC       DF    PH    ED     TKW   NKR    CD    CL    CW
#>   <chr> <dbl> <dbl> <dbl>   <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1 PC1      14 4.47  212.  104276. 337.   48.2 321.  2224.
#> 2 PC2      12 0.653 135.   33361. 192.   20.7 129.   503.
#> 3 PC3      10 0.277  51.3  26935.  73.1  12.6  77.5  330.

# Mean squares
get_model_data(AMMI, "ipca_ms")

#> # A tibble: 3 x 9
#>   PC       DF     PH    ED   TKW   NKR    CD    CL    CW
#>   <chr> <dbl>  <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1 PC1      14 0.319  15.2  7448. 24.1   3.45 22.9  159. 
#> 2 PC2      12 0.0545 11.2  2780. 16.0   1.73 10.8   41.9
#> 3 PC3      10 0.0277  5.13 2694.  7.31  1.26  7.75  33.0

# Examine the significance (p-value) of the IPCAs
get_model_data(AMMI, "ipca_pval")

#> # A tibble: 3 x 9
#>   PC       DF     PH     ED     TKW      NKR     CD    CL       CW
#>   <chr> <dbl>  <dbl>  <dbl>   <dbl>    <dbl>  <dbl> <dbl>    <dbl>
#> 1 PC1      14 0      0      0       0.000600 0          0 0       
#> 2 PC2      12 0.0084 0      0.00120 0.0285   0.0423     0 0.000300
#> 3 PC3      10 0.276  0.0438 0.003   0.509    0.192      0 0.0052  

# Explained sum of square for each IPCA
get_model_data(AMMI, "ipca_expl")

#> # A tibble: 3 x 9
#>   PC       DF    PH    ED   TKW   NKR    CD    CL    CW
#>   <chr> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1 PC1      14  82.8  53.3  63.4  56    59.1  60.8  72.8
#> 2 PC2      12  12.1  33.8  20.3  31.9  25.4  24.5  16.4
#> 3 PC3      10   5.1  12.9  16.4  12.1  15.5  14.7  10.8

# Accumulated sum of square
get_model_data(AMMI, "ipca_accum")

#> # A tibble: 3 x 9
#>   PC       DF    PH    ED   TKW   NKR    CD    CL    CW
#>   <chr> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1 PC1      14  82.8  53.3  63.4  56    59.1  60.8  72.8
#> 2 PC2      12  94.9  87.1  83.6  87.9  84.5  85.3  89.2
#> 3 PC3      10 100   100   100   100   100   100   100  

### AMMI-based stability statistics ###
# Get the AMMI stability value
AMMI %>%
AMMI_indexes() %>%
get_model_data("ASV")

#> # A tibble: 13 x 8
#>    gen      PH    ED    TKW   NKR    CD    CL     CW
#>    <fct> <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl>  <dbl>
#>  1 H1    2.27  1.41   7.76  2.01  0.314 2.75   6.24 
#>  2 H10   1.61  2.11   6.58  1.00  0.725 2.67   4.63 
#>  3 H11   1.48  1.31   3.55  2.05  0.505 0.918  3.16 
#>  4 H12   2.62  0.610  0.981 0.547 1.08  1.36   0.881
#>  5 H13   2.53  1.56   2.45  1.86  1.55  2.11   5.63 
#>  6 H2    2.56  2.58  19.0   0.793 1.39  4.47  12.0  
#>  7 H3    2.80  0.863 17.3   0.513 1.45  2.19   2.97 
#>  8 H4    2.22  0.540 13.0   0.565 1.31  1.58   2.59 
#>  9 H5    1.27  1.22   0.984 1.59  0.744 1.04   3.68 
#> 10 H6    2.00  0.848  5.73  3.11  1.39  1.93   5.14 
#> 11 H7    1.72  0.717  9.61  1.41  1.78  1.25   7.93 
#> 12 H8    2.38  1.54  19.8   2.51  2.29  3.25  10.5  
#> 13 H9    0.872 1.95  21.3   2.44  1.81  2.06   8.42 

#################### WAASB model #####################
# Fitting the WAAS index
AMMI <- waas(data_ge2, ENV, GEN, REP,
             resp = c(PH, ED, TKW, NKR))

#> Evaluating variable PH 0 % 
#> Evaluating variable ED 33.3 % 
#> Evaluating variable TKW 66.7 % 
#> Evaluating variable NKR 100 % 
#> All variables with significant (p < 0.05) genotype-vs-environment interaction
#> Done!

# Getting the weighted average of absolute scores
get_model_data(AMMI, what = "WAASB")

#> # A tibble: 13 x 5
#>    gen      PH    ED   TKW   NKR
#>    <fct> <dbl> <dbl> <dbl> <dbl>
#>  1 H1    0.318 0.667 2.72  0.929
#>  2 H10   0.230 0.973 2.15  0.506
#>  3 H11   0.201 0.649 1.26  0.836
#>  4 H12   0.364 0.315 0.558 0.228
#>  5 H13   0.363 0.838 0.514 0.946
#>  6 H2    0.342 1.08  4.41  0.404
#>  7 H3    0.374 0.486 4.10  0.252
#>  8 H4    0.294 0.378 3.07  0.281
#>  9 H5    0.168 0.567 0.738 0.611
#> 10 H6    0.270 0.409 1.64  1.60 
#> 11 H7    0.228 0.384 3.44  0.518
#> 12 H8    0.315 0.653 4.91  0.941
#> 13 H9    0.146 0.907 5.50  0.888

# And the rank for the WAASB inde.
get_model_data(AMMI, what = "OrWAASB")

#> # A tibble: 13 x 5
#>    gen      PH    ED   TKW   NKR
#>    <fct> <dbl> <dbl> <dbl> <dbl>
#>  1 H1        9     9     7    10
#>  2 H10       5    12     6     5
#>  3 H11       3     7     4     8
#>  4 H12      12     1     2     1
#>  5 H13      11    10     1    12
#>  6 H2       10    13    11     4
#>  7 H3       13     5    10     2
#>  8 H4        7     2     8     3
#>  9 H5        2     6     3     7
#> 10 H6        6     4     5    13
#> 11 H7        4     3     9     6
#> 12 H8        8     8    12    11
#> 13 H9        1    11    13     9

#################### BLUP model #####################
# Fitting a mixed-effect model
blup <- waasb(data_ge2, ENV, GEN, REP,
              resp = c(PH, ED, TKW, NKR))

#> Evaluating variable PH 0 % 
#> Evaluating variable ED 33.3 % 
#> Evaluating variable TKW 66.7 % 
#> Evaluating variable NKR 100 % 
#> All variables with significant (p < 0.05) genotype-vs-environment interaction
#> Done!

# Getting p-values for likelihood-ratio test
get_model_data(blup, what = "pval_lrt")

#> # A tibble: 3 x 5
#>   model           PH            ED       TKW      NKR
#>   <chr>        <dbl>         <dbl>     <dbl>    <dbl>
#> 1 Complete NA        NA            NA        NA      
#> 2 Genotype  9.39e- 1  0.299         1.00e+ 0  0.787  
#> 3 Gen:Env   1.09e-13  0.0000000169  4.21e-10  0.00404

# Getting the variance components
get_model_data(blup, what = "vcomp")

#> # A tibble: 3 x 5
#>   Group          PH    ED     TKW   NKR
#>   <chr>       <dbl> <dbl>   <dbl> <dbl>
#> 1 GEN      0.000456 0.556 8.27e-7 0.187
#> 2 GEN:ENV  0.0425   2.82  1.15e+3 2.96 
#> 3 Residual 0.0224   2.59  9.19e+2 7.85 

# Getting the genetic parameters
get_model_data(blup, what = "genpar")

#> # A tibble: 15 x 5
#>    Parameters                   PH      ED      TKW     NKR
#>    <chr>                     <dbl>   <dbl>    <dbl>   <dbl>
#>  1 GEI variance           0.0425    2.82   1.15e+ 3  2.96  
#>  2 GEI (%)               65.0      47.2    5.55e+ 1 26.9   
#>  3 Genotypic variance     0.000456  0.556  8.27e- 7  0.187 
#>  4 Gen (%)                0.698     9.32   4.00e- 8  1.70  
#>  5 Residual variance      0.0224    2.59   9.19e+ 2  7.85  
#>  6 Res (%)               34.3      43.4    4.45e+ 1 71.4   
#>  7 Phenotypic variance    0.0654    5.97   2.06e+ 3 11.0   
#>  8 Heritability           0.00698   0.0932 4.00e-10  0.0170
#>  9 GEIr2                  0.650     0.472  5.55e- 1  0.269 
#> 10 Heribatility of means  0.0352    0.376  2.28e- 9  0.118 
#> 11 Accuracy               0.188     0.614  4.77e- 5  0.344 
#> 12 rge                    0.655     0.521  5.55e- 1  0.274 
#> 13 CVg                    0.860     1.51   2.68e- 4  1.34  
#> 14 CVr                    6.03      3.25   8.95e+ 0  8.69  
#> 15 CV ratio               0.143     0.463  3.00e- 5  0.154 

### BLUP-based stability indexes ###
blup %>%
Resende_indexes() %>%
get_model_data("HMRPGV")

#> # A tibble: 13 x 5
#>    gen      PH    ED   TKW   NKR
#>    <chr> <dbl> <dbl> <dbl> <dbl>
#>  1 H1    1.04  1.03  1.06  0.998
#>  2 H10   0.937 0.979 0.946 1.00 
#>  3 H11   0.964 0.986 0.984 1.01 
#>  4 H12   0.981 0.984 0.946 0.966
#>  5 H13   1.02  1.02  1.00  0.976
#>  6 H2    1.04  1.02  1.03  0.993
#>  7 H3    1.03  0.998 1.01  0.984
#>  8 H4    1.03  0.994 1.01  1.05 
#>  9 H5    1.03  1.01  1.00  1.03 
#> 10 H6    1.02  1.03  1.05  1.01 
#> 11 H7    0.970 0.999 1.01  0.985
#> 12 H8    0.939 0.978 0.945 0.980
#> 13 H9    0.953 0.965 0.916 1.00 

# }