Linear and Partial Correlation Coefficients

Estimates the linear and partial correlation coefficients using as input a data frame or a correlation matrix.

lpcor(.data, n = NULL, method = "pearson", verbose = TRUE)

Arguments

.data
The data to be analyzed. Must be a symmetric correlation matrix or, a dataframe containing the predictor variables, or an object of class split_factors.

n
If a correlation matrix is provided, then n is the number of objects used to compute the correlation coefficients.

method
a character string indicating which correlation coefficient is to be computed. One of 'pearson' (default), 'kendall', or 'spearman'.

verbose
If verbose = TRUE then some results are shown in the console.

Value

If a grouping factor is used then a list is returned with the following values.

linear.mat

The matrix of linear correlation.

partial.mat

The matrix of partial correlations.

results

Hypothesis testing for each pairwise comparison.

Examples

library(metan)
partial1 = lpcor(iris)

#>                                 linear    partial         t         prob
#> Sepal.Length x Sepal.Width  -0.1175698  0.6285707  9.765380 0.000000e+00
#> Sepal.Length x Petal.Length  0.8717538  0.7190656 12.502483 0.000000e+00
#> Sepal.Length x Petal.Width   0.8179411 -0.3396174 -4.362929 2.412876e-05
#> Sepal.Width x Petal.Length  -0.4284401 -0.6152919 -9.431189 0.000000e+00
#> Sepal.Width x Petal.Width   -0.3661259  0.3526260  4.553279 1.104958e-05
#> Petal.Length x Petal.Width   0.9628654  0.8707698 21.398708 0.000000e+00

#> The factors  where excluded to perform the analysis. If you want to perform an analysis for each level of a factor, use the function 'split_factors() before.' 

# Alternatively using the pipe operator %>%
partial2 = iris %>% lpcor()

#>                                 linear    partial         t         prob
#> Sepal.Length x Sepal.Width  -0.1175698  0.6285707  9.765380 0.000000e+00
#> Sepal.Length x Petal.Length  0.8717538  0.7190656 12.502483 0.000000e+00
#> Sepal.Length x Petal.Width   0.8179411 -0.3396174 -4.362929 2.412876e-05
#> Sepal.Width x Petal.Length  -0.4284401 -0.6152919 -9.431189 0.000000e+00
#> Sepal.Width x Petal.Width   -0.3661259  0.3526260  4.553279 1.104958e-05
#> Petal.Length x Petal.Width   0.9628654  0.8707698 21.398708 0.000000e+00

#> The factors  where excluded to perform the analysis. If you want to perform an analysis for each level of a factor, use the function 'split_factors() before.' 

partial3 = cor(iris[1:4]) %>%
           lpcor(n = nrow(iris),
                 verbose = FALSE)

partial4 = iris %>%
           split_factors(Species) %>%
           lpcor()

#> The columns Species where deleted. Use 'keep_factors = TRUE' to keep this columns in the grouped data. 

#> 
#> ----------------------------------------------------------------------------
#> Level: setosa 
#> ----------------------------------------------------------------------------
#>                                linear     partial          t         prob
#> Sepal.Length x Sepal.Width  0.7425467  0.72228639  7.0833236 6.834434e-09
#> Sepal.Length x Petal.Length 0.2671758  0.16604293  1.1420107 2.593594e-01
#> Sepal.Length x Petal.Width  0.2780984  0.10656131  0.7268727 4.709870e-01
#> Sepal.Width x Petal.Length  0.1777000 -0.04521791 -0.3069968 7.602313e-01
#> Sepal.Width x Petal.Width   0.2327520  0.05178061  0.3516649 7.266954e-01
#> Petal.Length x Petal.Width  0.3316300  0.27968723  1.9757822 5.419574e-02
#> 
#> ----------------------------------------------------------------------------
#> Level: versicolor 
#> ----------------------------------------------------------------------------
#>                                linear    partial          t         prob
#> Sepal.Length x Sepal.Width  0.5259107  0.2686100  1.8913091 6.488965e-02
#> Sepal.Length x Petal.Length 0.7540490  0.6292137  5.4906811 1.666695e-06
#> Sepal.Length x Petal.Width  0.5464611 -0.2241652 -1.5600639 1.255990e-01
#> Sepal.Width x Petal.Length  0.5605221 -0.1071007 -0.7305949 4.687316e-01
#> Sepal.Width x Petal.Width   0.6639987  0.4697561  3.6090336 7.555314e-04
#> Petal.Length x Petal.Width  0.7866681  0.6480512  5.7711604 6.371049e-07
#> 
#> ----------------------------------------------------------------------------
#> Level: virginica 
#> ----------------------------------------------------------------------------
#>                                linear     partial          t         prob
#> Sepal.Length x Sepal.Width  0.4572278  0.26908423  1.8949086 6.439972e-02
#> Sepal.Length x Petal.Length 0.8642247  0.83815645 10.4223360 1.074696e-13
#> Sepal.Length x Petal.Width  0.2811077 -0.12536431 -0.8570233 3.958750e-01
#> Sepal.Width x Petal.Length  0.4010446 -0.07558634 -0.5141223 6.096277e-01
#> Sepal.Width x Petal.Width   0.5377280  0.48387140  3.7500043 4.933633e-04
#> Petal.Length x Petal.Width  0.3221082  0.17966044  1.2386712 2.217511e-01