Make a two-way table

This function help users to easely make a two-way table from a "long format" data.

make_mat(.data, row, col, value)

Arguments

.data
The dataset. Must contains at least two categorical columns.

row
The column of data in which the mean of each level will correspond to one line in the output.

col
The column of data in which the mean of each level will correspond to one column in the output.

value
The column of data that contains the values to fill the two-way table.

Examples

library(metan)
library(dplyr)
matrix = data_ge %>% make_mat(row = GEN, col = ENV, val = GY)
matrix

#>           E1      E10      E11      E12      E13      E14       E2       E3
#> G1  2.365787 2.307633 1.355667 1.342333 2.998259 1.533333 3.040690 4.080333
#> G10 1.974073 1.536253 0.899100 1.022136 1.830667 1.859210 3.145840 4.114333
#> G2  2.901747 2.297053 1.490720 1.993333 3.028309 1.426580 3.231053 4.568667
#> G3  2.888587 2.336727 1.567727 1.764358 3.469148 2.059654 3.608280 4.132333
#> G4  2.588567 2.167353 1.370000 1.527333 2.635716 1.861333 3.190080 3.854333
#> G5  2.188260 2.139833 1.325667 1.689000 2.568099 1.784296 3.140137 3.741333
#> G6  2.300800 2.208147 1.500667 1.385667 2.911667 1.798852 3.291173 3.434667
#> G7  2.773727 2.441893 1.364333 1.951630 3.183840 1.936333 2.612677 4.101333
#> G8  2.899360 2.574867 1.683433 2.000000 3.519975 1.991864 3.444797 4.110000
#> G9  2.325927 1.742333 1.125333 1.409333 2.950630 1.567877 3.094863 4.506000
#>           E4       E5       E6       E7       E8       E9
#> G1  3.494333 4.170531 2.809086 1.897880 2.271933 2.783667
#> G10 4.271000 3.365667 2.484877 2.243627 2.700310 3.150000
#> G2  3.715573 3.827432 2.541333 1.987627 2.045297 3.363333
#> G3  4.134000 4.126654 2.984716 2.158547 2.850420 3.293667
#> G4  3.302167 3.782605 2.701185 1.980887 2.300590 3.724667
#> G5  3.379113 3.469000 2.433321 1.656767 2.708330 3.299667
#> G6  3.403473 3.568926 2.335556 1.757787 2.539380 3.039667
#> G7  3.018953 4.047914 2.665901 2.550813 2.583383 3.140000
#> G8  4.137773 4.812099 2.908123 2.261780 2.878853 2.828000
#> G9  3.896340 3.934173 2.768037 1.393247 2.485260 1.943333