Multi-trait mean performance and stability index

Computes the multi-trait stability index proposed by Olivoto et al. (2019) considering different parametric and non-parametric stability indexes.

mtmps(model, SI = 15, mineval = 1, verbose = TRUE)

Arguments

model
An object of class mtmps

SI
An integer (0-100). The selection intensity in percentage of the total number of genotypes.

mineval
The minimum value so that an eigenvector is retained in the factor analysis.

verbose
If verbose = TRUE (Default), some results are shown in the console.

Value

An object of class mtmps with the following items:

data The data used to compute the factor analysis.

cormat The correlation matrix among the environments.

PCA The eigenvalues and explained variance.

FA The factor analysis.

KMO The result for the Kaiser-Meyer-Olkin test.

MSA The measure of sampling adequacy for individual variable.

communalities The communalities.

communalities_mean The communalities' mean.

initial_loadings The initial loadings.

finish_loadings The final loadings after varimax rotation.

canonical_loadings The canonical loadings.

scores_gen The scores for genotypes in all retained factors.

scores_ide The scores for the ideotype in all retained factors.

MTSI The multi-trait mean performance and stability index.

contri_fac The relative contribution of each factor on the MTSI value. The lower the contribution of a factor, the close of the ideotype the variables in such factor are.

contri_fac_rank, contri_fac_rank_sel The rank for the contribution of each factor for all genotypes and selected genotypes, respectively.

sel_dif_trait, sel_dif_stab, sel_dif_mps A data frame containing the selection differential (gains) for the mean performance, stability index, and mean performance and stability index, respectively. The following variables are shown.

VAR: the trait's name.

Factor: The factor that traits where grouped into.

Xo: The original population mean.

Xs: The mean of selected genotypes.

SD and SDperc: The selection differential and selection differential in percentage, respectively.

h2: The broad-sense heritability.

SG and SGperc: The selection gains and selection gains in percentage, respectively.

sense: The desired selection sense.

goal: selection gains match desired sense? 100 for yes and 0 for no.

stat_dif_trait, stat_dif_stab, stat_dif_mps A data frame with the descriptive statistic for the selection gains for the mean performance, stability index, and mean performance and stability index, respectively. The following columns are shown by sense.

sense: The desired selection sense.

variable: the trait's name.

min: the minimum value for the selection gain.

mean: the mean value for the selection gain.

ci: the confidence interval for the selection gain.

sd.amo: the standard deviation for the selection gain.

max: the maximum value for the selection gain.

sum: the sum of the selection gain.

sel_gen The selected genotypes.

References

Olivoto, T., A.D.C. L\'ucio, J.A.G. da silva, B.G. Sari, and M.I. Diel. 2019. Mean performance and stability in multi-environment trials II: Selection based on multiple traits. Agron. J. 111:2961-2969. doi: 10.2134/agronj2019.03.0220

See also

Author

Tiago Olivoto tiagoolivoto@gmail.com

Examples

# \donttest{
library(metan)
# The same approach as mtsi()
# mean performance and stability for GY and HM
# mean performance: The genotype's BLUP
# stability: the WAASB index (lower is better)
# weights: equal for mean performance and stability

model <-
mps(data_ge,
    env = ENV,
    gen = GEN,
    rep = REP,
    resp = everything())

#> Evaluating trait GY |======================                      | 50% 00:00:00 Evaluating trait HM |============================================| 100% 00:00:01 

#> Method: REML/BLUP

#> Random effects: GEN, GEN:ENV

#> Fixed effects: ENV, REP(ENV)

#> Denominador DF: Satterthwaite's method

#> ---------------------------------------------------------------------------
#> P-values for Likelihood Ratio Test of the analyzed traits
#> ---------------------------------------------------------------------------
#>     model       GY       HM
#>  COMPLETE       NA       NA
#>       GEN 1.11e-05 5.07e-03
#>   GEN:ENV 2.15e-11 2.27e-15
#> ---------------------------------------------------------------------------
#> All variables with significant (p < 0.05) genotype-vs-environment interaction

#> Mean performance: blupg

#> Stability: waasb

selection <- mtmps(model)

#> 
#> -------------------------------------------------------------------------------
#> Principal Component Analysis
#> -------------------------------------------------------------------------------
#> # A tibble: 2 x 4
#>   PC    Eigenvalues `Variance (%)` `Cum. variance (%)`
#>   <chr>       <dbl>          <dbl>               <dbl>
#> 1 PC1         1.37            68.5                68.5
#> 2 PC2         0.631           31.5               100  
#> -------------------------------------------------------------------------------
#> Factor Analysis - factorial loadings after rotation-
#> -------------------------------------------------------------------------------
#> # A tibble: 2 x 4
#>   VAR     FA1 Communality Uniquenesses
#>   <chr> <dbl>       <dbl>        <dbl>
#> 1 GY    0.827       0.685        0.315
#> 2 HM    0.827       0.685        0.315
#> -------------------------------------------------------------------------------
#> Comunalit Mean: 0.6846623 
#> -------------------------------------------------------------------------------
#> Selection differential for the mean performance and stability index
#> -------------------------------------------------------------------------------
#> # A tibble: 2 x 6
#>   VAR   Factor    Xo    Xs    SD SDperc
#>   <chr> <chr>  <dbl> <dbl> <dbl>  <dbl>
#> 1 GY    FA 1    48.3  86.4  38.0   78.7
#> 2 HM    FA 1    58.3  79.2  21.0   36.0
#> -------------------------------------------------------------------------------
#> Selection differential for the mean of the variables
#> -------------------------------------------------------------------------------
#> # A tibble: 2 x 11
#>   VAR   Factor    Xo    Xs    SD SDperc    h2    SG SGperc sense     goal
#>   <chr> <chr>  <dbl> <dbl> <dbl>  <dbl> <dbl> <dbl>  <dbl> <chr>    <dbl>
#> 1 GY    FA 1    2.67  2.98 0.305 11.4   0.815 0.249  9.31  increase   100
#> 2 HM    FA 1   48.1  48.4  0.265  0.551 0.686 0.182  0.378 increase   100
#> ------------------------------------------------------------------------------
#> Selected genotypes
#> -------------------------------------------------------------------------------
#> G8 G3
#> -------------------------------------------------------------------------------


# gains for stability
selection$sel_dif_stab

#> # A tibble: 2 x 7
#>   VAR      Xo    Xs    SD SDperc sense     goal
#>   <chr> <dbl> <dbl> <dbl>  <dbl> <chr>    <dbl>
#> 1 GY     58.6  77.3  18.8   32.0 decrease     0
#> 2 HM     62.4  94.3  31.9   51.1 decrease     0


# gains for mean performance
selection$sel_dif_trait

#> # A tibble: 2 x 11
#>   VAR   Factor    Xo    Xs    SD SDperc    h2    SG SGperc sense     goal
#>   <chr> <chr>  <dbl> <dbl> <dbl>  <dbl> <dbl> <dbl>  <dbl> <chr>    <dbl>
#> 1 GY    FA 1    2.67  2.98 0.305 11.4   0.815 0.249  9.31  increase   100
#> 2 HM    FA 1   48.1  48.4  0.265  0.551 0.686 0.182  0.378 increase   100

# }