Multi-trait stability index

Computes the multi-trait stability index proposed by Olivoto et al. (2019)

mtsi(.data, index = "waasb", SI = 15, mineval = 1, verbose = TRUE)

Arguments

.data
An object of class waasb or waas.

index
If index = 'waasb' (Default) the multi-trait index will be computed considering the stability of genotypes only. If index = 'waasby' both stability and mean performance are considered. More details can be seen in waasb and waas functions.

SI
An integer (0-100). The selection intensity in percentage of the total number of genotypes.

mineval
The minimum value so that an eigenvector is retained in the factor analysis.

verbose
If verbose = TRUE (Default) then some results are shown in the console.

Value

An object of class ge_factanal with the following items:

data The data used to compute the factor analysis.

cormat The correlation matrix among the environments.

PCA The eigenvalues and explained variance.

FA The factor analysis.

KMO The result for the Kaiser-Meyer-Olkin test.

MSA The measure of sampling adequacy for individual variable.

communalities The communalities.

communalities.mean The communalities' mean.

initial.loadings The initial loadings.

finish.loadings The final loadings after varimax rotation.

canonical.loadings The canonical loadings.

scores.gen The scores for genotypes in all retained factors.

scores.ide The scores for the ideotype in all retained factors.

MTSI The multi-trait stability index.

contri.fac The relative contribution of each factor on the MTSI value. The lower the contribution of a factor, the close of the ideotype the variables in such factor are.

sel.dif The selection differential for the WAASBY or WAASB index.

mean.sd The mean for the differential selection.

sel.dif.var The selection differential for the variables.

Selected The selected genotypes.

References

Olivoto, T., A.D.C. L'ucio, J.A.G. da silva, B.G. Sari, and M.I. Diel. 2019. Mean performance and stability in multi-environment trials II: Selection based on multiple traits. Agron. J. (in press).

Examples

library(metan)

# Based on stability only, for both GY and HM, higher is better
mtsi_model = waasb(data_ge,
                   env = ENV,
                   gen = GEN,
                   rep = REP,
                   resp = c(GY, HM))

#> Evaluating variable GY 50 % 
#> Evaluating variable HM 100 % 
#> All variables with significant (p < 0.05) genotype-vs-environment interaction
#> Done!

mtsi_index = mtsi(mtsi_model)

#> 
#> -------------------------------------------------------------------------------
#> Principal Component Analysis
#> -------------------------------------------------------------------------------
#> # A tibble: 2 x 4
#>   PC    Eigenvalues `Variance (%)` `Cum. variance (%)`
#>   <chr>       <dbl>          <dbl>               <dbl>
#> 1 PC1         1.66            82.8                82.8
#> 2 PC2         0.343           17.2               100  
#> -------------------------------------------------------------------------------
#> Factor Analysis - factorial loadings after rotation-
#> -------------------------------------------------------------------------------
#> # A tibble: 2 x 4
#>   VAR     FA1 Communality Uniquenesses
#>   <chr> <dbl>       <dbl>        <dbl>
#> 1 GY    0.910       0.828        0.172
#> 2 HM    0.910       0.828        0.172
#> -------------------------------------------------------------------------------
#> Comunalit Mean: 0.8283139 
#> -------------------------------------------------------------------------------
#> Multitrait stability index
#> -------------------------------------------------------------------------------
#>     G3     G1     G6     G8     G4     G2     G9     G5     G7    G10 
#> 0.0000 0.2303 0.4846 1.0066 1.2954 1.6022 1.8318 1.8374 1.8450 3.4178 
#> -------------------------------------------------------------------------------
#> Selection differential for the  waasb index
#> -------------------------------------------------------------------------------
#> # A tibble: 2 x 6
#>   VAR   Factor    Xo    Xs     SD SDperc
#>   <chr> <chr>  <dbl> <dbl>  <dbl>  <dbl>
#> 1 GY    FA 1   0.250 0.117 -0.134  -53.5
#> 2 HM    FA 1   0.614 0.373 -0.241  -39.2
#> ------------------------------------------------------------------------------
#> Mean of selection differential
#> -------------------------------------------------------------------------------
#>          Xo          Xs          SD      SDperc 
#>   0.4323192   0.2448716  -0.1874476 -46.3574863 
#> -------------------------------------------------------------------------------
#> Selection differential for the mean of the variables
#> -------------------------------------------------------------------------------
#> # A tibble: 2 x 6
#>   VAR   Factor    xo    Xs     SD SDperc
#>   <chr> <chr>  <dbl> <dbl>  <dbl>  <dbl>
#> 1 GY    FA 1    2.67  2.78  0.105   3.94
#> 2 HM    FA 1   48.1  47.3  -0.747  -1.55
#> ------------------------------------------------------------------------------
#> Selected genotypes
#> -------------------------------------------------------------------------------
#> G3 G1
#> -------------------------------------------------------------------------------

# Based on mean performance and stability (using pipe operator %>%)
# GY: higher is better
# HM: lower is better
mtsi_index2 = data_ge %>%
              waasb(ENV, GEN, REP, c(GY, HM), mresp = c(100, 0)) %>%
              mtsi(index = 'waasby')

#> Evaluating variable GY 50 % 
#> Evaluating variable HM 100 % 
#> All variables with significant (p < 0.05) genotype-vs-environment interaction
#> Done!
#> 
#> -------------------------------------------------------------------------------
#> Principal Component Analysis
#> -------------------------------------------------------------------------------
#> # A tibble: 2 x 4
#>   PC    Eigenvalues `Variance (%)` `Cum. variance (%)`
#>   <chr>       <dbl>          <dbl>               <dbl>
#> 1 PC1         1.58            79.2                79.2
#> 2 PC2         0.416           20.8               100  
#> -------------------------------------------------------------------------------
#> Factor Analysis - factorial loadings after rotation-
#> -------------------------------------------------------------------------------
#> # A tibble: 2 x 4
#>   VAR     FA1 Communality Uniquenesses
#>   <chr> <dbl>       <dbl>        <dbl>
#> 1 GY    0.890       0.792        0.208
#> 2 HM    0.890       0.792        0.208
#> -------------------------------------------------------------------------------
#> Comunalit Mean: 0.7922483 
#> -------------------------------------------------------------------------------
#> Multitrait stability index
#> -------------------------------------------------------------------------------
#>     G3     G1     G2     G8     G6     G4     G7     G9     G5    G10 
#> 0.4901 1.0853 1.5365 1.6202 2.0663 2.0779 2.1950 2.5718 3.2045 3.9391 
#> -------------------------------------------------------------------------------
#> Selection differential for the  waasby index
#> -------------------------------------------------------------------------------
#> # A tibble: 2 x 6
#>   VAR   Factor    Xo    Xs    SD SDperc
#>   <chr> <chr>  <dbl> <dbl> <dbl>  <dbl>
#> 1 GY    FA 1    48.3  76.5  28.2   58.3
#> 2 HM    FA 1    54.1  86.2  32.1   59.3
#> ------------------------------------------------------------------------------
#> Mean of selection differential
#> -------------------------------------------------------------------------------
#>       Xo       Xs       SD   SDperc 
#> 51.22793 81.37443 30.14650 58.82081 
#> -------------------------------------------------------------------------------
#> Selection differential for the mean of the variables
#> -------------------------------------------------------------------------------
#> # A tibble: 2 x 6
#>   VAR   Factor    xo    Xs     SD SDperc
#>   <chr> <chr>  <dbl> <dbl>  <dbl>  <dbl>
#> 1 GY    FA 1    2.67  2.78  0.105   3.94
#> 2 HM    FA 1   48.1  47.3  -0.747  -1.55
#> ------------------------------------------------------------------------------
#> Selected genotypes
#> -------------------------------------------------------------------------------
#> G3 G1
#> -------------------------------------------------------------------------------