Mantel test for a set of correlation matrices

This function generate a pairwise matrix of plots to compare the similarity of two or more correlation matrices. In the upper diagonal are presented the plots and in the lower diagonal the result of Mantel test based on permutations.

pairs_mantel(
  ...,
  type = 1,
  nrepet = 1000,
  names = NULL,
  prob = 0.05,
  diag = FALSE,
  export = FALSE,
  main = "auto",
  file.type = "pdf",
  file.name = NULL,
  width = 8,
  height = 7,
  resolution = 300,
  size.point = 0.5,
  shape.point = 19,
  alpha.point = 1,
  fill.point = NULL,
  col.point = "black",
  minsize = 2,
  maxsize = 3,
  signcol = "green",
  alpha = 0.15,
  diagcol = "gray",
  col.up.panel = "gray",
  col.lw.panel = "gray",
  col.dia.panel = "gray",
  pan.spacing = 0.15,
  digits = 2
)

Arguments

...
The input matrices. May be an output generated by the function lpcor or a coerced list generated by the function as.lpcor

type
The type of correlation if an obect generated by the function lpcor is used. 1 = Linear correlation matrices, or 2 = partial correlation matrices.

nrepet
The number of permutations. Default is 1000

names
An optional vector of names of the same length of ... .

prob
The error probability for Mantel test.

diag
Logical argument. If TRUE, the Kernel density is shown in the diagonal of plot.

export
Logical argument. If TRUE, then the plot is exported to the current directory.

main
The title of the plot, set to 'auto'.

file.type
The format of the file if export = TRUE. Set to 'pdf'. Other possible values are *.tiff using file.type = 'tiff'.

file.name
The name of the plot when exported. Set to NULL, i.e., automatically.

width
The width of the plot, set to 8.

height
The height of the plot, set to 7.

resolution
The resolution of the plot if file.type = 'tiff' is used. Set to 300 (300 dpi).

size.point
The size of the points in the plot. Set to 0.5.

shape.point
The shape of the point, set to 19.

alpha.point
The value for transparency of the points: 1 = full color.

fill.point
The color to fill the points. Valid argument if points are between 21 and 25.

col.point
The color for the edge of the point, set to black.

minsize
The size of the letter that will represent the smallest correlation coefficient.

maxsize
The size of the letter that will represent the largest correlation coefficient.

signcol
The colour that indicate significant correlations (based on the prob value.), set to 'green'.

alpha
The value for transparency of the color informed in signcol, when 1 = full color. Set to 0.15.

diagcol
The color in the kernel distribution. Set to 'gray'.

col.up.panel, col.lw.panel, col.dia.panel
The color for the opper, lower and diagonal pannels. Set to 'gray', 'gray', and 'gray', respectively.

pan.spacing
The space between the pannels. Set to 0.15.

digits
The number of digits to show in the plot.

Value

An object of class gg, ggmatrix.

Examples

# \donttest{
library(metan)
# iris dataset
lpc <- iris %>%
       group_by(Species) %>%
       lpcor() %>%
       pairs_mantel(names = c('setosa', 'versicolor', 'virginica'))

#>                         Pairs    linear     partial          t         prob
#> 1  Sepal.Length x Sepal.Width 0.7425467  0.72228639  7.0833236 6.834434e-09
#> 2 Sepal.Length x Petal.Length 0.2671758  0.16604293  1.1420107 2.593594e-01
#> 3  Sepal.Length x Petal.Width 0.2780984  0.10656131  0.7268727 4.709870e-01
#> 4  Sepal.Width x Petal.Length 0.1777000 -0.04521791 -0.3069968 7.602313e-01
#> 5   Sepal.Width x Petal.Width 0.2327520  0.05178061  0.3516649 7.266954e-01
#> 6  Petal.Length x Petal.Width 0.3316300  0.27968723  1.9757822 5.419574e-02
#>                         Pairs    linear    partial          t         prob
#> 1  Sepal.Length x Sepal.Width 0.5259107  0.2686100  1.8913091 6.488965e-02
#> 2 Sepal.Length x Petal.Length 0.7540490  0.6292137  5.4906811 1.666695e-06
#> 3  Sepal.Length x Petal.Width 0.5464611 -0.2241652 -1.5600639 1.255990e-01
#> 4  Sepal.Width x Petal.Length 0.5605221 -0.1071007 -0.7305949 4.687316e-01
#> 5   Sepal.Width x Petal.Width 0.6639987  0.4697561  3.6090336 7.555314e-04
#> 6  Petal.Length x Petal.Width 0.7866681  0.6480512  5.7711604 6.371049e-07
#>                         Pairs    linear     partial          t         prob
#> 1  Sepal.Length x Sepal.Width 0.4572278  0.26908423  1.8949086 6.439972e-02
#> 2 Sepal.Length x Petal.Length 0.8642247  0.83815645 10.4223360 1.074696e-13
#> 3  Sepal.Length x Petal.Width 0.2811077 -0.12536431 -0.8570233 3.958750e-01
#> 4  Sepal.Width x Petal.Length 0.4010446 -0.07558634 -0.5141223 6.096277e-01
#> 5   Sepal.Width x Petal.Width 0.5377280  0.48387140  3.7500043 4.933633e-04
#> 6  Petal.Length x Petal.Width 0.3221082  0.17966044  1.2386712 2.217511e-01

# mtcars dataset
mt_num <- select_numeric_cols(mtcars)
lpdata <- as.lpcor(cor(mt_num[1:5]),
                   cor(mt_num[1:5]),
                   cor(mt_num[2:6]),
                   cor(mt_num[4:8])) %>%
          pairs_mantel()
# }