Predict a two-way table based on GGE model

Predict the means for a genotype-vs-environment trial based on a Genotype plus Genotype-vs-Environment interction (GGE) model.

# S3 method for gge
predict(object, naxis = 2, output = "wide", ...)

Arguments

object
An object of class gge.

naxis
The the number of principal components to be used in the prediction. Generaly, two axis may be used. In this case, the estimated values will be those shown in the biplot.

output
The type of output. It must be one of the 'long' (default) returning a long-format table with the columns for environment (ENV), genotypes (GEN) and response variable (Y); or 'wide' to return a two-way table with genotypes in the row, environments in the columns, filled by the estimated values.

...
Additional parameter for the function

Details

This function is used to predict the response variable of a two-way table (for examples the yielding of g genotypes in e environments) based on GGE model. This prediction is based on the number of principal components used. For more detais see Yan and Kang (2007).

References

Yan, W., and M.S. Kang. 2003. GGE biplot analysis: a graphical tool for breeders, geneticists, and agronomists. CRC Press.

Examples

library(metan)
mod = gge(data_ge, GEN, ENV, GY)
predict(mod)

#>           G1       G10       G2       G3       G4       G5       G6       G7
#> E1  2.518587 2.0306120 2.618031 2.804462 2.419847 2.319841 2.379029 2.696073
#> E10 2.137539 1.6512896 2.243120 2.439484 2.094831 2.012357 2.076414 2.402064
#> E11 1.255415 0.9131181 1.391875 1.621847 1.401209 1.359064 1.410244 1.713249
#> E12 1.547003 1.0348687 1.658692 1.868286 1.579136 1.523884 1.600417 1.948089
#> E13 3.072438 1.9711141 3.090197 3.215874 2.635681 2.511312 2.668825 3.159141
#> E14 1.570777 1.7687767 1.766690 2.029515 1.897739 1.840094 1.793766 1.924890
#> E2  3.068640 3.2954517 3.244670 3.471146 3.190296 3.063671 2.992058 3.078529
#> E3  4.100825 4.0069075 4.222583 4.398186 3.937946 3.764411 3.734143 3.893428
#> E4  3.575485 4.0095739 3.768318 3.997471 3.710638 3.560311 3.446057 3.457343
#> E5  4.063913 3.3376109 4.110914 4.239519 3.640770 3.470090 3.549445 3.901946
#> E6  2.617414 2.4216597 2.750131 2.956175 2.627072 2.521984 2.529407 2.755208
#> E7  1.796229 1.9607399 1.984643 2.239129 2.076401 2.008443 1.964836 2.100743
#> E8  2.372309 2.6006988 2.559366 2.803875 2.595697 2.501113 2.438670 2.541347
#> E9  2.754893 3.5946706 3.008826 3.289849 3.179554 3.066246 2.893116 2.800679
#>           G8        G9
#> E1  2.967991 2.4310905
#> E10 2.593563 2.0009150
#> E11 1.707902 0.9983921
#> E12 2.017135 1.3356030
#> E13 3.594542 3.0854534
#> E14 1.944778 1.3286002
#> E2  3.411822 3.0178833
#> E3  4.468669 4.1982051
#> E4  3.881619 3.5800900
#> E5  4.517647 4.1866751
#> E6  3.025660 2.5287145
#> E7  2.170688 1.5846660
#> E8  2.727800 2.2318512
#> E9  3.021108 2.6345466