Print an object of class AMMI_indexes

Print the AMMI_indexes object in two ways. By default, the results are shown in the R console. The results can also be exported to the directory into a *.txt file.

# S3 method for AMMI_indexes
print(x, which = "stats", export = FALSE, file.name = NULL, digits = 3, ...)

Arguments

x
An object of class AMMI_indexes.

which
Which should be printed. Defaults to "stats". Other possible values are "ranks" for genotype ranking and "ssi" for the simultaneous selection index.

export
A logical argument. If TRUE, a *.txt file is exported to the working directory.

file.name
The name of the file if export = TRUE

digits
The significant digits to be shown.

...
Options used by the tibble package to format the output. See tibble::print() for more details.

Examples

# \donttest{
library(metan)
model <- performs_ammi(data_ge, ENV, GEN, REP, GY) %>%
         AMMI_indexes()

#> variable GY 
#> ---------------------------------------------------------------------------
#> AMMI analysis table
#> ---------------------------------------------------------------------------
#>     Source  Df  Sum Sq Mean Sq F value   Pr(>F) Proportion Accumulated
#>        ENV  13 279.574 21.5057   62.33 0.00e+00          .           .
#>   REP(ENV)  28   9.662  0.3451    3.57 3.59e-08          .           .
#>        GEN   9  12.995  1.4439   14.93 2.19e-19          .           .
#>    GEN:ENV 117  31.220  0.2668    2.76 1.01e-11          .           .
#>        PC1  21  10.749  0.5119    5.29 0.00e+00       34.4        34.4
#>        PC2  19   9.924  0.5223    5.40 0.00e+00       31.8        66.2
#>        PC3  17   4.039  0.2376    2.46 1.40e-03       12.9        79.2
#>        PC4  15   3.074  0.2049    2.12 9.60e-03        9.8          89
#>        PC5  13   1.446  0.1113    1.15 3.18e-01        4.6        93.6
#>        PC6  11   0.932  0.0848    0.88 5.61e-01          3        96.6
#>        PC7   9   0.567  0.0630    0.65 7.53e-01        1.8        98.4
#>        PC8   7   0.362  0.0518    0.54 8.04e-01        1.2        99.6
#>        PC9   5   0.126  0.0252    0.26 9.34e-01        0.4         100
#>  Residuals 252  24.367  0.0967      NA       NA          .           .
#>      Total 536 389.036  0.7258      NA       NA       <NA>        <NA>
#> ---------------------------------------------------------------------------
#> 
#> All variables with significant (p < 0.05) genotype-vs-environment interaction
#> Done!

print(model)

#> Variable GY 
#> ---------------------------------------------------------------------------
#> AMMI-based stability indexes
#> ---------------------------------------------------------------------------
#> # A tibble: 10 x 7
#>    GEN       Y   ASV  SIPC     EV     ZA  WAAS
#>    <chr> <dbl> <dbl> <dbl>  <dbl>  <dbl> <dbl>
#>  1 G1     2.60 0.346 0.463 0.0149 0.100  0.151
#>  2 G10    2.47 1.23  2.07  0.210  0.437  0.652
#>  3 G2     2.74 0.249 1.54  0.179  0.216  0.283
#>  4 G3     2.96 0.113 0.552 0.0207 0.0797 0.106
#>  5 G4     2.64 0.594 1.04  0.0521 0.219  0.326
#>  6 G5     2.54 0.430 0.997 0.0433 0.186  0.270
#>  7 G6     2.53 0.265 1.14  0.0911 0.172  0.233
#>  8 G7     2.74 0.663 1.79  0.191  0.303  0.428
#>  9 G8     3.00 0.574 1.18  0.0669 0.225  0.327
#> 10 G9     2.51 0.983 1.50  0.131  0.336  0.507

# }