Print an object of class anova_ind

Print the anova_ind object in two ways. By default, the results are shown in the R console. The results can also be exported to the directory into a *.txt file.

# S3 method for anova_ind
print(x, export = FALSE, file.name = NULL, digits = 3, ...)

Arguments

x
An object of class anova_ind.

export
A logical argument. If TRUE, a *.txt file is exported to the working directory.

file.name
The name of the file if export = TRUE

digits
The significant digits to be shown.

...
Options used by the tibble package to format the output. See tibble::print() for more details.

Examples

library(metan)
model <- data_ge %>% anova_ind(ENV, GEN, REP, c(GY, HM))

#> Evaluating variable GY 0 % 
#> Evaluating variable HM 100 % 

print(model)

#> Variable GY 
#> ---------------------------------------------------------------------------
#> # A tibble: 14 x 12
#>    ENV    MEAN    MSB   MSG    MSR    FCB    PRFB   FCG    PRFG    CV    h2
#>    <chr> <dbl>  <dbl> <dbl>  <dbl>  <dbl>   <dbl> <dbl>   <dbl> <dbl> <dbl>
#>  1 E1     2.52 0.0652 0.337 0.144   0.453 6.43e-1  2.34 5.94e-2 15.1  0.573
#>  2 E10    2.18 0.654  0.296 0.0267 24.5   7.28e-6 11.1  1.10e-5  7.51 0.910
#>  3 E11    1.37 0.377  0.151 0.105   3.59  4.86e-2  1.44 2.44e-1 23.7  0.304
#>  4 E12    1.61 0.0919 0.320 0.0535  1.72  2.08e-1  5.98 6.47e-4 14.4  0.833
#>  5 E13    2.91 0.0767 0.713 0.0994  0.772 4.77e-1  7.18 2.10e-4 10.8  0.861
#>  6 E14    1.78 0.104  0.131 0.0753  1.37  2.78e-1  1.73 1.53e-1 15.4  0.423
#>  7 E2     3.18 0.698  0.207 0.179   3.91  3.88e-2  1.16 3.76e-1 13.3  0.136
#>  8 E3     4.06 0.489  0.335 0.179   2.73  9.21e-2  1.87 1.23e-1 10.4  0.466
#>  9 E4     3.68 0.116  0.531 0.138   0.846 4.46e-1  3.86 7.12e-3 10.1  0.741
#> 10 E5     3.91 0.219  0.526 0.0664  3.30  6.02e-2  7.93 1.10e-4  6.59 0.874
#> 11 E6     2.66 0.160  0.135 0.0586  2.73  9.22e-2  2.30 6.35e-2  9.09 0.565
#> 12 E7     1.99 0.381  0.337 0.0910  4.19  3.22e-2  3.70 8.73e-3 15.2  0.730
#> 13 E8     2.54 0.817  0.215 0.0278 29.4   2.15e-6  7.72 1.31e-4  6.57 0.870
#> 14 E9     3.06 0.583  0.679 0.111   5.25  1.60e-2  6.12 5.62e-4 10.9  0.837
#> # ... with 1 more variable: AS <dbl>
#> ---------------------------------------------------------------------------
#> 
#> 
#> 
#> Variable HM 
#> ---------------------------------------------------------------------------
#> # A tibble: 14 x 12
#>    ENV    MEAN    MSB   MSG   MSR    FCB    PRFB   FCG    PRFG    CV     h2
#>    <chr> <dbl>  <dbl> <dbl> <dbl>  <dbl>   <dbl> <dbl>   <dbl> <dbl>  <dbl>
#>  1 E1     47.4  3.11   5.22  1.85  1.68  2.15e-1  2.82 2.93e-2  2.87 0.645 
#>  2 E10    44.3  9.1   28.8   2.17  4.19  3.22e-2 13.2  3.02e-6  3.33 0.924 
#>  3 E11    54.2  2.81   6.06  3.09  0.910 4.20e-1  1.96 1.07e-1  3.24 0.490 
#>  4 E12    49.6  2.77  10.4   5.32  0.520 6.03e-1  1.96 1.08e-1  4.65 0.489 
#>  5 E13    46.6 25.5   13.9   1.34 19.0   3.71e-5 10.3  1.80e-5  2.49 0.903 
#>  6 E14    41.0 10.0    5.28  1.22  8.23  2.89e-3  4.34 3.93e-3  2.69 0.769 
#>  7 E2     44.1  6.70  13.2   3.05  2.20  1.40e-1  4.32 4.02e-3  3.96 0.768 
#>  8 E3     52.9  7.3    8.00  2.78  2.62  1.00e-1  2.88 2.69e-2  3.15 0.652 
#>  9 E4     50    2.10   7.93  1.25  1.68  2.14e-1  6.35 4.51e-4  2.23 0.843 
#> 10 E5     52.2  6.86  11.8   6.37  1.08  3.62e-1  1.86 1.26e-1  4.83 0.462 
#> 11 E6     45.9 12.9   15.8   2.12  6.06  9.71e-3  7.46 1.63e-4  3.17 0.866 
#> 12 E7     48.5  0.705  5.00  4.74  0.149 8.63e-1  1.05 4.38e-1  4.49 0.0518
#> 13 E8     45.2 12.0   11.6   2.17  5.51  1.36e-2  5.34 1.27e-3  3.26 0.813 
#> 14 E9     51.3  5.73   9.33  2.29  2.50  1.10e-1  4.08 5.41e-3  2.95 0.755 
#> # ... with 1 more variable: AS <dbl>
#> ---------------------------------------------------------------------------
#> 
#> 
#>