Print an object of class ge_reg

Print the ge_reg object in two ways. By default, the results are shown in the R console. The results can also be exported to the directory into a *.txt file.

Usage

# S3 method for ge_reg
print(x, export = FALSE, file.name = NULL, digits = 3, ...)

Arguments

x

An object of class ge_reg.

export

A logical argument. If TRUE, a *.txt file is exported to the working directory.

file.name

The name of the file if export = TRUE

digits

The significant digits to be shown.

...

Options used by the tibble package to format the output. See tibble::print() for more details.

Author

Tiago Olivoto tiagoolivoto@gmail.com

Examples

# \donttest{

library(metan)
model <- ge_reg(data_ge2, ENV, GEN, REP, PH)
#> Evaluating trait PH |============================================| 100% 00:00:00 
print(model)
#> Variable PH 
#> ---------------------------------------------------------------------------
#> Joint-regression Analysis of variance
#> ---------------------------------------------------------------------------
#> # A tibble: 20 x 6
#>    SV                       Df `Sum Sq` `Mean Sq` `F value`    `Pr(>F)`
#>    <chr>                 <dbl>    <dbl>     <dbl>     <dbl>       <dbl>
#>  1 "Total"                  51   15.0      0.294     NA     NA         
#>  2 "GEN"                    12    1.86     0.155      0.870  0.585     
#>  3 "ENV + (GEN x ENV)"      39   13.1      0.336     NA     NA         
#>  4 "ENV (linear)"            1    7.72     7.72      NA     NA         
#>  5 " GEN x ENV (linear)"    12    0.755    0.0629     0.352  0.969     
#>  6 "Pooled deviation"       26    4.64     0.179     NA     NA         
#>  7 "H1"                      2    0.471    0.235     10.5    0.0000756 
#>  8 "H10"                     2    0.236    0.118      5.25   0.00685   
#>  9 "H11"                     2    0.211    0.106      4.71   0.0112    
#> 10 "H12"                     2    0.450    0.225     10.0    0.000110  
#> 11 "H13"                     2    0.410    0.205      9.15   0.000231  
#> 12 "H2"                      2    0.534    0.267     11.9    0.0000241 
#> 13 "H3"                      2    0.584    0.292     13.0    0.00000995
#> 14 "H4"                      2    0.354    0.177      7.89   0.000671  
#> 15 "H5"                      2    0.140    0.0701     3.13   0.0484    
#> 16 "H6"                      2    0.334    0.167      7.44   0.000994  
#> 17 "H7"                      2    0.241    0.121      5.38   0.00609   
#> 18 "H8"                      2    0.505    0.253     11.3    0.0000403 
#> 19 "H9"                      2    0.172    0.0861     3.84   0.0249    
#> 20 "Pooled error"           96    2.15     0.0224    NA     NA         
#> ---------------------------------------------------------------------------
#> Regression parameters
#> ---------------------------------------------------------------------------
#> # A tibble: 13 x 10
#>    GEN      b0    b1 `t(b1=1)`[,1] pval_t[,1]   s2di `F(s2di=0)`    pval_f  RMSE
#>    <chr> <dbl> <dbl>         <dbl>      <dbl>  <dbl>       <dbl>     <dbl> <dbl>
#>  1 H1     2.62 0.806       -0.997    0.321    0.0710       10.5    7.56e-5 0.198
#>  2 H10    2.31 1.22         1.12     0.265    0.0318        5.25   6.85e-3 0.140
#>  3 H11    2.39 1.08         0.411    0.682    0.0278        4.71   1.12e-2 0.133
#>  4 H12    2.44 0.465       -2.75     0.00711  0.0676       10.0    1.10e-4 0.194
#>  5 H13    2.54 0.306       -3.57     0.000558 0.0609        9.15   2.31e-4 0.185
#>  6 H2     2.60 0.963       -0.188    0.851    0.0815       11.9    2.41e-5 0.211
#>  7 H3     2.59 1.35         1.83     0.0711   0.0898       13.0    9.95e-6 0.221
#>  8 H4     2.58 1.27         1.38     0.171    0.0515        7.89   6.71e-4 0.172
#>  9 H5     2.57 1.17         0.887    0.378    0.0159        3.13   4.84e-2 0.108
#> 10 H6     2.56 0.936       -0.330    0.742    0.0481        7.44   9.94e-4 0.167
#> 11 H7     2.40 0.992       -0.0393   0.969    0.0328        5.38   6.09e-3 0.142
#> 12 H8     2.33 1.01         0.0720   0.943    0.0767       11.3    4.03e-5 0.205
#> 13 H9     2.36 1.42         2.18     0.0317   0.0212        3.84   2.49e-2 0.120
#> # ... with 1 more variable: R2 <dbl>
#> ---------------------------------------------------------------------------
#> Variance of b0: 0.001868774 
#> Variance of b1: 0.03777002 
#> 
#> 
#> 
# }