Print an object of class ge_stats

Print the ge_stats object in two ways. By default, the results are shown in the R console. The results can also be exported to the directory into a *.txt file.

# S3 method for ge_stats
print(x, export = FALSE, file.name = NULL,
  digits = 3, ...)

Arguments

x
An object of class ge_stats.

export
A logical argument. If TRUE, a *.txt file is exported to the working directory.

file.name
The name of the file if export = TRUE

digits
The significant digits to be shown.

...
Options used by the tibble package to format the output. See trunc_mat for more details.

Examples

library(metan)
model = ge_stats(.data = data_ge2,
                 env = ENV,
                 gen = GEN,
                 rep = REP,
                 resp = PH)
print(model)

#> Variable PH 
#> ---------------------------------------------------------------------------
#> Individual analysis of variance
#> ---------------------------------------------------------------------------
#> # A tibble: 4 x 12
#>   ENV    MEAN     MSB    MSG    MSR   FCB  PRFB    FCG     PRFG    CV     h2
#>   <chr> <dbl>   <dbl>  <dbl>  <dbl> <dbl> <dbl>  <dbl>    <dbl> <dbl>  <dbl>
#> 1 A1     2.79 0.00437 0.0185 0.0146 0.300 0.743  1.27  2.98e- 1  4.32  0.212
#> 2 A2     2.46 0.00747 0.477  0.0128 0.585 0.565 37.4   1.43e-12  4.59  0.973
#> 3 A3     2.17 0.0507  0.0840 0.0328 1.55  0.233  2.56  2.39e- 2  8.35  0.610
#> 4 A4     2.52 0.0179  0.0254 0.0296 0.603 0.555  0.858 5.96e- 1  6.84 -0.166
#> # ... with 1 more variable: AS <dbl>
#> ---------------------------------------------------------------------------
#> Genotype-vs-environment mean
#> ---------------------------------------------------------------------------
#> # A tibble: 13 x 4
#>       A1    A2    A3    A4
#>  * <dbl> <dbl> <dbl> <dbl>
#>  1  2.72  2.93  2.20  2.64
#>  2  2.78  2.05  2.04  2.39
#>  3  2.75  2.15  2.10  2.56
#>  4  2.69  2.09  2.43  2.53
#>  5  2.77  2.23  2.60  2.56
#>  6  2.79  2.95  2.15  2.52
#>  7  2.94  2.94  2.04  2.46
#>  8  2.87  2.85  2.05  2.56
#>  9  2.83  2.70  2.09  2.64
#> 10  2.77  2.82  2.15  2.49
#> 11  2.77  2.14  2.18  2.53
#> 12  2.69  1.96  2.1   2.55
#> 13  2.93  2.20  2.03  2.30
#> ---------------------------------------------------------------------------
#> Genotype-vs-environment effects
#> ---------------------------------------------------------------------------
#> # A tibble: 13 x 4
#>         A1     A2      A3      A4
#>  *   <dbl>  <dbl>   <dbl>   <dbl>
#>  1 -0.207   0.332 -0.106  -0.0189
#>  2  0.160  -0.243  0.0414  0.0409
#>  3  0.0504 -0.220  0.0296  0.140 
#>  4 -0.0531 -0.321  0.311   0.0636
#>  5 -0.0749 -0.288  0.378  -0.0149
#>  6 -0.120   0.365 -0.132  -0.113 
#>  7  0.0322  0.369 -0.233  -0.168 
#>  8 -0.0196  0.289 -0.212  -0.0569
#>  9 -0.0416  0.160 -0.155   0.0371
#> 10 -0.0976  0.290 -0.0864 -0.106 
#> 11  0.0613 -0.245  0.0913  0.0921
#> 12  0.0572 -0.344  0.0917  0.195 
#> 13  0.253  -0.142 -0.0181 -0.0926
#> ---------------------------------------------------------------------------
#> Genotype + Genotype-vs-environment effects
#> ---------------------------------------------------------------------------
#> # A tibble: 13 x 4
#>          A1     A2      A3       A4
#>  *    <dbl>  <dbl>   <dbl>    <dbl>
#>  1 -0.0704   0.468  0.0308  0.118  
#>  2 -0.00973 -0.413 -0.129  -0.129  
#>  3 -0.0444  -0.315 -0.0652  0.0456 
#>  4 -0.100   -0.369  0.263   0.0163 
#>  5 -0.0211  -0.235  0.432   0.0389 
#>  6 -0.00107  0.485 -0.0126  0.00628
#>  7  0.142    0.479 -0.123  -0.0576 
#>  8  0.0756   0.384 -0.117   0.0383 
#>  9  0.0416   0.243 -0.0719  0.120  
#> 10 -0.0244   0.363 -0.0132 -0.0324 
#> 11 -0.0192  -0.325  0.0108  0.0116 
#> 12 -0.101   -0.503 -0.0666  0.0372 
#> 13  0.132   -0.263 -0.139  -0.213  
#> ---------------------------------------------------------------------------
#> Genotype-vs-environment statistics
#> ---------------------------------------------------------------------------
#> # A tibble: 13 x 7
#>    GEN    Mean Variance SQE_Gi MSE_Gi  Fcal    P_val
#>    <chr> <dbl>    <dbl>  <dbl>  <dbl> <dbl>    <dbl>
#>  1 H1     2.62    0.286  0.857  0.286 12.7  4.49e- 7
#>  2 H10    2.31    0.372  1.12   0.372 16.6  9.20e- 9
#>  3 H11    2.39    0.301  0.904  0.301 13.4  2.17e- 7
#>  4 H12    2.44    0.193  0.579  0.193  8.60 4.09e- 5
#>  5 H13    2.54    0.155  0.466  0.155  6.93 2.86e- 4
#>  6 H2     2.60    0.362  1.08   0.362 16.1  1.45e- 8
#>  7 H3     2.59    0.558  1.67   0.558 24.9  5.41e-12
#>  8 H4     2.58    0.436  1.31   0.436 19.5  6.18e-10
#>  9 H5     2.57    0.319  0.956  0.319 14.2  9.74e- 8
#> 10 H6     2.56    0.285  0.854  0.285 12.7  4.70e- 7
#> 11 H7     2.40    0.275  0.826  0.275 12.3  7.23e- 7
#> 12 H8     2.33    0.372  1.12   0.372 16.6  9.27e- 9
#> 13 H9     2.36    0.459  1.38   0.459 20.5  2.50e-10
#> ---------------------------------------------------------------------------
#> Regression-based stability (anova)
#> ---------------------------------------------------------------------------
#> # A tibble: 20 x 6
#>    SV                       Df `Sum Sq` `Mean Sq` `F value`    `Pr(>F)`
#>    <chr>                 <dbl>    <dbl>     <dbl>     <dbl>       <dbl>
#>  1 Total                    51   15.0      0.294     NA     NA         
#>  2 GEN                      12    1.86     0.155      0.870  0.585     
#>  3 ENV + (GEN x ENV)        39   13.1      0.336     NA     NA         
#>  4 ENV (linear)              1    7.72     7.72      NA     NA         
#>  5 " GEN x ENV (linear)"    12    0.755    0.0629     0.352  0.969     
#>  6 Pooled deviation         26    4.64     0.179     NA     NA         
#>  7 H1                        2    0.471    0.235     11.4    0.0000343 
#>  8 H10                       2    0.236    0.118      5.69   0.00452   
#>  9 H11                       2    0.211    0.106      5.11   0.00767   
#> 10 H12                       2    0.450    0.225     10.9    0.0000515 
#> 11 H13                       2    0.410    0.205      9.91   0.000115  
#> 12 H2                        2    0.534    0.267     12.9    0.00000994
#> 13 H3                        2    0.584    0.292     14.1    0.00000381
#> 14 H4                        2    0.354    0.177      8.55   0.000365  
#> 15 H5                        2    0.140    0.0701     3.39   0.0376    
#> 16 H6                        2    0.334    0.167      8.06   0.000559  
#> 17 H7                        2    0.241    0.121      5.83   0.00398   
#> 18 H8                        2    0.505    0.253     12.2    0.0000174 
#> 19 H9                        2    0.172    0.0861     4.16   0.0183    
#> 20 Pooled error            104    2.15     0.0207    NA     NA         
#> ---------------------------------------------------------------------------
#> Regression-based stability (parameters)
#> ---------------------------------------------------------------------------
#> # A tibble: 13 x 6
#>    GEN    Mean bij[,1] sdij[,1]  RMSE    R2
#>    <chr> <dbl>   <dbl>    <dbl> <dbl> <dbl>
#>  1 H1     2.62   0.806   0.0710 0.198 0.451
#>  2 H10    2.31   1.22    0.0318 0.140 0.789
#>  3 H11    2.39   1.08    0.0278 0.133 0.766
#>  4 H12    2.44   0.465   0.0676 0.194 0.222
#>  5 H13    2.54   0.306   0.0609 0.185 0.119
#>  6 H2     2.60   0.963   0.0815 0.211 0.508
#>  7 H3     2.59   1.35    0.0898 0.221 0.651
#>  8 H4     2.58   1.27    0.0515 0.172 0.730
#>  9 H5     2.57   1.17    0.0159 0.108 0.853
#> 10 H6     2.56   0.936   0.0481 0.167 0.609
#> 11 H7     2.40   0.992   0.0328 0.142 0.708
#> 12 H8     2.33   1.01    0.0767 0.205 0.547
#> 13 H9     2.36   1.42    0.0212 0.120 0.875
#> ---------------------------------------------------------------------------
#> Genotypic confidence index (all environments)
#> ---------------------------------------------------------------------------
#> # A tibble: 13 x 6
#>    Genotype  Mean Mean_rp Sd_rp    Wi  rank
#>    <chr>    <dbl>   <dbl> <dbl> <dbl> <dbl>
#>  1 H1        2.62   106.   9.38  90.2     2
#>  2 H10       2.31    93.0  6.93  81.6    11
#>  3 H11       2.39    96.1  6.27  85.8     8
#>  4 H12       2.44    98.6 11.2   80.1    12
#>  5 H13       2.54   103.  12.4   82.4    10
#>  6 H2        2.60   105.   9.91  88.5     5
#>  7 H3        2.59   104.  11.2   85.8     6
#>  8 H4        2.58   104.   8.76  89.2     4
#>  9 H5        2.57   103.   5.55  94.1     1
#> 10 H6        2.56   103.   7.83  90.1     3
#> 11 H7        2.40    96.8  6.67  85.8     7
#> 12 H8        2.33    93.6  9.61  77.8    13
#> 13 H9        2.36    94.8  6.85  83.5     9
#> ---------------------------------------------------------------------------
#> Wricke's Ecovalence
#> ---------------------------------------------------------------------------
#> # A tibble: 13 x 7
#>         A1     A2      A3      A4 Ecoval Ecov_perc  rank
#>  *   <dbl>  <dbl>   <dbl>   <dbl>  <dbl>     <dbl> <dbl>
#>  1 -0.207   0.332 -0.106  -0.0189  0.493      9.13     8
#>  2  0.160  -0.243  0.0414  0.0409  0.264      4.89     4
#>  3  0.0504 -0.220  0.0296  0.140   0.215      3.99     2
#>  4 -0.0531 -0.321  0.311   0.0636  0.620     11.5     11
#>  5 -0.0749 -0.288  0.378  -0.0149  0.696     12.9     13
#>  6 -0.120   0.365 -0.132  -0.113   0.535      9.91    10
#>  7  0.0322  0.369 -0.233  -0.168   0.659     12.2     12
#>  8 -0.0196  0.289 -0.212  -0.0569  0.397      7.35     7
#>  9 -0.0416  0.160 -0.155   0.0371  0.158      2.92     1
#> 10 -0.0976  0.290 -0.0864 -0.106   0.336      6.23     6
#> 11  0.0613 -0.245  0.0913  0.0921  0.241      4.47     3
#> 12  0.0572 -0.344  0.0917  0.195   0.505      9.36     9
#> 13  0.253  -0.142 -0.0181 -0.0926  0.279      5.17     5
#> ---------------------------------------------------------------------------
#> Nonparametric Superiority index
#> ---------------------------------------------------------------------------
#> # A tibble: 13 x 7
#>    Genotypes Pi_all  Or_a Pi_favorable  Or_f Pi_unfavorable  Or_u
#>    <chr>      <dbl> <dbl>        <dbl> <dbl>          <dbl> <dbl>
#>  1 H1        0.0258     1      0.0113      8         0.0403     1
#>  2 H10       0.151     12      0.0213     12         0.280     12
#>  3 H11       0.116     10      0.0101      7         0.222     11
#>  4 H12       0.103      8      0.0174     11         0.189      8
#>  5 H13       0.0688     7      0.00832     4         0.129      7
#>  6 H2        0.0289     2      0.00839     5         0.0494     2
#>  7 H3        0.0425     6      0.00791     3         0.0771     4
#>  8 H4        0.0404     5      0.00279     2         0.0780     5
#>  9 H5        0.0403     4      0.00253     1         0.0781     6
#> 10 H6        0.0330     3      0.0128      9         0.0533     3
#> 11 H7        0.109      9      0.00947     6         0.208      9
#> 12 H8        0.161     13      0.0165     10         0.306     13
#> 13 H9        0.124     11      0.0278     13         0.221     10