Print an object of class ge_stats

Print the ge_stats object in two ways. By default, the results are shown in the R console. The results can also be exported to the directory into a *.txt file.

# S3 method for ge_stats
print(x, what = "all", export = FALSE, file.name = NULL, digits = 3, ...)

Arguments

x
An object of class ge_stats.

what
What should be printed. what = "all" for both statistics and ranks, what = "stats" for statistics, and what = "ranks" for ranks.

export
A logical argument. If TRUE, a *.txt file is exported to the working directory.

file.name
The name of the file if export = TRUE

digits
The significant digits to be shown.

...
Options used by the tibble package to format the output. See tibble::print() for more details.

Author

Tiago Olivoto tiagoolivoto@gmail.com

Examples

# \donttest{
library(metan)
model <- ge_stats(data_ge, ENV, GEN, REP, GY)
#> Evaluating trait GY |============================================| 100% 00:00:04 
print(model)
#> Variable GY 
#> ---------------------------------------------------------------------------
#> Stability statistics and ranks
#> ---------------------------------------------------------------------------
#> # A tibble: 10 x 84
#>    GEN       Y   Y_R    CV  CV_R   ACV ACV_R   POLAR POLAR_R   Var Var_R Shukla
#>    <chr> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>   <dbl>   <dbl> <dbl> <dbl>  <dbl>
#>  1 G1     2.60     6  35.2     8  34.1     6  0.0298       6 10.9      7 0.0280
#>  2 G10    2.47    10  42.3     9  38.6     9  0.136        9 14.2      9 0.244 
#>  3 G2     2.74     3  34.0     7  35.2     8  0.0570       8 11.3      8 0.0861
#>  4 G3     2.96     2  29.9     3  33.8     5  0.0216       5 10.1      5 0.0121
#>  5 G4     2.64     5  31.4     6  31.0     4 -0.0537       4  8.93     4 0.0640
#>  6 G5     2.54     7  30.6     5  28.8     3 -0.119        3  7.82     3 0.0480
#>  7 G6     2.53     8  29.7     2  27.8     1 -0.147        1  7.34     2 0.0468
#>  8 G7     2.74     4  27.4     1  28.3     2 -0.133        2  7.33     1 0.122 
#>  9 G8     3.00     1  30.4     4  35.1     7  0.0531       7 10.8      6 0.0712
#> 10 G9     2.51     9  42.4    10  39.4    10  0.154       10 14.7     10 0.167 
#> # ... with 72 more variables: Shukla_R <dbl>, Wi_g <dbl>, Wi_g_R <dbl>,
#> #   Wi_f <dbl>, Wi_f_R <dbl>, Wi_u <dbl>, Wi_u_R <dbl>, Ecoval <dbl>,
#> #   Ecoval_R <dbl>, bij <dbl>, Sij <dbl>, Sij_R <dbl>, R2 <dbl>, R2_R <dbl>,
#> #   ASTAB <dbl>, ASTAB_R <dbl>, ASI <dbl>, ASI_R <dbl>, ASV <dbl>, ASV_R <dbl>,
#> #   AVAMGE <dbl>, AVAMGE_R <dbl>, DA <dbl>, DA_R <dbl>, DZ <dbl>, DZ_R <dbl>,
#> #   EV <dbl>, EV_R <dbl>, FA <dbl>, FA_R <dbl>, MASI <dbl>, MASI_R <dbl>,
#> #   MASV <dbl>, MASV_R <dbl>, SIPC <dbl>, SIPC_R <dbl>, ZA <dbl>, ...
#> ---------------------------------------------------------------------------
#> 
#> 
#> 
# }