Print an object of class mgidi Print a mgidi object in two ways. By default, the results are shown in the R console. The results can also be exported to the directory.

Print an object of class mgidi Print a mgidi object in two ways. By default, the results are shown in the R console. The results can also be exported to the directory.

# S3 method for mgidi
print(x, export = FALSE, file.name = NULL, digits = 4, ...)

Arguments

x
An object of class mgidi.

export
A logical argument. If TRUE|T, a *.txt file is exported to the working directory

file.name
The name of the file if export = TRUE

digits
The significant digits to be shown.

...
Options used by the tibble package to format the output. See tibble::print() for more details.

Examples

# \donttest{
library(metan)
model <- gamem(data_g,
               gen = GEN,
               rep = REP,
               resp = c(KW, NR, NKE, NKR))

#> Model:  Y ~ REP + (1 | GEN) 
#> ---------------------------------------------------------------------------
#> P-values for Likelihood Ratio Test of the analyzed traits
#> ---------------------------------------------------------------------------
#>     model     KW     NR     NKE   NKR
#>  Complete     NA     NA      NA    NA
#>  Genotype 0.0253 0.0056 0.00952 0.216
#> ---------------------------------------------------------------------------
#> Variables with nonsignificant Genotype effect
#> NKR 
#> ---------------------------------------------------------------------------

mgidi_index <- mgidi(model)

#> 
#> -------------------------------------------------------------------------------
#> Principal Component Analysis
#> -------------------------------------------------------------------------------
#> # A tibble: 4 x 4
#>   PC    Eigenvalues `Variance (%)` `Cum. variance (%)`
#>   <chr>       <dbl>          <dbl>               <dbl>
#> 1 PC1        2.42            60.6                 60.6
#> 2 PC2        1.19            29.8                 90.3
#> 3 PC3        0.320            8.00                98.3
#> 4 PC4        0.0663           1.66               100  
#> -------------------------------------------------------------------------------
#> Factor Analysis - factorial loadings after rotation-
#> -------------------------------------------------------------------------------
#> # A tibble: 4 x 5
#>   VAR       FA1     FA2 Communality Uniquenesses
#>   <chr>   <dbl>   <dbl>       <dbl>        <dbl>
#> 1 KW    -0.904   0.0400       0.819       0.181 
#> 2 NR    -0.923  -0.118        0.867       0.133 
#> 3 NKE   -0.697  -0.686        0.956       0.0439
#> 4 NKR    0.0463 -0.985        0.972       0.0284
#> -------------------------------------------------------------------------------
#> Comunalit Mean: 0.9033994 
#> -------------------------------------------------------------------------------
#> Multi-trait genotype-ideotype distance index
#> -------------------------------------------------------------------------------
#>    H13     H5    H12    H11     H2     H7     H1     H4    H10     H3     H6 
#> 0.5181 1.1120 2.0926 2.2855 2.6419 2.7675 3.0717 3.1662 3.2112 3.3374 3.6255 
#>     H8     H9 
#> 3.6525 3.7268 
#> -------------------------------------------------------------------------------
#> Selection differential 
#> -------------------------------------------------------------------------------
#> # A tibble: 4 x 6
#>   VAR   Factor    xo    Xs     SD SDperc
#>   <chr> <chr>  <dbl> <dbl>  <dbl>  <dbl>
#> 1 KW    FA 1   147.  163.  16.2    11.0 
#> 2 NR    FA 1    15.8  17.4  1.63   10.3 
#> 3 NKE   FA 1   468.  532.  64.0    13.7 
#> 4 NKR   FA 2    30.4  31.2  0.814   2.68
#> ------------------------------------------------------------------------------
#> Selected genotypes
#> -------------------------------------------------------------------------------
#> H13 H5
#> -------------------------------------------------------------------------------

print(mgidi_index)

#> -------------------------------------------------------------------------------
#> Correlation matrix used used in factor analysis 
#> -------------------------------------------------------------------------------
#>              KW         NR       NKE         NKR
#> KW   1.00000000 0.68774136 0.5601661 -0.02740098
#> NR   0.68774136 1.00000000 0.7310000  0.04724335
#> NKE  0.56016613 0.73100001 1.0000000  0.61024279
#> NKR -0.02740098 0.04724335 0.6102428  1.00000000
#> -------------------------------------------------------------------------------
#> Principal component analysis 
#> -------------------------------------------------------------------------------
#> # A tibble: 4 x 4
#>   PC    Eigenvalues `Variance (%)` `Cum. variance (%)`
#>   <chr>       <dbl>          <dbl>               <dbl>
#> 1 PC1       2.422           60.55                60.55
#> 2 PC2       1.192           29.79                90.34
#> 3 PC3       0.3201           8.002               98.34
#> 4 PC4       0.06631          1.658              100   
#> -------------------------------------------------------------------------------
#> Initial loadings 
#> -------------------------------------------------------------------------------
#> # A tibble: 4 x 3
#>   VAR       PC1     PC2
#>   <chr>   <dbl>   <dbl>
#> 1 KW    -0.7827  0.4545
#> 2 NR    -0.8731  0.3232
#> 3 NKE   -0.9353 -0.2852
#> 4 NKR   -0.4153 -0.8939
#> -------------------------------------------------------------------------------
#> Loadings after varimax rotation 
#> -------------------------------------------------------------------------------
#> # A tibble: 4 x 3
#>   VAR        FA1      FA2
#>   <chr>    <dbl>    <dbl>
#> 1 KW    -0.9042   0.04004
#> 2 NR    -0.9235  -0.1182 
#> 3 NKE   -0.6966  -0.6862 
#> 4 NKR    0.04631 -0.9846 
#> -------------------------------------------------------------------------------
#> Scores for genotypes-ideotype 
#> -------------------------------------------------------------------------------
#> # A tibble: 14 x 3
#>    GEN        FA1     FA2
#>    <chr>    <dbl>   <dbl>
#>  1 H1    -1.615   -0.2874
#>  2 H10   -0.2495  -2.888 
#>  3 H11   -1.220   -2.278 
#>  4 H12   -1.813   -1.450 
#>  5 H13   -3.589   -2.244 
#>  6 H2    -1.816   -0.6748
#>  7 H3    -0.5856  -1.045 
#>  8 H4    -0.2922  -2.673 
#>  9 H5    -2.432   -2.315 
#> 10 H6    -1.405    0.2438
#> 11 H7    -1.449   -0.8412
#> 12 H8    -0.6321  -0.4303
#> 13 H9     0.01742 -1.398 
#> 14 ID1   -3.458   -2.745 
#> -------------------------------------------------------------------------------
#> Multi-trait genotype-ideotype distance index 
#> -------------------------------------------------------------------------------
#>       H13        H5       H12       H11        H2        H7        H1        H4 
#> 0.5181182 1.1119885 2.0926188 2.2855330 2.6418504 2.7674832 3.0716616 3.1661570 
#>       H10        H3        H6        H8        H9 
#> 3.2112047 3.3373673 3.6255180 3.6525397 3.7268215 
#> -------------------------------------------------------------------------------
#> Selection differential 
#> -------------------------------------------------------------------------------
#> # A tibble: 4 x 6
#>   VAR   Factor     xo     Xs      SD SDperc
#>   <chr> <chr>   <dbl>  <dbl>   <dbl>  <dbl>
#> 1 KW    FA 1   146.8  163.0  16.19   11.03 
#> 2 NR    FA 1    15.78  17.42  1.630  10.33 
#> 3 NKE   FA 1   467.9  531.9  63.97   13.67 
#> 4 NKR   FA 2    30.40  31.21  0.8142  2.678
#> -------------------------------------------------------------------------------
#> Selected genotypes 
#> -------------------------------------------------------------------------------
#> H13 H5

# }