Print an object of class mtsi

Print a mtsi object in two ways. By default, the results are shown in the R console. The results can also be exported to the directory.

# S3 method for mtsi
print(x, export = FALSE, file.name = NULL, digits = 4,
  ...)

Arguments

x
An object of class mtsi.

export
A logical argument. If TRUE|T, a *.txt file is exported to the working directory

file.name
The name of the file if export = TRUE

digits
The significant digits to be shown.

...
Options used by the tibble package to format the output. See trunc_mat for more details.

Examples

library(metan)
# Based on stability only
MTSI_MODEL <- waasb(data_ge,
  resp = c(GY, HM),
  gen = GEN,
  env = ENV,
  rep = REP
)

#> Evaluating variable GY 50 % 
#> Evaluating variable HM 100 % 
#> All variables with significant (p < 0.05) genotype-vs-environment interaction
#> Done!

MTSI_index <- mtsi(MTSI_MODEL)

#> 
#> -------------------------------------------------------------------------------
#> Principal Component Analysis
#> -------------------------------------------------------------------------------
#> # A tibble: 2 x 4
#>   PC    Eigenvalues `Variance (%)` `Cum. variance (%)`
#>   <chr>       <dbl>          <dbl>               <dbl>
#> 1 PC1         1.66            82.8                82.8
#> 2 PC2         0.343           17.2               100  
#> -------------------------------------------------------------------------------
#> Factor Analysis - factorial loadings after rotation-
#> -------------------------------------------------------------------------------
#> # A tibble: 2 x 4
#>   VAR     FA1 Communality Uniquenesses
#>   <chr> <dbl>       <dbl>        <dbl>
#> 1 GY    0.910       0.828        0.172
#> 2 HM    0.910       0.828        0.172
#> -------------------------------------------------------------------------------
#> Comunalit Mean: 0.8283139 
#> -------------------------------------------------------------------------------
#> Multitrait stability index
#> -------------------------------------------------------------------------------
#>     G3     G1     G6     G8     G4     G2     G9     G5     G7    G10 
#> 0.0000 0.2303 0.4846 1.0066 1.2954 1.6022 1.8318 1.8374 1.8450 3.4178 
#> -------------------------------------------------------------------------------
#> Selection differential for the  waasb index
#> -------------------------------------------------------------------------------
#> # A tibble: 2 x 6
#>   VAR   Factor    Xo    Xs     SD SDperc
#>   <chr> <chr>  <dbl> <dbl>  <dbl>  <dbl>
#> 1 GY    FA 1   0.250 0.117 -0.134  -53.5
#> 2 HM    FA 1   0.614 0.373 -0.241  -39.2
#> ------------------------------------------------------------------------------
#> Mean of selection differential
#> -------------------------------------------------------------------------------
#>          Xo          Xs          SD      SDperc 
#>   0.4323192   0.2448716  -0.1874476 -46.3574863 
#> -------------------------------------------------------------------------------
#> Selection differential for the mean of the variables
#> -------------------------------------------------------------------------------
#> # A tibble: 2 x 6
#>   VAR   Factor    xo    Xs     SD SDperc
#>   <chr> <chr>  <dbl> <dbl>  <dbl>  <dbl>
#> 1 GY    FA 1    2.67  2.78  0.105   3.94
#> 2 HM    FA 1   48.1  47.3  -0.747  -1.55
#> ------------------------------------------------------------------------------
#> Selected genotypes
#> -------------------------------------------------------------------------------
#> G3 G1
#> -------------------------------------------------------------------------------

print(MTSI_index)

#> -------------------- Correlation matrix used used in factor analysis -----------------
#>           GY        HM
#> GY 1.0000000 0.6566277
#> HM 0.6566277 1.0000000
#> 
#> ---------------------------- Principal component analysis -----------------------------
#> # A tibble: 2 x 4
#>   PC    Eigenvalues `Variance (%)` `Cum. variance (%)`
#>   <chr>       <dbl>          <dbl>               <dbl>
#> 1 PC1        1.657           82.83               82.83
#> 2 PC2        0.3434          17.17              100   
#> 
#> --------------------------------- Initial loadings -----------------------------------
#> # A tibble: 2 x 2
#>   VAR   initial.loadings
#>   <chr>            <dbl>
#> 1 GY              0.9101
#> 2 HM              0.9101
#> 
#> -------------------------- Loadings after varimax rotation ---------------------------
#> # A tibble: 2 x 2
#>   VAR      FA1
#>   <chr>  <dbl>
#> 1 GY    0.9101
#> 2 HM    0.9101
#> 
#> --------------------------- Scores for genotypes-ideotype -----------------------------
#> # A tibble: 11 x 2
#>    GEN     FA1
#>    <chr> <dbl>
#>  1 G1    1.585
#>  2 G10   4.772
#>  3 G2    2.957
#>  4 G3    1.355
#>  5 G4    2.650
#>  6 G5    3.192
#>  7 G6    1.839
#>  8 G7    3.200
#>  9 G8    2.361
#> 10 G9    3.186
#> 11 ID1   1.355
#> 
#> ---------------------------- Multitrait stability index ------------------------------
#>        G3        G1        G6        G8        G4        G2        G9        G5 
#> 0.0000000 0.2302629 0.4845581 1.0066251 1.2954053 1.6021844 1.8318275 1.8374089 
#>        G7       G10 
#> 1.8450285 3.4177760 
#> 
#> --------------------------- Selection differential (index) ----------------------------
#> # A tibble: 2 x 6
#>   VAR   Factor     Xo     Xs      SD SDperc
#>   <chr> <chr>   <dbl>  <dbl>   <dbl>  <dbl>
#> 1 GY    FA 1   0.2505 0.1165 -0.1340 -53.49
#> 2 HM    FA 1   0.6142 0.3732 -0.2409 -39.23
#> 
#> -------------------------- Mean of Selection differential -----------------------------
#>          Xo          Xs          SD      SDperc 
#>   0.4323192   0.2448716  -0.1874476 -46.3574863 
#> 
#> ------------------------- Selection differential (variables) --------------------------
#> # A tibble: 2 x 6
#>   VAR   Factor     xo     Xs      SD SDperc
#>   <chr> <chr>   <dbl>  <dbl>   <dbl>  <dbl>
#> 1 GY    FA 1    2.674  2.780  0.1053  3.936
#> 2 HM    FA 1   48.09  47.34  -0.7465 -1.552
#> 
#> -------------------------------- Selected genotypes -----------------------------------
#> G3 G1