Print an object of class path_coeff

Print an object generated by the function 'path_coeff()'. By default, the results are shown in the R console. The results can also be exported to the directory.

# S3 method for path_coeff
print(x, export = FALSE, file.name = NULL, digits = 4, ...)

Arguments

x
An object of class path_coeff or group_path.

export
A logical argument. If TRUE, a *.txt file is exported to the working directory

file.name
The name of the file if export = TRUE

digits
The significant digits to be shown.

...
Options used by the tibble package to format the output. See tibble::print() for more details.

Examples

# \donttest{
library(metan)

# KW as dependent trait and all others as predictors
pcoeff <- path_coeff(data_ge2, resp = KW)

#> Severe multicollinearity. 
#> Condition Number = 7865.84
#> Please, consider using a correction factor, or use 'brutstep = TRUE'. 

print(pcoeff)

#> ----------------------------------------------------------------------------------------------
#> Correlation matrix between the predictor traits
#> ----------------------------------------------------------------------------------------------
#> # A tibble: 14 x 14
#>          PH       EH       EP       EL       ED       CL       CD      CW
#>  *    <dbl>    <dbl>    <dbl>    <dbl>    <dbl>    <dbl>    <dbl>   <dbl>
#>  1  1        0.9318   0.6384   0.3802   0.6613   0.3252   0.3154   0.5047
#>  2  0.9318   1        0.8695   0.3627   0.6303   0.3972   0.2805   0.5193
#>  3  0.6384   0.8695   1        0.2634   0.4580   0.3908   0.1750   0.4248
#>  4  0.3802   0.3627   0.2634   1        0.3851   0.2554   0.9119   0.4582
#>  5  0.6613   0.6303   0.4580   0.3851   1        0.6975   0.3897   0.7371
#>  6  0.3252   0.3972   0.3908   0.2554   0.6975   1        0.3004   0.7383
#>  7  0.3154   0.2805   0.1750   0.9119   0.3897   0.3004   1        0.4840
#>  8  0.5047   0.5193   0.4248   0.4582   0.7371   0.7383   0.4840   1     
#>  9  0.3286   0.2648   0.1404  -0.01387  0.5525   0.2619  -0.03585  0.1657
#> 10  0.3530   0.3311   0.2588   0.6172   0.2221  -0.1149   0.5933   0.3403
#> 11 -0.1920  -0.06591  0.08966 -0.01258 -0.01004  0.7080   0.04531  0.2999
#> 12  0.04081 -0.02135 -0.08709  0.03526 -0.2244  -0.5731  -0.04820 -0.6811
#> 13  0.5685   0.5624   0.4263   0.4421   0.6420   0.6187   0.4433   0.6735
#> 14  0.4584   0.3881   0.2331   0.4657   0.5051   0.04894  0.4156   0.3463
#> # ... with 6 more variables: NR <dbl>, NKR <dbl>, CDED <dbl>, PERK <dbl>,
#> #   TKW <dbl>, NKE <dbl>
#> ----------------------------------------------------------------------------------------------
#> Vector of correlations between dependent and each predictor
#> ----------------------------------------------------------------------------------------------
#>           PH        EH        EP        EL        ED       CL        CD
#> KW 0.7534439 0.7029469 0.4974193 0.6685601 0.8241426 0.470931 0.6259806
#>           CW        NR       NKR      CDED        PERK       TKW       NKE
#> KW 0.7348622 0.3621447 0.5973701 -0.147029 -0.02683251 0.6730371 0.6810756
#> ----------------------------------------------------------------------------------------------
#> Multicollinearity diagnosis and goodness-of-fit
#> ----------------------------------------------------------------------------------------------
#> Condition number:  7865.84 
#> Determinant:       0 
#> R-square:          0.9889 
#> Residual:          0.0111 
#> Response:          KW 
#> Predictors:        PH EH EP EL ED CL CD CW NR NKR CDED PERK TKW NKE 
#> ----------------------------------------------------------------------------------------------
#> Variance inflation factors
#> ----------------------------------------------------------------------------------------------
#> # A tibble: 14 x 2
#>    VAR       VIF
#>    <chr>   <dbl>
#>  1 PH    123.6  
#>  2 EH    278.3  
#>  3 EP     60.31 
#>  4 EL      7.569
#>  5 ED    351.3  
#>  6 CL    665.1  
#>  7 CD      7.219
#>  8 CW     46.46 
#>  9 NR      5.768
#> 10 NKR     6.451
#> 11 CDED  327.1  
#> 12 PERK   18.24 
#> 13 TKW    19.69 
#> 14 NKE    26.07 
#> ----------------------------------------------------------------------------------------------
#> Eigenvalues and eigenvectors
#> ----------------------------------------------------------------------------------------------
#> # A tibble: 14 x 15
#>    Eigenvalues       PH        EH        EP        EL       ED       CL
#>          <dbl>    <dbl>     <dbl>     <dbl>     <dbl>    <dbl>    <dbl>
#>  1   5.702     -0.3344  -3.424e-1 -0.2805   -0.2756   -0.3578  -0.2791 
#>  2   2.960      0.1284   6.995e-2 -0.007025  0.1463   -0.02762 -0.3756 
#>  3   1.819     -0.2329  -2.507e-1 -0.2220    0.4533   -0.1981  -0.06882
#>  4   1.364      0.2629   3.518e-1  0.3861   -0.006280 -0.2013  -0.2008 
#>  5   0.7747    -0.03192  1.044e-1  0.2783   -0.2332   -0.2196  -0.1746 
#>  6   0.6654    -0.1246   1.278e-1  0.4317    0.1863   -0.2731   0.1965 
#>  7   0.2473     0.5967   2.245e-1 -0.3831    0.1622   -0.3374  -0.1759 
#>  8   0.2214    -0.2918   4.341e-4  0.4050    0.2985    0.1087  -0.1990 
#>  9   0.09185    0.1205   7.788e-2 -0.05355  -0.4210    0.4094   0.2413 
#> 10   0.08453    0.01898  2.316e-2 -0.07535   0.4072    0.2085   0.1458 
#> 11   0.05241    0.1106   2.531e-3 -0.1027    0.3856    0.2729   0.1481 
#> 12   0.01359   -0.01388 -3.245e-2  0.04255  -0.02922   0.03773 -0.08825
#> 13   0.002298   0.5063  -7.688e-1  0.3532    0.004010  0.08191 -0.09833
#> 14   0.0007249 -0.08362  1.219e-1 -0.05200  -0.002194  0.4992  -0.6910 
#> # ... with 8 more variables: CD <dbl>, CW <dbl>, NR <dbl>, NKR <dbl>,
#> #   CDED <dbl>, PERK <dbl>, TKW <dbl>, NKE <dbl>
#> ----------------------------------------------------------------------------------------------
#> Variables with the largest weight in the eigenvalue of smallest magnitude
#> ----------------------------------------------------------------------------------------------
#> CL > ED > CDED > EH > CW > PH > NKE > EP > TKW > PERK > NR > EL > NKR > CD 
#> ----------------------------------------------------------------------------------------------
#> Direct (diagonal) and indirect (off-diagonal) effects
#> ----------------------------------------------------------------------------------------------
#> # A tibble: 14 x 14
#>           PH        EH        EP        EL        ED        CL        CD
#>  *     <dbl>     <dbl>     <dbl>     <dbl>     <dbl>     <dbl>     <dbl>
#>  1  0.1344    0.1252    0.08580   5.109e-2  0.08887   0.04370   4.239e-2
#>  2 -0.1712   -0.1837   -0.1598   -6.664e-2 -0.1158   -0.07298  -5.154e-2
#>  3  0.06188   0.08428   0.09692   2.553e-2  0.04439   0.03788   1.697e-2
#>  4  0.01088   0.01038   0.007541  2.863e-2  0.01103   0.007311  2.610e-2
#>  5  0.4466    0.4256    0.3093    2.601e-1  0.6754    0.4710    2.632e-1
#>  6 -0.2934   -0.3583   -0.3526   -2.304e-1 -0.6293   -0.9022   -2.710e-1
#>  7 -0.005267 -0.004684 -0.002923 -1.523e-2 -0.006508 -0.005016 -1.670e-2
#>  8  0.2655    0.2732    0.2235    2.410e-1  0.3878    0.3884    2.546e-1
#>  9 -0.005372 -0.004329 -0.002296  2.268e-4 -0.009033 -0.004282  5.861e-4
#> 10  0.003360  0.003151  0.002464  5.874e-3  0.002114 -0.001094  5.647e-3
#> 11 -0.1196   -0.04106   0.05586  -7.836e-3 -0.006251  0.4410    2.822e-2
#> 12  0.01420  -0.007429 -0.03030   1.227e-2 -0.07808  -0.1994   -1.677e-2
#> 13  0.2268    0.2244    0.1701    1.764e-1  0.2561    0.2468    1.769e-1
#> 14  0.1846    0.1563    0.09385   1.875e-1  0.2034    0.01971   1.674e-1
#> # ... with 7 more variables: CW <dbl>, NR <dbl>, NKR <dbl>, CDED <dbl>,
#> #   PERK <dbl>, TKW <dbl>, NKE <dbl>
#> ----------------------------------------------------------------------------------------------

# Call the algorithm for selecting a set of predictors
# With minimal multicollinearity (no VIF larger than 5)
pcoeff2 <- path_coeff(data_ge2,
                      resp = KW,
                      brutstep = TRUE,
                      maxvif = 5)

#> --------------------------------------------------------------------------
#> The algorithm has selected a set of 8 predictors with largest VIF = 3.346. 
#> Selected predictors: NR PERK EP CDED EL NKR TKW PH 
#> A forward stepwise-based selection procedure will fit 6 models.
#> --------------------------------------------------------------------------
#> Adjusting the model 1 with 7 predictors (16.67% concluded)
#> Adjusting the model 2 with 6 predictors (33.33% concluded)
#> Adjusting the model 3 with 5 predictors (50% concluded)
#> Adjusting the model 4 with 4 predictors (66.67% concluded)
#> Adjusting the model 5 with 3 predictors (83.33% concluded)
#> Adjusting the model 6 with 2 predictors (100% concluded)
#> Done!
#> --------------------------------------------------------------------------
#> Summary of the adjusted models 
#> --------------------------------------------------------------------------
#>    Model  AIC Numpred    CN Determinant    R2 Residual maxVIF
#>  MODEL_1 1127       7 13.67      0.0841 0.933   0.0669   2.59
#>  MODEL_2 1125       6 12.26      0.1383 0.933   0.0670   2.46
#>  MODEL_3 1126       5 12.05      0.1989 0.932   0.0683   2.31
#>  MODEL_4 1138       4  6.40      0.4595 0.925   0.0747   2.13
#>  MODEL_5 1148       3  1.34      0.9787 0.919   0.0808   1.02
#>  MODEL_6 1329       2  2.23      0.8555 0.738   0.2616   1.17
#> --------------------------------------------------------------------------
#> 

print(pcoeff2)

#> # A tibble: 6 x 8
#>   Model     AIC Numpred     CN Determinant     R2 Residual maxVIF
#>   <chr>   <dbl>   <dbl>  <dbl>       <dbl>  <dbl>    <dbl>  <dbl>
#> 1 MODEL_1 1127.       7 13.67      0.08406 0.9331  0.06694  2.592
#> 2 MODEL_2 1125.       6 12.26      0.1383  0.9330  0.06695  2.461
#> 3 MODEL_3 1126.       5 12.05      0.1989  0.9317  0.06825  2.310
#> 4 MODEL_4 1138.       4  6.402     0.4595  0.9253  0.07474  2.130
#> 5 MODEL_5 1148.       3  1.339     0.9787  0.9192  0.08077  1.021
#> 6 MODEL_6 1329.       2  2.227     0.8555  0.7384  0.2616   1.169

#> Go to 'pcoeff2 > s' to select a specific model

# }