Print an object of class performs_ammi

Print the performs_ammi object in two ways. By default, the results are shown in the R console. The results can also be exported to the directory.

# S3 method for performs_ammi
print(x, export = FALSE, file.name = NULL, digits = 4, ...)

Arguments

x
An object of class performs_ammi.

export
A logical argument. If TRUE, a *.txt file is exported to the working directory

file.name
The name of the file if export = TRUE

digits
The significant digits to be shown.

...
Options used by the tibble package to format the output. See tibble::print() for more details.

Examples

library(metan)
model = performs_ammi(data_ge, ENV, GEN, REP,
                      resp = c(GY, HM))

#> variable GY 
#> ---------------------------------------------------------------------------
#> AMMI analysis table
#> ---------------------------------------------------------------------------
#>     Source  Df  Sum Sq Mean Sq F value   Pr(>F) Percent Accumul
#>        ENV  13 279.574 21.5057   62.33 0.00e+00       .       .
#>   REP(ENV)  28   9.662  0.3451    3.57 3.59e-08       .       .
#>   GEN        9  12.995  1.4439   14.93 2.19e-19       .       .
#>    ENV:GEN 117  31.220  0.2668    2.76 1.01e-11       .       .
#>        PC1  21  10.749  0.5119    5.29 0.00e+00    34.4    34.4
#>        PC2  19   9.924  0.5223    5.40 0.00e+00    31.8    66.2
#>        PC3  17   4.039  0.2376    2.46 1.40e-03    12.9    79.2
#>        PC4  15   3.074  0.2049    2.12 9.60e-03     9.8      89
#>        PC5  13   1.446  0.1113    1.15 3.18e-01     4.6    93.6
#>        PC6  11   0.932  0.0848    0.88 5.61e-01       3    96.6
#>        PC7   9   0.567  0.0630    0.65 7.53e-01     1.8    98.4
#>        PC8   7   0.362  0.0518    0.54 8.04e-01     1.2    99.6
#>        PC9   5   0.126  0.0252    0.26 9.34e-01     0.4     100
#>  Residuals 252  24.367  0.0967      NA       NA       .       .
#>      Total 419 357.816  0.8540      NA       NA       .       .
#> ---------------------------------------------------------------------------
#> 
#> variable HM 
#> ---------------------------------------------------------------------------
#> AMMI analysis table
#> ---------------------------------------------------------------------------
#>     Source  Df  Sum Sq Mean Sq F value   Pr(>F) Percent Accumul
#>        ENV  13 5710.32 439.255   57.22 1.11e-16       .       .
#>   REP(ENV)  28  214.93   7.676    2.70 2.20e-05       .       .
#>   GEN        9  269.81  29.979   10.56 7.41e-14       .       .
#>    ENV:GEN 117 1100.73   9.408    3.31 1.06e-15       .       .
#>        PC1  21  381.13  18.149    6.39 0.00e+00    34.6    34.6
#>        PC2  19  319.43  16.812    5.92 0.00e+00      29    63.6
#>        PC3  17  114.26   6.721    2.37 2.10e-03    10.4      74
#>        PC4  15   81.96   5.464    1.92 2.18e-02     7.4    81.5
#>        PC5  13   68.11   5.240    1.84 3.77e-02     6.2    87.7
#>        PC6  11   59.07   5.370    1.89 4.10e-02     5.4      93
#>        PC7   9   46.69   5.188    1.83 6.33e-02     4.2    97.3
#>        PC8   7   26.65   3.808    1.34 2.32e-01     2.4    99.7
#>        PC9   5    3.41   0.682    0.24 9.45e-01     0.3     100
#>  Residuals 252  715.69   2.840      NA       NA       .       .
#>      Total 419 8011.48  19.120      NA       NA       .       .
#> ---------------------------------------------------------------------------
#> 
#> All variables with significant (p < 0.05) genotype-vs-environment interaction
#> Done!

print(model)

#> Variable GY 
#> ---------------------------------------------------------------------------
#> AMMI analysis table
#> ---------------------------------------------------------------------------
#> # A tibble: 15 x 8
#>    Source         Df `Sum Sq` `Mean Sq` `F value`   `Pr(>F)` Percent Accumul
#>    <chr>       <dbl>    <dbl>     <dbl>     <dbl>      <dbl> <chr>   <chr>  
#>  1 "ENV"          13 279.6     21.51       62.33   0.        .       .      
#>  2 "REP(ENV)"     28   9.662    0.3451      3.569  3.593e- 8 .       .      
#>  3 "GEN     "      9  13.00     1.444      14.93   2.190e-19 .       .      
#>  4 "ENV:GEN"     117  31.22     0.2668      2.760  1.005e-11 .       .      
#>  5 "PC1"          21  10.75     0.5119      5.29   0.        34.4    34.4   
#>  6 "PC2"          19   9.924    0.5223      5.4    0.        31.8    66.2   
#>  7 "PC3"          17   4.039    0.2376      2.46   1.400e- 3 12.9    79.2   
#>  8 "PC4"          15   3.074    0.2049      2.12   9.600e- 3 9.8     89     
#>  9 "PC5"          13   1.446    0.1113      1.150  3.176e- 1 4.6     93.6   
#> 10 "PC6"          11   0.9322   0.08475     0.88   5.606e- 1 3       96.6   
#> 11 "PC7"           9   0.5667   0.06297     0.65   7.535e- 1 1.8     98.4   
#> 12 "PC8"           7   0.3623   0.05176     0.54   8.037e- 1 1.2     99.6   
#> 13 "PC9"           5   0.1259   0.02517     0.26   9.345e- 1 0.4     100    
#> 14 "Residuals"   252  24.37     0.09669    NA     NA         .       .      
#> 15 "Total"       419 357.8      0.8540     NA     NA         .       .      
#> ---------------------------------------------------------------------------
#> Scores for genotypes and environments
#> ---------------------------------------------------------------------------
#> # A tibble: 24 x 12
#>    type  Code      Y      PC1      PC2      PC3      PC4      PC5      PC6
#>    <chr> <fct> <dbl>    <dbl>    <dbl>    <dbl>    <dbl>    <dbl>    <dbl>
#>  1 GEN   G1    2.604  0.3166  -0.04417 -0.03600 -0.06595 -0.3125   0.4272 
#>  2 GEN   G10   2.471 -1.001   -0.5718  -0.1652  -0.3309  -0.1243  -0.1064 
#>  3 GEN   G2    2.744  0.1390   0.1988  -0.7331   0.4735  -0.04816 -0.2841 
#>  4 GEN   G3    2.955  0.04340 -0.1028   0.2284   0.1769  -0.1270  -0.1400 
#>  5 GEN   G4    2.642 -0.3251   0.4782  -0.09073  0.1417  -0.1924   0.3550 
#>  6 GEN   G5    2.537 -0.3260   0.2461   0.2452   0.1794   0.4662   0.03315
#>  7 GEN   G6    2.534 -0.09836  0.2429   0.5607   0.2377   0.05094 -0.1011 
#>  8 GEN   G7    2.741  0.2849   0.5871  -0.2068  -0.7085   0.2315  -0.08406
#>  9 GEN   G8    3.004  0.4995  -0.1916   0.3191  -0.1676  -0.3261  -0.2886 
#> 10 GEN   G9    2.510  0.4668  -0.8427  -0.1217   0.06385  0.3819   0.1889 
#> # ... with 14 more rows, and 3 more variables: PC7 <dbl>, PC8 <dbl>, PC9 <dbl>
#> 
#> 
#> 
#> Variable HM 
#> ---------------------------------------------------------------------------
#> AMMI analysis table
#> ---------------------------------------------------------------------------
#> # A tibble: 15 x 8
#>    Source         Df `Sum Sq` `Mean Sq` `F value`   `Pr(>F)` Percent Accumul
#>    <chr>       <dbl>    <dbl>     <dbl>     <dbl>      <dbl> <chr>   <chr>  
#>  1 "ENV"          13 5710.     439.3       57.22   1.110e-16 .       .      
#>  2 "REP(ENV)"     28  214.9      7.676      2.703  2.197e- 5 .       .      
#>  3 "GEN     "      9  269.8     29.98      10.56   7.415e-14 .       .      
#>  4 "ENV:GEN"     117 1101.       9.408      3.313  1.064e-15 .       .      
#>  5 "PC1"          21  381.1     18.15       6.39   0.        34.6    34.6   
#>  6 "PC2"          19  319.4     16.81       5.92   0.        29      63.6   
#>  7 "PC3"          17  114.3      6.721      2.37   2.100e- 3 10.4    74     
#>  8 "PC4"          15   81.96     5.464      1.92   2.180e- 2 7.4     81.5   
#>  9 "PC5"          13   68.11     5.240      1.84   3.770e- 2 6.2     87.7   
#> 10 "PC6"          11   59.07     5.370      1.89   4.100e- 2 5.4     93     
#> 11 "PC7"           9   46.69     5.188      1.83   6.330e- 2 4.2     97.3   
#> 12 "PC8"           7   26.65     3.808      1.34   2.318e- 1 2.4     99.7   
#> 13 "PC9"           5    3.409    0.6817     0.24   9.445e- 1 0.3     100    
#> 14 "Residuals"   252  715.7      2.840     NA     NA         .       .      
#> 15 "Total"       419 8011.      19.12      NA     NA         .       .      
#> ---------------------------------------------------------------------------
#> Scores for genotypes and environments
#> ---------------------------------------------------------------------------
#> # A tibble: 24 x 12
#>    type  Code      Y      PC1      PC2       PC3       PC4      PC5       PC6
#>    <chr> <fct> <dbl>    <dbl>    <dbl>     <dbl>     <dbl>    <dbl>     <dbl>
#>  1 GEN   G1    47.08  0.2800   0.4635   0.1740   -1.369    -1.135    0.03658 
#>  2 GEN   G10   48.51 -1.779    1.866   -0.006219  0.9219    0.1096  -0.009745
#>  3 GEN   G2    46.66  1.563    0.5518  -0.9357    0.4913    0.2843   1.184   
#>  4 GEN   G3    47.60  0.3417  -0.2012  -0.8001    0.3753   -0.4979  -1.294   
#>  5 GEN   G4    48.03 -0.2020  -1.841    0.2801    0.005954  0.8201   0.2734  
#>  6 GEN   G5    49.30  1.580    1.030    1.078    -0.2789    1.005   -0.7368  
#>  7 GEN   G6    48.73  0.5474  -0.2453   0.5324    0.4603   -1.008    0.5861  
#>  8 GEN   G7    47.97 -1.218   -0.4680   1.254    -0.05482  -0.03429  0.3366  
#>  9 GEN   G8    49.10 -0.04176 -1.241   -0.4105    0.6394   -0.1785  -0.5149  
#> 10 GEN   G9    47.90 -1.072    0.08563 -1.166    -1.191     0.6351   0.1393  
#> # ... with 14 more rows, and 3 more variables: PC7 <dbl>, PC8 <dbl>, PC9 <dbl>
#> 
#> 
#>