Print an object of class waas

Print the waas object in two ways. By default, the results are shown in the R console. The results can also be exported to the directory.

# S3 method for waas
print(x, export = FALSE, file.name = NULL, digits = 4, ...)

Arguments

x
An object of class waas.

export
A logical argument. If TRUE, a *.txt file is exported to the working directory

file.name
The name of the file if export = TRUE

digits
The significant digits to be shown.

...
Options used by the tibble package to format the output. See trunc_mat for more details.

Examples

# \donttest{
library(metan)
model <- waas(data_ge,
  resp = c(GY, HM),
  gen = GEN,
  env = ENV,
  rep = REP
)

#> Evaluating variable GY 0 % 
#> Evaluating variable HM 100 % 
#> All variables with significant (p < 0.05) genotype-vs-environment interaction
#> Done!

print(model)

#> Variable GY 
#> ---------------------------------------------------------------------------
#> Individual analysis of variance
#> ---------------------------------------------------------------------------
#> # A tibble: 14 x 12
#>    ENV    MEAN     MSB    MSG     MSR     FCB     PRFB    FCG     PRFG     CV
#>    <chr> <dbl>   <dbl>  <dbl>   <dbl>   <dbl>    <dbl>  <dbl>    <dbl>  <dbl>
#>  1 E1    2.521 0.06522 0.3374 0.1440   0.4529 6.428e-1  2.343 5.936e-2 15.05 
#>  2 E10   2.175 0.6537  0.2958 0.02667 24.51   7.275e-6 11.09  1.097e-5  7.508
#>  3 E11   1.368 0.3773  0.1509 0.1050   3.594  4.860e-2  1.438 2.443e-1 23.68 
#>  4 E12   1.609 0.09190 0.3202 0.05352  1.717  2.077e-1  5.982 6.474e-4 14.38 
#>  5 E13   2.910 0.07674 0.7134 0.09937  0.7722 4.767e-1  7.179 2.095e-4 10.83 
#>  6 E14   1.782 0.1035  0.1305 0.07533  1.374  2.784e-1  1.732 1.535e-1 15.40 
#>  7 E2    3.180 0.6983  0.2066 0.1785   3.912  3.885e-2  1.157 3.763e-1 13.29 
#>  8 E3    4.064 0.4886  0.3351 0.1789   2.731  9.209e-2  1.872 1.230e-1 10.41 
#>  9 E4    3.675 0.1165  0.5313 0.1377   0.8458 4.456e-1  3.859 7.123e-3 10.10 
#> 10 E5    3.910 0.2188  0.5261 0.06636  3.297  6.024e-2  7.929 1.101e-4  6.587
#> 11 E6    2.663 0.1599  0.1346 0.05857  2.729  9.219e-2  2.299 6.345e-2  9.087
#> 12 E7    1.989 0.3810  0.3369 0.09103  4.185  3.217e-2  3.701 8.727e-3 15.17 
#> 13 E8    2.536 0.8168  0.2146 0.02781 29.37   2.149e-6  7.717 1.314e-4  6.575
#> 14 E9    3.057 0.5826  0.6793 0.1109   5.253  1.596e-2  6.124 5.622e-4 10.90 
#> # … with 2 more variables: h2 <dbl>, AS <dbl>
#> ---------------------------------------------------------------------------
#> AMMI analysis table
#> ---------------------------------------------------------------------------
#> # A tibble: 15 x 8
#>    Source         Df `Sum Sq` `Mean Sq` `F value`   `Pr(>F)` Percent Accumul
#>    <chr>       <dbl>    <dbl>     <dbl>     <dbl>      <dbl> <chr>   <chr>  
#>  1 "ENV"          13 279.6     21.51       62.33   0.        .       .      
#>  2 "REP(ENV)"     28   9.662    0.3451      3.569  3.593e- 8 .       .      
#>  3 "GEN     "      9  13.00     1.444      14.93   2.190e-19 .       .      
#>  4 "ENV:GEN"     117  31.22     0.2668      2.760  1.005e-11 .       .      
#>  5 "PC1"          21  10.75     0.5119      5.29   0.        34.4    34.4   
#>  6 "PC2"          19   9.924    0.5223      5.4    0.        31.8    66.2   
#>  7 "PC3"          17   4.039    0.2376      2.46   1.400e- 3 12.9    79.2   
#>  8 "PC4"          15   3.074    0.2049      2.12   9.600e- 3 9.8     89     
#>  9 "PC5"          13   1.446    0.1113      1.150  3.176e- 1 4.6     93.6   
#> 10 "PC6"          11   0.9322   0.08475     0.88   5.606e- 1 3       96.6   
#> 11 "PC7"           9   0.5667   0.06297     0.65   7.535e- 1 1.8     98.4   
#> 12 "PC8"           7   0.3623   0.05176     0.54   8.037e- 1 1.2     99.6   
#> 13 "PC9"           5   0.1259   0.02517     0.26   9.345e- 1 0.4     100    
#> 14 "Residuals"   252  24.37     0.09669    NA     NA         .       .      
#> 15 "Total"       419 357.8      0.8540     NA     NA         .       .      
#> ---------------------------------------------------------------------------
#> Weighted average of the absolute scores
#> ---------------------------------------------------------------------------
#> # A tibble: 24 x 22
#>    type  Code      Y      PC1      PC2      PC3      PC4      PC5      PC6
#>    <chr> <fct> <dbl>    <dbl>    <dbl>    <dbl>    <dbl>    <dbl>    <dbl>
#>  1 GEN   G1    2.604  0.3166  -0.04417 -0.03600 -0.06595 -0.3125   0.4272 
#>  2 GEN   G10   2.471 -1.001   -0.5718  -0.1652  -0.3309  -0.1243  -0.1064 
#>  3 GEN   G2    2.744  0.1390   0.1988  -0.7331   0.4735  -0.04816 -0.2841 
#>  4 GEN   G3    2.955  0.04340 -0.1028   0.2284   0.1769  -0.1270  -0.1400 
#>  5 GEN   G4    2.642 -0.3251   0.4782  -0.09073  0.1417  -0.1924   0.3550 
#>  6 GEN   G5    2.537 -0.3260   0.2461   0.2452   0.1794   0.4662   0.03315
#>  7 GEN   G6    2.534 -0.09836  0.2429   0.5607   0.2377   0.05094 -0.1011 
#>  8 GEN   G7    2.741  0.2849   0.5871  -0.2068  -0.7085   0.2315  -0.08406
#>  9 GEN   G8    3.004  0.4995  -0.1916   0.3191  -0.1676  -0.3261  -0.2886 
#> 10 GEN   G9    2.510  0.4668  -0.8427  -0.1217   0.06385  0.3819   0.1889 
#> # … with 14 more rows, and 13 more variables: PC7 <dbl>, PC8 <dbl>, PC9 <dbl>,
#> #   WAAS <dbl>, PctResp <dbl>, PctWAAS <dbl>, wRes <dbl>, wWAAS <dbl>,
#> #   OrResp <dbl>, OrWAAS <dbl>, OrPC1 <dbl>, WAASY <dbl>, OrWAASY <dbl>
#> ---------------------------------------------------------------------------
#> Some information regarding the analysis
#> ---------------------------------------------------------------------------
#> # A tibble: 10 x 2
#>    Parameters Values               
#>    <chr>      <chr>                
#>  1 Ngen       "10"                 
#>  2 Nenv       "14"                 
#>  3 OVmean     "2.6742"             
#>  4 Min        "0.8991 (G10 in E11)"
#>  5 Max        "4.8121 (G8 in E5)"  
#>  6 MinENV     "E11 (1.368)"        
#>  7 MaxENV     "E3 (4.064)"         
#>  8 MinGEN     "G10 (2.471) "       
#>  9 MaxGEN     "G8 (3.004) "        
#> 10 SigPC      "4"                  
#> 
#> 
#> 
#> Variable HM 
#> ---------------------------------------------------------------------------
#> Individual analysis of variance
#> ---------------------------------------------------------------------------
#> # A tibble: 14 x 12
#>    ENV    MEAN     MSB    MSG   MSR     FCB     PRFB    FCG     PRFG    CV
#>    <chr> <dbl>   <dbl>  <dbl> <dbl>   <dbl>    <dbl>  <dbl>    <dbl> <dbl>
#>  1 E1    47.42  3.106   5.221 1.852  1.677  2.148e-1  2.819 2.928e-2 2.870
#>  2 E10   44.3   9.1    28.77  2.174  4.186  3.216e-2 13.24  3.024e-6 3.328
#>  3 E11   54.20  2.810   6.057 3.087  0.9103 4.202e-1  1.962 1.069e-1 3.242
#>  4 E12   49.60  2.766  10.40  5.322  0.5198 6.033e-1  1.955 1.081e-1 4.651
#>  5 E13   46.56 25.45   13.88  1.342 18.96   3.705e-5 10.34  1.798e-5 2.488
#>  6 E14   41.03 10.02    5.278 1.217  8.230  2.894e-3  4.338 3.935e-3 2.688
#>  7 E2    44.12  6.699  13.17  3.049  2.197  1.400e-1  4.319 4.025e-3 3.958
#>  8 E3    52.9   7.3     8.004 2.781  2.625  9.996e-2  2.877 2.692e-2 3.153
#>  9 E4    50     2.100   7.926 1.248  1.682  2.139e-1  6.350 4.510e-4 2.234
#> 10 E5    52.23  6.860  11.83  6.371  1.077  3.616e-1  1.858 1.260e-1 4.833
#> 11 E6    45.89 12.86   15.84  2.122  6.063  9.706e-3  7.463 1.633e-4 3.174
#> 12 E7    48.50  0.7052  4.997 4.738  0.1488 8.627e-1  1.055 4.382e-1 4.489
#> 13 E8    45.16 11.95   11.57  2.168  5.514  1.356e-2  5.335 1.266e-3 3.260
#> 14 E9    51.33  5.733   9.333 2.289  2.505  1.097e-1  4.078 5.408e-3 2.947
#> # … with 2 more variables: h2 <dbl>, AS <dbl>
#> ---------------------------------------------------------------------------
#> AMMI analysis table
#> ---------------------------------------------------------------------------
#> # A tibble: 15 x 8
#>    Source         Df `Sum Sq` `Mean Sq` `F value`   `Pr(>F)` Percent Accumul
#>    <chr>       <dbl>    <dbl>     <dbl>     <dbl>      <dbl> <chr>   <chr>  
#>  1 "ENV"          13 5710.     439.3       57.22   1.110e-16 .       .      
#>  2 "REP(ENV)"     28  214.9      7.676      2.703  2.197e- 5 .       .      
#>  3 "GEN     "      9  269.8     29.98      10.56   7.415e-14 .       .      
#>  4 "ENV:GEN"     117 1101.       9.408      3.313  1.064e-15 .       .      
#>  5 "PC1"          21  381.1     18.15       6.39   0.        34.6    34.6   
#>  6 "PC2"          19  319.4     16.81       5.92   0.        29      63.6   
#>  7 "PC3"          17  114.3      6.721      2.37   2.100e- 3 10.4    74     
#>  8 "PC4"          15   81.96     5.464      1.92   2.180e- 2 7.4     81.5   
#>  9 "PC5"          13   68.11     5.240      1.84   3.770e- 2 6.2     87.7   
#> 10 "PC6"          11   59.07     5.370      1.89   4.100e- 2 5.4     93     
#> 11 "PC7"           9   46.69     5.188      1.83   6.330e- 2 4.2     97.3   
#> 12 "PC8"           7   26.65     3.808      1.34   2.318e- 1 2.4     99.7   
#> 13 "PC9"           5    3.409    0.6817     0.24   9.445e- 1 0.3     100    
#> 14 "Residuals"   252  715.7      2.840     NA     NA         .       .      
#> 15 "Total"       419 8011.      19.12      NA     NA         .       .      
#> ---------------------------------------------------------------------------
#> Weighted average of the absolute scores
#> ---------------------------------------------------------------------------
#> # A tibble: 24 x 22
#>    type  Code      Y      PC1      PC2       PC3       PC4      PC5       PC6
#>    <chr> <fct> <dbl>    <dbl>    <dbl>     <dbl>     <dbl>    <dbl>     <dbl>
#>  1 GEN   G1    47.08  0.2800   0.4635   0.1740   -1.369    -1.135    0.03658 
#>  2 GEN   G10   48.51 -1.779    1.866   -0.006219  0.9219    0.1096  -0.009745
#>  3 GEN   G2    46.66  1.563    0.5518  -0.9357    0.4913    0.2843   1.184   
#>  4 GEN   G3    47.60  0.3417  -0.2012  -0.8001    0.3753   -0.4979  -1.294   
#>  5 GEN   G4    48.03 -0.2020  -1.841    0.2801    0.005954  0.8201   0.2734  
#>  6 GEN   G5    49.30  1.580    1.030    1.078    -0.2789    1.005   -0.7368  
#>  7 GEN   G6    48.73  0.5474  -0.2453   0.5324    0.4603   -1.008    0.5861  
#>  8 GEN   G7    47.97 -1.218   -0.4680   1.254    -0.05482  -0.03429  0.3366  
#>  9 GEN   G8    49.10 -0.04176 -1.241   -0.4105    0.6394   -0.1785  -0.5149  
#> 10 GEN   G9    47.90 -1.072    0.08563 -1.166    -1.191     0.6351   0.1393  
#> # … with 14 more rows, and 13 more variables: PC7 <dbl>, PC8 <dbl>, PC9 <dbl>,
#> #   WAAS <dbl>, PctResp <dbl>, PctWAAS <dbl>, wRes <dbl>, wWAAS <dbl>,
#> #   OrResp <dbl>, OrWAAS <dbl>, OrPC1 <dbl>, WAASY <dbl>, OrWAASY <dbl>
#> ---------------------------------------------------------------------------
#> Some information regarding the analysis
#> ---------------------------------------------------------------------------
#> # A tibble: 10 x 2
#>    Parameters Values               
#>    <chr>      <chr>                
#>  1 Ngen       "10"                 
#>  2 Nenv       "14"                 
#>  3 OVmean     "48.0883"            
#>  4 Min        "39.35 (G2 in E14)"  
#>  5 Max        "56.3333 (G6 in E11)"
#>  6 MinENV     "E14 (41.032)"       
#>  7 MaxENV     "E11 (54.198)"       
#>  8 MinGEN     "G2 (46.66) "        
#>  9 MaxGEN     "G5 (49.298) "       
#> 10 SigPC      "6"                  
#> 
#> 
#> 

# }