Print an object of class waasb

Print a waasb object in two ways. By default, the results are shown in the R console. The results can also be exported to the directory.

# S3 method for waasb
print(x, export = FALSE, blup = FALSE,
  file.name = NULL, digits = 4, ...)

Arguments

x
An object of class waasb.

export
A logical argument. If TRUE|T, a *.txt file is exported to the working directory

blup
A logical argument. If TRUE|T, the blups are shown.

file.name
The name of the file if export = TRUE

digits
The significant digits to be shown.

...
Options used by the tibble package to format the output. See trunc_mat for more details.

Examples

library(metan)
model <- waasb(data_ge,
  resp = c(GY, HM),
  gen = GEN,
  env = ENV,
  rep = REP
)

#> Evaluating variable GY 0 % 
#> Evaluating variable HM 100 % 
#> All variables with significant (p < 0.05) genotype-vs-environment interaction
#> Done!

print(model)

#> Variable GY 
#> ---------------------------------------------------------------------------
#> Individual fixed-model analysis of variance
#> ---------------------------------------------------------------------------
#> # A tibble: 14 x 12
#>    ENV    MEAN     MSB    MSG     MSR     FCB     PRFB    FCG     PRFG     CV
#>    <chr> <dbl>   <dbl>  <dbl>   <dbl>   <dbl>    <dbl>  <dbl>    <dbl>  <dbl>
#>  1 E1    2.521 0.06522 0.3374 0.1440   0.4529 6.428e-1  2.343 5.936e-2 15.05 
#>  2 E10   2.175 0.6537  0.2958 0.02667 24.51   7.275e-6 11.09  1.097e-5  7.508
#>  3 E11   1.368 0.3773  0.1509 0.1050   3.594  4.860e-2  1.438 2.443e-1 23.68 
#>  4 E12   1.609 0.09190 0.3202 0.05352  1.717  2.077e-1  5.982 6.474e-4 14.38 
#>  5 E13   2.910 0.07674 0.7134 0.09937  0.7722 4.767e-1  7.179 2.095e-4 10.83 
#>  6 E14   1.782 0.1035  0.1305 0.07533  1.374  2.784e-1  1.732 1.535e-1 15.40 
#>  7 E2    3.180 0.6983  0.2066 0.1785   3.912  3.885e-2  1.157 3.763e-1 13.29 
#>  8 E3    4.064 0.4886  0.3351 0.1789   2.731  9.209e-2  1.872 1.230e-1 10.41 
#>  9 E4    3.675 0.1165  0.5313 0.1377   0.8458 4.456e-1  3.859 7.123e-3 10.10 
#> 10 E5    3.910 0.2188  0.5261 0.06636  3.297  6.024e-2  7.929 1.101e-4  6.587
#> 11 E6    2.663 0.1599  0.1346 0.05857  2.729  9.219e-2  2.299 6.345e-2  9.087
#> 12 E7    1.989 0.3810  0.3369 0.09103  4.185  3.217e-2  3.701 8.727e-3 15.17 
#> 13 E8    2.536 0.8168  0.2146 0.02781 29.37   2.149e-6  7.717 1.314e-4  6.575
#> 14 E9    3.057 0.5826  0.6793 0.1109   5.253  1.596e-2  6.124 5.622e-4 10.90 
#> # ... with 2 more variables: h2 <dbl>, AS <dbl>
#> ---------------------------------------------------------------------------
#> Fixed effects
#> ---------------------------------------------------------------------------
#> # A tibble: 2 x 7
#>   SOURCE  `Sum Sq` `Mean Sq` NumDF DenDF `F value`  `Pr(>F)`
#>   <chr>      <dbl>     <dbl> <int> <dbl>     <dbl>     <dbl>
#> 1 ENV       66.09     5.084     13 279.8    52.58  2.954e-67
#> 2 REP:ENV    9.662    0.3451    28 252.0     3.569 3.593e- 8
#> ---------------------------------------------------------------------------
#> Random effects
#> ---------------------------------------------------------------------------
#> # A tibble: 3 x 2
#>   Group    Variance
#>   <chr>       <dbl>
#> 1 GEN       0.02803
#> 2 GEN:ENV   0.05671
#> 3 Residual  0.09669
#> ---------------------------------------------------------------------------
#> Likelihood ratio test
#> ---------------------------------------------------------------------------
#> # A tibble: 3 x 7
#>   model     npar logLik   AIC   LRT    Df `Pr(>Chisq)`
#>   <chr>    <int>  <dbl> <dbl> <dbl> <dbl>        <dbl>
#> 1 Complete    45 -214.7 519.4 NA       NA   NA        
#> 2 Genotype    44 -224.4 536.7 19.32     1    1.108e- 5
#> 3 Gen:Env     44 -237.1 562.3 44.83     1    2.147e-11
#> ---------------------------------------------------------------------------
#> Variance components and genetic parameters
#> ---------------------------------------------------------------------------
#> # A tibble: 15 x 2
#>    Parameters              Values
#>    <chr>                    <dbl>
#>  1 GEI variance           0.05671
#>  2 GEI (%)               31.26   
#>  3 Genotypic variance     0.02803
#>  4 Gen (%)               15.45   
#>  5 Residual variance      0.09669
#>  6 Res (%)               53.29   
#>  7 Phenotypic variance    0.1814 
#>  8 Heritability           0.1545 
#>  9 GEIr2                  0.3126 
#> 10 Heribatility of means  0.8152 
#> 11 Accuracy               0.9029 
#> 12 rge                    0.3697 
#> 13 CVg                    6.260  
#> 14 CVr                   11.63   
#> 15 CV ratio               0.5384 
#> ---------------------------------------------------------------------------
#>  Principal component analysis of the G x E interaction matrix
#> ---------------------------------------------------------------------------
#> # A tibble: 9 x 4
#>   PC    Eigenvalue Proportion Accumulated
#>   <chr>      <dbl>      <dbl>       <dbl>
#> 1 PC1      1.472      34.31         34.31
#> 2 PC2      1.347      31.38         65.69
#> 3 PC3      0.5479     12.77         78.46
#> 4 PC4      0.4167      9.710        88.17
#> 5 PC5      0.2126      4.955        93.12
#> 6 PC6      0.1397      3.256        96.38
#> 7 PC7      0.07912     1.844        98.22
#> 8 PC8      0.05673     1.322        99.55
#> 9 PC9      0.01947     0.4537      100   
#> ---------------------------------------------------------------------------
#> Some information regarding the analysis
#> ---------------------------------------------------------------------------
#> # A tibble: 9 x 2
#>   Parameters Values             
#>   <chr>      <chr>              
#> 1 Ngen       10                 
#> 2 Nenv       14                 
#> 3 OVmean     2.6742             
#> 4 Min        0.8991 (G10 in E11)
#> 5 Max        4.8121 (G8 in E5)  
#> 6 MinENV     E11 (1.368)        
#> 7 MaxENV     E3 (4.064)         
#> 8 MinGEN     "G10 (2.471) "     
#> 9 MaxGEN     "G8 (3.004) "      
#> 
#> 
#> 
#> Variable HM 
#> ---------------------------------------------------------------------------
#> Individual fixed-model analysis of variance
#> ---------------------------------------------------------------------------
#> # A tibble: 14 x 12
#>    ENV    MEAN     MSB    MSG   MSR     FCB     PRFB    FCG     PRFG    CV
#>    <chr> <dbl>   <dbl>  <dbl> <dbl>   <dbl>    <dbl>  <dbl>    <dbl> <dbl>
#>  1 E1    47.42  3.106   5.221 1.852  1.677  2.148e-1  2.819 2.928e-2 2.870
#>  2 E10   44.3   9.1    28.77  2.174  4.186  3.216e-2 13.24  3.024e-6 3.328
#>  3 E11   54.20  2.810   6.057 3.087  0.9103 4.202e-1  1.962 1.069e-1 3.242
#>  4 E12   49.60  2.766  10.40  5.322  0.5198 6.033e-1  1.955 1.081e-1 4.651
#>  5 E13   46.56 25.45   13.88  1.342 18.96   3.705e-5 10.34  1.798e-5 2.488
#>  6 E14   41.03 10.02    5.278 1.217  8.230  2.894e-3  4.338 3.935e-3 2.688
#>  7 E2    44.12  6.699  13.17  3.049  2.197  1.400e-1  4.319 4.025e-3 3.958
#>  8 E3    52.9   7.3     8.004 2.781  2.625  9.996e-2  2.877 2.692e-2 3.153
#>  9 E4    50     2.100   7.926 1.248  1.682  2.139e-1  6.350 4.510e-4 2.234
#> 10 E5    52.23  6.860  11.83  6.371  1.077  3.616e-1  1.858 1.260e-1 4.833
#> 11 E6    45.89 12.86   15.84  2.122  6.063  9.706e-3  7.463 1.633e-4 3.174
#> 12 E7    48.50  0.7052  4.997 4.738  0.1488 8.627e-1  1.055 4.382e-1 4.489
#> 13 E8    45.16 11.95   11.57  2.168  5.514  1.356e-2  5.335 1.266e-3 3.260
#> 14 E9    51.33  5.733   9.333 2.289  2.505  1.097e-1  4.078 5.408e-3 2.947
#> # ... with 2 more variables: h2 <dbl>, AS <dbl>
#> ---------------------------------------------------------------------------
#> Fixed effects
#> ---------------------------------------------------------------------------
#> # A tibble: 2 x 7
#>   SOURCE  `Sum Sq` `Mean Sq` NumDF DenDF `F value`  `Pr(>F)`
#>   <chr>      <dbl>     <dbl> <int> <dbl>     <dbl>     <dbl>
#> 1 ENV       1166.     89.69     13 257.4    31.58  5.684e-46
#> 2 REP:ENV    214.9     7.676    28 252.0     2.703 2.197e- 5
#> ---------------------------------------------------------------------------
#> Random effects
#> ---------------------------------------------------------------------------
#> # A tibble: 3 x 2
#>   Group    Variance
#>   <chr>       <dbl>
#> 1 GEN        0.4899
#> 2 GEN:ENV    2.189 
#> 3 Residual   2.840 
#> ---------------------------------------------------------------------------
#> Likelihood ratio test
#> ---------------------------------------------------------------------------
#> # A tibble: 3 x 7
#>   model     npar logLik   AIC    LRT    Df `Pr(>Chisq)`
#>   <chr>    <int>  <dbl> <dbl>  <dbl> <dbl>        <dbl>
#> 1 Complete    45 -862.7 1815. NA        NA   NA        
#> 2 Genotype    44 -866.6 1821.  7.855     1    5.068e- 3
#> 3 Gen:Env     44 -894.1 1876. 62.82      1    2.266e-15
#> ---------------------------------------------------------------------------
#> Variance components and genetic parameters
#> ---------------------------------------------------------------------------
#> # A tibble: 15 x 2
#>    Parameters              Values
#>    <chr>                    <dbl>
#>  1 GEI variance           2.189  
#>  2 GEI (%)               39.67   
#>  3 Genotypic variance     0.4899 
#>  4 Gen (%)                8.876  
#>  5 Residual variance      2.840  
#>  6 Res (%)               51.46   
#>  7 Phenotypic variance    5.519  
#>  8 Heritability           0.08876
#>  9 GEIr2                  0.3967 
#> 10 Heribatility of means  0.6862 
#> 11 Accuracy               0.8284 
#> 12 rge                    0.4353 
#> 13 CVg                    1.456  
#> 14 CVr                    3.504  
#> 15 CV ratio               0.4153 
#> ---------------------------------------------------------------------------
#>  Principal component analysis of the G x E interaction matrix
#> ---------------------------------------------------------------------------
#> # A tibble: 9 x 4
#>   PC    Eigenvalue Proportion Accumulated
#>   <chr>      <dbl>      <dbl>       <dbl>
#> 1 PC1      61.96      33.83         33.83
#> 2 PC2      51.90      28.34         62.18
#> 3 PC3      19.82      10.82         73.00
#> 4 PC4      13.90       7.588        80.59
#> 5 PC5      11.50       6.282        86.87
#> 6 PC6      10.37       5.665        92.53
#> 7 PC7       8.391      4.582        97.11
#> 8 PC8       4.583      2.502        99.62
#> 9 PC9       0.7015     0.3830      100   
#> ---------------------------------------------------------------------------
#> Some information regarding the analysis
#> ---------------------------------------------------------------------------
#> # A tibble: 9 x 2
#>   Parameters Values             
#>   <chr>      <chr>              
#> 1 Ngen       10                 
#> 2 Nenv       14                 
#> 3 OVmean     48.0883            
#> 4 Min        39.35 (G2 in E14)  
#> 5 Max        56.3333 (G6 in E11)
#> 6 MinENV     E14 (41.032)       
#> 7 MaxENV     E11 (54.198)       
#> 8 MinGEN     "G2 (46.66) "      
#> 9 MaxGEN     "G5 (49.298) "     
#> 
#> 
#>