Rescale a variable to have specified minimum and maximum values

Helper function that rescales a continuous variable to have specified minimum and maximum values.

resca(.data = NULL, ..., values = NULL, new_min = 0, new_max = 100,
  keep = TRUE)

Arguments

.data
The dataset. Grouped data is allowed.

...
Comma-separated list of unquoted variable names that will be rescaled.

values
Optional vector of values to rescale

new_min
The minimum value of the new scale. Default is 0.

new_max
The maximum value of the new scale. Default is 100

keep
Should all variables be kept after rescaling? If false, only rescaled variables will be kept.

Value

A numeric vector if values is used as input data or a tibble if a data frame is used as input in .data.

Details

The function rescale a continuous variable as follows: $$Rv_i = (Nmax - Nmin)/(Omax - Omin) * (O_i - Omax) + Nmax$$ Where \(Rv_i\) is the rescaled value of the ith position of the variable/ vector; \(Nmax\) and \(Nmin\) are the new maximum and minimum values; \(Omax and Omin\) are the maximum and minimum values of the original data, and \(O_i\) is the ith value of the original data.

There are basically two options to use resca to rescale a variable. The first is passing a data frame to .data argument and selecting one or more variables to be scaled using .... The function will return the original variables in .data plus the rescaled variable(s) with the prefix _res. By using the function group_by from dplyr package it is possible to rescale the variable(s) within each level of the grouping factor. The second option is pass a numeric vector in the argument values. The output, of course, will be a numeric vector of rescaled values.

Examples

library(metan)
library(dplyr)
# Rescale a numeric vector
resca(values = c(1:5))

#> [1]   0  25  50  75 100

 # Using a data frame
head(
 resca(data_ge, GY, HM, new_min = 0, new_max = 1)
)

#> # A tibble: 6 x 7
#>   ENV   GEN   REP      GY    HM GY_res HM_res
#>   <fct> <fct> <fct> <dbl> <dbl>  <dbl>  <dbl>
#> 1 E1    G1    1     2.167 44.93 0.3383 0.3465
#> 2 E1    G1    2     2.503 46.9  0.4143 0.4450
#> 3 E1    G1    3     2.427 47.75 0.3972 0.4875
#> 4 E1    G2    1     3.208 45.2  0.5736 0.36  
#> 5 E1    G2    2     2.933 45.3  0.5115 0.3650
#> 6 E1    G2    3     2.565 45.49 0.4283 0.3745

# Rescale within factors;
# Select variables that stats with 'N' and ends with 'L';
# Compute the mean of these variables by ENV and GEN;
# Rescale the variables that ends with 'L' whithin ENV;
data_ge2 %>%
  select(ENV, GEN, starts_with("N"), ends_with("L")) %>%
  group_by(ENV, GEN) %>%
  summarise_all(mean) %>%
  group_by(ENV) %>%
  resca(ends_with("L")) %>%
  head(n = 13)

#> # A tibble: 13 x 9
#>    ENV   GEN      NR   NKR   NKE    EL    CL EL_res    CL_res
#>    <fct> <fct> <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl>     <dbl>
#>  1 A1    H1    16.27 33.27 526.9 15.43 28.05  34.40 2.501e+ 1
#>  2 A1    H10   16.67 31.2  515.4 16.1  31.37  69.86 8.175e+ 1
#>  3 A1    H11   15.2  34.6  529.5 16.63 28.98  98.23 4.094e+ 1
#>  4 A1    H12   17.33 32.73 553.3 15.18 29.82  20.92 5.521e+ 1
#>  5 A1    H13   18.67 32.93 611.1 14.79 31.17   0    7.846e+ 1
#>  6 A1    H2    19.2  33.53 621.7 14.99 26.59  10.99 1.421e-14
#>  7 A1    H3    18.53 34.13 609.9 15.45 28.00  35.46 2.410e+ 1
#>  8 A1    H4    16.4  38.27 602.8 15.97 27.42  63.12 1.418e+ 1
#>  9 A1    H5    14.53 37.4  538.8 15.76 28.31  51.77 2.941e+ 1
#> 10 A1    H6    16.8  35.53 568.5 16.67 31.70 100    8.739e+ 1
#> 11 A1    H7    17.87 31.53 544   15.44 30.64  34.75 6.937e+ 1
#> 12 A1    H8    17.2  32.8  541.8 15.14 32.43  18.79 1.000e+ 2
#> 13 A1    H9    14.93 32.33 488.5 15.51 31.95  38.65 9.174e+ 1